AET01 Table 1 (Default) Safety Summary Table 1

AET01 Table 1 (Default) Safety Summary Table 1

aet01_1(
  adsl,
  adae,
  armvar = .study$armvar,
  lbl_overall = .study$lbl_overall,
  deco = std_deco("AET01"),
  .study = list(armvar = "ACTARM", totals_for = c("DTHFL", "DCSREAS"), lbl_overall =
    NULL)
)

Arguments

adsl
(data.frame) Subject Level Analysis Dataset

adae
(data.frame) Adverse Events Analysis Dataset

armvar
(character) variable used for column splitting

lbl_overall
(character) label used for overall column, if set to NULL the overall column is omitted

deco
(character) decoration with title, subtitles and main_footer content

.study
(list) with default values for the arguments of the function

Examples

library(scda)

#> 

library(dplyr)

#> 
#> Attaching package: ‘dplyr’

#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag

#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

sd <- synthetic_cdisc_data("rcd_2021_03_22")
adsl <- sd$adsl
adae <- sd$adae %>%
 mutate(ANL01FL = 'Y')

adsl <- df_explicit_na(adsl)
adae <- df_explicit_na(
 adae,
  omit_columns = c("SMQ01NAM", "SMQ01SC", "SMQ02NAM", "SMQ02SC", "CQ01NAM", "STUDYID", "USUBJID")
)

set.seed(99)

adae <- adae %>%
  mutate(
    AEDECOD = as.character(AEDECOD),
    AESDTH = sample(c("N", "Y"), size = nrow(adae), replace = TRUE, prob = c(0.99, 0.01)),
    AEACN = sample(
      c("DOSE NOT CHANGED", "DOSE INCREASED", "DRUG INTERRUPTED", "DRUG WITHDRAWN"),
      size = nrow(adae),
      replace = TRUE, prob = c(0.68, 0.02, 0.25, 0.05)
    ),
    FATAL = AESDTH == "Y",
    SER = AESER == "Y",
    SERWD = AESER == "Y" & AEACN == "DRUG WITHDRAWN",
    SERDSM = AESER == "Y" & AEACN %in% c("DRUG INTERRUPTED", "DOSE INCREASED", "DOSE REDUCED"),
    RELSER = AESER == "Y" & AEREL == "Y",
    WD = AEACN == "DRUG WITHDRAWN",
    DSM = AEACN %in% c("DRUG INTERRUPTED", "DOSE INCREASED", "DOSE REDUCED"),
    REL = AEREL == "Y",
    RELWD = AEREL == "Y" & AEACN == "DRUG WITHDRAWN",
    RELDSM = AEREL == "Y" & AEACN %in% c("DRUG INTERRUPTED", "DOSE INCREASED", "DOSE REDUCED"),
    CTC35 = AETOXGR %in% c("3", "4", "5"),
    CTC45 = AETOXGR %in% c("4", "5"),
    SMQ01 = SMQ01NAM != "",
    SMQ02 = SMQ02NAM != "",
    CQ01 = CQ01NAM != "",
    USUBJID_AESEQ = paste(USUBJID, AESEQ, sep = "@") # Create unique ID per AE in dataset.
  ) %>%
  var_relabel(
    AEDECOD = "Dictionary-Derived Term",
    AESDTH = "Results in Death",
    AEACN = "Action Taken with Study Treatment",
    FATAL = "AE with fatal outcome",
    SER = "Serious AE",
    SERWD = "Serious AE leading to withdrawal from treatment",
    SERDSM = "Serious AE leading to dose modification/interruption",
    RELSER = "Related Serious AE",
    WD = "AE leading to withdrawal from treatment",
    DSM = "AE leading to dose modification/interruption",
    REL = "Related AE",
    RELWD = "Related AE leading to withdrawal from treatment",
    RELDSM = "Related AE leading to dose modification/interruption",
    CTC35 = "Grade 3-5 AE",
    CTC45 = "Grade 4/5 AE",
    SMQ01 =  aesi_label(adae$SMQ01NAM, adae$SMQ01SC),
    SMQ02 = aesi_label(adae$SMQ02NAM, adae$SMQ02SC),
    CQ01 = aesi_label(adae$CQ01NAM)
  )

aet01_1(adsl, adae)

#>                                                               A: Drug X    B: Placebo    C: Combination
#>                                                                (N=134)       (N=134)        (N=132)    
#> -------------------------------------------------------------------------------------------------------
#> Total number of patients with at least one adverse event      122 (91%)    123 (91.8%)    120 (90.9%)  
#> Total AEs                                                        609           622            703      
#> Total number of deaths                                       22 (16.42%)   26 (19.4%)     19 (14.39%)  
#> Total number of patients withdrawn from study due to an AE    6 (4.48%)     1 (0.75%)      2 (1.52%)   
#>   AE with fatal outcome                                       6 (4.9%)      9 (7.3%)         6 (5%)    
#>   Serious AE                                                 104 (85.2%)   101 (82.1%)     99 (82.5%)  
#>   Serious AE leading to withdrawal from treatment             6 (4.9%)      16 (13%)       11 (9.2%)   
#>   Serious AE leading to dose modification/interruption       43 (35.2%)     48 (39%)       52 (43.3%)  
#>   Related Serious AE                                         76 (62.3%)    70 (56.9%)      75 (62.5%)  
#>   AE leading to withdrawal from treatment                    25 (20.5%)    31 (25.2%)      31 (25.8%)  
#>   AE leading to dose modification/interruption               79 (64.8%)    89 (72.4%)       90 (75%)   
#>   Related AE                                                 105 (86.1%)   108 (87.8%)    109 (90.8%)  
#>   Related AE leading to withdrawal from treatment            14 (11.5%)    14 (11.4%)      15 (12.5%)  
#>   Related AE leading to dose modification/interruption       56 (45.9%)    56 (45.5%)       60 (50%)   
#>   Grade 3-5 AE                                               109 (89.3%)   104 (84.6%)    109 (90.8%)