AET02 Table 1 (Default) Adverse Events by System Organ Class and Preferred Term Table 1

The AET02 table provides an overview of the number of patients experiencing advert events and the number of advert events categorized by Body System and Dictionary-Derived Term.

aet02_1(
  adam_db,
  armvar = .study$armvar,
  lbl_overall = .study$lbl_overall,
  prune_0 = TRUE,
  deco = std_deco("AET02"),
  .study = list(armvar = "ACTARM", lbl_overall = NULL)
)

aet02_1_lyt(
  armvar = .study$armvar,
  lbl_overall = .study$lbl_overall,
  lbl_AEBODSYS = "Body System or Organ Class",
  lbl_AEDECOD = "Dictionary-Derived Term",
  deco = std_deco("AET02"),
  .study = list(armvar = "ACTARM", lbl_overall = NULL)
)

Arguments

adam_db
(dm) object containing the adam Datasets.

armvar
(character) variable used for column splitting

lbl_overall
(character) label used for overall column, if set to NULL the overall column is omitted

prune_0
(logical) remove 0 count rows

deco
(character) decoration with title, subtitles and main_footer content

.study
(list) with default values for the arguments of the function

lbl_AEBODSYS
(string) text label for AEBODSYS.

lbl_AEDECOD
(string) text label for AEDECOD.

Value

Details

Numbers represent absolute numbers of patients and fraction of N, or absolute number of event when specified.

Remove zero-count rows unless overridden with prune_0 = FALSE.

Split columns by arm.

Does not include a total column by default.

Sort Dictionary-Derived Code (AEDECOD) by highest overall frequencies.

Functions

aet02_1_lyt:

Examples

library(dm)
#> 
#> Attaching package: ‘dm’
#> The following object is masked from ‘package:stats’:
#> 
#>     filter
library(rtables)
#> Loading required package: magrittr

db <- syn_test_data() %>%
   dm_select_tbl(adsl, adae)
#> 

db <- db %>%
  (std_filter("aet02_1"))() %>%
  (std_mutate("aet02_1"))()

aet02_1(adam_db = db) %>% head(15)
#> Adverse Events: {Specify Population}
#> Protocol: {{protocol}}, Snapshot: {{snapshot}}, Snapshot Date: {{snapshot-date}}, Cutoff Date: {{cutoff-date}}
#> 
#> —————————————————————————————————————————————————————————————————————————————————————————————————————
#>                                                              A: Drug X    B: Placebo   C: Combination
#>                                                               (N=122)      (N=123)        (N=120)    
#> —————————————————————————————————————————————————————————————————————————————————————————————————————
#> Total number of patients with at least one adverse event     122 (100%)   123 (100%)     120 (100%)  
#> Overall total number of events                                  609          622            703      
#> cl A.1                                                                                               
#>   Total number of patients with at least one adverse event   78 (63.9%)    75 (61%)      89 (74.2%)  
#>   Total number of events                                        132          130            160      
#>   dcd A.1.1.1.1                                               50 (41%)    45 (36.6%)     63 (52.5%)  
#>   dcd A.1.1.1.2                                              48 (39.3%)    48 (39%)      50 (41.7%)  
#> cl B.2                                                                                               
#>   Total number of patients with at least one adverse event   79 (64.8%)   74 (60.2%)     85 (70.8%)  
#>   Total number of events                                        129          138            143      
#>   dcd B.2.2.3.1                                              48 (39.3%)   54 (43.9%)     51 (42.5%)  
#>   dcd B.2.1.2.1                                              49 (40.2%)   44 (35.8%)     52 (43.3%)  
#> cl D.1                                                                                               
#>   Total number of patients with at least one adverse event   79 (64.8%)   67 (54.5%)     80 (66.7%)  
#>   Total number of events                                        127          106            135      
#> —————————————————————————————————————————————————————————————————————————————————————————————————————
#> 
#> Investigator text for AEs encoded using MedDRA version xx.x.  Percentages are based on N in the column headings. For frequency counts by preferred term, multiple occurrences of the same AE in an individual are counted only once.  For frequency counts of "Total number of events" rows, multiple occurrences of the same AE in an individual are counted separately. Standard footnote defining AEs included in table (e.g., includes AEs with onset from first dose of study drug through 7 days after last dose of study drug).

# Additional Examples
db_s <- db %>%
  dm_filter(adsl, SEX == "F")

aet02_1(adam_db = db_s) %>% head()
#> Adverse Events: {Specify Population}
#> Protocol: {{protocol}}, Snapshot: {{snapshot}}, Snapshot Date: {{snapshot-date}}, Cutoff Date: {{cutoff-date}}
#> 
#> —————————————————————————————————————————————————————————————————————————————————————————————————————
#>                                                              A: Drug X    B: Placebo   C: Combination
#>                                                                (N=72)       (N=77)         (N=65)    
#> —————————————————————————————————————————————————————————————————————————————————————————————————————
#> Total number of patients with at least one adverse event     72 (100%)    77 (100%)      65 (100%)   
#> Overall total number of events                                  377          419            378      
#> cl B.2                                                                                               
#>   Total number of patients with at least one adverse event   46 (63.9%)   45 (58.4%)     44 (67.7%)  
#>   Total number of events                                         81           86             64      
#>   dcd B.2.2.3.1                                              30 (41.7%)   32 (41.6%)      26 (40%)   
#> —————————————————————————————————————————————————————————————————————————————————————————————————————
#> 
#> Investigator text for AEs encoded using MedDRA version xx.x.  Percentages are based on N in the column headings. For frequency counts by preferred term, multiple occurrences of the same AE in an individual are counted only once.  For frequency counts of "Total number of events" rows, multiple occurrences of the same AE in an individual are counted separately. Standard footnote defining AEs included in table (e.g., includes AEs with onset from first dose of study drug through 7 days after last dose of study drug).

# alternatively adam_db also accepts a names list
aet02_1(adam_db = list(adsl = db$adsl, adae = db$adae)) %>% head()
#> Adverse Events: {Specify Population}
#> Protocol: {{protocol}}, Snapshot: {{snapshot}}, Snapshot Date: {{snapshot-date}}, Cutoff Date: {{cutoff-date}}
#> 
#> —————————————————————————————————————————————————————————————————————————————————————————————————————
#>                                                              A: Drug X    B: Placebo   C: Combination
#>                                                               (N=122)      (N=123)        (N=120)    
#> —————————————————————————————————————————————————————————————————————————————————————————————————————
#> Total number of patients with at least one adverse event     122 (100%)   123 (100%)     120 (100%)  
#> Overall total number of events                                  609          622            703      
#> cl A.1                                                                                               
#>   Total number of patients with at least one adverse event   78 (63.9%)    75 (61%)      89 (74.2%)  
#>   Total number of events                                        132          130            160      
#>   dcd A.1.1.1.1                                               50 (41%)    45 (36.6%)     63 (52.5%)  
#> —————————————————————————————————————————————————————————————————————————————————————————————————————
#> 
#> Investigator text for AEs encoded using MedDRA version xx.x.  Percentages are based on N in the column headings. For frequency counts by preferred term, multiple occurrences of the same AE in an individual are counted only once.  For frequency counts of "Total number of events" rows, multiple occurrences of the same AE in an individual are counted separately. Standard footnote defining AEs included in table (e.g., includes AEs with onset from first dose of study drug through 7 days after last dose of study drug).

aet02_1(db, lbl_overall = "All Patients") %>% head()
#> Adverse Events: {Specify Population}
#> Protocol: {{protocol}}, Snapshot: {{snapshot}}, Snapshot Date: {{snapshot-date}}, Cutoff Date: {{cutoff-date}}
#> 
#> ————————————————————————————————————————————————————————————————————————————————————————————————————————————————————
#>                                                              A: Drug X    B: Placebo   C: Combination   All Patients
#>                                                               (N=122)      (N=123)        (N=120)         (N=365)   
#> ————————————————————————————————————————————————————————————————————————————————————————————————————————————————————
#> Total number of patients with at least one adverse event     122 (100%)   123 (100%)     120 (100%)      365 (100%) 
#> Overall total number of events                                  609          622            703             1934    
#> cl A.1                                                                                                              
#>   Total number of patients with at least one adverse event   78 (63.9%)    75 (61%)      89 (74.2%)     242 (66.3%) 
#>   Total number of events                                        132          130            160             422     
#>   dcd A.1.1.1.1                                               50 (41%)    45 (36.6%)     63 (52.5%)     158 (43.3%) 
#> ————————————————————————————————————————————————————————————————————————————————————————————————————————————————————
#> 
#> Investigator text for AEs encoded using MedDRA version xx.x.  Percentages are based on N in the column headings. For frequency counts by preferred term, multiple occurrences of the same AE in an individual are counted only once.  For frequency counts of "Total number of events" rows, multiple occurrences of the same AE in an individual are counted separately. Standard footnote defining AEs included in table (e.g., includes AEs with onset from first dose of study drug through 7 days after last dose of study drug).

db_m <- db %>%
  dm_zoom_to(adae) %>%
  mutate(AEBODSYS = with_label(AEBODSYS, "Medra System Organ Class")) %>%
  dm_update_zoomed()

aet02_1(db_m) %>% head()
#> Adverse Events: {Specify Population}
#> Protocol: {{protocol}}, Snapshot: {{snapshot}}, Snapshot Date: {{snapshot-date}}, Cutoff Date: {{cutoff-date}}
#> 
#> —————————————————————————————————————————————————————————————————————————————————————————————————————
#>                                                              A: Drug X    B: Placebo   C: Combination
#>                                                               (N=122)      (N=123)        (N=120)    
#> —————————————————————————————————————————————————————————————————————————————————————————————————————
#> Total number of patients with at least one adverse event     122 (100%)   123 (100%)     120 (100%)  
#> Overall total number of events                                  609          622            703      
#> cl A.1                                                                                               
#>   Total number of patients with at least one adverse event   78 (63.9%)    75 (61%)      89 (74.2%)  
#>   Total number of events                                        132          130            160      
#>   dcd A.1.1.1.1                                               50 (41%)    45 (36.6%)     63 (52.5%)  
#> —————————————————————————————————————————————————————————————————————————————————————————————————————
#> 
#> Investigator text for AEs encoded using MedDRA version xx.x.  Percentages are based on N in the column headings. For frequency counts by preferred term, multiple occurrences of the same AE in an individual are counted only once.  For frequency counts of "Total number of events" rows, multiple occurrences of the same AE in an individual are counted separately. Standard footnote defining AEs included in table (e.g., includes AEs with onset from first dose of study drug through 7 days after last dose of study drug).

aet02_1_lyt(
 armvar = "ACTARM",
 lbl_overall = NULL,
 deco = std_deco("AET02")
)
#> A Pre-data Table Layout
#> 
#> Column-Split Structure:
#> ACTARM (lvls) 
#> 
#> Row-Split Structure:
#> AEBODSYS (lvls) -> AEDECOD (** analysis **) 
#>