Adverse Events by System Organ Class and Dictionary-Derived Term Table

The AET02_2 table provides an overview of the number of patients experiencing advert events and the number of advert events categorized by Body System, High Level Term and Dictionary-Derived Term.

aet02_2(
  adsl,
  adae,
  armvar = .study$armvar,
  lbl_overall = .study$lbl_overall,
  prune_0 = TRUE,
  deco = std_deco("AET02"),
  .study = list(armvar = "ACTARM", lbl_overall = "")
)

Arguments

adsl
Subject Level Analysis Dataset

adae
Adverse Events Analysis Dataset

armvar
variable used for column splitting

lbl_overall
label used for overall column, if set to NULL the overall column is omitted

prune_0
remove 0 count rows

deco
decoration list with title, subtitles and main_footer content

.study
a list with default values for the arguments of the function

Details

Numbers represent absolute numbers of patients and fraction of N, or absolute number of event when specified.

Remove zero-count rows unless overridden with prune_0 = FALSE.

Split columns by arm.

Does not include a total column by default.

Sort Body System or Organ Class, High Level Term and Dictionary-Derived Term hierarchically by highest overall frequencies.

Examples

library(scda)
library(dplyr)
sd <- synthetic_cdisc_data("rcd_2021_03_22")
adsl <- sd$adsl
adae <- sd$adae %>%
 mutate(ANL01FL = 'Y')

aet02_2(adsl, adae)

#> Body System or Organ Class                                                                              
#>   High Level Term                                              A: Drug X    B: Placebo    C: Combination
#>  Dictionary-Derived Term                                        (N=134)       (N=134)        (N=132)    
#> --------------------------------------------------------------------------------------------------------
#> Total number of patients with at least one adverse event       122 (91%)    123 (91.8%)    120 (90.9%)  
#> Overall total number of events                                    609           622            703      
#> cl A.1                                                                                                  
#>   Total number of patients with at least one adverse event     78 (58.2%)    75 (56%)       89 (67.4%)  
#>   Total number of events                                          132           130            160      
#>   hlt A.1.1.1                                                                                           
#>     Total number of patients with at least one adverse event   78 (58.2%)    75 (56%)       89 (67.4%)  
#>     Total number of events                                        132           130            160      
#>     dcd A.1.1.1.1                                              50 (37.3%)   45 (33.6%)      63 (47.7%)  
#>     dcd A.1.1.1.2                                              48 (35.8%)   48 (35.8%)      50 (37.9%)  
#> cl B.2                                                                                                  
#>   Total number of patients with at least one adverse event      79 (59%)    74 (55.2%)      85 (64.4%)  
#>   Total number of events                                          129           138            143      
#>   hlt B.2.2.3                                                                                           
#>     Total number of patients with at least one adverse event   48 (35.8%)   54 (40.3%)      51 (38.6%)  
#>     Total number of events                                         64           76              77      
#>     dcd B.2.2.3.1                                              48 (35.8%)   54 (40.3%)      51 (38.6%)  
#>   hlt B.2.1.2                                                                                           
#>     Total number of patients with at least one adverse event   49 (36.6%)   44 (32.8%)      52 (39.4%)  
#>     Total number of events                                         65           62              66      
#>     dcd B.2.1.2.1                                              49 (36.6%)   44 (32.8%)      52 (39.4%)  
#> cl D.1                                                                                                  
#>   Total number of patients with at least one adverse event      79 (59%)     67 (50%)       80 (60.6%)  
#>   Total number of events                                          127           106            135      
#>   hlt D.1.1.1                                                                                           
#>     Total number of patients with at least one adverse event   50 (37.3%)   42 (31.3%)      51 (38.6%)  
#>     Total number of events                                         61           51              71      
#>     dcd D.1.1.1.1                                              50 (37.3%)   42 (31.3%)      51 (38.6%)  
#>   hlt D.1.1.4                                                                                           
#>     Total number of patients with at least one adverse event   48 (35.8%)   42 (31.3%)      50 (37.9%)  
#>     Total number of events                                         66           55              64      
#>     dcd D.1.1.4.2                                              48 (35.8%)   42 (31.3%)      50 (37.9%)  
#> cl D.2                                                                                                  
#>   Total number of patients with at least one adverse event     47 (35.1%)   58 (43.3%)      57 (43.2%)  
#>   Total number of events                                           62           72              74      
#>   hlt D.2.1.5                                                                                           
#>     Total number of patients with at least one adverse event   47 (35.1%)   58 (43.3%)      57 (43.2%)  
#>     Total number of events                                         62           72              74      
#>     dcd D.2.1.5.3                                              47 (35.1%)   58 (43.3%)      57 (43.2%)  
#> cl B.1                                                                                                  
#>   Total number of patients with at least one adverse event     47 (35.1%)   49 (36.6%)      43 (32.6%)  
#>   Total number of events                                           56           60              62      
#>   hlt B.1.1.1                                                                                           
#>     Total number of patients with at least one adverse event   47 (35.1%)   49 (36.6%)      43 (32.6%)  
#>     Total number of events                                         56           60              62      
#>     dcd B.1.1.1.1                                              47 (35.1%)   49 (36.6%)      43 (32.6%)  
#> cl C.2                                                                                                  
#>   Total number of patients with at least one adverse event     35 (26.1%)   48 (35.8%)      55 (41.7%)  
#>   Total number of events                                           48           53              65      
#>   hlt C.2.1.2                                                                                           
#>     Total number of patients with at least one adverse event   35 (26.1%)   48 (35.8%)      55 (41.7%)  
#>     Total number of events                                         48           53              65      
#>     dcd C.2.1.2.1                                              35 (26.1%)   48 (35.8%)      55 (41.7%)  
#> cl C.1                                                                                                  
#>   Total number of patients with at least one adverse event     43 (32.1%)   46 (34.3%)      43 (32.6%)  
#>   Total number of events                                           55           63              64      
#>   hlt C.1.1.1                                                                                           
#>     Total number of patients with at least one adverse event   43 (32.1%)   46 (34.3%)      43 (32.6%)  
#>     Total number of events                                         55           63              64      
#>     dcd C.1.1.1.3                                              43 (32.1%)   46 (34.3%)      43 (32.6%)  
#> --------------------------------------------------------------------------------------------------------
#> 
#> Footnotes go here

aet02_2(adsl, adae, lbl_overall = "All Patients")

#> Main Title
#> Protocol: {{protocol}}, Snapshot: {{snapshot}}, Snapshot Date: {{snapshot-date}}, Cutoff Date: {{cutoff-date}}
#> 
#> -----------------------------------------------------------------------------------------------------------------------
#> Body System or Organ Class                                                                                             
#>   High Level Term                                              A: Drug X    B: Placebo    C: Combination   All Patients
#>  Dictionary-Derived Term                                        (N=134)       (N=134)        (N=132)         (N=400)   
#> -----------------------------------------------------------------------------------------------------------------------
#> Total number of patients with at least one adverse event       122 (91%)    123 (91.8%)    120 (90.9%)     365 (91.2%) 
#> Overall total number of events                                    609           622            703             1934    
#> cl A.1                                                                                                                 
#>   Total number of patients with at least one adverse event     78 (58.2%)    75 (56%)       89 (67.4%)     242 (60.5%) 
#>   Total number of events                                          132           130            160             422     
#>   hlt A.1.1.1                                                                                                          
#>     Total number of patients with at least one adverse event   78 (58.2%)    75 (56%)       89 (67.4%)     242 (60.5%) 
#>     Total number of events                                        132           130            160             422     
#>     dcd A.1.1.1.1                                              50 (37.3%)   45 (33.6%)      63 (47.7%)     158 (39.5%) 
#>     dcd A.1.1.1.2                                              48 (35.8%)   48 (35.8%)      50 (37.9%)     146 (36.5%) 
#> cl B.2                                                                                                                 
#>   Total number of patients with at least one adverse event      79 (59%)    74 (55.2%)      85 (64.4%)     238 (59.5%) 
#>   Total number of events                                          129           138            143             410     
#>   hlt B.2.2.3                                                                                                          
#>     Total number of patients with at least one adverse event   48 (35.8%)   54 (40.3%)      51 (38.6%)     153 (38.2%) 
#>     Total number of events                                         64           76              77             217     
#>     dcd B.2.2.3.1                                              48 (35.8%)   54 (40.3%)      51 (38.6%)     153 (38.2%) 
#>   hlt B.2.1.2                                                                                                          
#>     Total number of patients with at least one adverse event   49 (36.6%)   44 (32.8%)      52 (39.4%)     145 (36.2%) 
#>     Total number of events                                         65           62              66             193     
#>     dcd B.2.1.2.1                                              49 (36.6%)   44 (32.8%)      52 (39.4%)     145 (36.2%) 
#> cl D.1                                                                                                                 
#>   Total number of patients with at least one adverse event      79 (59%)     67 (50%)       80 (60.6%)     226 (56.5%) 
#>   Total number of events                                          127           106            135             368     
#>   hlt D.1.1.1                                                                                                          
#>     Total number of patients with at least one adverse event   50 (37.3%)   42 (31.3%)      51 (38.6%)     143 (35.8%) 
#>     Total number of events                                         61           51              71             183     
#>     dcd D.1.1.1.1                                              50 (37.3%)   42 (31.3%)      51 (38.6%)     143 (35.8%) 
#>   hlt D.1.1.4                                                                                                          
#>     Total number of patients with at least one adverse event   48 (35.8%)   42 (31.3%)      50 (37.9%)      140 (35%)  
#>     Total number of events                                         66           55              64             185     
#>     dcd D.1.1.4.2                                              48 (35.8%)   42 (31.3%)      50 (37.9%)      140 (35%)  
#> cl D.2                                                                                                                 
#>   Total number of patients with at least one adverse event     47 (35.1%)   58 (43.3%)      57 (43.2%)     162 (40.5%) 
#>   Total number of events                                           62           72              74             208     
#>   hlt D.2.1.5                                                                                                          
#>     Total number of patients with at least one adverse event   47 (35.1%)   58 (43.3%)      57 (43.2%)     162 (40.5%) 
#>     Total number of events                                         62           72              74             208     
#>     dcd D.2.1.5.3                                              47 (35.1%)   58 (43.3%)      57 (43.2%)     162 (40.5%) 
#> cl B.1                                                                                                                 
#>   Total number of patients with at least one adverse event     47 (35.1%)   49 (36.6%)      43 (32.6%)     139 (34.8%) 
#>   Total number of events                                           56           60              62             178     
#>   hlt B.1.1.1                                                                                                          
#>     Total number of patients with at least one adverse event   47 (35.1%)   49 (36.6%)      43 (32.6%)     139 (34.8%) 
#>     Total number of events                                         56           60              62             178     
#>     dcd B.1.1.1.1                                              47 (35.1%)   49 (36.6%)      43 (32.6%)     139 (34.8%) 
#> cl C.2                                                                                                                 
#>   Total number of patients with at least one adverse event     35 (26.1%)   48 (35.8%)      55 (41.7%)     138 (34.5%) 
#>   Total number of events                                           48           53              65             166     
#>   hlt C.2.1.2                                                                                                          
#>     Total number of patients with at least one adverse event   35 (26.1%)   48 (35.8%)      55 (41.7%)     138 (34.5%) 
#>     Total number of events                                         48           53              65             166     
#>     dcd C.2.1.2.1                                              35 (26.1%)   48 (35.8%)      55 (41.7%)     138 (34.5%) 
#> cl C.1                                                                                                                 
#>   Total number of patients with at least one adverse event     43 (32.1%)   46 (34.3%)      43 (32.6%)      132 (33%)  
#>   Total number of events                                           55           63              64             182     
#>   hlt C.1.1.1                                                                                                          
#>     Total number of patients with at least one adverse event   43 (32.1%)   46 (34.3%)      43 (32.6%)      132 (33%)  
#>     Total number of events                                         55           63              64             182     
#>     dcd C.1.1.1.3                                              43 (32.1%)   46 (34.3%)      43 (32.6%)      132 (33%)  
#> -----------------------------------------------------------------------------------------------------------------------
#> 
#> Footnotes go here

adae <- adae %>% var_relabel(AEBODSYS = "Medra System Organ Class")
aet02_2(adsl, adae)

#> Medra System Organ Class                                                                                
#>   High Level Term                                              A: Drug X    B: Placebo    C: Combination
#>  Dictionary-Derived Term                                        (N=134)       (N=134)        (N=132)    
#> --------------------------------------------------------------------------------------------------------
#> Total number of patients with at least one adverse event       122 (91%)    123 (91.8%)    120 (90.9%)  
#> Overall total number of events                                    609           622            703      
#> cl A.1                                                                                                  
#>   Total number of patients with at least one adverse event     78 (58.2%)    75 (56%)       89 (67.4%)  
#>   Total number of events                                          132           130            160      
#>   hlt A.1.1.1                                                                                           
#>     Total number of patients with at least one adverse event   78 (58.2%)    75 (56%)       89 (67.4%)  
#>     Total number of events                                        132           130            160      
#>     dcd A.1.1.1.1                                              50 (37.3%)   45 (33.6%)      63 (47.7%)  
#>     dcd A.1.1.1.2                                              48 (35.8%)   48 (35.8%)      50 (37.9%)  
#> cl B.2                                                                                                  
#>   Total number of patients with at least one adverse event      79 (59%)    74 (55.2%)      85 (64.4%)  
#>   Total number of events                                          129           138            143      
#>   hlt B.2.2.3                                                                                           
#>     Total number of patients with at least one adverse event   48 (35.8%)   54 (40.3%)      51 (38.6%)  
#>     Total number of events                                         64           76              77      
#>     dcd B.2.2.3.1                                              48 (35.8%)   54 (40.3%)      51 (38.6%)  
#>   hlt B.2.1.2                                                                                           
#>     Total number of patients with at least one adverse event   49 (36.6%)   44 (32.8%)      52 (39.4%)  
#>     Total number of events                                         65           62              66      
#>     dcd B.2.1.2.1                                              49 (36.6%)   44 (32.8%)      52 (39.4%)  
#> cl D.1                                                                                                  
#>   Total number of patients with at least one adverse event      79 (59%)     67 (50%)       80 (60.6%)  
#>   Total number of events                                          127           106            135      
#>   hlt D.1.1.1                                                                                           
#>     Total number of patients with at least one adverse event   50 (37.3%)   42 (31.3%)      51 (38.6%)  
#>     Total number of events                                         61           51              71      
#>     dcd D.1.1.1.1                                              50 (37.3%)   42 (31.3%)      51 (38.6%)  
#>   hlt D.1.1.4                                                                                           
#>     Total number of patients with at least one adverse event   48 (35.8%)   42 (31.3%)      50 (37.9%)  
#>     Total number of events                                         66           55              64      
#>     dcd D.1.1.4.2                                              48 (35.8%)   42 (31.3%)      50 (37.9%)  
#> cl D.2                                                                                                  
#>   Total number of patients with at least one adverse event     47 (35.1%)   58 (43.3%)      57 (43.2%)  
#>   Total number of events                                           62           72              74      
#>   hlt D.2.1.5                                                                                           
#>     Total number of patients with at least one adverse event   47 (35.1%)   58 (43.3%)      57 (43.2%)  
#>     Total number of events                                         62           72              74      
#>     dcd D.2.1.5.3                                              47 (35.1%)   58 (43.3%)      57 (43.2%)  
#> cl B.1                                                                                                  
#>   Total number of patients with at least one adverse event     47 (35.1%)   49 (36.6%)      43 (32.6%)  
#>   Total number of events                                           56           60              62      
#>   hlt B.1.1.1                                                                                           
#>     Total number of patients with at least one adverse event   47 (35.1%)   49 (36.6%)      43 (32.6%)  
#>     Total number of events                                         56           60              62      
#>     dcd B.1.1.1.1                                              47 (35.1%)   49 (36.6%)      43 (32.6%)  
#> cl C.2                                                                                                  
#>   Total number of patients with at least one adverse event     35 (26.1%)   48 (35.8%)      55 (41.7%)  
#>   Total number of events                                           48           53              65      
#>   hlt C.2.1.2                                                                                           
#>     Total number of patients with at least one adverse event   35 (26.1%)   48 (35.8%)      55 (41.7%)  
#>     Total number of events                                         48           53              65      
#>     dcd C.2.1.2.1                                              35 (26.1%)   48 (35.8%)      55 (41.7%)  
#> cl C.1                                                                                                  
#>   Total number of patients with at least one adverse event     43 (32.1%)   46 (34.3%)      43 (32.6%)  
#>   Total number of events                                           55           63              64      
#>   hlt C.1.1.1                                                                                           
#>     Total number of patients with at least one adverse event   43 (32.1%)   46 (34.3%)      43 (32.6%)  
#>     Total number of events                                         55           63              64      
#>     dcd C.1.1.1.3                                              43 (32.1%)   46 (34.3%)      43 (32.6%)  
#> --------------------------------------------------------------------------------------------------------
#> 
#> Footnotes go here