AET02 Table 3 (Supplementary) Adverse Events by Dictionary-Derived Term

The AET02_3 table provides an overview of the number of patients experiencing advert events and the number of advert events categorized by Dictionary-Derived Term.

aet02_3(
  adsl,
  adae,
  armvar = .study$armvar,
  lbl_overall = .study$lbl_overall,
  prune_0 = TRUE,
  deco = std_deco("AET02"),
  .study = list(armvar = "ACTARM", lbl_overall = "")
)

aet02_3_lyt(
  armvar = .study$armvar,
  lbl_overall = .study$lbl_overall,
  lbl_AEDECOD = "AEDECOD",
  deco = std_deco("AET02"),
  .study = list(armvar = "ACTARM", lbl_overall = "")
)

Arguments

adsl
Subject Level Analysis Dataset

adae
Adverse Events Analysis Dataset

armvar
variable used for column splitting

lbl_overall
label used for overall column, if set to NULL the overall column is omitted

prune_0
remove 0 count rows

deco
decoration list with title, subtitles and main_footer content

.study
a list with default values for the arguments of the function

lbl_AEDECOD
(string) text label for AEDECOD.

Value

Details

Numbers represent absolute numbers of patients and fraction of N, or absolute number of event when specified.

Remove zero-count rows unless overridden with prune_0 = FALSE.

Split columns by arm.

Does not include a total column by default.

Sort Dictionary-Derived Code by highest overall frequencies.

Functions

aet02_3_lyt:

Examples

library(scda)
library(dplyr)
sd <- synthetic_cdisc_data("rcd_2021_03_22")
adsl <- sd$adsl
adae <- sd$adae %>%
 mutate(ANL01FL = 'Y')

aet02_3(adsl, adae)

#> Error in is.data.frame(df): object 'adae' not found

aet02_3(adsl, adae, lbl_overall = "All Patients")

#> Error in is.data.frame(df): object 'adae' not found

aet02_3_lyt(armvar = "ACTARM",
 lbl_overall = "",
 lbl_AEDECOD = "Dictionary-Derived Term",
 deco = std_deco("AET02"))

#> Error in is.data.frame(df): object 'adae' not found