AET03 Table 1 (Default) Advert Events by Greatest Intensity Table 1

An adverse events table categorized by System Organ Class, Dictionary-Derived Term and Greatest intensity.

aet03_1(
  adam_db,
  armvar = .study$armvar,
  prune_0 = TRUE,
  lbl_overall = .study$lbl_overall,
  deco = std_deco("AET03"),
  .study = list(armvar = "ACTARM", lbl_overall = NULL)
)

aet03_1_lyt(
  armvar = .study$armvar,
  lbl_AEBODSYS = "",
  lbl_AEDECOD = "",
  severity_grade = .study$severity_grade,
  lbl_overall = .study$lbl_overall,
  deco = std_deco("AET03"),
  .study = list(armvar = "ACTARM", lbl_overall = NULL, severity_grade = c("MILD",
    "MODERATE", "SEVERE", "LIFE THREATENING"))
)

Arguments

adam_db
(dm) object containing the ADaM datasets

armvar
(character) variable used for column splitting

prune_0
(logical) remove 0 count rows

lbl_overall
(character) label used for overall column, if set to NULL the overall column is omitted

deco
(character) decoration with title, subtitles and main_footer content

.study
(list) with default values for the arguments of the function

lbl_AEBODSYS
(string) text label for AEBODSYS.

lbl_AEDECOD
(string) text label for AEDECOD.

severity_grade
(vector of strings) describing the severity levels present in the data set.

Details

Default Adverse Events by Greatest Intensity table

Numbers represent absolute numbers of patients and fraction of N

Remove zero-count rows unless overridden with prune_0 = FALSE.

Split columns by arm

Does not include a total column by default

Sort by Body System or Organ Class (SOC) and Dictionary-Derived Term (PT)

Functions

aet03_1_lyt:

Examples

library(dm)
library(rtables)

db <- syn_test_data() %>%
  preprocess_data("aet03_1")

aet03_1(db)
#> Adverse Events by Greatest Intensity: {Specify Population}
#> Protocol: {{protocol}}, Snapshot: {{snapshot}}, Snapshot Date: {{snapshot-date}}, Cutoff Date: {{cutoff-date}}
#> 
#> ————————————————————————————————————————————————————————————————————
#> MedDRA System Organ Class   A: Drug X    B: Placebo   C: Combination
#>   MedDRA Preferred Term       (N=97)      (N=103)        (N=102)    
#> ————————————————————————————————————————————————————————————————————
#> - Any Intensity -           97 (100%)    103 (100%)     102 (100%)  
#> MILD                         5 (5.2%)     6 (5.8%)       4 (3.9%)   
#> MODERATE                    19 (19.6%)   23 (22.3%)     20 (19.6%)  
#> SEVERE                      73 (75.3%)   74 (71.8%)     78 (76.5%)  
#> cl A.1                                                              
#>   - Any Intensity -         62 (63.9%)   64 (62.1%)     77 (75.5%)  
#>   MILD                      25 (25.8%)   20 (19.4%)     34 (33.3%)  
#>   MODERATE                  37 (38.1%)   44 (42.7%)     43 (42.2%)  
#>   dcd A.1.1.1.1                                                     
#>     - Any Intensity -       42 (43.3%)    36 (35%)      54 (52.9%)  
#>     MILD                    42 (43.3%)    36 (35%)      54 (52.9%)  
#>   dcd A.1.1.1.2                                                     
#>     - Any Intensity -       37 (38.1%)   44 (42.7%)     43 (42.2%)  
#>     MODERATE                37 (38.1%)   44 (42.7%)     43 (42.2%)  
#> cl B.2                                                              
#>   - Any Intensity -         59 (60.8%)   59 (57.3%)     73 (71.6%)  
#>   MILD                      23 (23.7%)   22 (21.4%)     28 (27.5%)  
#>   MODERATE                  36 (37.1%)   37 (35.9%)     45 (44.1%)  
#>   dcd B.2.2.3.1                                                     
#>     - Any Intensity -       38 (39.2%)   42 (40.8%)     42 (41.2%)  
#>     MILD                    38 (39.2%)   42 (40.8%)     42 (41.2%)  
#>   dcd B.2.1.2.1                                                     
#>     - Any Intensity -       36 (37.1%)   37 (35.9%)     45 (44.1%)  
#>     MODERATE                36 (37.1%)   37 (35.9%)     45 (44.1%)  
#> cl D.1                                                              
#>   - Any Intensity -         60 (61.9%)   56 (54.4%)     67 (65.7%)  
#>   MODERATE                  20 (20.6%)   24 (23.3%)     24 (23.5%)  
#>   SEVERE                    40 (41.2%)   32 (31.1%)     43 (42.2%)  
#>   dcd D.1.1.1.1                                                     
#>     - Any Intensity -       40 (41.2%)   32 (31.1%)     43 (42.2%)  
#>     SEVERE                  40 (41.2%)   32 (31.1%)     43 (42.2%)  
#>   dcd D.1.1.4.2                                                     
#>     - Any Intensity -       37 (38.1%)   37 (35.9%)     41 (40.2%)  
#>     MODERATE                37 (38.1%)   37 (35.9%)     41 (40.2%)  
#> cl D.2                                                              
#>   - Any Intensity -         35 (36.1%)   50 (48.5%)      49 (48%)   
#>   MILD                      35 (36.1%)   50 (48.5%)      49 (48%)   
#>   dcd D.2.1.5.3                                                     
#>     - Any Intensity -       35 (36.1%)   50 (48.5%)      49 (48%)   
#>     MILD                    35 (36.1%)   50 (48.5%)      49 (48%)   
#> cl B.1                                                              
#>   - Any Intensity -         37 (38.1%)   40 (38.8%)     36 (35.3%)  
#>   SEVERE                    37 (38.1%)   40 (38.8%)     36 (35.3%)  
#>   dcd B.1.1.1.1                                                     
#>     - Any Intensity -       37 (38.1%)   40 (38.8%)     36 (35.3%)  
#>     SEVERE                  37 (38.1%)   40 (38.8%)     36 (35.3%)  
#> cl C.2                                                              
#>   - Any Intensity -         26 (26.8%)   38 (36.9%)      49 (48%)   
#>   MODERATE                  26 (26.8%)   38 (36.9%)      49 (48%)   
#>   dcd C.2.1.2.1                                                     
#>     - Any Intensity -       26 (26.8%)   38 (36.9%)      49 (48%)   
#>     MODERATE                26 (26.8%)   38 (36.9%)      49 (48%)   
#> cl C.1                                                              
#>   - Any Intensity -          32 (33%)    38 (36.9%)     35 (34.3%)  
#>   SEVERE                     32 (33%)    38 (36.9%)     35 (34.3%)  
#>   dcd C.1.1.1.3                                                     
#>     - Any Intensity -        32 (33%)    38 (36.9%)     35 (34.3%)  
#>     SEVERE                   32 (33%)    38 (36.9%)     35 (34.3%)  
#> ————————————————————————————————————————————————————————————————————
#> 
#> Investigator text for AEs encoded using MedDRA version xx.x.  All counts represent patients. Multiple occurrences of the same AE in one individual are counted once at the greatest intensity for this preferred term. To the SOC Overall row counts, a patient contributes only with the AE occuring with the greatest intensity within the SOC. Standard footnote defining AEs included in table (e.g., includes AEs with onset from first dose of study drug through 7 days after last dose of study drug).
aet03_1(db, lbl_overall = "All Patients")
#> Adverse Events by Greatest Intensity: {Specify Population}
#> Protocol: {{protocol}}, Snapshot: {{snapshot}}, Snapshot Date: {{snapshot-date}}, Cutoff Date: {{cutoff-date}}
#> 
#> ———————————————————————————————————————————————————————————————————————————————————
#> MedDRA System Organ Class   A: Drug X    B: Placebo   C: Combination   All Patients
#>   MedDRA Preferred Term       (N=97)      (N=103)        (N=102)         (N=302)   
#> ———————————————————————————————————————————————————————————————————————————————————
#> - Any Intensity -           97 (100%)    103 (100%)     102 (100%)      302 (100%) 
#> MILD                         5 (5.2%)     6 (5.8%)       4 (3.9%)        15 (5%)   
#> MODERATE                    19 (19.6%)   23 (22.3%)     20 (19.6%)      62 (20.5%) 
#> SEVERE                      73 (75.3%)   74 (71.8%)     78 (76.5%)     225 (74.5%) 
#> cl A.1                                                                             
#>   - Any Intensity -         62 (63.9%)   64 (62.1%)     77 (75.5%)     203 (67.2%) 
#>   MILD                      25 (25.8%)   20 (19.4%)     34 (33.3%)      79 (26.2%) 
#>   MODERATE                  37 (38.1%)   44 (42.7%)     43 (42.2%)     124 (41.1%) 
#>   dcd A.1.1.1.1                                                                    
#>     - Any Intensity -       42 (43.3%)    36 (35%)      54 (52.9%)     132 (43.7%) 
#>     MILD                    42 (43.3%)    36 (35%)      54 (52.9%)     132 (43.7%) 
#>   dcd A.1.1.1.2                                                                    
#>     - Any Intensity -       37 (38.1%)   44 (42.7%)     43 (42.2%)     124 (41.1%) 
#>     MODERATE                37 (38.1%)   44 (42.7%)     43 (42.2%)     124 (41.1%) 
#> cl B.2                                                                             
#>   - Any Intensity -         59 (60.8%)   59 (57.3%)     73 (71.6%)     191 (63.2%) 
#>   MILD                      23 (23.7%)   22 (21.4%)     28 (27.5%)      73 (24.2%) 
#>   MODERATE                  36 (37.1%)   37 (35.9%)     45 (44.1%)     118 (39.1%) 
#>   dcd B.2.2.3.1                                                                    
#>     - Any Intensity -       38 (39.2%)   42 (40.8%)     42 (41.2%)     122 (40.4%) 
#>     MILD                    38 (39.2%)   42 (40.8%)     42 (41.2%)     122 (40.4%) 
#>   dcd B.2.1.2.1                                                                    
#>     - Any Intensity -       36 (37.1%)   37 (35.9%)     45 (44.1%)     118 (39.1%) 
#>     MODERATE                36 (37.1%)   37 (35.9%)     45 (44.1%)     118 (39.1%) 
#> cl D.1                                                                             
#>   - Any Intensity -         60 (61.9%)   56 (54.4%)     67 (65.7%)     183 (60.6%) 
#>   MODERATE                  20 (20.6%)   24 (23.3%)     24 (23.5%)      68 (22.5%) 
#>   SEVERE                    40 (41.2%)   32 (31.1%)     43 (42.2%)     115 (38.1%) 
#>   dcd D.1.1.1.1                                                                    
#>     - Any Intensity -       40 (41.2%)   32 (31.1%)     43 (42.2%)     115 (38.1%) 
#>     SEVERE                  40 (41.2%)   32 (31.1%)     43 (42.2%)     115 (38.1%) 
#>   dcd D.1.1.4.2                                                                    
#>     - Any Intensity -       37 (38.1%)   37 (35.9%)     41 (40.2%)     115 (38.1%) 
#>     MODERATE                37 (38.1%)   37 (35.9%)     41 (40.2%)     115 (38.1%) 
#> cl D.2                                                                             
#>   - Any Intensity -         35 (36.1%)   50 (48.5%)      49 (48%)      134 (44.4%) 
#>   MILD                      35 (36.1%)   50 (48.5%)      49 (48%)      134 (44.4%) 
#>   dcd D.2.1.5.3                                                                    
#>     - Any Intensity -       35 (36.1%)   50 (48.5%)      49 (48%)      134 (44.4%) 
#>     MILD                    35 (36.1%)   50 (48.5%)      49 (48%)      134 (44.4%) 
#> cl B.1                                                                             
#>   - Any Intensity -         37 (38.1%)   40 (38.8%)     36 (35.3%)     113 (37.4%) 
#>   SEVERE                    37 (38.1%)   40 (38.8%)     36 (35.3%)     113 (37.4%) 
#>   dcd B.1.1.1.1                                                                    
#>     - Any Intensity -       37 (38.1%)   40 (38.8%)     36 (35.3%)     113 (37.4%) 
#>     SEVERE                  37 (38.1%)   40 (38.8%)     36 (35.3%)     113 (37.4%) 
#> cl C.2                                                                             
#>   - Any Intensity -         26 (26.8%)   38 (36.9%)      49 (48%)      113 (37.4%) 
#>   MODERATE                  26 (26.8%)   38 (36.9%)      49 (48%)      113 (37.4%) 
#>   dcd C.2.1.2.1                                                                    
#>     - Any Intensity -       26 (26.8%)   38 (36.9%)      49 (48%)      113 (37.4%) 
#>     MODERATE                26 (26.8%)   38 (36.9%)      49 (48%)      113 (37.4%) 
#> cl C.1                                                                             
#>   - Any Intensity -          32 (33%)    38 (36.9%)     35 (34.3%)     105 (34.8%) 
#>   SEVERE                     32 (33%)    38 (36.9%)     35 (34.3%)     105 (34.8%) 
#>   dcd C.1.1.1.3                                                                    
#>     - Any Intensity -        32 (33%)    38 (36.9%)     35 (34.3%)     105 (34.8%) 
#>     SEVERE                   32 (33%)    38 (36.9%)     35 (34.3%)     105 (34.8%) 
#> ———————————————————————————————————————————————————————————————————————————————————
#> 
#> Investigator text for AEs encoded using MedDRA version xx.x.  All counts represent patients. Multiple occurrences of the same AE in one individual are counted once at the greatest intensity for this preferred term. To the SOC Overall row counts, a patient contributes only with the AE occuring with the greatest intensity within the SOC. Standard footnote defining AEs included in table (e.g., includes AEs with onset from first dose of study drug through 7 days after last dose of study drug).
aet03_1_lyt(
  armvar = "ACTARM",
  lbl_AEBODSYS = "Body System or Organ Class",
  lbl_AEDECOD = "Dictionary-Derived Term",
  lbl_overall = NULL,
  deco = std_deco("AET03")
)
#> A Pre-data Table Layout
#> 
#> Column-Split Structure:
#> ACTARM (lvls) 
#> 
#> Row-Split Structure:
#> AEBODSYS (lvls) -> AEDECOD (lvls) 
#>