AET04 Table 1 (Default) Adverse Events by Highest NCI CTACAE AE Grade Table 1.

The AET04 table provides an overview of adverse event with the highest NCI CTCAE grade per individual.

Usage

aet04_1_main(
  adam_db,
  arm_var = "ACTARM",
  lbl_overall = NULL,
  lbl_aebodsys = "MedDRA System Organ Class",
  lbl_aedecod = "MedDRA Preferred Term",
  grade_groups = NULL,
  deco = std_deco("AET04"),
  ...
)

aet04_1_lyt(
  arm_var,
  lbl_overall,
  lbl_aebodsys,
  lbl_aedecod,
  toxicity_grade,
  grade_groups,
  deco
)

aet04_1_pre(adam_db, ...)

aet04_1_post(tlg, prune_0 = TRUE, ...)

aet04_1

Format

An object of class chevron_t of length 1.

Arguments

adam_db

(dm) object containing the ADaM datasets

arm_var

(string) variable used for column splitting

lbl_overall

(string) label used for overall column, if set to NULL the overall column is omitted

lbl_aebodsys

(string) text label for AEBODSYS.

lbl_aedecod

(string) text label for AEDECOD.

grade_groups

(list) putting in correspondence toxicity grades and labels.

deco

(character) decoration with title, subtitles and main_footer content

...

not used.

toxicity_grade

(character) putting in correspondence toxicity levels.

tlg

(TableTree, Listing or ggplot) object typically produced by a main function.

prune_0

(flag) remove 0 count rows

Details

Numbers represent absolute numbers of patients and fraction of N, or absolute number of event when specified.

Remove zero-count rows unless overridden with prune_0 = FALSE.

Events with missing grading values are excluded.

Split columns by arm, typically ACTARM.

Does not include a total column by default.

Sort Body System or Organ Class and Dictionary-Derived Term by highest overall frequencies. Analysis Toxicity Grade is sorted by severity.

Functions

aet04_1_main(): Main TLG function

aet04_1_lyt(): Layout

aet04_1_pre(): Preprocessing

aet04_1_post(): Postprocessing

Note

adam_db object must contain an adae table with the columns "AEBODSYS", "AEDECOD" and "ATOXGR".

Examples

grade_groups <- list(
  "Grade 1-2" = c("1", "2"),
  "Grade 3-4" = c("3", "4"),
  "Grade 5" = c("5")
)

grade_groups <- list(
  "Grade 1-2" = c("1", "2"),
  "Grade 3-5" = c("3", "4", "5")
)

run(aet04_1, syn_data, grade_groups = grade_groups)
#>   MedDRA System Organ Class                                                             
#>     MedDRA Preferred Term                      A: Drug X    B: Placebo    C: Combination
#>                               Grade             (N=134)       (N=134)        (N=132)    
#>   ——————————————————————————————————————————————————————————————————————————————————————
#>   - Any adverse events -                                                                
#>                               - Any Grade -   122 (91.0%)   123 (91.8%)    120 (90.9%)  
#>                               Grade 1-2        13 (9.7%)    19 (14.2%)      11 (8.3%)   
#>                               1                7 (5.2%)      9 (6.7%)        4 (3.0%)   
#>                               2                6 (4.5%)      10 (7.5%)       7 (5.3%)   
#>                               Grade 3-5       109 (81.3%)   104 (77.6%)    109 (82.6%)  
#>                               3               18 (13.4%)    14 (10.4%)      16 (12.1%)  
#>                               4               15 (11.2%)    20 (14.9%)      18 (13.6%)  
#>                               5               76 (56.7%)    70 (52.2%)      75 (56.8%)  
#>   cl A.1                                                                                
#>     - Overall -                                                                         
#>                               - Any Grade -   78 (58.2%)    75 (56.0%)      89 (67.4%)  
#>                               Grade 1-2       78 (58.2%)    75 (56.0%)      89 (67.4%)  
#>                               1               30 (22.4%)    27 (20.1%)      39 (29.5%)  
#>                               2               48 (35.8%)    48 (35.8%)      50 (37.9%)  
#>     dcd A.1.1.1.1                                                                       
#>                               - Any Grade -   50 (37.3%)    45 (33.6%)      63 (47.7%)  
#>                               Grade 1-2       50 (37.3%)    45 (33.6%)      63 (47.7%)  
#>                               1               50 (37.3%)    45 (33.6%)      63 (47.7%)  
#>     dcd A.1.1.1.2                                                                       
#>                               - Any Grade -   48 (35.8%)    48 (35.8%)      50 (37.9%)  
#>                               Grade 1-2       48 (35.8%)    48 (35.8%)      50 (37.9%)  
#>                               2               48 (35.8%)    48 (35.8%)      50 (37.9%)  
#>   cl B.2                                                                                
#>     - Overall -                                                                         
#>                               - Any Grade -   79 (59.0%)    74 (55.2%)      85 (64.4%)  
#>                               Grade 1-2       30 (22.4%)    30 (22.4%)      33 (25.0%)  
#>                               1               30 (22.4%)    30 (22.4%)      33 (25.0%)  
#>                               Grade 3-5       49 (36.6%)    44 (32.8%)      52 (39.4%)  
#>                               3               49 (36.6%)    44 (32.8%)      52 (39.4%)  
#>     dcd B.2.2.3.1                                                                       
#>                               - Any Grade -   48 (35.8%)    54 (40.3%)      51 (38.6%)  
#>                               Grade 1-2       48 (35.8%)    54 (40.3%)      51 (38.6%)  
#>                               1               48 (35.8%)    54 (40.3%)      51 (38.6%)  
#>     dcd B.2.1.2.1                                                                       
#>                               - Any Grade -   49 (36.6%)    44 (32.8%)      52 (39.4%)  
#>                               Grade 3-5       49 (36.6%)    44 (32.8%)      52 (39.4%)  
#>                               3               49 (36.6%)    44 (32.8%)      52 (39.4%)  
#>   cl D.1                                                                                
#>     - Overall -                                                                         
#>                               - Any Grade -   79 (59.0%)    67 (50.0%)      80 (60.6%)  
#>                               Grade 3-5       79 (59.0%)    67 (50.0%)      80 (60.6%)  
#>                               3               29 (21.6%)    25 (18.7%)      29 (22.0%)  
#>                               5               50 (37.3%)    42 (31.3%)      51 (38.6%)  
#>     dcd D.1.1.1.1                                                                       
#>                               - Any Grade -   50 (37.3%)    42 (31.3%)      51 (38.6%)  
#>                               Grade 3-5       50 (37.3%)    42 (31.3%)      51 (38.6%)  
#>                               5               50 (37.3%)    42 (31.3%)      51 (38.6%)  
#>     dcd D.1.1.4.2                                                                       
#>                               - Any Grade -   48 (35.8%)    42 (31.3%)      50 (37.9%)  
#>                               Grade 3-5       48 (35.8%)    42 (31.3%)      50 (37.9%)  
#>                               3               48 (35.8%)    42 (31.3%)      50 (37.9%)  
#>   cl D.2                                                                                
#>     - Overall -                                                                         
#>                               - Any Grade -   47 (35.1%)    58 (43.3%)      57 (43.2%)  
#>                               Grade 1-2       47 (35.1%)    58 (43.3%)      57 (43.2%)  
#>                               1               47 (35.1%)    58 (43.3%)      57 (43.2%)  
#>     dcd D.2.1.5.3                                                                       
#>                               - Any Grade -   47 (35.1%)    58 (43.3%)      57 (43.2%)  
#>                               Grade 1-2       47 (35.1%)    58 (43.3%)      57 (43.2%)  
#>                               1               47 (35.1%)    58 (43.3%)      57 (43.2%)  
#>   cl B.1                                                                                
#>     - Overall -                                                                         
#>                               - Any Grade -   47 (35.1%)    49 (36.6%)      43 (32.6%)  
#>                               Grade 3-5       47 (35.1%)    49 (36.6%)      43 (32.6%)  
#>                               5               47 (35.1%)    49 (36.6%)      43 (32.6%)  
#>     dcd B.1.1.1.1                                                                       
#>                               - Any Grade -   47 (35.1%)    49 (36.6%)      43 (32.6%)  
#>                               Grade 3-5       47 (35.1%)    49 (36.6%)      43 (32.6%)  
#>                               5               47 (35.1%)    49 (36.6%)      43 (32.6%)  
#>   cl C.2                                                                                
#>     - Overall -                                                                         
#>                               - Any Grade -   35 (26.1%)    48 (35.8%)      55 (41.7%)  
#>                               Grade 1-2       35 (26.1%)    48 (35.8%)      55 (41.7%)  
#>                               2               35 (26.1%)    48 (35.8%)      55 (41.7%)  
#>     dcd C.2.1.2.1                                                                       
#>                               - Any Grade -   35 (26.1%)    48 (35.8%)      55 (41.7%)  
#>                               Grade 1-2       35 (26.1%)    48 (35.8%)      55 (41.7%)  
#>                               2               35 (26.1%)    48 (35.8%)      55 (41.7%)  
#>   cl C.1                                                                                
#>     - Overall -                                                                         
#>                               - Any Grade -   43 (32.1%)    46 (34.3%)      43 (32.6%)  
#>                               Grade 3-5       43 (32.1%)    46 (34.3%)      43 (32.6%)  
#>                               4               43 (32.1%)    46 (34.3%)      43 (32.6%)  
#>     dcd C.1.1.1.3                                                                       
#>                               - Any Grade -   43 (32.1%)    46 (34.3%)      43 (32.6%)  
#>                               Grade 3-5       43 (32.1%)    46 (34.3%)      43 (32.6%)  
#>                               4               43 (32.1%)    46 (34.3%)      43 (32.6%)