AET05 Table 1 (Default) Adverse Event Rate Adjusted for Patient-Years at Risk - First Occurrence.

The AET05 table produces the standard adverse event rate adjusted for patient-years at risk summary considering first occurrence.

Usage

aet05_main(adam_db, arm_var = "ACTARM", ...)

aet05_pre(adam_db, ...)

aet05_post(tlg, prune_0 = FALSE, ...)

aet05

Format

An object of class chevron_t of length 1.

Arguments

adam_db

(list of data.frames) object containing the ADaM datasets

arm_var

(string) the arm variable used for arm splitting.

...

Further arguments passed to tern::control_incidence_rate().

tlg

(TableTree, Listing or ggplot) object typically produced by a main function.

prune_0

(flag) remove 0 count rows

Details

Total patient-years at risk is the sum over all patients of the time intervals (in years).

Split columns by arm, typically ACTARM.

Split rows by parameter code.

AVAL is patient-years at risk.

n_events is the number of adverse events observed.

The table allows confidence level to be adjusted, default is 95%.

Keep zero count rows by default.

Functions

aet05_main(): Main TLG function

aet05_pre(): Preprocessing

aet05_post(): Postprocessing

Note

adam_db object must contain an adaette table with the columns "PARAMCD", "PARAM", "AVAL", and "CNSR".

Examples

library(dplyr)
#> 
#> Attaching package: ‘dplyr’
#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag
#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union
library(dunlin)

proc_data <- log_filter(syn_data, PARAMCD == "AETTE1", "adaette")

run(aet05, proc_data)
#>                                                     A: Drug X        B: Placebo      C: Combination
#>                                                      (N=134)           (N=134)          (N=132)    
#>   —————————————————————————————————————————————————————————————————————————————————————————————————
#>   Time to first occurrence of any adverse event                                                    
#>     Total patient-years at risk                       162.4             103.8            172.6     
#>     Number of adverse events observed                   78               104               67      
#>     AE rate per 100 patient-years                     48.03            100.15            38.82     
#>     95% CI                                        (37.37, 58.69)   (80.90, 119.40)   (29.53, 48.12)

run(aet05, proc_data, conf_level = 0.90, conf_type = "exact")
#>                                                     A: Drug X        B: Placebo      C: Combination
#>                                                      (N=134)           (N=134)          (N=132)    
#>   —————————————————————————————————————————————————————————————————————————————————————————————————
#>   Time to first occurrence of any adverse event                                                    
#>     Total patient-years at risk                       162.4             103.8            172.6     
#>     Number of adverse events observed                   78               104               67      
#>     AE rate per 100 patient-years                     48.03            100.15            38.82     
#>     90% CI                                        (39.45, 57.99)   (84.56, 117.87)   (31.36, 47.58)