chevron_tlg Class

The chevron_tlg class associates a preprocess function, a main tlg function and AdAM tables names and optionally a postprocess function.

chevron_tlg(
  main = function(adam_db, ...) adam_db,
  preprocess = function(adam_db, ...) adam_db,
  postprocess = function(tlg, ...) tlg,
  adam_datasets = NA_character_
)

aet01_1

aet01_2

aet02_1

aet02_2

aet02_3

aet03_1

aet04_1

cmt01a_1

cmt01a_2

cmt01a_3

cmt02_pt_1

dmt01_1

dst01_1

dst01_2

dst01_3

dtht01_1

egt01_1

egt02_1

egt02_2

ext01_1

ext01_2

lbt01_1

mht01_1

vst01_1

vst02_1

vst02_2

Format

An object of class chevron_tlg of length 1.

An object of class chevron_tlg of length 1.

An object of class chevron_tlg of length 1.

An object of class chevron_tlg of length 1.

An object of class chevron_tlg of length 1.

An object of class chevron_tlg of length 1.

An object of class chevron_tlg of length 1.

An object of class chevron_tlg of length 1.

An object of class chevron_tlg of length 1.

An object of class chevron_tlg of length 1.

An object of class chevron_tlg of length 1.

An object of class chevron_tlg of length 1.

An object of class chevron_tlg of length 1.

An object of class chevron_tlg of length 1.

An object of class chevron_tlg of length 1.

An object of class chevron_tlg of length 1.

An object of class chevron_tlg of length 1.

An object of class chevron_tlg of length 1.

An object of class chevron_tlg of length 1.

An object of class chevron_tlg of length 1.

An object of class chevron_tlg of length 1.

An object of class chevron_tlg of length 1.

An object of class chevron_tlg of length 1.

An object of class chevron_tlg of length 1.

An object of class chevron_tlg of length 1.

An object of class chevron_tlg of length 1.

Arguments

main

(function) returning a tlg. Typically one of the _main function of chevron.

preprocess

(function) returning a pre processed dm object amenable to tlg creation. Typically one of the _pre function of chevron.

postprocess

(function) returning a post processed tlg.

adam_datasets

(character) representing the name of the table from an AdAM data set required for tlg creation.

Functions

chevron_tlg: Default Constructor

Slots

main

(function) returning a tlg. Typically one of the *_main function from chevron.

preprocess

(function) returning a pre processed dm object amenable to tlg creation. Typically one of the *_pre function from chevron.

postprocess

(function) returning a post processed tlg.

adam_datasets

(character) representing the name of the tables from an AdAM data set required for tlg creation.

Note

To ensure the correct execution of the pipeline workflow additional validation criteria are:

the first argument of the main function must be adam_db, the input dm object to preprocess.

the first argument of the preprocess function must be adam_db, the input dm object to create tlg output.

the first argument of the postprocess function must be tlg, the input TableTree object to post process.

See also

Examples

x <- chevron_tlg(aet01_1_main, aet01_1_pre, adam_datasets = c("adsl", "adae"))