CMT01A Table 3 (Supplementary) Concomitant Medication by Medication Class and Preferred Name (Total number of treatments per medication class suppressed).

A concomitant medication table with the number of subjects and the total number of treatments by medication class sorted alphabetically and medication name sorted by frequencies presented without the total number of treatments per medication.

Usage

cmt01a_3_main(
  adam_db,
  arm_var = "ARM",
  medcat_var = "ATC2",
  lbl_medcat_var = "ATC Class Level 2",
  medname_var = "CMDECOD",
  lbl_medname_var = "Other Treatment",
  lbl_overall = NULL,
  deco = std_deco("CMT01A"),
  ...
)

cmt01a_3_lyt(
  arm_var = "ARM",
  lbl_overall = NULL,
  medcat_var = "ATC2",
  lbl_medcat_var = "ATC Class Level 2",
  medname_var = "CMDECOD",
  lbl_medname_var = "Other Treatment",
  deco = std_deco("CMT01A")
)

cmt01a_3_pre(adam_db, medcat_var = "ATC2", medname_var = "CMDECOD", ...)

cmt01a_3_post(
  tlg,
  prune_0 = TRUE,
  medcat_var = "ATC2",
  medname_var = "CMDECOD",
  ...
)

cmt01a_3

Format

An object of class chevron_t of length 1.

Arguments

adam_db

(list of data.frames) object containing the ADaM datasets

arm_var

(string) variable used for column splitting

medcat_var

(string) the variable defining the medication category. By default ATC2.

lbl_medcat_var

(string) the label for the medication category.

medname_var

(string) the variable defining the medication name. By default CMDECOD.

lbl_medname_var

(string) the label for the medication name.

lbl_overall

(string) label used for overall column, if set to NULL the overall column is omitted

deco

(character) decoration with title, subtitles and main_footer content

...

not used.

tlg

(TableTree, Listing or ggplot) object typically produced by a main function.

prune_0

(flag) remove 0 count rows

Details

Data should be filtered for concomitant medication. (ATIREL == "CONCOMITANT").

Numbers represent absolute numbers of subjects and fraction of N, or absolute numbers when specified.

Remove zero-count rows unless overridden with prune_0 = FALSE.

Split columns by arm.

Does not include a total column by default.

Sort by medication class alphabetically and within medication class by decreasing total number of patients with the specific medication.

Functions

cmt01a_3_main(): Main TLG function

cmt01a_3_lyt(): Layout

cmt01a_3_pre(): Preprocessing

cmt01a_3_post(): Preprocessing

Note

adam_db object must contain an adcm table with the columns specified in medcat_var and medname_var as well as "CMSEQ".

Examples

library(dplyr)

proc_data <- syn_data
proc_data$adcm <- proc_data$adcm %>%
  filter(ATIREL == "CONCOMITANT")

run(cmt01a_3, proc_data)
#>   ATC Class Level 2                                         A: Drug X    B: Placebo    C: Combination
#>     Other Treatment                                          (N=134)       (N=134)        (N=132)    
#>   ———————————————————————————————————————————————————————————————————————————————————————————————————
#>   Total number of patients with at least one treatment     117 (87.3%)   116 (86.6%)    116 (87.9%)  
#>   Total number of treatments                                   415           414            460      
#>   ATCCLAS2 A                                                                                         
#>     Total number of patients with at least one treatment   75 (56.0%)    79 (59.0%)      81 (61.4%)  
#>     medname A_2/3                                          53 (39.6%)    50 (37.3%)      56 (42.4%)  
#>     medname A_3/3                                          45 (33.6%)    54 (40.3%)      48 (36.4%)  
#>   ATCCLAS2 A p2                                                                                      
#>     Total number of patients with at least one treatment   45 (33.6%)    54 (40.3%)      48 (36.4%)  
#>     medname A_3/3                                          45 (33.6%)    54 (40.3%)      48 (36.4%)  
#>   ATCCLAS2 B                                                                                         
#>     Total number of patients with at least one treatment   83 (61.9%)    74 (55.2%)      88 (66.7%)  
#>     medname B_1/4                                          52 (38.8%)    57 (42.5%)      59 (44.7%)  
#>     medname B_4/4                                          50 (37.3%)    45 (33.6%)      55 (41.7%)  
#>   ATCCLAS2 B p2                                                                                      
#>     Total number of patients with at least one treatment   52 (38.8%)    57 (42.5%)      59 (44.7%)  
#>     medname B_1/4                                          52 (38.8%)    57 (42.5%)      59 (44.7%)  
#>   ATCCLAS2 B p3                                                                                      
#>     Total number of patients with at least one treatment   52 (38.8%)    57 (42.5%)      59 (44.7%)  
#>     medname B_1/4                                          52 (38.8%)    57 (42.5%)      59 (44.7%)  
#>   ATCCLAS2 C                                                                                         
#>     Total number of patients with at least one treatment   82 (61.2%)    84 (62.7%)      89 (67.4%)  
#>     medname C_2/2                                          52 (38.8%)    58 (43.3%)      60 (45.5%)  
#>     medname C_1/2                                          51 (38.1%)    50 (37.3%)      56 (42.4%)  
#>   ATCCLAS2 C p2                                                                                      
#>     Total number of patients with at least one treatment   82 (61.2%)    84 (62.7%)      89 (67.4%)  
#>     medname C_2/2                                          52 (38.8%)    58 (43.3%)      60 (45.5%)  
#>     medname C_1/2                                          51 (38.1%)    50 (37.3%)      56 (42.4%)  
#>   ATCCLAS2 C p3                                                                                      
#>     Total number of patients with at least one treatment   52 (38.8%)    58 (43.3%)      60 (45.5%)  
#>     medname C_2/2                                          52 (38.8%)    58 (43.3%)      60 (45.5%)