DST01 Table 1 (Default) Patient Disposition Table 1.

The DST01 Disposition Table provides an overview of patients study completion. For patients who discontinued the study a reason is provided.

Usage

dst01_main(
  adam_db,
  arm_var = "ARM",
  study_status_var = "EOSSTT",
  detail_vars = list(Discontinued = c("DCSREAS")),
  trt_status_var = NULL,
  lbl_overall = "All Patients",
  ...
)

dst01_pre(adam_db, ...)

dst01_post(tlg, prune_0 = TRUE, ...)

dst01

Format

An object of class chevron_t of length 1.

Arguments

adam_db

(list of data.frames) object containing the ADaM datasets

arm_var

(string) variable. Usually one of ARM, ACTARM, TRT01A, or TRT01A.

study_status_var

(string) variable used to define patient status. Default is EOSSTT, however can also be a variable name with the pattern EOPxxSTT where xx must be substituted by 2 digits referring to the analysis period.

detail_vars

Named (list) of grouped display of study_status_var. The names must be subset of unique levels of study_status_var.

trt_status_var

(string) variable of treatment status.

lbl_overall

(string) label used for overall column, if set to NULL the overall column is omitted

...

not used.

tlg

(TableTree, Listing or ggplot) object typically produced by a main function.

prune_0

(flag) remove 0 count rows

Details

Default patient disposition table summarizing the reasons for patients withdrawal.

Numbers represent absolute numbers of patients and fraction of N.

Remove zero-count rows.

Split columns by arm.

Include a total column by default.

Sort withdrawal reasons by alphabetic order.

Functions

dst01_main(): Main TLG function

dst01_pre(): Preprocessing

dst01_post(): Postprocessing

Note

adam_db object must contain an adsl table with the columns specified by status_var and disc_reason_var.

Examples

run(dst01, syn_data, detail_vars = list(Ongoing = "STDONS"))
#>                           A: Drug X    B: Placebo   C: Combination   All Patients
#>                            (N=134)      (N=134)        (N=132)         (N=400)   
#>   ———————————————————————————————————————————————————————————————————————————————
#>   Completed               68 (50.7%)   66 (49.3%)     73 (55.3%)     207 (51.7%) 
#>   Ongoing                 24 (17.9%)   28 (20.9%)     21 (15.9%)      73 (18.2%) 
#>     Alive: Ongoing         8 (6.0%)     6 (4.5%)       8 (6.1%)       22 (5.5%)  
#>     Alive: In Follow-up   16 (11.9%)   22 (16.4%)     13 (9.8%)       51 (12.8%) 
run(dst01, syn_data, detail_vars = list(Discontinued = "DCSREAS", Ongoing = "STDONS"))
#>                                     A: Drug X    B: Placebo   C: Combination   All Patients
#>                                      (N=134)      (N=134)        (N=132)         (N=400)   
#>   —————————————————————————————————————————————————————————————————————————————————————————
#>   Completed                         68 (50.7%)   66 (49.3%)     73 (55.3%)     207 (51.7%) 
#>   Ongoing                           24 (17.9%)   28 (20.9%)     21 (15.9%)      73 (18.2%) 
#>     Alive: Ongoing                   8 (6.0%)     6 (4.5%)       8 (6.1%)       22 (5.5%)  
#>     Alive: In Follow-up             16 (11.9%)   22 (16.4%)     13 (9.8%)       51 (12.8%) 
#>   Discontinued                      42 (31.3%)   40 (29.9%)     38 (28.8%)     120 (30.0%) 
#>     ADVERSE EVENT                    3 (2.2%)     6 (4.5%)       5 (3.8%)       14 (3.5%)  
#>     DEATH                           25 (18.7%)   23 (17.2%)     22 (16.7%)      70 (17.5%) 
#>     LACK OF EFFICACY                 2 (1.5%)     2 (1.5%)       3 (2.3%)        7 (1.8%)  
#>     PHYSICIAN DECISION               2 (1.5%)     3 (2.2%)       2 (1.5%)        7 (1.8%)  
#>     PROTOCOL VIOLATION               5 (3.7%)     3 (2.2%)       4 (3.0%)       12 (3.0%)  
#>     WITHDRAWAL BY PARENT/GUARDIAN    4 (3.0%)     2 (1.5%)       1 (0.8%)        7 (1.8%)  
#>     WITHDRAWAL BY SUBJECT            1 (0.7%)     1 (0.7%)       1 (0.8%)        3 (0.8%)  
run(
  dst01, syn_data,
  detail_vars = list(
    Discontinued = c("DCSREASGP", "DCSREAS"),
    Ongoing = "STDONS"
  )
)
#>                                       A: Drug X    B: Placebo   C: Combination   All Patients
#>                                        (N=134)      (N=134)        (N=132)         (N=400)   
#>   ———————————————————————————————————————————————————————————————————————————————————————————
#>   Completed                           68 (50.7%)   66 (49.3%)     73 (55.3%)     207 (51.7%) 
#>   Ongoing                             24 (17.9%)   28 (20.9%)     21 (15.9%)      73 (18.2%) 
#>     Alive: Ongoing                     8 (6.0%)     6 (4.5%)       8 (6.1%)       22 (5.5%)  
#>     Alive: In Follow-up               16 (11.9%)   22 (16.4%)     13 (9.8%)       51 (12.8%) 
#>   Discontinued                        42 (31.3%)   40 (29.9%)     38 (28.8%)     120 (30.0%) 
#>     Safety                                                                                   
#>       ADVERSE EVENT                    3 (2.2%)     6 (4.5%)       5 (3.8%)       14 (3.5%)  
#>       DEATH                           25 (18.7%)   23 (17.2%)     22 (16.7%)      70 (17.5%) 
#>     Non-Safety                                                                               
#>       LACK OF EFFICACY                 2 (1.5%)     2 (1.5%)       3 (2.3%)        7 (1.8%)  
#>       PHYSICIAN DECISION               2 (1.5%)     3 (2.2%)       2 (1.5%)        7 (1.8%)  
#>       PROTOCOL VIOLATION               5 (3.7%)     3 (2.2%)       4 (3.0%)       12 (3.0%)  
#>       WITHDRAWAL BY PARENT/GUARDIAN    4 (3.0%)     2 (1.5%)       1 (0.8%)        7 (1.8%)  
#>       WITHDRAWAL BY SUBJECT            1 (0.7%)     1 (0.7%)       1 (0.8%)        3 (0.8%)