Patient Disposition Table

The DST01 Disposition Table provides an overview of patients study completion. For patients who discontinued the study a reason is provided.

dst01_1(
  adsl,
  adae,
  armvar = .study$armvar,
  lbl_overall = .study$lbl_overall,
  status = .study$status,
  reason = .study$reason,
  deco = std_deco("DST01"),
  .study = list(armvar = "ARM", lbl_overall = "All patients", status = "EOSSTT", reason
    = "DCSREAS")
)

Arguments

adsl
Subject Level Analysis Dataset

adae
Adverse Events Analysis Dataset

armvar
(string) variable. Usually one of ARM, ACTARM, TRT01A, or TRT01A.

lbl_overall
label used for overall column, if set to NULL the overall column is omitted

status
(string) variable used to define patient status. Default is EOSSTT, however can also be a variable name with the pattern EOPxxSTT where xx must be substituted by 2 digits referring to the analysis period.

reason
(string) variable used to define reason for patient withdrawal. Default is DCSREAS, however can also be a variable with the pattern DCPxxRS where xx must be substituted by 2 digits referring to the analysis period.

deco
decoration list with title, subtitles and main_footer content

.study
a list with default values for the arguments of the function

Details

Default patient disposition table summarizing the reasons for patients withdrawal

Numbers represent absolute numbers of patients and fraction of N

Remove zero-count rows

Split columns by arm

Include a total column by default

Sort withdrawal reasons by alphabetic order

Examples

library(scda)
library(dplyr)

sd <- synthetic_cdisc_data("rcd_2021_03_22")
adsl <- sd$adsl
adae <- sd$adae %>%
 mutate(ANL01FL = 'Y')

dst01_1(adsl, adae)

#> Main Title
#> Protocol: {{protocol}}, Snapshot: {{snapshot}}, Snapshot Date: {{snapshot-date}}, Cutoff Date: {{cutoff-date}}
#> 
#> ---------------------------------------------------------------------------------------------
#>                                    A: Drug X     B: Placebo    C: Combination   All patients 
#>                                     (N=609)        (N=622)        (N=703)         (N=1934)   
#> ---------------------------------------------------------------------------------------------
#> Completed Study                   306 (50.25%)   306 (49.2%)    421 (59.89%)    1033 (53.41%)
#> Discontinued Study                166 (27.3%)    207 (33.3%)    179 (25.5%)      552 (28.5%) 
#>   ADVERSE EVENT                    21 (3.4%)      8 (1.3%)       13 (1.8%)        42 (2.2%)  
#>   DEATH                           106 (17.4%)     118 (19%)      94 (13.4%)      318 (16.4%) 
#>   LACK OF EFFICACY                  12 (2%)       3 (0.5%)        4 (0.6%)         19 (1%)   
#>   PHYSICIAN DECISION                   0          10 (1.6%)      38 (5.4%)        48 (2.5%)  
#>   PROTOCOL VIOLATION               16 (2.6%)      35 (5.6%)      19 (2.7%)        70 (3.6%)  
#>   WITHDRAWAL BY PARENT/GUARDIAN     9 (1.5%)      23 (3.7%)      11 (1.6%)        43 (2.2%)  
#>   WITHDRAWAL BY SUBJECT             2 (0.3%)      10 (1.6%)          0            12 (0.6%)  
#> ---------------------------------------------------------------------------------------------
#> 
#> Footnotes go here

dst01_1(adsl, adae, lbl_overall = "")

#>                                    A: Drug X     B: Placebo    C: Combination
#>                                     (N=609)        (N=622)        (N=703)    
#> -----------------------------------------------------------------------------
#> Completed Study                   306 (50.25%)   306 (49.2%)    421 (59.89%) 
#> Discontinued Study                166 (27.3%)    207 (33.3%)    179 (25.5%)  
#>   ADVERSE EVENT                    21 (3.4%)      8 (1.3%)       13 (1.8%)   
#>   DEATH                           106 (17.4%)     118 (19%)      94 (13.4%)  
#>   LACK OF EFFICACY                  12 (2%)       3 (0.5%)        4 (0.6%)   
#>   PHYSICIAN DECISION                   0          10 (1.6%)      38 (5.4%)   
#>   PROTOCOL VIOLATION               16 (2.6%)      35 (5.6%)      19 (2.7%)   
#>   WITHDRAWAL BY PARENT/GUARDIAN     9 (1.5%)      23 (3.7%)      11 (1.6%)   
#>   WITHDRAWAL BY SUBJECT             2 (0.3%)      10 (1.6%)          0       
#> -----------------------------------------------------------------------------
#> 
#> Footnotes go here