DST01 Table 3 (Supplementary) Patient Disposition Table 3

The DST01_3 Disposition Table provides an overview of patients study treatment status. For patients who discontinued the study, the reason provided is categorized as "Safety" or "Non-Safety" issue.

dst01_3(
  adsl,
  armvar = .study$armvar,
  lbl_overall = .study$lbl_overall,
  status = .study$status,
  reason = .study$reason,
  prune_0 = TRUE,
  status_treatment = .study$status_treatment,
  deco = std_deco("DST01"),
  .study = list(armvar = "ARM", lbl_overall = "All patients", status = "EOSSTT", reason
    = "DCSREAS", status_treatment = "EOTSTT")
)

dst01_3_lyt(
  armvar = .study$armvar,
  lbl_overall = .study$lbl_overall,
  status_treatment = .study$status,
  deco = std_deco("DST01"),
  completed_lbl = "COMPLETED",
  discontinued_lbl = "DISCONTINUED",
  ongoing_lbl = "ONGOING",
  .study = list(armvar = "ARM", lbl_overall = "All patients", status_treatment =
    "EOTSTT")
)

Arguments

adsl
(data.frame) Subject Level Analysis Dataset

armvar
(string) variable used for column splitting

lbl_overall
(string) label used for overall column, if set to NULL the overall column is omitted

status
(string) variable used to define patient status. Default is EOSSTT, however can also be a variable name with the pattern EOPxxSTT where xx must be substituted by 2 digits referring to the analysis period.

reason
(string) variable used to define reason for patient withdrawal. Default is DCSREAS, however can also be a variable with the pattern DCPxxRS where xx must be substituted by 2 digits referring to the analysis period.

prune_0
(logical) remove 0 count rows

status_treatment
(string) variable used to define the treatment status of the patients. Default is EOTSTT, however can also be a variable with the pattern EOTxxSTT where xx must be substituted by 2 digits referring to the analysis period.

deco
(list) decoration with title, subtitles and main_footer content

.study
(list) with default values for the arguments of the function

completed_lbl
(string) associated with completed treatment and found in the columns given by status_treatment. By Default COMPLETED.

discontinued_lbl
(string) associated with discontinued treatment and found in the columns given by status_treatment. By Default DISCONTINUED.

ongoing_lbl
(string) associated with ongoing treatment and found in the columns given by status_treatment. By Default `ONGOING.

Value

Details

Non-standard disposition table summarizing the reasons for patient withdrawal and treatment status.

Withdrawal reasons are grouped into Safety and Non-Safety issues.

Safety issues include Death and Adverse Events.

Numbers represent absolute numbers of patients and fraction of N.

Remove zero-count rows.

Split columns by arm.

Include a total column by default.

Sort withdrawal reasons by alphabetic order.

Functions

dst01_3_lyt:

Examples

library(scda)
library(dplyr)
sd <- synthetic_cdisc_data("rcd_2021_03_22")

adsl <- sd$adsl %>%
 mutate(ANL01FL = 'Y',
        EOTSTT = as.factor(sample(c("ONGOING", "COMPLETED", "DISCONTINUED"),
                        nrow(sd$adsl),
                        replace = TRUE)))

dst01_3(adsl)

#>                                      A: Drug X    B: Placebo    C: Combination   All patients
#>                                       (N=134)       (N=134)        (N=132)         (N=400)   
#> ---------------------------------------------------------------------------------------------
#> Completed Study                     69 (51.49%)   69 (51.49%)    72 (54.55%)     210 (52.5%) 
#> Ongoing Study                       27 (20.15%)   22 (16.42%)    21 (15.91%)      70 (17.5%) 
#> Discontinued Study                  38 (28.4%)    43 (32.1%)      39 (29.5%)      120 (30%)  
#>   Safety                            28 (20.9%)    27 (20.1%)      21 (15.9%)       76 (19%)  
#>     ADVERSE EVENT                    6 (4.5%)      1 (0.7%)        2 (1.5%)        9 (2.2%)  
#>     DEATH                           22 (16.4%)    26 (19.4%)      19 (14.4%)      67 (16.8%) 
#>   Non Safety                         10 (7.5%)    16 (11.9%)      18 (13.6%)       44 (11%)  
#>     LACK OF EFFICACY                 2 (1.5%)      1 (0.7%)        2 (1.5%)        5 (1.2%)  
#>     PHYSICIAN DECISION                   0         1 (0.7%)        7 (5.3%)         8 (2%)   
#>     PROTOCOL VIOLATION                4 (3%)       7 (5.2%)        6 (4.5%)       17 (4.2%)  
#>     WITHDRAWAL BY PARENT/GUARDIAN    2 (1.5%)      5 (3.7%)        3 (2.3%)       10 (2.5%)  
#>     WITHDRAWAL BY SUBJECT            2 (1.5%)      2 (1.5%)           0             4 (1%)   
#> Completed Treatment                 50 (37.31%)   42 (31.34%)    46 (34.85%)     138 (34.5%) 
#> Ongoing Treatment                   45 (33.58%)   43 (32.09%)    50 (37.88%)     138 (34.5%) 
#> Discontinued Treatment              39 (29.1%)    49 (36.57%)    36 (27.27%)      124 (31%)  

dst01_3(adsl, lbl_overall = "")

#>                                      A: Drug X    B: Placebo    C: Combination
#>                                       (N=134)       (N=134)        (N=132)    
#> ------------------------------------------------------------------------------
#> Completed Study                     69 (51.49%)   69 (51.49%)    72 (54.55%)  
#> Ongoing Study                       27 (20.15%)   22 (16.42%)    21 (15.91%)  
#> Discontinued Study                  38 (28.4%)    43 (32.1%)      39 (29.5%)  
#>   Safety                            28 (20.9%)    27 (20.1%)      21 (15.9%)  
#>     ADVERSE EVENT                    6 (4.5%)      1 (0.7%)        2 (1.5%)   
#>     DEATH                           22 (16.4%)    26 (19.4%)      19 (14.4%)  
#>   Non Safety                         10 (7.5%)    16 (11.9%)      18 (13.6%)  
#>     LACK OF EFFICACY                 2 (1.5%)      1 (0.7%)        2 (1.5%)   
#>     PHYSICIAN DECISION                   0         1 (0.7%)        7 (5.3%)   
#>     PROTOCOL VIOLATION                4 (3%)       7 (5.2%)        6 (4.5%)   
#>     WITHDRAWAL BY PARENT/GUARDIAN    2 (1.5%)      5 (3.7%)        3 (2.3%)   
#>     WITHDRAWAL BY SUBJECT            2 (1.5%)      2 (1.5%)           0       
#> Completed Treatment                 50 (37.31%)   42 (31.34%)    46 (34.85%)  
#> Ongoing Treatment                   45 (33.58%)   43 (32.09%)    50 (37.88%)  
#> Discontinued Treatment              39 (29.1%)    49 (36.57%)    36 (27.27%)  

adsl <- adsl %>% filter(EOSSTT != "COMPLETED")
dst01_3(adsl)

#>                                      A: Drug X    B: Placebo    C: Combination   All patients
#>                                       (N=65)        (N=65)          (N=60)         (N=190)   
#> ---------------------------------------------------------------------------------------------
#> Completed Study                       0 (0%)        0 (0%)          0 (0%)          0 (0%)   
#> Ongoing Study                       27 (41.54%)   22 (33.85%)      21 (35%)      70 (36.84%) 
#> Discontinued Study                  38 (58.5%)    43 (66.2%)       39 (65%)      120 (63.2%) 
#>   Safety                            28 (43.1%)    27 (41.5%)       21 (35%)        76 (40%)  
#>     ADVERSE EVENT                    6 (9.2%)      1 (1.5%)        2 (3.3%)        9 (4.7%)  
#>     DEATH                           22 (33.8%)     26 (40%)       19 (31.7%)      67 (35.3%) 
#>   Non Safety                        10 (15.4%)    16 (24.6%)       18 (30%)       44 (23.2%) 
#>     LACK OF EFFICACY                 2 (3.1%)      1 (1.5%)        2 (3.3%)        5 (2.6%)  
#>     PHYSICIAN DECISION                   0         1 (1.5%)       7 (11.7%)        8 (4.2%)  
#>     PROTOCOL VIOLATION               4 (6.2%)      7 (10.8%)       6 (10%)        17 (8.9%)  
#>     WITHDRAWAL BY PARENT/GUARDIAN    2 (3.1%)      5 (7.7%)         3 (5%)        10 (5.3%)  
#>     WITHDRAWAL BY SUBJECT            2 (3.1%)      2 (3.1%)           0            4 (2.1%)  
#> Completed Treatment                 23 (35.38%)   16 (24.62%)    19 (31.67%)     58 (30.53%) 
#> Ongoing Treatment                   25 (38.46%)   24 (36.92%)      21 (35%)      70 (36.84%) 
#> Discontinued Treatment              17 (26.15%)   25 (38.46%)    20 (33.33%)     62 (32.63%) 

dst01_3_lyt(armvar = "ACTARM",
status_treatment = "EOTxx01")

#> A Pre-data Table Layout
#> 
#> Column-Split Structure:
#> ACTARM (lvls) 
#>  (all obs) 
#> 
#> Row-Split Structure:
#> ( (** multivar analysis **) -> EOTxx01, EOTxx01, EOTxx01 (** multivar analysis **) -> ) (** multivar analysis **) 
#>