DTHT01 Table 1 (Default) Death Table.

A description of the causes of death optionally with the breakdown of the OTHER category and/or post-study reporting of death.

dtht01_1(
  adam_db,
  armvar = .study$actualarm,
  time_since_last_dose = FALSE,
  other_category = FALSE,
  lbl_overall = .study$lbl_overall,
  prune_0 = TRUE,
  deco = std_deco("DTHT01"),
  .study = list(actualarm = "ACTARM", lbl_overall = NULL)
)

dtht01_1_lyt(
  armvar = .study$actualarm,
  lbl_overall = .study$lbl_overall,
  other_category = TRUE,
  deco = std_deco("DTHT01"),
  .study = list(actualarm = "ACTARM", lbl_overall = NULL)
)

dtht01_1_opt_lyt(
  armvar = .study$actualarm,
  lbl_overall = .study$lbl_overall,
  deco = std_deco("DTHT01"),
  .study = list(actualarm = "ACTARM", lbl_overall = NULL)
)

Arguments

adam_db

(dm) object containing the ADaM datasets

armvar

(character) variable used for column splitting

time_since_last_dose

(logical) should the time to event information be displayed.

other_category

(logical) should the breakdown of the OTHER category be displayed.

lbl_overall

(character) label used for overall column, if set to NULL the overall column is omitted

prune_0

(logical) remove 0 count rows

deco

(character) decoration with title, subtitles and main_footer content

.study

(list) with default values for the arguments of the function

Details

Numbers represent absolute numbers of subjects and fraction of N, or absolute numbers when specified.

Remove zero-count rows unless overridden with prune_0 = FALSE.

Does not include a total column by default.

Functions

dtht01_1_lyt:

dtht01_1_opt_lyt:

Examples

library(dm)

db <- syn_test_data() %>%
  dm_zoom_to(adsl) %>%
  mutate(DTHCAT = tern::explicit_na(DTHCAT)) %>%
  mutate(LDDTHGR1 = tern::explicit_na(LDDTHGR1)) %>%
  dm_update_zoomed() %>%
  preprocess_data("dtht01_1")

dtht01_1(adam_db = db)
#> Deaths: {Specify Population}
#> Protocol: {{protocol}}, Snapshot: {{snapshot}}, Snapshot Date: {{snapshot-date}}, Cutoff Date: {{cutoff-date}}
#> 
#> —————————————————————————————————————————————————————————————————
#>                          A: Drug X    B: Placebo   C: Combination
#>                           (N=134)      (N=134)        (N=132)    
#> —————————————————————————————————————————————————————————————————
#> Total number of deaths   22 (16.4%)   26 (19.4%)     19 (14.4%)  
#> Primary cause of death                                           
#>   n                          22           26             19      
#>   ADVERSE EVENT          6 (27.3%)    12 (46.2%)     7 (36.8%)   
#>   PROGRESSIVE DISEASE    7 (31.8%)    9 (34.6%)      8 (42.1%)   
#>   OTHER                  9 (40.9%)    5 (19.2%)      4 (21.1%)   
#> —————————————————————————————————————————————————————————————————
#> 
#> If applicable, include the footnote highlighted in grey for post-study deaths collected from public records.
dtht01_1(adam_db = db, other_category = FALSE)
#> Deaths: {Specify Population}
#> Protocol: {{protocol}}, Snapshot: {{snapshot}}, Snapshot Date: {{snapshot-date}}, Cutoff Date: {{cutoff-date}}
#> 
#> —————————————————————————————————————————————————————————————————
#>                          A: Drug X    B: Placebo   C: Combination
#>                           (N=134)      (N=134)        (N=132)    
#> —————————————————————————————————————————————————————————————————
#> Total number of deaths   22 (16.4%)   26 (19.4%)     19 (14.4%)  
#> Primary cause of death                                           
#>   n                          22           26             19      
#>   ADVERSE EVENT          6 (27.3%)    12 (46.2%)     7 (36.8%)   
#>   PROGRESSIVE DISEASE    7 (31.8%)    9 (34.6%)      8 (42.1%)   
#>   OTHER                  9 (40.9%)    5 (19.2%)      4 (21.1%)   
#> —————————————————————————————————————————————————————————————————
#> 
#> If applicable, include the footnote highlighted in grey for post-study deaths collected from public records.
dtht01_1(adam_db = db, time_since_last_dose = TRUE)
#> Deaths: {Specify Population}
#> Protocol: {{protocol}}, Snapshot: {{snapshot}}, Snapshot Date: {{snapshot-date}}, Cutoff Date: {{cutoff-date}}
#> 
#> ————————————————————————————————————————————————————————————————————————————————————————————————————
#>                                                             A: Drug X    B: Placebo   C: Combination
#>                                                              (N=134)      (N=134)        (N=132)    
#> ————————————————————————————————————————————————————————————————————————————————————————————————————
#> Total number of deaths                                      22 (16.4%)   26 (19.4%)     19 (14.4%)  
#> Primary cause of death                                                                              
#>   n                                                             22           26             19      
#>   ADVERSE EVENT                                             6 (27.3%)    12 (46.2%)     7 (36.8%)   
#>   PROGRESSIVE DISEASE                                       7 (31.8%)    9 (34.6%)      8 (42.1%)   
#>   OTHER                                                     9 (40.9%)    5 (19.2%)      4 (21.1%)   
#> Days from last drug administration                                                                  
#>   n                                                             22           26             19      
#>   <=30                                                      12 (54.5%)   16 (61.5%)     10 (52.6%)  
#>   >30                                                       10 (45.5%)   10 (38.5%)     9 (47.4%)   
#> Primary cause by days from last study drug administration                                           
#>   <=30                                                                                              
#>     n                                                           12           16             10      
#>     ADVERSE EVENT                                            3 (25%)     9 (56.2%)       4 (40%)    
#>     PROGRESSIVE DISEASE                                      3 (25%)      4 (25%)        3 (30%)    
#>     OTHER                                                    6 (50%)     3 (18.8%)       3 (30%)    
#>   >30                                                                                               
#>     n                                                           10           10             9       
#>     ADVERSE EVENT                                            3 (30%)      3 (30%)       3 (33.3%)   
#>     PROGRESSIVE DISEASE                                      4 (40%)      5 (50%)       5 (55.6%)   
#>     OTHER                                                    3 (30%)      2 (20%)       1 (11.1%)   
#> ————————————————————————————————————————————————————————————————————————————————————————————————————
#> 
#> If applicable, include the footnote highlighted in grey for post-study deaths collected from public records.
dtht01_1(adam_db = db, time_since_last_dose = TRUE, other_category = FALSE)
#> Deaths: {Specify Population}
#> Protocol: {{protocol}}, Snapshot: {{snapshot}}, Snapshot Date: {{snapshot-date}}, Cutoff Date: {{cutoff-date}}
#> 
#> ————————————————————————————————————————————————————————————————————————————————————————————————————
#>                                                             A: Drug X    B: Placebo   C: Combination
#>                                                              (N=134)      (N=134)        (N=132)    
#> ————————————————————————————————————————————————————————————————————————————————————————————————————
#> Total number of deaths                                      22 (16.4%)   26 (19.4%)     19 (14.4%)  
#> Primary cause of death                                                                              
#>   n                                                             22           26             19      
#>   ADVERSE EVENT                                             6 (27.3%)    12 (46.2%)     7 (36.8%)   
#>   PROGRESSIVE DISEASE                                       7 (31.8%)    9 (34.6%)      8 (42.1%)   
#>   OTHER                                                     9 (40.9%)    5 (19.2%)      4 (21.1%)   
#> Days from last drug administration                                                                  
#>   n                                                             22           26             19      
#>   <=30                                                      12 (54.5%)   16 (61.5%)     10 (52.6%)  
#>   >30                                                       10 (45.5%)   10 (38.5%)     9 (47.4%)   
#> Primary cause by days from last study drug administration                                           
#>   <=30                                                                                              
#>     n                                                           12           16             10      
#>     ADVERSE EVENT                                            3 (25%)     9 (56.2%)       4 (40%)    
#>     PROGRESSIVE DISEASE                                      3 (25%)      4 (25%)        3 (30%)    
#>     OTHER                                                    6 (50%)     3 (18.8%)       3 (30%)    
#>   >30                                                                                               
#>     n                                                           10           10             9       
#>     ADVERSE EVENT                                            3 (30%)      3 (30%)       3 (33.3%)   
#>     PROGRESSIVE DISEASE                                      4 (40%)      5 (50%)       5 (55.6%)   
#>     OTHER                                                    3 (30%)      2 (20%)       1 (11.1%)   
#> ————————————————————————————————————————————————————————————————————————————————————————————————————
#> 
#> If applicable, include the footnote highlighted in grey for post-study deaths collected from public records.
dtht01_1_lyt(
  armvar = "ACTARM",
  lbl_overall = NULL
)
#> A Pre-data Table Layout
#> 
#> Column-Split Structure:
#> ACTARM (lvls) 
#> 
#> Row-Split Structure:
#> DTHFL:DTHCAT (** multivar analysis **) 
#> DTHCAT (lvls) -> DTHCAUS (** analysis **) 
#> 
dtht01_1_opt_lyt(
  armvar = "ACTARM",
  lbl_overall = NULL
)
#> A Pre-data Table Layout
#> 
#> Column-Split Structure:
#> ACTARM (lvls) 
#> 
#> Row-Split Structure:
#> LDDTHGR1 (** analysis **) 
#> LDDTHGR1 (lvls) -> DTHCAT (** analysis **) 
#>