EGT01 Table 1 (Default) ECG Assessments and Change from Baseline by Visit Table 1.

The EGT01 table 1 summarizes several electrocardiogram parameters and their evolution throughout the study.

Usage

egt01_1_main(
  adam_db,
  arm_var = "ACTARM",
  summaryvars = list(`Value at Visit` = "AVAL", `Change from \nBaseline` = "CHG"),
  visitvar = "AVISIT",
  deco = std_deco("EGT01"),
  ...
)

egt01_1_lyt(
  arm_var,
  summaryvars,
  summaryvars_lbls,
  visitvar,
  lbl_avisit,
  lbl_param,
  deco
)

egt01_1_pre(adam_db, ...)

egt01_1_post(tlg, prune_0 = TRUE, ...)

egt01_1

Format

An object of class chevron_t of length 1.

Arguments

adam_db

(dm) object containing the ADaM datasets

arm_var

(character) variable used for column splitting

summaryvars

(vector of character) the variables to be analyzed. For this table, AVAL and CHG by default.

visitvar

(character) typically one of "AVISIT" (Default) or "ATPTN" depending on the type of time point to be displayed.

deco

(character) decoration with title, subtitles and main_footer content

...

not used.

summaryvars_lbls

(vector of character) the label of the variables to be analyzed.

lbl_avisit

(character) label of the visitvar variable.

lbl_param

(character) label of the PARAM variable.

tlg

(TableTree, Listing or ggplot) object typically produced by a main function.

prune_0

(logical) remove 0 count rows

Details

The Analysis Value column, displays the number of patients, the mean, standard deviation, median and range of the analysis value for each visit.

The Change from Baseline column, displays the number of patient and the mean, standard deviation, median and range of changes relative to the baseline.

Remove zero-count rows unless overridden with prune_0 = FALSE.

Split columns by arm, typically ACTARM.

Does not include a total column by default.

Sorted based on factor level; first by PARAM labels in alphabetic order then by chronological time point given by AVISIT. Re-level to customize order.

Functions

egt01_1_main(): Main TLG function

egt01_1_lyt(): Layout

egt01_1_pre(): Preprocessing

egt01_1_post(): Postprocessing

Note

adam_db object must contain an adeg table with a "PARAM" column as well as columns specified in summaryvars and visitvar.

Examples

db <- syn_data
run(egt01_1, db)
#>   Parameter                                                                                                                   
#>     Analysis Visit               A: Drug X                          B: Placebo                        C: Combination          
#>                                         Change from                         Change from                         Change from   
#>                      Value at Visit       Baseline       Value at Visit       Baseline       Value at Visit       Baseline    
#>   ————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————
#>   Heart Rate                                                                                                                  
#>     BASELINE                                                                                                                  
#>       n                   134                0                134                0                132                0        
#>       Mean (SD)       71.4 (17.9)            NA           69.8 (20.7)            NA           69.3 (20.9)            NA       
#>       Median              72.7               NA               73.4               NA               72.0               NA       
#>       Min - Max       9.1 - 106.9            NA           13.5 - 115.5           NA           11.6 - 115.5           NA       
#>     WEEK 1 DAY 8                                                                                                              
#>       n                   134               134               134               134               132               132       
#>       Mean (SD)       70.2 (20.3)       -1.3 (26.0)       69.4 (20.8)       -0.4 (29.7)       68.7 (18.1)       -0.7 (27.8)   
#>       Median              70.7              -2.2              70.5              -0.5              68.6              -0.8      
#>       Min - Max       8.5 - 127.5       -51.0 - 89.2      16.9 - 129.1      -68.2 - 88.4      16.9 - 115.6      -68.2 - 67.2  
#>     WEEK 2 DAY 15                                                                                                             
#>       n                   134               134               134               134               132               132       
#>       Mean (SD)       70.3 (19.6)       -1.1 (26.3)       71.1 (20.6)        1.3 (31.4)       69.2 (18.1)       -0.1 (28.6)   
#>       Median              70.9              -1.1              72.5              -0.3              69.3              -1.5      
#>       Min - Max       17.1 - 116.3      -85.0 - 67.5      9.2 - 120.5       -73.1 - 81.4      29.6 - 120.5      -67.7 - 66.3  
#>     WEEK 3 DAY 22                                                                                                             
#>       n                   134               134               134               134               132               132       
#>       Mean (SD)       68.3 (20.2)       -3.2 (28.7)       68.9 (20.6)       -0.9 (30.1)       70.9 (20.3)        1.5 (27.3)   
#>       Median              68.8              -2.2              67.8              1.2               70.1              0.1       
#>       Min - Max       13.3 - 131.7      -81.2 - 72.6      24.0 - 130.4     -73.0 - 103.3      20.9 - 116.8      -64.5 - 72.5  
#>     WEEK 4 DAY 29                                                                                                             
#>       n                   134               134               134               134               132               132       
#>       Mean (SD)       70.1 (18.4)       -1.4 (26.6)       71.6 (20.3)        1.8 (30.3)       71.0 (20.6)        1.7 (29.6)   
#>       Median              69.3              -4.5              72.2              -0.9              70.5              3.8       
#>       Min - Max       22.3 - 116.5      -58.1 - 77.3      17.5 - 129.1      -64.3 - 85.3      10.4 - 117.3     -65.9 - 105.7  
#>     WEEK 5 DAY 36                                                                                                             
#>       n                   134               134               134               134               132               132       
#>       Mean (SD)       66.4 (19.7)       -5.0 (27.1)       71.3 (18.9)        1.4 (27.1)       71.3 (18.6)        1.9 (29.6)   
#>       Median              65.4              -6.8              70.9              0.6               73.3              1.7       
#>       Min - Max       23.9 - 110.4      -73.3 - 57.2      25.9 - 125.7      -71.2 - 67.4      23.9 - 117.2      -58.2 - 68.9  
#>   QT Duration                                                                                                                 
#>     BASELINE                                                                                                                  
#>       n                   134                0                134                0                132                0        
#>       Mean (SD)      336.8 (118.0)           NA           351.0 (98.4)           NA          352.6 (105.1)           NA       
#>       Median             344.4               NA              351.5               NA              348.0               NA       
#>       Min - Max       87.0 - 665.4           NA           98.9 - 641.9           NA          105.7 - 628.1           NA       
#>     WEEK 1 DAY 8                                                                                                              
#>       n                   134               134               134               134               132               132       
#>       Mean (SD)      342.6 (101.1)      5.7 (159.9)      363.9 (102.3)      12.9 (140.7)     359.4 (105.2)      6.8 (147.1)   
#>       Median             347.3              -1.8             356.5              13.7             363.8              8.7       
#>       Min - Max       91.6 - 591.4     -346.4 - 452.8    114.9 - 656.5     -317.5 - 416.4     51.9 - 611.9     -473.2 - 358.0 
#>     WEEK 2 DAY 15                                                                                                             
#>       n                   134               134               134               134               132               132       
#>       Mean (SD)       353.2 (93.9)      16.4 (162.1)      345.9 (96.8)      -5.1 (152.5)      335.7 (98.6)     -16.9 (138.4)  
#>       Median             351.1              16.7             346.8              -9.7             321.0             -22.3      
#>       Min - Max      138.0 - 587.3     -414.1 - 389.2    146.4 - 556.1     -440.3 - 364.8    104.9 - 562.3     -326.5 - 325.3 
#>     WEEK 3 DAY 22                                                                                                             
#>       n                   134               134               134               134               132               132       
#>       Mean (SD)      370.1 (101.6)      33.3 (152.8)     343.4 (102.6)      -7.6 (138.8)      346.9 (95.0)      -5.7 (144.7)  
#>       Median             378.2              39.6             335.1             -15.9             352.2              9.1       
#>       Min - Max      118.1 - 615.2     -391.7 - 520.1     63.4 - 566.5     -311.3 - 293.8    126.1 - 580.8     -412.1 - 410.0 
#>     WEEK 4 DAY 29                                                                                                             
#>       n                   134               134               134               134               132               132       
#>       Mean (SD)       345.8 (96.0)      8.9 (145.6)      354.4 (108.3)      3.4 (142.4)      341.4 (106.7)     -11.2 (145.6)  
#>       Median             340.1              5.6              346.7             -17.1             352.3             -11.9      
#>       Min - Max      110.1 - 616.6     -393.3 - 456.0     80.8 - 687.7     -439.9 - 364.8     4.9 - 570.6      -480.9 - 330.7 
#>     WEEK 5 DAY 36                                                                                                             
#>       n                   134               134               134               134               132               132       
#>       Mean (SD)       358.9 (97.7)      22.1 (155.7)      338.6 (99.3)     -12.4 (130.5)      349.3 (95.8)      -3.3 (136.0)  
#>       Median             352.0              5.9              344.2              0.4              350.3              12.7      
#>       Min - Max       88.4 - 661.1     -353.3 - 539.8     31.3 - 563.9     -338.9 - 352.7    119.0 - 581.8     -311.5 - 295.5 
#>   RR Duration                                                                                                                 
#>     BASELINE                                                                                                                  
#>       n                   134                0                134                0                132                0        
#>       Mean (SD)      1028.4 (286.4)          NA          1027.5 (324.0)          NA          1074.6 (277.3)          NA       
#>       Median             1041.9              NA              1047.3              NA              1080.9              NA       
#>       Min - Max      34.3 - 1783.7           NA           5.3 - 1877.2           NA          289.6 - 1617.1          NA       
#>     WEEK 1 DAY 8                                                                                                              
#>       n                   134               134               134               134               132               132       
#>       Mean (SD)      990.2 (318.7)     -38.3 (446.4)     1061.6 (269.1)     34.1 (401.8)     1039.6 (284.4)    -35.0 (406.0)  
#>       Median             963.9             -86.5             1061.5             66.9             1014.8            -97.8      
#>       Min - Max      110.8 - 2014.6   -1014.8 - 1389.4   276.8 - 1712.0    -942.2 - 993.8    498.2 - 1937.5    -958.6 - 908.9 
#>     WEEK 2 DAY 15                                                                                                             
#>       n                   134               134               134               134               132               132       
#>       Mean (SD)      1013.4 (304.1)    -15.0 (425.8)     1109.4 (318.2)     81.9 (453.1)     1045.2 (277.5)    -29.4 (380.8)  
#>       Median             1040.7             40.5             1117.8             69.3             1034.6            -15.3      
#>       Min - Max      164.2 - 1677.1   -1216.1 - 1053.2   160.7 - 2048.7   -1127.1 - 1148.6   252.8 - 1722.3    -859.3 - 871.5 
#>     WEEK 3 DAY 22                                                                                                             
#>       n                   134               134               134               134               132               132       
#>       Mean (SD)      1014.0 (305.0)    -14.4 (406.5)     1118.3 (296.5)     90.9 (409.6)     1036.6 (268.5)    -38.0 (421.5)  
#>       Median             1027.2            -50.5             1134.6             52.0             1030.7            -60.7      
#>       Min - Max      357.0 - 1798.7   -882.9 - 1080.3    97.1 - 1825.4    -887.1 - 1166.2    446.0 - 1713.4    -984.8 - 902.4 
#>     WEEK 4 DAY 29                                                                                                             
#>       n                   134               134               134               134               132               132       
#>       Mean (SD)      1033.3 (313.9)     4.9 (483.2)      1079.8 (298.9)     52.3 (429.3)     1027.5 (295.9)    -47.1 (378.2)  
#>       Median             1050.4             5.6              1051.6             69.6             1023.4            -49.5      
#>       Min - Max      54.3 - 1979.4    -1345.9 - 1195.4   347.9 - 1762.0   -826.4 - 1191.8    341.6 - 2144.9   -1002.0 - 1048.7
#>     WEEK 5 DAY 36                                                                                                             
#>       n                   134               134               134               134               132               132       
#>       Mean (SD)      1072.3 (282.9)     43.9 (379.0)     1058.5 (271.0)     31.0 (436.0)     1029.0 (271.2)    -45.6 (405.5)  
#>       Median             1067.3             55.9             1068.9             33.6             1065.3            -34.4      
#>       Min - Max      353.0 - 2000.6   -1028.8 - 1418.6   208.8 - 1794.7   -979.0 - 1365.1    436.3 - 1794.1   -962.2 - 1329.9 
run(egt01_1, db, summaryvars = list("AVAL", "Change" = "CHG"))
#>   Parameter                                                                                                                   
#>     Analysis Visit               A: Drug X                          B: Placebo                        C: Combination          
#>                      Analysis Value        Change        Analysis Value        Change        Analysis Value        Change     
#>   ————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————
#>   Heart Rate                                                                                                                  
#>     BASELINE                                                                                                                  
#>       n                   134                0                134                0                132                0        
#>       Mean (SD)       71.4 (17.9)            NA           69.8 (20.7)            NA           69.3 (20.9)            NA       
#>       Median              72.7               NA               73.4               NA               72.0               NA       
#>       Min - Max       9.1 - 106.9            NA           13.5 - 115.5           NA           11.6 - 115.5           NA       
#>     WEEK 1 DAY 8                                                                                                              
#>       n                   134               134               134               134               132               132       
#>       Mean (SD)       70.2 (20.3)       -1.3 (26.0)       69.4 (20.8)       -0.4 (29.7)       68.7 (18.1)       -0.7 (27.8)   
#>       Median              70.7              -2.2              70.5              -0.5              68.6              -0.8      
#>       Min - Max       8.5 - 127.5       -51.0 - 89.2      16.9 - 129.1      -68.2 - 88.4      16.9 - 115.6      -68.2 - 67.2  
#>     WEEK 2 DAY 15                                                                                                             
#>       n                   134               134               134               134               132               132       
#>       Mean (SD)       70.3 (19.6)       -1.1 (26.3)       71.1 (20.6)        1.3 (31.4)       69.2 (18.1)       -0.1 (28.6)   
#>       Median              70.9              -1.1              72.5              -0.3              69.3              -1.5      
#>       Min - Max       17.1 - 116.3      -85.0 - 67.5      9.2 - 120.5       -73.1 - 81.4      29.6 - 120.5      -67.7 - 66.3  
#>     WEEK 3 DAY 22                                                                                                             
#>       n                   134               134               134               134               132               132       
#>       Mean (SD)       68.3 (20.2)       -3.2 (28.7)       68.9 (20.6)       -0.9 (30.1)       70.9 (20.3)        1.5 (27.3)   
#>       Median              68.8              -2.2              67.8              1.2               70.1              0.1       
#>       Min - Max       13.3 - 131.7      -81.2 - 72.6      24.0 - 130.4     -73.0 - 103.3      20.9 - 116.8      -64.5 - 72.5  
#>     WEEK 4 DAY 29                                                                                                             
#>       n                   134               134               134               134               132               132       
#>       Mean (SD)       70.1 (18.4)       -1.4 (26.6)       71.6 (20.3)        1.8 (30.3)       71.0 (20.6)        1.7 (29.6)   
#>       Median              69.3              -4.5              72.2              -0.9              70.5              3.8       
#>       Min - Max       22.3 - 116.5      -58.1 - 77.3      17.5 - 129.1      -64.3 - 85.3      10.4 - 117.3     -65.9 - 105.7  
#>     WEEK 5 DAY 36                                                                                                             
#>       n                   134               134               134               134               132               132       
#>       Mean (SD)       66.4 (19.7)       -5.0 (27.1)       71.3 (18.9)        1.4 (27.1)       71.3 (18.6)        1.9 (29.6)   
#>       Median              65.4              -6.8              70.9              0.6               73.3              1.7       
#>       Min - Max       23.9 - 110.4      -73.3 - 57.2      25.9 - 125.7      -71.2 - 67.4      23.9 - 117.2      -58.2 - 68.9  
#>   QT Duration                                                                                                                 
#>     BASELINE                                                                                                                  
#>       n                   134                0                134                0                132                0        
#>       Mean (SD)      336.8 (118.0)           NA           351.0 (98.4)           NA          352.6 (105.1)           NA       
#>       Median             344.4               NA              351.5               NA              348.0               NA       
#>       Min - Max       87.0 - 665.4           NA           98.9 - 641.9           NA          105.7 - 628.1           NA       
#>     WEEK 1 DAY 8                                                                                                              
#>       n                   134               134               134               134               132               132       
#>       Mean (SD)      342.6 (101.1)      5.7 (159.9)      363.9 (102.3)      12.9 (140.7)     359.4 (105.2)      6.8 (147.1)   
#>       Median             347.3              -1.8             356.5              13.7             363.8              8.7       
#>       Min - Max       91.6 - 591.4     -346.4 - 452.8    114.9 - 656.5     -317.5 - 416.4     51.9 - 611.9     -473.2 - 358.0 
#>     WEEK 2 DAY 15                                                                                                             
#>       n                   134               134               134               134               132               132       
#>       Mean (SD)       353.2 (93.9)      16.4 (162.1)      345.9 (96.8)      -5.1 (152.5)      335.7 (98.6)     -16.9 (138.4)  
#>       Median             351.1              16.7             346.8              -9.7             321.0             -22.3      
#>       Min - Max      138.0 - 587.3     -414.1 - 389.2    146.4 - 556.1     -440.3 - 364.8    104.9 - 562.3     -326.5 - 325.3 
#>     WEEK 3 DAY 22                                                                                                             
#>       n                   134               134               134               134               132               132       
#>       Mean (SD)      370.1 (101.6)      33.3 (152.8)     343.4 (102.6)      -7.6 (138.8)      346.9 (95.0)      -5.7 (144.7)  
#>       Median             378.2              39.6             335.1             -15.9             352.2              9.1       
#>       Min - Max      118.1 - 615.2     -391.7 - 520.1     63.4 - 566.5     -311.3 - 293.8    126.1 - 580.8     -412.1 - 410.0 
#>     WEEK 4 DAY 29                                                                                                             
#>       n                   134               134               134               134               132               132       
#>       Mean (SD)       345.8 (96.0)      8.9 (145.6)      354.4 (108.3)      3.4 (142.4)      341.4 (106.7)     -11.2 (145.6)  
#>       Median             340.1              5.6              346.7             -17.1             352.3             -11.9      
#>       Min - Max      110.1 - 616.6     -393.3 - 456.0     80.8 - 687.7     -439.9 - 364.8     4.9 - 570.6      -480.9 - 330.7 
#>     WEEK 5 DAY 36                                                                                                             
#>       n                   134               134               134               134               132               132       
#>       Mean (SD)       358.9 (97.7)      22.1 (155.7)      338.6 (99.3)     -12.4 (130.5)      349.3 (95.8)      -3.3 (136.0)  
#>       Median             352.0              5.9              344.2              0.4              350.3              12.7      
#>       Min - Max       88.4 - 661.1     -353.3 - 539.8     31.3 - 563.9     -338.9 - 352.7    119.0 - 581.8     -311.5 - 295.5 
#>   RR Duration                                                                                                                 
#>     BASELINE                                                                                                                  
#>       n                   134                0                134                0                132                0        
#>       Mean (SD)      1028.4 (286.4)          NA          1027.5 (324.0)          NA          1074.6 (277.3)          NA       
#>       Median             1041.9              NA              1047.3              NA              1080.9              NA       
#>       Min - Max      34.3 - 1783.7           NA           5.3 - 1877.2           NA          289.6 - 1617.1          NA       
#>     WEEK 1 DAY 8                                                                                                              
#>       n                   134               134               134               134               132               132       
#>       Mean (SD)      990.2 (318.7)     -38.3 (446.4)     1061.6 (269.1)     34.1 (401.8)     1039.6 (284.4)    -35.0 (406.0)  
#>       Median             963.9             -86.5             1061.5             66.9             1014.8            -97.8      
#>       Min - Max      110.8 - 2014.6   -1014.8 - 1389.4   276.8 - 1712.0    -942.2 - 993.8    498.2 - 1937.5    -958.6 - 908.9 
#>     WEEK 2 DAY 15                                                                                                             
#>       n                   134               134               134               134               132               132       
#>       Mean (SD)      1013.4 (304.1)    -15.0 (425.8)     1109.4 (318.2)     81.9 (453.1)     1045.2 (277.5)    -29.4 (380.8)  
#>       Median             1040.7             40.5             1117.8             69.3             1034.6            -15.3      
#>       Min - Max      164.2 - 1677.1   -1216.1 - 1053.2   160.7 - 2048.7   -1127.1 - 1148.6   252.8 - 1722.3    -859.3 - 871.5 
#>     WEEK 3 DAY 22                                                                                                             
#>       n                   134               134               134               134               132               132       
#>       Mean (SD)      1014.0 (305.0)    -14.4 (406.5)     1118.3 (296.5)     90.9 (409.6)     1036.6 (268.5)    -38.0 (421.5)  
#>       Median             1027.2            -50.5             1134.6             52.0             1030.7            -60.7      
#>       Min - Max      357.0 - 1798.7   -882.9 - 1080.3    97.1 - 1825.4    -887.1 - 1166.2    446.0 - 1713.4    -984.8 - 902.4 
#>     WEEK 4 DAY 29                                                                                                             
#>       n                   134               134               134               134               132               132       
#>       Mean (SD)      1033.3 (313.9)     4.9 (483.2)      1079.8 (298.9)     52.3 (429.3)     1027.5 (295.9)    -47.1 (378.2)  
#>       Median             1050.4             5.6              1051.6             69.6             1023.4            -49.5      
#>       Min - Max      54.3 - 1979.4    -1345.9 - 1195.4   347.9 - 1762.0   -826.4 - 1191.8    341.6 - 2144.9   -1002.0 - 1048.7
#>     WEEK 5 DAY 36                                                                                                             
#>       n                   134               134               134               134               132               132       
#>       Mean (SD)      1072.3 (282.9)     43.9 (379.0)     1058.5 (271.0)     31.0 (436.0)     1029.0 (271.2)    -45.6 (405.5)  
#>       Median             1067.3             55.9             1068.9             33.6             1065.3            -34.4      
#>       Min - Max      353.0 - 2000.6   -1028.8 - 1418.6   208.8 - 1794.7   -979.0 - 1365.1    436.3 - 1794.1   -962.2 - 1329.9