EGT02 Table 1 (Default) ECG Abnormalities Table 1

Assessments Outside Normal Limits Regardless of Abnormality at Baseline Table.

egt02_1(
  adam_db,
  armvar = .study$actualarm,
  lbl_vs_assessment = "Assessment",
  lbl_vs_abnormality = "Abnormality",
  lbl_overall = .study$lbl_overall,
  prune_0 = FALSE,
  deco = std_deco("EGT02"),
  .study = list(actualarm = "ACTARM", lbl_overall = NULL)
)

egt02_1_lyt(
  armvar = .study$actualarm,
  lbl_vs_assessment = "Assessment",
  lbl_vs_abnormality = "Abnormality",
  lbl_overall = .study$lbl_overall,
  deco = std_deco("EGT02"),
  .study = list(actualarm = "ACTARM", lbl_overall = NULL)
)

Arguments

adam_db

(dm) object containing the ADaM datasets

armvar

(character) variable used for column splitting

lbl_vs_assessment

(character) the label of the assessment variable.

lbl_vs_abnormality

(character) the label of the abnormality variable.

lbl_overall

(character) label used for overall column, if set to NULL the overall column is omitted

prune_0

(logical) remove 0 count rows

deco

(character) decoration with title, subtitles and main_footer content

.study

(list) with default values for the arguments of the function

Details

Only count LOW or HIGH values.

Results of "LOW LOW" are treated as the same as "LOW", and "HIGH HIGH" the same as "HIGH".

Does not include a total column by default.

Does not remove zero-count rows unless overridden with prune_0 = TRUE.

Functions

egt02_1_lyt:

Examples

library(dm)

db <- syn_test_data() %>%
  dm_zoom_to("adeg") %>%
  filter(PARAM %in% c("Heart Rate", "QT Duration", "RR Duration")) %>%
  dm_update_zoomed() %>%
  preprocess_data("egt02_1")

egt02_1(adam_db = db)
#> ECG Assessments Outside Normal Limits Regardless of Abnormality at Baseline: {Specify Population}
#> Protocol: {{protocol}}, Snapshot: {{snapshot}}, Snapshot Date: {{snapshot-date}}, Cutoff Date: {{cutoff-date}}
#> 
#> ———————————————————————————————————————————————————————————————
#> Assessment       A: Drug X        B: Placebo     C: Combination
#>  Abnormality      (N=134)          (N=134)          (N=132)    
#> ———————————————————————————————————————————————————————————————
#> Heart Rate                                                     
#>   Low          48/134 (35.8%)   57/134 (42.5%)   55/132 (41.7%)
#>   High         48/134 (35.8%)   50/134 (37.3%)   48/132 (36.4%)
#> QT Duration                                                    
#>   Low          53/134 (39.6%)   56/134 (41.8%)   55/132 (41.7%)
#>   High         45/134 (33.6%)   51/134 (38.1%)   46/132 (34.8%)
#> RR Duration                                                    
#>   Low          56/134 (41.8%)   41/134 (30.6%)   50/132 (37.9%)
#>   High         39/134 (29.1%)   61/134 (45.5%)   41/132 (31.1%)
#> ———————————————————————————————————————————————————————————————
#> 
#> Abnormalities are based on the {specify range used} ranges.  Table entries provide the number of patients with a during-treatment assessment abnormality in the direction specified regardless of abnormality at baseline.  Abnormalities reported in patients with missing baseline values are included. Any values collected more than {specify time interval} after the last dose of study drug are not included.
egt02_1_lyt(
  armvar = "ACTARM",
  lbl_overall = NULL
)
#> A Pre-data Table Layout
#> 
#> Column-Split Structure:
#> ACTARM (lvls) 
#> 
#> Row-Split Structure:
#> PARAM (lvls) -> ANRIND (** analysis **) 
#>