EGT02 Table 2 (Supplementary) ECG Abnormalities Table 2

Assessments Outside Normal Limits Among Subject Without Abnormality at Baseline.

egt02_2(
  adam_db,
  armvar = .study$actualarm,
  lbl_vs_assessment = "Assessment",
  lbl_vs_abnormality = "Abnormality",
  lbl_overall = .study$lbl_overall,
  prune_0 = FALSE,
  deco = std_deco("EGT02_2"),
  .study = list(actualarm = "ACTARM", lbl_overall = NULL)
)

egt02_2_lyt(
  armvar = .study$actualarm,
  lbl_vs_assessment = "Assessment",
  lbl_vs_abnormality = "Abnormality",
  lbl_overall = .study$lbl_overall,
  deco = std_deco("EGT02_2"),
  .study = list(actualarm = "ACTARM", lbl_overall = NULL)
)

egt02_2_pre(adam_db)

Arguments

adam_db

(dm) object containing the ADaM datasets

armvar

(character) variable used for column splitting

lbl_vs_assessment

(character) the label of the assessment variable.

lbl_vs_abnormality

(character) the label of the abnormality variable.

lbl_overall

(character) label used for overall column, if set to NULL the overall column is omitted

prune_0

(logical) remove 0 count rows

deco

(character) decoration with title, subtitles and main_footer content

.study

(list) with default values for the arguments of the function

Details

Only count LOW or HIGH values.

Results of "LOW LOW" are treated as the same as "LOW", and "HIGH HIGH" the same as "HIGH".

Does not include a total column by default.

Does not remove zero-count rows unless overridden with prune_0 = TRUE.

Functions

egt02_2_lyt: egt02_2 Layout

egt02_2_pre: egt02_2 Preprocessing

Examples

library(dm)

db <- syn_test_data() %>%
  dm_zoom_to("adeg") %>%
  filter(PARAM %in% c("Heart Rate", "QT Duration", "RR Duration")) %>%
  dm_update_zoomed() %>%
  egt02_2_pre()

egt02_2(adam_db = db)
#> ECG Assessments Outside Normal Limits Among Subjects Without Abnormality at Baseline: {Specify Population}
#> Protocol: {{protocol}}
#> 
#> ———————————————————————————————————————————————————————————————
#> Assessment       A: Drug X        B: Placebo     C: Combination
#>  Abnormality      (N=134)          (N=134)          (N=132)    
#> ———————————————————————————————————————————————————————————————
#> Heart Rate                                                     
#>   Low          42/128 (32.8%)   47/124 (37.9%)   40/117 (34.2%)
#>   High         41/127 (32.3%)   45/129 (34.9%)   38/122 (31.1%)
#> QT Duration                                                    
#>   Low          32/113 (28.3%)   48/126 (38.1%)   47/124 (37.9%)
#>   High         36/125 (28.8%)   47/130 (36.2%)   33/119 (27.7%)
#> RR Duration                                                    
#>   Low          51/129 (39.5%)   30/123 (24.4%)   44/126 (34.9%)
#>   High         32/127 (25.2%)   49/122 (40.2%)   34/125 (27.2%)
#> ———————————————————————————————————————————————————————————————
#> 
#> Abnormalities are based on the {specify range used} ranges.  Table entries provide the number of patients with a during-treatment assessment abnormality in the direction specified  among patient without abnormality at baseline.  Abnormalities reported in patients with missing baseline values are included. Any values collected more than {specify time interval} after the last dose of study drug are not included.
egt02_2_lyt(
  armvar = "ACTARM",
  lbl_overall = NULL
)
#> A Pre-data Table Layout
#> 
#> Column-Split Structure:
#> ACTARM (lvls) 
#> 
#> Row-Split Structure:
#> PARAM (lvls) -> ANRIND (** analysis **) 
#> 
syn_test_data() %>%
  egt02_2_pre()
#> ── Metadata ────────────────────────────────────────────────────────────────────
#> Tables: `adsl`, `adae`, `adaette`, `adcm`, `addv`, … (15 total)
#> Columns: 846
#> Primary keys: 2
#> Foreign keys: 1