EGT05_QTCAT ECG Actual Values and Changes from Baseline by Visit Table.

The EGT05_QTCAT table summarizes several electrocardiogram parameters and their evolution throughout the study.

Usage

egt05_qtcat_main(
  adam_db,
  arm_var = "ACTARM",
  summaryvars = c("AVALCAT1", "CHGCAT1"),
  lbl_overall = NULL,
  visitvar = "AVISIT",
  ...
)

egt05_qtcat_pre(adam_db, ...)

egt05_qtcat_post(tlg, prune_0 = TRUE, ...)

egt05_qtcat

Format

An object of class chevron_t of length 1.

Arguments

adam_db

(list of data.frames) object containing the ADaM datasets

arm_var

(string) variable used for column splitting

summaryvars

(character) variables to be analyzed. The label attribute of the corresponding column in adeg table of adam_db is used as name.

lbl_overall

(string) label used for overall column, if set to NULL the overall column is omitted

visitvar

(string) typically "AVISIT" (Default) or "ATPTN".

...

not used.

tlg

(TableTree, Listing or ggplot) object typically produced by a main function.

prune_0

(flag) remove 0 count rows

Details

The Value at Visit column, displays the categories of the specific "PARAMCD" value for patients.

The Change from Baseline column, displays the categories of the specific "PARAMCD" value change from baseline for patients.

Remove zero-count rows unless overridden with prune_0 = FALSE.

Split columns by arm, typically "ACTARM".

Does not include a total column by default.

Sorted based on factor level; by chronological time point given by one of "AVISIT" (Default) or "ATPTN". Re-level to customize order.

Functions

egt05_qtcat_main(): Main TLG function

egt05_qtcat_pre(): Preprocessing

egt05_qtcat_post(): Postprocessing

Note

adam_db object must contain an adeg table with column specified in visitvar. For summaryvars, please make sure AVALCAT1 and CHGCAT1 columns existed in input data sets.

Examples

run(egt05_qtcat, syn_data)
#>   Parameter                                                            
#>     Analysis Visit            A: Drug X    B: Placebo    C: Combination
#>       Category                 (N=134)       (N=134)        (N=132)    
#>   —————————————————————————————————————————————————————————————————————
#>   QT Duration                                                          
#>     BASELINE                                                           
#>       Value at Visit                                                   
#>         n                        134           134            132      
#>         <=450 msec           115 (85.8%)   117 (87.3%)    104 (78.8%)  
#>         >450 to <=480 msec    6 (4.5%)      10 (7.5%)       9 (6.8%)   
#>         >480 to <=500 msec    4 (3.0%)      3 (2.2%)        6 (4.5%)   
#>         >500 msec             9 (6.7%)      4 (3.0%)       13 (9.8%)   
#>     WEEK 1 DAY 8                                                       
#>       Value at Visit                                                   
#>         n                        134           134            132      
#>         <=450 msec           113 (84.3%)   106 (79.1%)    106 (80.3%)  
#>         >450 to <=480 msec    10 (7.5%)     10 (7.5%)      11 (8.3%)   
#>         >480 to <=500 msec    4 (3.0%)      4 (3.0%)        3 (2.3%)   
#>         >500 msec             7 (5.2%)     14 (10.4%)      12 (9.1%)   
#>       Change from Baseline                                             
#>         n                        134           134            132      
#>         <=30 msec            76 (56.7%)    75 (56.0%)      75 (56.8%)  
#>         >30 to <=60 msec      7 (5.2%)      13 (9.7%)      11 (8.3%)   
#>         >60 msec             51 (38.1%)    46 (34.3%)      46 (34.8%)  
#>     WEEK 2 DAY 15                                                      
#>       Value at Visit                                                   
#>         n                        134           134            132      
#>         <=450 msec           111 (82.8%)   114 (85.1%)    112 (84.8%)  
#>         >450 to <=480 msec    10 (7.5%)     9 (6.7%)        9 (6.8%)   
#>         >480 to <=500 msec    7 (5.2%)      2 (1.5%)        5 (3.8%)   
#>         >500 msec             6 (4.5%)      9 (6.7%)        6 (4.5%)   
#>       Change from Baseline                                             
#>         n                        134           134            132      
#>         <=30 msec            71 (53.0%)    87 (64.9%)      89 (67.4%)  
#>         >30 to <=60 msec      11 (8.2%)     9 (6.7%)        9 (6.8%)   
#>         >60 msec             52 (38.8%)    38 (28.4%)      34 (25.8%)  
#>     WEEK 3 DAY 22                                                      
#>       Value at Visit                                                   
#>         n                        134           134            132      
#>         <=450 msec           106 (79.1%)   112 (83.6%)    118 (89.4%)  
#>         >450 to <=480 msec    13 (9.7%)     7 (5.2%)        3 (2.3%)   
#>         >480 to <=500 msec    4 (3.0%)      5 (3.7%)        2 (1.5%)   
#>         >500 msec             11 (8.2%)     10 (7.5%)       9 (6.8%)   
#>       Change from Baseline                                             
#>         n                        134           134            132      
#>         <=30 msec            63 (47.0%)    80 (59.7%)      81 (61.4%)  
#>         >30 to <=60 msec     14 (10.4%)     8 (6.0%)       11 (8.3%)   
#>         >60 msec             57 (42.5%)    46 (34.3%)      40 (30.3%)  
#>     WEEK 4 DAY 29                                                      
#>       Value at Visit                                                   
#>         n                        134           134            132      
#>         <=450 msec           117 (87.3%)   103 (76.9%)    114 (86.4%)  
#>         >450 to <=480 msec    7 (5.2%)     14 (10.4%)       6 (4.5%)   
#>         >480 to <=500 msec    4 (3.0%)      7 (5.2%)        3 (2.3%)   
#>         >500 msec             6 (4.5%)      10 (7.5%)       9 (6.8%)   
#>       Change from Baseline                                             
#>         n                        134           134            132      
#>         <=30 msec            79 (59.0%)    80 (59.7%)      79 (59.8%)  
#>         >30 to <=60 msec      11 (8.2%)     7 (5.2%)       10 (7.6%)   
#>         >60 msec             44 (32.8%)    47 (35.1%)      43 (32.6%)  
#>     WEEK 5 DAY 36                                                      
#>       Value at Visit                                                   
#>         n                        134           134            132      
#>         <=450 msec           107 (79.9%)   117 (87.3%)    112 (84.8%)  
#>         >450 to <=480 msec   16 (11.9%)     5 (3.7%)       13 (9.8%)   
#>         >480 to <=500 msec    5 (3.7%)      9 (6.7%)        3 (2.3%)   
#>         >500 msec             6 (4.5%)      3 (2.2%)        4 (3.0%)   
#>       Change from Baseline                                             
#>         n                        134           134            132      
#>         <=30 msec            72 (53.7%)    82 (61.2%)      73 (55.3%)  
#>         >30 to <=60 msec      10 (7.5%)     11 (8.2%)      11 (8.3%)   
#>         >60 msec             52 (38.8%)    41 (30.6%)      48 (36.4%)