EXT01 Table 1 (Default) Exposure Summary Table

The EXT01 table provides an overview of the of the exposure of the patients in terms of Total dose administered or missed, and treatment duration.

ext01_1(
  adex,
  adsl,
  armvar = .study$armvar,
  summaryvars = .study$analysis_var,
  lbl_overall = .study$lbl_overall,
  prune_0 = TRUE,
  deco = std_deco("EXT01"),
  .study = list(armvar = "ACTARM", lbl_overall = "", analysis_var = "AVAL")
)

ext01_1_lyt(
  armvar = .study$armvar,
  summaryvars = .study$analysis_var,
  summaryvars_lbls = .study$lbl_analysis_var,
  lbl_overall = .study$lbl_overall,
  deco = std_deco("EXT01"),
  .study = list(armvar = "ACTARM", overall_col = "", analysis_var = "AVAL",
    lbl_analysis_var = "Analysis Value")
)

Arguments

adex
(data.frame) Exposure Analysis Dataset

adsl
(data.frame) Subject Level Analysis Dataset

armvar
(string) variable used for column splitting

summaryvars
(string) the name of the variable to be analyzed. By default "AVAL".

lbl_overall
(string) label used for overall column, if set to NULL the overall column is omitted

prune_0
(logical) remove 0 count rows

deco
(list) decoration with title, subtitles and main_footer content

.study
(list) with default values for the arguments of the function

summaryvars_lbls
(string) the label associated with the analyzed variable.

Value

Details

Default Exposure table

The n row provides the number of non-missing values. The percentages for categorical variables is based on n. The percentages for Total number of patients with at least one dose modification are based on the number of patients in the corresponding analysis population given by N.

Split columns by arm, typically ACTARM.

Does not include a total column by default.

Sorted by alphabetic order of the PARAM value. Transform to factor and re-level for custom order.

Functions

ext01_1_lyt:

Examples

library(scda)
library(dplyr)
sd <- synthetic_cdisc_data("rcd_2021_03_22")
adsl <- sd$adsl
adex <- sd$adex %>%
 mutate(ANL01FL = 'Y')

ext01_1(adex, adsl)

#>                                                                   A: Drug X        B: Placebo      C: Combination 
#>                                                                    (N=134)           (N=134)           (N=132)    
#> ------------------------------------------------------------------------------------------------------------------
#> Dose administered during constant dosing interval                                                                 
#>   n                                                                  938               938               924      
#>   Mean (SD)                                                     950.8 (232.9)     914.5 (240.4)     968.1 (229.2) 
#>   Median                                                             960               960               960      
#>   Min - Max                                                      480 - 1440        480 - 1440        480 - 1440   
#> Number of doses administered during constant dosing interval                                                      
#>   n                                                                  938               938               924      
#>   Mean (SD)                                                         1 (0)             1 (0)             1 (0)     
#>   Median                                                              1                 1                 1       
#>   Min - Max                                                         1 - 1             1 - 1             1 - 1     
#> Total dose administered                                                                                           
#>   n                                                                  134               134               132      
#>   Mean (SD)                                                    6655.5 (1209.6)   6401.2 (1269.9)   6776.4 (1212.5)
#>   Median                                                            6720              6360              6720      
#>   Min - Max                                                      4320 - 9360       4080 - 9360       4320 - 9360  
#> Total number of doses administered                                                                                
#>   n                                                                  134               134               132      
#>   Mean (SD)                                                         7 (0)             7 (0)             7 (0)     
#>   Median                                                              7                 7                 7       
#>   Min - Max                                                         7 - 7             7 - 7             7 - 7     
#> ------------------------------------------------------------------------------------------------------------------
#> 
#> Footnotes go here

adex2 <- adex %>%
   relevel_params(paramcd_levels = c("TNDOSE", "DOSE", "NDOSE", "TDOSE"))

ext01_1(adex2, adsl)

#>                                                                   A: Drug X        B: Placebo      C: Combination 
#>                                                                    (N=134)           (N=134)           (N=132)    
#> ------------------------------------------------------------------------------------------------------------------
#> Total number of doses administered                                                                                
#>   n                                                                  938               938               924      
#>   Mean (SD)                                                     950.8 (232.9)     914.5 (240.4)     968.1 (229.2) 
#>   Median                                                             960               960               960      
#>   Min - Max                                                      480 - 1440        480 - 1440        480 - 1440   
#> Dose administered during constant dosing interval                                                                 
#>   n                                                                  938               938               924      
#>   Mean (SD)                                                         1 (0)             1 (0)             1 (0)     
#>   Median                                                              1                 1                 1       
#>   Min - Max                                                         1 - 1             1 - 1             1 - 1     
#> Number of doses administered during constant dosing interval                                                      
#>   n                                                                  134               134               132      
#>   Mean (SD)                                                    6655.5 (1209.6)   6401.2 (1269.9)   6776.4 (1212.5)
#>   Median                                                            6720              6360              6720      
#>   Min - Max                                                      4320 - 9360       4080 - 9360       4320 - 9360  
#> Total dose administered                                                                                           
#>   n                                                                  134               134               132      
#>   Mean (SD)                                                         7 (0)             7 (0)             7 (0)     
#>   Median                                                              7                 7                 7       
#>   Min - Max                                                         7 - 7             7 - 7             7 - 7     
#> ------------------------------------------------------------------------------------------------------------------
#> 
#> Footnotes go here