EXT01 Table 2 (Supplementary) Exposure Summary Table with grouping options

EXT01 Table 2 (Supplementary) Exposure Summary Table with grouping options

Usage

ext01_2_main(
  adam_db,
  lyt_ls = list(ext01_2_lyt),
  armvar = .study$actualarm,
  lbl_overall = .study$lbl_overall,
  prune_0 = TRUE,
  deco = std_deco("EXT01"),
  .study = list(actualarm = "ACTARM", lbl_overall = NULL),
  ...
)

ext01_2_lyt(
  armvar = .study$actualarm,
  summaryvars = c("AVAL", "AVALCAT1"),
  summaryvars_lbls = c("Summary", "Categories"),
  lbl_overall = .study$lbl_overall,
  deco = std_deco("EXT01"),
  .study = list(actualarm = "ACTARM", lbl_overall = NULL),
  ...
)

ext01_2_pre(
  adam_db,
  show_stats = .study$show_cont_stats,
  show_bins = .study$show_cat_stats,
  .study = list(show_cont_stats = c("ALL"), show_cat_stats = c("ALL")),
  ...
)

ext01_2

Format

An object of class chevron_t of length 1.

Arguments

adam_db

(dm) object containing the ADaM datasets

lyt_ls

(list of functions returning PreDataTableLayout object. Typically one of the _lyt function of chevron stored in a list.

armvar

(character) variable used for column splitting

lbl_overall

(character) label used for overall column, if set to NULL the overall column is omitted

prune_0

(logical) remove 0 count rows

deco

(character) decoration with title, subtitles and main_footer content

.study

(list) with default values for the arguments of the function

...

not used.

summaryvars

(string) the name of the variable to be analyzed. By default "AVAL".

summaryvars_lbls

(string) the label associated with the analyzed variable.

show_stats

(vector of character) providing the name of the parameters whose statistical summary should be presented. To analyze all, provide show_stats = "ALL" (Default), to analyze none, provide show_stats = "".

show_bins

(vector of character) providing the name of the parameters whose categorical summary should be presented. To analyze all, provide show_bins = "ALL" (Default), to analyze none, provide show_bins = "".

Details

Supplementary Exposure table with binning of desired analysis values.

The n row provides the number of non-missing values. The percentages for categorical variables is based on n. The percentages for Total number of patients with at least one dose modification are based on the number of patients in the corresponding analysis population given by N.

Split columns by arm, typically ACTARM.

Does not include a total column by default.

Sorted by alphabetic order of the PARAM value. Transform to factor and re-level for custom order.

ANL01FL is not relevant subset

Functions

ext01_2_main(): Main TLG function

ext01_2_lyt(): Layout

ext01_2_pre(): Preprocessing

Note

adam_db object must contain an adex table with columns specified in summaryvars.

Examples

run(ext01_2, syn_test_data())
#>                                         A: Drug X        B: Placebo      C: Combination 
#>                                          (N=134)           (N=134)           (N=132)    
#> ————————————————————————————————————————————————————————————————————————————————————————
#> Total dose administered                                                                 
#>   n                                        134               134               132      
#>   Mean (SD)                          6655.5 (1209.6)   6401.2 (1269.9)   6776.4 (1212.5)
#>   Median                                 6720.0            6360.0            6720.0     
#>   Min - Max                          4320.0 - 9360.0   4080.0 - 9360.0   4320.0 - 9360.0
#>   n                                        134               134               132      
#>   5000-7000                           72.0 (53.7%)      71.0 (53.0%)      71.0 (53.8%)  
#>   7000-9000                           50.0 (37.3%)      42.0 (31.3%)      51.0 (38.6%)  
#>   <5000                                10.0 (7.5%)      18.0 (13.4%)       7.0 (5.3%)   
#>   >9000                                2.0 (1.5%)        3.0 (2.2%)        3.0 (2.3%)   
#> Total number of doses administered                                                      
#>   n                                        134               134               132      
#>   Mean (SD)                             7.0 (0.0)         7.0 (0.0)         7.0 (0.0)   
#>   Median                                   7.0               7.0               7.0      
#>   Min - Max                             7.0 - 7.0         7.0 - 7.0         7.0 - 7.0