LBT04 Table 1 (Default) Laboratory Abnormalities Not Present at Baseline.

The LBT04 table provides an overview of laboratory abnormalities not present at baseline.

Usage

lbt04_1_main(
  adam_db,
  arm_var = "ACTARM",
  lbl_param = "Laboratory Test",
  lbl_anrind = "Direction of Abnormality",
  deco = std_deco("LBT04"),
  ...
)

lbt04_1_lyt(arm_var, lbl_param, lbl_anrind, deco, ...)

lbt04_1_pre(adam_db, ...)

lbt04_1_check(adam_db, req_tables = c("adsl", "adlb"), arm_var = "ACTARM", ...)

lbt04_1_post(tlg, ...)

lbt04_1

Format

An object of class chevron_t of length 1.

Arguments

adam_db

(dm) object containing the ADaM datasets

arm_var

(character) variable used for column splitting

lbl_param

(character) label of the PARAM variable.

lbl_anrind

(character) label of the ANRIND variable.

deco

(character) decoration with title, subtitles and main_footer content

...

not used.

req_tables

(character) names of the required tables.

tlg

(TableTree, Listing or ggplot) object typically produced by a main function.

Details

Only count LOW or HIGH values.

Lab test results with missing ANRIND values are excluded.

Split columns by arm, typically ACTARM.

Does not include a total column by default.

Functions

lbt04_1_main(): Main TLG function

lbt04_1_lyt(): Layout

lbt04_1_pre(): Preprocessing

lbt04_1_check(): Checks

lbt04_1_post(): Postprocessing

Note

adam_db object must contain an adlb table with columns "PARCAT1", "PARCAT2", "PARAM", "ANRIND", and column specified by arm_var.

Examples

run(lbt04_1, syn_data)
#>     Laboratory Test                          A: Drug X      B: Placebo     C: Combination
#>       Direction of Abnormality                (N=134)         (N=134)         (N=132)    
#>   ———————————————————————————————————————————————————————————————————————————————————————
#>   CHEMISTRY                                                                              
#>     Alanine Aminotransferase Measurement                                                 
#>       Low                                   4/54 (7.4%)    6/54 (11.1%)     6/50 (12.0%) 
#>       High                                 6/55 (10.9%)    10/54 (18.5%)    8/48 (16.7%) 
#>     C-Reactive Protein Measurement                                                       
#>       Low                                  8/50 (16.0%)    7/52 (13.5%)     5/50 (10.0%) 
#>       High                                 7/46 (15.2%)     3/49 (6.1%)     3/49 (6.1%)  
#>     Immunoglobulin A Measurement                                                         
#>       Low                                   4/51 (7.8%)     3/54 (5.6%)     9/61 (14.8%) 
#>       High                                 5/48 (10.4%)    8/56 (14.3%)     9/60 (15.0%) 
#>   COAGULATION                                                                            
#>     Alanine Aminotransferase Measurement                                                 
#>       Low                                   5/55 (9.1%)     4/54 (7.4%)     7/50 (14.0%) 
#>       High                                  4/53 (7.5%)     3/52 (5.8%)     8/52 (15.4%) 
#>     C-Reactive Protein Measurement                                                       
#>       Low                                  7/55 (12.7%)    5/47 (10.6%)    11/68 (16.2%) 
#>       High                                 8/55 (14.5%)    7/44 (15.9%)     7/64 (10.9%) 
#>     Immunoglobulin A Measurement                                                         
#>       Low                                   2/46 (4.3%)    10/55 (18.2%)    5/48 (10.4%) 
#>       High                                 8/48 (16.7%)    6/59 (10.2%)     6/47 (12.8%) 
#>   HEMATOLOGY                                                                             
#>     Alanine Aminotransferase Measurement                                                 
#>       Low                                  5/43 (11.6%)     3/49 (6.1%)     3/52 (5.8%)  
#>       High                                 7/41 (17.1%)     2/43 (4.7%)     7/58 (12.1%) 
#>     C-Reactive Protein Measurement                                                       
#>       Low                                  9/48 (18.8%)     6/63 (9.5%)     3/48 (6.2%)  
#>       High                                  3/49 (6.1%)    7/55 (12.7%)     2/50 (4.0%)  
#>     Immunoglobulin A Measurement                                                         
#>       Low                                   6/61 (9.8%)    10/54 (18.5%)    6/52 (11.5%) 
#>       High                                 13/60 (21.7%)    4/58 (6.9%)     6/53 (11.3%)