LBT06 Table 1 (Default) Laboratory Abnormalities by Visit and Baseline Status Table 1.

The LBT06 table produces the standard laboratory abnormalities by visit and baseline status summary.

Usage

lbt06_main(adam_db, arm_var = "ACTARM", ...)

lbt06_pre(adam_db, ...)

lbt06_post(tlg, prune_0 = FALSE, ...)

lbt06

Format

An object of class chevron_t of length 1.

Arguments

adam_db

(list of data.frames) object containing the ADaM datasets

arm_var

(string) the arm variable used for arm splitting.

...

not used.

tlg

(TableTree, Listing or ggplot) object typically produced by a main function.

prune_0

(flag) remove 0 count rows

Details

Only count "LOW" or "HIGH" values for ANRIND and BNRIND.

Lab test results with missing ANRIND values are excluded.

Split columns by arm, typically ACTARM.

Keep zero count rows by default.

Functions

lbt06_main(): Main TLG function

lbt06_pre(): Preprocessing

lbt06_post(): Postprocessing

Note

adam_db object must contain an adlb table with columns "AVISIT", "ANRIND", "BNRIND", "ONTRTFL", and "PARCAT2", and column specified by arm_var.

Examples

run(lbt06, syn_data)
#>   Parameter                                                                            
#>     Visit                                                                              
#>       Abnormality at Visit                  A: Drug X      B: Placebo    C: Combination
#>               Baseline Status                (N=134)        (N=134)         (N=132)    
#>   —————————————————————————————————————————————————————————————————————————————————————
#>   Alanine Aminotransferase Measurement                                                 
#>     WEEK 1 DAY 8                                                                       
#>       Low                                                                              
#>                 Not low baseline status    3/39 (7.7%)     2/40 (5%)      1/37 (2.7%)  
#>                 Low baseline status         1/4 (25%)         0/2             0/7      
#>                 Total                      4/43 (9.3%)    2/42 (4.8%)     1/44 (2.3%)  
#>       High                                                                             
#>                 Not high baseline status   3/42 (7.1%)    3/35 (8.6%)     4/41 (9.8%)  
#>                 High baseline status           0/1            0/7             0/3      
#>                 Total                       3/43 (7%)     3/42 (7.1%)     4/44 (9.1%)  
#>     WEEK 2 DAY 15                                                                      
#>       Low                                                                              
#>                 Not low baseline status    5/29 (17.2%)   2/30 (6.7%)     7/45 (15.6%) 
#>                 Low baseline status         1/2 (50%)         0/5             0/7      
#>                 Total                      6/31 (19.4%)   2/35 (5.7%)     7/52 (13.5%) 
#>       High                                                                             
#>                 Not high baseline status   1/26 (3.8%)    4/33 (12.1%)    4/45 (8.9%)  
#>                 High baseline status           0/5            0/2         1/7 (14.3%)  
#>                 Total                      1/31 (3.2%)    4/35 (11.4%)    5/52 (9.6%)  
#>     WEEK 3 DAY 22                                                                      
#>       Low                                                                              
#>                 Not low baseline status    2/34 (5.9%)    3/34 (8.8%)     1/35 (2.9%)  
#>                 Low baseline status        1/3 (33.3%)        0/0          1/4 (25%)   
#>                 Total                      3/37 (8.1%)    3/34 (8.8%)     2/39 (5.1%)  
#>       High                                                                             
#>                 Not high baseline status   5/33 (15.2%)   5/29 (17.2%)    3/34 (8.8%)  
#>                 High baseline status        1/4 (25%)         0/5             0/5      
#>                 Total                      6/37 (16.2%)   5/34 (14.7%)    3/39 (7.7%)  
#>     WEEK 4 DAY 29                                                                      
#>       Low                                                                              
#>                 Not low baseline status     2/40 (5%)     2/39 (5.1%)     1/38 (2.6%)  
#>                 Low baseline status            0/2         1/4 (25%)          0/5      
#>                 Total                      2/42 (4.8%)     3/43 (7%)      1/43 (2.3%)  
#>       High                                                                             
#>                 Not high baseline status   5/36 (13.9%)   1/34 (2.9%)     6/39 (15.4%) 
#>                 High baseline status       1/6 (16.7%)    1/9 (11.1%)         0/4      
#>                 Total                      6/42 (14.3%)   2/43 (4.7%)      6/43 (14%)  
#>     WEEK 5 DAY 36                                                                      
#>       Low                                                                              
#>                 Not low baseline status    2/35 (5.7%)    6/35 (17.1%)    4/44 (9.1%)  
#>                 Low baseline status        1/3 (33.3%)     2/8 (25%)          0/3      
#>                 Total                      3/38 (7.9%)    8/43 (18.6%)    4/47 (8.5%)  
#>       High                                                                             
#>                 Not high baseline status   4/38 (10.5%)   7/38 (18.4%)    1/41 (2.4%)  
#>                 High baseline status           0/0            0/5             0/6      
#>                 Total                      4/38 (10.5%)   7/43 (16.3%)    1/47 (2.1%)  
#>   C-Reactive Protein Measurement                                                       
#>     WEEK 1 DAY 8                                                                       
#>       Low                                                                              
#>                 Not low baseline status    4/46 (8.7%)     2/40 (5%)      4/42 (9.5%)  
#>                 Low baseline status            0/5            0/2             0/2      
#>                 Total                      4/51 (7.8%)    2/42 (4.8%)     4/44 (9.1%)  
#>       High                                                                             
#>                 Not high baseline status   6/44 (13.6%)   4/37 (10.8%)    1/35 (2.9%)  
#>                 High baseline status           0/7         1/5 (20%)          0/9      
#>                 Total                      6/51 (11.8%)   5/42 (11.9%)    1/44 (2.3%)  
#>     WEEK 2 DAY 15                                                                      
#>       Low                                                                              
#>                 Not low baseline status    4/39 (10.3%)   3/29 (10.3%)    2/33 (6.1%)  
#>                 Low baseline status        1/3 (33.3%)        0/2             0/3      
#>                 Total                      5/42 (11.9%)   3/31 (9.7%)     2/36 (5.6%)  
#>       High                                                                             
#>                 Not high baseline status   2/36 (5.6%)    3/27 (11.1%)    3/27 (11.1%) 
#>                 High baseline status           0/6         1/4 (25%)          0/9      
#>                 Total                      2/42 (4.8%)    4/31 (12.9%)    3/36 (8.3%)  
#>     WEEK 3 DAY 22                                                                      
#>       Low                                                                              
#>                 Not low baseline status    3/37 (8.1%)     6/46 (13%)     8/44 (18.2%) 
#>                 Low baseline status            0/3            0/5             0/6      
#>                 Total                      3/40 (7.5%)    6/51 (11.8%)     8/50 (16%)  
#>       High                                                                             
#>                 Not high baseline status   2/34 (5.9%)    2/44 (4.5%)     1/43 (2.3%)  
#>                 High baseline status           0/6            0/7             0/7      
#>                 Total                       2/40 (5%)     2/51 (3.9%)      1/50 (2%)   
#>     WEEK 4 DAY 29                                                                      
#>       Low                                                                              
#>                 Not low baseline status    3/42 (7.1%)    1/36 (2.8%)     4/34 (11.8%) 
#>                 Low baseline status            0/4         1/2 (50%)          0/5      
#>                 Total                      3/46 (6.5%)    2/38 (5.3%)     4/39 (10.3%) 
#>       High                                                                             
#>                 Not high baseline status   3/40 (7.5%)    1/35 (2.9%)     5/34 (14.7%) 
#>                 High baseline status       1/6 (16.7%)    1/3 (33.3%)         0/5      
#>                 Total                      4/46 (8.7%)    2/38 (5.3%)     5/39 (12.8%) 
#>     WEEK 5 DAY 36                                                                      
#>       Low                                                                              
#>                 Not low baseline status    5/34 (14.7%)   4/38 (10.5%)    5/41 (12.2%) 
#>                 Low baseline status        1/6 (16.7%)        0/4          1/5 (20%)   
#>                 Total                       6/40 (15%)    4/42 (9.5%)      6/46 (13%)  
#>       High                                                                             
#>                 Not high baseline status   4/35 (11.4%)   4/35 (11.4%)    3/39 (7.7%)  
#>                 High baseline status           0/5            0/7         1/7 (14.3%)  
#>                 Total                       4/40 (10%)    4/42 (9.5%)     4/46 (8.7%)  
#>   Immunoglobulin A Measurement                                                         
#>     WEEK 1 DAY 8                                                                       
#>       Low                                                                              
#>                 Not low baseline status    2/38 (5.3%)    3/38 (7.9%)      3/30 (10%)  
#>                 Low baseline status        1/7 (14.3%)     1/4 (25%)       1/4 (25%)   
#>                 Total                      3/45 (6.7%)    4/42 (9.5%)     4/34 (11.8%) 
#>       High                                                                             
#>                 Not high baseline status   2/38 (5.3%)    3/37 (8.1%)     1/30 (3.3%)  
#>                 High baseline status       2/7 (28.6%)     1/5 (20%)       1/4 (25%)   
#>                 Total                      4/45 (8.9%)    4/42 (9.5%)     2/34 (5.9%)  
#>     WEEK 2 DAY 15                                                                      
#>       Low                                                                              
#>                 Not low baseline status    2/36 (5.6%)    5/45 (11.1%)    5/43 (11.6%) 
#>                 Low baseline status         1/4 (25%)         0/4         1/6 (16.7%)  
#>                 Total                      3/40 (7.5%)    5/49 (10.2%)    6/49 (12.2%) 
#>       High                                                                             
#>                 Not high baseline status   7/39 (17.9%)   4/44 (9.1%)     5/46 (10.9%) 
#>                 High baseline status           0/1            0/5             0/3      
#>                 Total                      7/40 (17.5%)   4/49 (8.2%)     5/49 (10.2%) 
#>     WEEK 3 DAY 22                                                                      
#>       Low                                                                              
#>                 Not low baseline status    7/43 (16.3%)    4/40 (10%)     2/38 (5.3%)  
#>                 Low baseline status            0/5            0/10         2/4 (50%)   
#>                 Total                      7/48 (14.6%)    4/50 (8%)      4/42 (9.5%)  
#>       High                                                                             
#>                 Not high baseline status   2/41 (4.9%)    5/48 (10.4%)    5/36 (13.9%) 
#>                 High baseline status           0/7            0/2             0/6      
#>                 Total                      2/48 (4.2%)     5/50 (10%)     5/42 (11.9%) 
#>     WEEK 4 DAY 29                                                                      
#>       Low                                                                              
#>                 Not low baseline status    2/29 (6.9%)    7/43 (16.3%)    4/46 (8.7%)  
#>                 Low baseline status            0/5            0/10            0/2      
#>                 Total                      2/34 (5.9%)    7/53 (13.2%)    4/48 (8.3%)  
#>       High                                                                             
#>                 Not high baseline status   6/29 (20.7%)   4/49 (8.2%)     3/44 (6.8%)  
#>                 High baseline status           0/5         1/4 (25%)       1/4 (25%)   
#>                 Total                      6/34 (17.6%)   5/53 (9.4%)     4/48 (8.3%)  
#>     WEEK 5 DAY 36                                                                      
#>       Low                                                                              
#>                 Not low baseline status    4/41 (9.8%)    6/39 (15.4%)    5/38 (13.2%) 
#>                 Low baseline status            0/2            0/5             0/2      
#>                 Total                      4/43 (9.3%)    6/44 (13.6%)    5/40 (12.5%) 
#>       High                                                                             
#>                 Not high baseline status   3/37 (8.1%)    5/39 (12.8%)    4/38 (10.5%) 
#>                 High baseline status           0/6            0/5          1/2 (50%)   
#>                 Total                       3/43 (7%)     5/44 (11.4%)    5/40 (12.5%)