LBT07 Table 1 (Default) Laboratory Test Results and Change from Baseline by Visit.

The LBT07 table provides an overview of the analysis values and its change from baseline of each respective arm over the course of the trial.

Usage

lbt07_1_main(
  adam_db,
  arm_var = "ACTARM",
  lbl_overall = NULL,
  grade_var = c("PARAM", "GRADE_DIR", "GRADE_ANL"),
  deco = std_deco("LBT07"),
  lbl_grade_var = c("Parameter", "Direction of Abnormality", "Toxicity Grade"),
  req_tables = c("adsl", "adlb"),
  ...
)

lbt07_1_lyt(
  arm_var,
  lbl_overall,
  lbl_gradedir,
  lbl_param,
  grade_var,
  lbl_grade_var,
  deco,
  map
)

lbt07_1_pre(adam_db, ...)

lbt07_1_check(adam_db, req_tables = c("adsl", "adlb"), arm_var = "ACTARM")

lbt07_1_post(tlg, prune_0 = TRUE, ...)

lbt07_1

Format

An object of class chevron_t of length 1.

Arguments

adam_db

(dm) object containing the ADaM datasets

arm_var

(character) variable used for column splitting

lbl_overall

(character) label used for overall column, if set to NULL the overall column is omitted

grade_var

(character) PARAM and variables derived from the standard lab grade variable ATOXGR:

A grade direction variable (GRADE_DIR) is required in order to obtain the correct denominators when building the layout as it is used to define row splitting.

A toxicity grade variable (e.g. GRADE_ANL) where all negative values from ATOXGR are replaced by their absolute values.

deco

(character) decoration with title, subtitles and main_footer content

lbl_grade_var

(character) label of the variables in grade_var. If NULL, uses the label attribute of the columns selected in grade_var.

req_tables

(character) names of the required tables.

...

not used.

lbl_gradedir

(character) label of the GRADE_DIR variable.

lbl_param

(character) label of the PARAM variable.

map

(data.frame) mapping of PARAMs to directions of abnormality.

tlg

(TableTree, Listing or ggplot) object typically produced by a main function.

prune_0

(logical) remove 0 count rows

Details

Split columns by arm, typically ACTARM.

Functions

lbt07_1_main(): Main TLG function

lbt07_1_lyt(): Layout

lbt07_1_pre(): Preprocessing

lbt07_1_check(): Checks

lbt07_1_post(): Postprocessing

Note

adam_db object must contain an adlb table with columns "USUBJID", "ATOXGR", "ONTRTFL" and column specified by arm_var.

Examples

run(lbt07_1, syn_data)
#>   Parameter                                                                          
#>     Direction of Abnormality                 A: Drug X    B: Placebo   C: Combination
#>               Highest NCI CTCAE Grade         (N=134)      (N=134)        (N=132)    
#>   ———————————————————————————————————————————————————————————————————————————————————
#>   Alanine Aminotransferase Measurement (n)      134          133            132      
#>     LOW                                                                              
#>               1                              14 (10.4%)   15 (11.3%)     10 (7.6%)   
#>               2                              13 (9.7%)    18 (13.5%)     11 (8.3%)   
#>               3                              20 (14.9%)   12 (9.0%)      10 (7.6%)   
#>               4                               7 (5.2%)     8 (6.0%)      10 (7.6%)   
#>               Any                            54 (40.3%)   53 (39.8%)     41 (31.1%)  
#>     HIGH                                                                             
#>               1                              18 (13.4%)   16 (12.0%)     18 (13.6%)  
#>               2                              15 (11.2%)   13 (9.8%)      17 (12.9%)  
#>               3                              10 (7.5%)    16 (12.0%)     13 (9.8%)   
#>               4                              12 (9.0%)    11 (8.3%)      13 (9.8%)   
#>               Any                            55 (41.0%)   56 (42.1%)     61 (46.2%)  
#>   C-Reactive Protein Measurement (n)            134          134            132      
#>     LOW                                                                              
#>               1                              16 (11.9%)   19 (14.2%)     15 (11.4%)  
#>               2                              21 (15.7%)   13 (9.7%)      16 (12.1%)  
#>               3                              12 (9.0%)     9 (6.7%)      18 (13.6%)  
#>               4                              11 (8.2%)     7 (5.2%)       5 (3.8%)   
#>               Any                            60 (44.8%)   48 (35.8%)     54 (40.9%)  
#>     HIGH                                                                             
#>               1                              17 (12.7%)   15 (11.2%)      9 (6.8%)   
#>               2                              15 (11.2%)   16 (11.9%)     13 (9.8%)   
#>               3                              16 (11.9%)   12 (9.0%)      14 (10.6%)  
#>               4                              12 (9.0%)    12 (9.0%)      10 (7.6%)   
#>               Any                            60 (44.8%)   55 (41.0%)     46 (34.8%)  
#>   Immunoglobulin A Measurement (n)              133          133            132      
#>     LOW                                                                              
#>               1                              17 (12.8%)   23 (17.3%)     17 (12.9%)  
#>               2                              12 (9.0%)    15 (11.3%)     14 (10.6%)  
#>               3                              15 (11.3%)   11 (8.3%)      14 (10.6%)  
#>               4                              10 (7.5%)    16 (12.0%)     10 (7.6%)   
#>               Any                            54 (40.6%)   65 (48.9%)     55 (41.7%)  
#>     HIGH                                                                             
#>               1                              25 (18.8%)   14 (10.5%)     13 (9.8%)   
#>               2                              14 (10.5%)   19 (14.3%)     16 (12.1%)  
#>               3                              12 (9.0%)    12 (9.0%)      17 (12.9%)  
#>               4                              11 (8.3%)    13 (9.8%)       4 (3.0%)   
#>               Any                            62 (46.6%)   58 (43.6%)     50 (37.9%)