LBT14 Laboratory Test Results Shift Table – Highest NCI-CTCAE Grade Post-Baseline by Baseline Grade (Low or High Direction).

LBT14 Laboratory Test Results Shift Table – Highest NCI-CTCAE Grade Post-Baseline by Baseline Grade (Low or High Direction).

Usage

lbt14_main(adam_db, arm_var = "ACTARM", gr_missing = "incl", ...)

lbt14_pre(adam_db, gr_missing = "incl", direction = "low", ...)

lbt14_post(tlg, prune_0 = TRUE, ...)

lbt14

Format

An object of class chevron_t of length 1.

Arguments

adam_db

(list of data.frames) object containing the ADaM datasets

arm_var

(string) variable used for column splitting

gr_missing

(string) how missing baseline grades should be handled. Defaults to "incl" to include the "Missing" level. Other options are "excl" to exclude patients with missing baseline grades and "gr_0" to convert missing baseline grades to grade 0.

...

not used.

direction

(string) one of "high" or "low" indicating which shift direction should be detailed.

tlg

(TableTree, Listing or ggplot) object typically produced by a main function.

prune_0

(flag) remove 0 count rows

Details

Only the worst grade recorded for each patient is included in the table.

If no missing baseline lab results, the "Missing" level of BTOXGR is excluded.

Grades 0, 1, 2, 3, and 4 are counted as "Not Low" when direction = "low". Conversely, when direction = "high", Grades 0, -1, -2, -3, and -4 are counted as `"Not High".

Remove zero-count rows unless overridden with prune_0 = FALSE.

Split columns by arm, typically ACTARM.

Functions

lbt14_main(): Main TLG function

lbt14_pre(): Preprocessing

lbt14_post(): Postprocessing

Note

adam_db object must contain an adlb table with columns "USUBJID", "PARAM", "BTOXGR", "ATOXGR", and the column specified by arm_var.

Examples

run(lbt14, syn_data)
#>   Parameter                                                                             
#>         Baseline NCI-CTCAE Grade                A: Drug X    B: Placebo   C: Combination
#>                 Post-baseline NCI-CTCAE Grade    (N=134)      (N=134)        (N=132)    
#>   ——————————————————————————————————————————————————————————————————————————————————————
#>   Alanine Aminotransferase Measurement                                                  
#>         Not Low                                    124          122            117      
#>                 Not Low                         74 (55.2%)   76 (56.7%)     80 (60.6%)  
#>                 1                               13 (9.7%)    12 (9.0%)       9 (6.8%)   
#>                 2                               13 (9.7%)    16 (11.9%)     11 (8.3%)   
#>                 3                               17 (12.7%)   11 (8.2%)       9 (6.8%)   
#>                 4                                7 (5.2%)     7 (5.2%)       8 (6.1%)   
#>         1                                           3            6              7       
#>                 Not Low                             0         2 (1.5%)       5 (3.8%)   
#>                 1                                1 (0.7%)     2 (1.5%)          0       
#>                 2                                   0         1 (0.7%)          0       
#>                 3                                2 (1.5%)     1 (0.7%)          0       
#>                 4                                   0            0           2 (1.5%)   
#>         2                                           2            1              4       
#>                 Not Low                          2 (1.5%)        0           3 (2.3%)   
#>                 2                                   0         1 (0.7%)          0       
#>                 3                                   0            0           1 (0.8%)   
#>         3                                           2            3              4       
#>                 Not Low                          2 (1.5%)     2 (1.5%)       3 (2.3%)   
#>                 1                                   0         1 (0.7%)       1 (0.8%)   
#>         4                                           3            2              0       
#>                 Not Low                          2 (1.5%)     1 (0.7%)          0       
#>                 3                                1 (0.7%)        0              0       
#>                 4                                   0         1 (0.7%)          0       
#>   C-Reactive Protein Measurement                                                        
#>         Not Low                                    122          125            120      
#>                 Not Low                         67 (50.0%)   81 (60.4%)     70 (53.0%)  
#>                 1                               14 (10.4%)   17 (12.7%)     13 (9.8%)   
#>                 2                               20 (14.9%)   11 (8.2%)      16 (12.1%)  
#>                 3                               12 (9.0%)     9 (6.7%)      17 (12.9%)  
#>                 4                                9 (6.7%)     7 (5.2%)       4 (3.0%)   
#>         1                                           2            5              5       
#>                 Not Low                             0         3 (2.2%)       4 (3.0%)   
#>                 1                                   0         1 (0.7%)       1 (0.8%)   
#>                 2                                1 (0.7%)     1 (0.7%)          0       
#>                 4                                1 (0.7%)        0              0       
#>         2                                           5            1              1       
#>                 Not Low                          3 (2.2%)        0           1 (0.8%)   
#>                 1                                2 (1.5%)        0              0       
#>                 2                                   0         1 (0.7%)          0       
#>         3                                           2            2              1       
#>                 Not Low                          2 (1.5%)     1 (0.7%)          0       
#>                 1                                   0         1 (0.7%)          0       
#>                 3                                   0            0           1 (0.8%)   
#>         4                                           3            1              5       
#>                 Not Low                          2 (1.5%)     1 (0.7%)       3 (2.3%)   
#>                 1                                   0            0           1 (0.8%)   
#>                 4                                1 (0.7%)        0           1 (0.8%)   
#>   Immunoglobulin A Measurement                                                          
#>         Not Low                                    120          119            120      
#>                 Not Low                         69 (51.5%)   58 (43.3%)     70 (53.0%)  
#>                 1                               16 (11.9%)   20 (14.9%)     16 (12.1%)  
#>                 2                               12 (9.0%)    14 (10.4%)     14 (10.6%)  
#>                 3                               14 (10.4%)   11 (8.2%)      13 (9.8%)   
#>                 4                                9 (6.7%)    16 (11.9%)      7 (5.3%)   
#>         1                                           8            8              3       
#>                 Not Low                          7 (5.2%)     4 (3.0%)       2 (1.5%)   
#>                 1                                1 (0.7%)     3 (2.2%)          0       
#>                 2                                   0         1 (0.7%)          0       
#>                 4                                   0            0           1 (0.8%)   
#>         2                                           2            1              2       
#>                 Not Low                          2 (1.5%)     1 (0.7%)       2 (1.5%)   
#>         3                                           3            3              4       
#>                 Not Low                          1 (0.7%)     3 (2.2%)       2 (1.5%)   
#>                 1                                   0            0           1 (0.8%)   
#>                 3                                1 (0.7%)        0              0       
#>                 4                                1 (0.7%)        0           1 (0.8%)   
#>         4                                           1            3              3       
#>                 Not Low                          1 (0.7%)     3 (2.2%)       1 (0.8%)   
#>                 3                                   0            0           1 (0.8%)   
#>                 4                                   0            0           1 (0.8%)