LBT14 Table 1 (Low Direction) Laboratory Test Results Shift Table – Highest NCI-CTCAE Grade Post-Baseline by Baseline Grade.

LBT14 Table 1 (Low Direction) Laboratory Test Results Shift Table – Highest NCI-CTCAE Grade Post-Baseline by Baseline Grade.

Usage

lbt14_1_main(
  adam_db,
  arm_var = "ACTARM",
  gr_missing = "incl",
  title = "LBT14 - Low Direction",
  deco = std_deco("LBT14")
)

lbt14_1_lyt(arm_var, gr_missing, title, deco)

lbt14_1_pre(
  adam_db,
  gr_missing = "incl",
  req_tables = c("adsl", "adlb"),
  arm_var = "ACTARM"
)

lbt14_1_check(adam_db, req_tables = c("adsl", "adlb"), arm_var = "ACTARM")

lbt14_1_post(tlg, prune_0 = TRUE)

lbt14_1

Format

An object of class chevron_t of length 1.

Arguments

adam_db

(dm) object containing the ADaM datasets

arm_var

(character) variable used for column splitting

gr_missing

how missing baseline grades should be handled. Defaults to "incl" to include the "Missing" level. Other options are "excl" to exclude patients with missing baseline grades and "gr_0" to convert missing baseline grades to grade 0.

title

table title to display. Defaults to "LBT14 - Low Direction".

deco

(character) decoration with title, subtitles and main_footer content

req_tables

(character) names of the required tables.

tlg

(TableTree, Listing or ggplot) object typically produced by a main function.

prune_0

(logical) remove 0 count rows

Details

Only the worst grade recorded for each patient is included in the table.

If no missing baseline lab results, the "Missing" level of BTOXGR is excluded.

Grades 0, 1, 2, 3, and 4 are counted as "Not Low".

Remove zero-count rows unless overridden with prune_0 = FALSE.

Split columns by arm, typically ACTARM.

Functions

lbt14_1_main(): Main TLG function

lbt14_1_lyt(): Layout

lbt14_1_pre(): Preprocessing

lbt14_1_check(): Checks

lbt14_1_post(): Postprocessing

Note

adam_db object must contain an adlb table with columns "USUBJID", "PARAM", "BTOXGR", "ATOXGR", and the column specified by arm_var.

Examples

run(lbt14_1, syn_data)
#> LBT14 - Low Direction
#> 
#> 
#>   ——————————————————————————————————————————————————————————————————————————————————————
#>   Parameter                                                                             
#>         Baseline NCI-CTCAE Grade                A: Drug X    B: Placebo   C: Combination
#>                 Post-baseline NCI-CTCAE Grade    (N=134)      (N=134)        (N=132)    
#>   ——————————————————————————————————————————————————————————————————————————————————————
#>   Alanine Aminotransferase Measurement                                                  
#>         Not Low                                    124          122            117      
#>                 Not Low                         74 (59.7%)   76 (62.3%)     80 (68.4%)  
#>                 1                               13 (10.5%)   12 (9.8%)       9 (7.7%)   
#>                 2                               13 (10.5%)   16 (13.1%)     11 (9.4%)   
#>                 3                               17 (13.7%)   11 (9.0%)       9 (7.7%)   
#>                 4                                7 (5.6%)     7 (5.7%)       8 (6.8%)   
#>         1                                           3            6              7       
#>                 Not Low                             0        2 (33.3%)      5 (71.4%)   
#>                 1                               1 (33.3%)    2 (33.3%)          0       
#>                 2                                   0        1 (16.7%)          0       
#>                 3                               2 (66.7%)    1 (16.7%)          0       
#>                 4                                   0            0          2 (28.6%)   
#>         2                                           2            1              4       
#>                 Not Low                          2 (100%)        0          3 (75.0%)   
#>                 2                                   0         1 (100%)          0       
#>                 3                                   0            0          1 (25.0%)   
#>         3                                           2            3              4       
#>                 Not Low                          2 (100%)    2 (66.7%)      3 (75.0%)   
#>                 1                                   0        1 (33.3%)      1 (25.0%)   
#>         4                                           3            2              0       
#>                 Not Low                         2 (66.7%)    1 (50.0%)          0       
#>                 3                               1 (33.3%)        0              0       
#>                 4                                   0        1 (50.0%)          0       
#>   C-Reactive Protein Measurement                                                        
#>         Not Low                                    122          125            120      
#>                 Not Low                         67 (54.9%)   81 (64.8%)     70 (58.3%)  
#>                 1                               14 (11.5%)   17 (13.6%)     13 (10.8%)  
#>                 2                               20 (16.4%)   11 (8.8%)      16 (13.3%)  
#>                 3                               12 (9.8%)     9 (7.2%)      17 (14.2%)  
#>                 4                                9 (7.4%)     7 (5.6%)       4 (3.3%)   
#>         1                                           2            5              5       
#>                 Not Low                             0        3 (60.0%)      4 (80.0%)   
#>                 1                                   0        1 (20.0%)      1 (20.0%)   
#>                 2                               1 (50.0%)    1 (20.0%)          0       
#>                 4                               1 (50.0%)        0              0       
#>         2                                           5            1              1       
#>                 Not Low                         3 (60.0%)        0           1 (100%)   
#>                 1                               2 (40.0%)        0              0       
#>                 2                                   0         1 (100%)          0       
#>         3                                           2            2              1       
#>                 Not Low                          2 (100%)    1 (50.0%)          0       
#>                 1                                   0        1 (50.0%)          0       
#>                 3                                   0            0           1 (100%)   
#>         4                                           3            1              5       
#>                 Not Low                         2 (66.7%)     1 (100%)      3 (60.0%)   
#>                 1                                   0            0          1 (20.0%)   
#>                 4                               1 (33.3%)        0          1 (20.0%)   
#>   Immunoglobulin A Measurement                                                          
#>         Not Low                                    120          119            120      
#>                 Not Low                         69 (57.5%)   58 (48.7%)     70 (58.3%)  
#>                 1                               16 (13.3%)   20 (16.8%)     16 (13.3%)  
#>                 2                               12 (10.0%)   14 (11.8%)     14 (11.7%)  
#>                 3                               14 (11.7%)   11 (9.2%)      13 (10.8%)  
#>                 4                                9 (7.5%)    16 (13.4%)      7 (5.8%)   
#>         1                                           8            8              3       
#>                 Not Low                         7 (87.5%)    4 (50.0%)      2 (66.7%)   
#>                 1                               1 (12.5%)    3 (37.5%)          0       
#>                 2                                   0        1 (12.5%)          0       
#>                 4                                   0            0          1 (33.3%)   
#>         2                                           2            1              2       
#>                 Not Low                          2 (100%)     1 (100%)       2 (100%)   
#>         3                                           3            3              4       
#>                 Not Low                         1 (33.3%)     3 (100%)      2 (50.0%)   
#>                 1                                   0            0          1 (25.0%)   
#>                 3                               1 (33.3%)        0              0       
#>                 4                               1 (33.3%)        0          1 (25.0%)   
#>         4                                           1            3              3       
#>                 Not Low                          1 (100%)     3 (100%)      1 (33.3%)   
#>                 3                                   0            0          1 (33.3%)   
#>                 4                                   0            0          1 (33.3%)