Remove specific AVAL and AVALCAT1 data from adex

Useful if some continuous AVAL data, or categorical AVALCAT1 data corresponding to specific PARAM should not be displayed.

remove_adex_aval(
  adam_db,
  show_stats = .study$show_cont_stats,
  show_bins = .study$show_cat_stats,
  .study = list(show_cont_stats = c("ALL"), show_cat_stats = c("ALL"))
)

Arguments

adam_db

(dm) object containing the ADaM datasets

show_stats

(vector of character) providing the name of the parameters whose statistical summary should be presented. To analyze all, provide show_stats = "ALL" (Default), to analyze none, provide show_stats = "".

show_bins

(vector of character) providing the name of the parameters whose categorical summary should be presented. To analyze all, provide show_bins = "ALL" (Default), to analyze none, provide show_bins = "".

.study

(list) with default values for the arguments of the function