VST01 Vital Sign Results and change from Baseline By Visit Table.

The VST01 table provides an overview of the Vital Sign values and its change from baseline of each respective arm over the course of the trial.

Usage

egt01_pre(adam_db, dataset = "adeg", ...)

vst01_main(
  adam_db,
  dataset = "advs",
  arm_var = "ACTARM",
  row_split_var = NULL,
  summaryvars = c("AVAL", "CHG"),
  visitvar = "AVISIT",
  precision = list(),
  default_precision = 2,
  page_by = TRUE,
  ...
)

vst01_pre(adam_db, dataset = "advs", ...)

vst01

Format

An object of class chevron_t of length 1.

Source

cfbt01.R

Arguments

adam_db

(list of data.frames) object containing the ADaM datasets

dataset

(string) the name of a table in the adam_db object.

...

not used.

arm_var

(string) variable used for column splitting

row_split_var

(character) row split variable other than PARAMCD.

summaryvars

(character) variables to be analyzed. The label attribute of the corresponding column in table of adam_db is used as label.

visitvar

(string) typically one of "AVISIT" (Default) or "ATPTN" depending on the type of time point to be displayed

precision

(named list of integer) where names are values found in the PARAMCD column and the the values indicate the number of digits that should be represented for min, max and median. Mean and sd are represented with one more decimal of precision.

default_precision

(integer) the default number of digits.

page_by

(flag) indicator whether the parameter row split is by page.

Details

The Analysis Value column, displays the number of patients, the mean, standard deviation, median and range of the analysis value for each visit.

The Change from Baseline column, displays the number of patient and the mean, standard deviation, median and range of changes relative to the baseline.

Remove zero-count rows unless overridden with prune_0 = FALSE.

Split columns by arm, typically ACTARM.

Does not include a total column by default.

Sorted based on factor level; first by PARAM labels in alphabetic order then by chronological time point given by AVISIT. Re-level to customize order

Functions

egt01_pre(): Preprocessing

vst01_main(): Main TLG function

vst01_pre(): Preprocessing

Note

adam_db object must contain table named as dataset with the columns specified in summaryvars.

Examples

run(vst01, syn_data)
#>     Analysis Visit                      A: Drug X                          B: Placebo                       C: Combination         
#>                                                 Change from                        Change from                        Change from  
#>                              Value at Visit      Baseline       Value at Visit      Baseline       Value at Visit      Baseline    
#>                                 (N=134)           (N=134)          (N=134)           (N=134)          (N=132)           (N=132)    
#>   —————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————
#>   Diastolic Blood Pressure                                                                                                         
#>     SCREENING                                                                                                                      
#>       n                           134                0               134                0               132                0       
#>       Mean (SD)              49.968 (7.197)       NE (NE)       50.753 (8.429)       NE (NE)       50.192 (7.627)       NE (NE)    
#>       Median                     49.706             NE              50.087             NE              49.641             NE       
#>       Min - Max              31.69 - 71.19        NE - NE       29.26 - 69.21        NE - NE       26.89 - 69.98        NE - NE    
#>     BASELINE                                                                                                                       
#>       n                           134                                134                                132                        
#>       Mean (SD)              48.602 (7.962)                     50.442 (7.948)                     51.107 (7.790)                  
#>       Median                     48.418                             50.179                             50.804                      
#>       Min - Max              27.71 - 64.64                      21.68 - 67.51                      29.75 - 71.40                   
#>     WEEK 1 DAY 8                                                                                                                   
#>       n                           134               134              134               134              132               132      
#>       Mean (SD)              50.260 (7.514)   1.658 (10.774)    49.674 (7.743)   -0.767 (10.947)   48.855 (7.888)   -2.251 (10.380)
#>       Median                     50.088            0.607            49.747           -1.114            47.677           -2.229     
#>       Min - Max              33.04 - 68.98    -20.37 - 29.94    33.71 - 66.49    -25.80 - 28.37    30.39 - 66.99    -22.06 - 22.44 
#>     WEEK 2 DAY 15                                                                                                                  
#>       n                           134               134              134               134              132               132      
#>       Mean (SD)              50.836 (7.850)   2.234 (11.752)    49.718 (8.445)   -0.724 (12.444)   49.981 (8.338)   -1.126 (11.639)
#>       Median                     51.437            3.126            50.227           -0.709            51.007           -2.271     
#>       Min - Max              31.89 - 70.34    -23.88 - 39.23    30.64 - 68.11    -28.40 - 36.97    24.82 - 65.57    -29.24 - 30.72 
#>     WEEK 3 DAY 22                                                                                                                  
#>       n                           134               134              134               134              132               132      
#>       Mean (SD)              50.712 (7.815)   2.110 (10.865)    49.092 (7.941)   -1.350 (11.104)   49.944 (7.662)   -1.162 (10.901)
#>       Median                     50.157            2.611            49.533           -2.317            50.025           -0.617     
#>       Min - Max              29.07 - 74.89    -25.78 - 30.66    24.50 - 67.11    -28.09 - 33.90    31.52 - 68.55    -29.72 - 26.98 
#>     WEEK 4 DAY 29                                                                                                                  
#>       n                           134               134              134               134              132               132      
#>       Mean (SD)              50.074 (8.071)   1.472 (10.885)    49.624 (7.102)   -0.818 (10.634)   49.707 (8.048)   -1.399 (11.047)
#>       Median                     49.269            1.202            49.193           -0.757            49.844           -0.182     
#>       Min - Max              29.41 - 71.36    -30.92 - 24.76    32.71 - 66.98    -33.18 - 35.95    28.75 - 68.45    -35.47 - 28.36 
#>     WEEK 5 DAY 36                                                                                                                  
#>       n                           134               134              134               134              132               132      
#>       Mean (SD)              50.569 (7.531)   1.967 (11.206)    48.366 (7.959)   -2.075 (11.736)   49.091 (7.570)   -2.015 (10.658)
#>       Median                     49.422            1.573            47.721           -2.871            49.237           -2.247     
#>       Min - Max              35.06 - 72.21    -25.22 - 31.47    30.42 - 70.61    -31.23 - 29.16    30.59 - 72.91    -30.45 - 34.57 
#>   Pulse Rate                                                                                                                       
#>     SCREENING                                                                                                                      
#>       n                           134                0               134                0               132                0       
#>       Mean (SD)              49.599 (8.981)       NE (NE)       49.408 (8.388)       NE (NE)       49.293 (8.593)       NE (NE)    
#>       Median                     49.871             NE              48.874             NE              48.603             NE       
#>       Min - Max              25.94 - 68.80        NE - NE       30.54 - 80.48        NE - NE       26.48 - 67.52        NE - NE    
#>     BASELINE                                                                                                                       
#>       n                           134                                134                                132                        
#>       Mean (SD)              51.861 (8.077)                     50.270 (8.808)                     50.255 (9.366)                  
#>       Median                     51.232                             51.248                             50.698                      
#>       Min - Max              28.47 - 72.90                      26.92 - 67.34                      27.78 - 71.18                   
#>     WEEK 1 DAY 8                                                                                                                   
#>       n                           134               134              134               134              132               132      
#>       Mean (SD)              50.065 (8.316)   -1.796 (11.589)   49.267 (7.520)   -1.002 (11.583)   49.796 (7.650)   -0.459 (12.609)
#>       Median                     49.930           -2.741            48.776           -2.188            49.478           -1.452     
#>       Min - Max              26.70 - 69.37    -26.42 - 29.59    30.70 - 67.76    -29.47 - 32.12    30.14 - 72.89    -25.12 - 33.32 
#>     WEEK 2 DAY 15                                                                                                                  
#>       n                           134               134              134               134              132               132      
#>       Mean (SD)              49.710 (8.757)   -2.150 (12.372)   50.832 (7.595)   0.562 (11.778)    49.058 (7.657)   -1.197 (11.326)
#>       Median                     49.224           -3.015            50.253            0.081            49.200           -1.144     
#>       Min - Max              25.97 - 69.62    -37.59 - 26.40    33.32 - 70.27    -28.92 - 37.68    32.20 - 73.64    -33.74 - 26.35 
#>     WEEK 3 DAY 22                                                                                                                  
#>       n                           134               134              134               134              132               132      
#>       Mean (SD)              50.450 (7.436)   -1.410 (11.076)   49.915 (7.818)   -0.355 (10.985)   49.803 (7.156)   -0.452 (12.082)
#>       Median                     50.600           -1.147            49.386           -0.791            49.648           -2.273     
#>       Min - Max              27.72 - 72.97    -30.00 - 26.46    27.84 - 71.02    -26.47 - 33.42    27.94 - 67.47    -28.68 - 28.02 
#>     WEEK 4 DAY 29                                                                                                                  
#>       n                           134               134              134               134              132               132      
#>       Mean (SD)              48.956 (8.315)   -2.905 (11.073)   49.959 (8.390)   -0.310 (11.956)   50.991 (7.662)   0.736 (11.571) 
#>       Median                     48.846           -3.039            49.948           -0.407            51.400            0.687     
#>       Min - Max              32.27 - 78.39    -28.35 - 30.35    30.43 - 79.82    -29.06 - 28.32    31.41 - 70.33    -30.57 - 25.08 
#>     WEEK 5 DAY 36                                                                                                                  
#>       n                           134               134              134               134              132               132      
#>       Mean (SD)              48.404 (8.500)   -3.457 (11.602)   50.309 (7.177)   0.039 (11.233)    49.724 (7.584)   -0.531 (12.428)
#>       Median                     47.703           -3.970            50.307           -1.452            49.061           -2.687     
#>       Min - Max              24.96 - 75.72    -41.33 - 31.50    29.38 - 67.65    -24.94 - 27.29    35.29 - 73.67    -24.59 - 34.05 
#>   Respiratory Rate                                                                                                                 
#>     SCREENING                                                                                                                      
#>       n                           134                0               134                0               132                0       
#>       Mean (SD)              49.176 (8.185)       NE (NE)       50.312 (8.326)       NE (NE)       49.623 (8.440)       NE (NE)    
#>       Median                     49.265             NE              49.873             NE              50.058             NE       
#>       Min - Max              26.02 - 74.45        NE - NE       31.43 - 69.21        NE - NE       24.29 - 67.89        NE - NE    
#>     BASELINE                                                                                                                       
#>       n                           134                                134                                132                        
#>       Mean (SD)              49.406 (8.087)                     51.116 (8.187)                     50.754 (8.413)                  
#>       Median                     49.781                             50.523                             51.107                      
#>       Min - Max              29.33 - 69.31                      31.19 - 74.52                      26.15 - 70.95                   
#>     WEEK 1 DAY 8                                                                                                                   
#>       n                           134               134              134               134              132               132      
#>       Mean (SD)              49.488 (8.043)   0.082 (11.647)    49.982 (7.919)   -1.134 (11.530)   50.967 (7.820)   0.213 (11.140) 
#>       Median                     49.276            0.244            49.147           -1.619            51.117            0.436     
#>       Min - Max              24.14 - 66.14    -31.62 - 32.15    25.05 - 67.71    -29.13 - 30.92    29.89 - 69.76    -23.60 - 27.67 
#>     WEEK 2 DAY 15                                                                                                                  
#>       n                           134               134              134               134              132               132      
#>       Mean (SD)              49.662 (7.678)    0.256 (9.707)    50.350 (8.663)   -0.766 (12.149)   49.309 (8.540)   -1.445 (12.012)
#>       Median                     49.211            1.369            49.736           -1.573            50.184           -1.611     
#>       Min - Max              30.81 - 71.33    -24.86 - 27.63    28.47 - 77.02    -29.01 - 28.89    22.40 - 67.65    -43.84 - 26.98 
#>     WEEK 3 DAY 22                                                                                                                  
#>       n                           134               134              134               134              132               132      
#>       Mean (SD)              50.605 (8.215)   1.199 (11.669)    50.155 (6.857)   -0.960 (10.139)   49.147 (7.348)   -1.607 (11.112)
#>       Median                     50.517            0.664            49.664           -0.707            49.563           -0.867     
#>       Min - Max              33.17 - 70.78    -33.52 - 33.26    32.30 - 67.83    -29.57 - 22.20    27.42 - 63.34    -25.43 - 33.62 
#>     WEEK 4 DAY 29                                                                                                                  
#>       n                           134               134              134               134              132               132      
#>       Mean (SD)              49.424 (7.669)   0.018 (10.853)    49.873 (7.786)   -1.243 (11.005)   50.806 (7.810)   0.052 (11.106) 
#>       Median                     48.900           -1.369            48.933           -2.448            50.364           -0.600     
#>       Min - Max              25.94 - 69.65    -25.87 - 40.32    33.70 - 68.55    -25.04 - 29.17    31.87 - 71.97    -27.64 - 31.06 
#>     WEEK 5 DAY 36                                                                                                                  
#>       n                           134               134              134               134              132               132      
#>       Mean (SD)              49.878 (7.731)   0.472 (11.474)    50.632 (8.216)   -0.484 (11.956)   49.945 (7.893)   -0.809 (10.574)
#>       Median                     49.847           -0.208            50.304           -0.565            49.622           -1.718     
#>       Min - Max              34.34 - 70.70    -30.30 - 37.18    28.42 - 69.84    -29.38 - 23.81    29.78 - 68.40    -21.30 - 22.32 
#>   Systolic Blood Pressure                                                                                                          
#>     SCREENING                                                                                                                      
#>       n                           134                0               134                0               132                0       
#>       Mean (SD)              49.572 (8.322)       NE (NE)       50.306 (8.343)       NE (NE)       50.902 (7.822)       NE (NE)    
#>       Median                     49.571             NE              50.155             NE              50.774             NE       
#>       Min - Max              23.97 - 70.85        NE - NE       26.22 - 79.12        NE - NE       27.63 - 67.45        NE - NE    
#>     BASELINE                                                                                                                       
#>       n                           134                                134                                132                        
#>       Mean (SD)              49.407 (8.467)                     50.245 (8.531)                     48.494 (7.184)                  
#>       Median                     48.335                             49.971                             49.217                      
#>       Min - Max              24.34 - 71.06                      24.38 - 71.07                      26.22 - 63.35                   
#>     WEEK 1 DAY 8                                                                                                                   
#>       n                           134               134              134               134              132               132      
#>       Mean (SD)              50.702 (9.166)   1.295 (12.814)    49.263 (8.693)   -0.982 (11.221)   49.647 (7.991)   1.153 (10.185) 
#>       Median                     49.526            0.129            48.302           -1.678            49.768            0.085     
#>       Min - Max              29.75 - 78.99    -29.51 - 37.65    32.99 - 73.97    -27.62 - 24.77    25.52 - 68.62    -22.79 - 31.30 
#>     WEEK 2 DAY 15                                                                                                                  
#>       n                           134               134              134               134              132               132      
#>       Mean (SD)              49.369 (8.469)   -0.038 (11.936)   49.377 (7.009)   -0.869 (10.968)   49.260 (7.683)    0.766 (9.716) 
#>       Median                     49.489           -0.243            49.386           -0.929            48.289           -0.216     
#>       Min - Max              24.18 - 65.44    -39.12 - 32.72    27.62 - 69.23    -37.10 - 33.85    29.39 - 74.75    -24.92 - 28.99 
#>     WEEK 3 DAY 22                                                                                                                  
#>       n                           134               134              134               134              132               132      
#>       Mean (SD)              49.682 (7.776)   0.275 (11.383)    49.034 (7.844)   -1.211 (12.123)   50.544 (8.303)   2.050 (11.059) 
#>       Median                     50.136            1.479            48.794           -1.190            51.672            1.951     
#>       Min - Max              31.30 - 70.31    -27.49 - 27.15    15.58 - 66.17    -35.56 - 36.36    29.62 - 69.40    -29.72 - 30.62 
#>     WEEK 4 DAY 29                                                                                                                  
#>       n                           134               134              134               134              132               132      
#>       Mean (SD)              49.424 (7.637)   0.017 (11.473)    49.400 (8.640)   -0.846 (12.599)   50.657 (7.395)   2.162 (10.495) 
#>       Median                     49.783            0.599            49.927           -1.288            50.823            2.965     
#>       Min - Max              22.92 - 69.57    -32.30 - 28.46    22.14 - 72.06    -29.71 - 38.11    29.72 - 65.12    -28.75 - 34.15 
#>     WEEK 5 DAY 36                                                                                                                  
#>       n                           134               134              134               134              132               132      
#>       Mean (SD)              49.395 (7.898)   -0.012 (11.876)   50.222 (8.653)   -0.023 (11.828)   50.367 (8.176)   1.873 (10.490) 
#>       Median                     49.565            0.730            49.934            0.192            50.811            2.431     
#>       Min - Max              26.11 - 69.66    -39.25 - 33.93    27.91 - 74.17    -27.70 - 36.16    34.27 - 74.52    -23.70 - 32.42 
#>   Temperature                                                                                                                      
#>     SCREENING                                                                                                                      
#>       n                           134                0               134                0               132                0       
#>       Mean (SD)              48.947 (9.437)       NE (NE)       50.080 (7.875)       NE (NE)       50.208 (8.409)       NE (NE)    
#>       Median                     49.553             NE              50.118             NE              49.836             NE       
#>       Min - Max              28.96 - 75.23        NE - NE       29.91 - 73.35        NE - NE       30.45 - 72.25        NE - NE    
#>     BASELINE                                                                                                                       
#>       n                           134                                134                                132                        
#>       Mean (SD)              49.748 (8.270)                     50.873 (7.423)                     49.118 (8.248)                  
#>       Median                     50.441                             51.506                             48.685                      
#>       Min - Max              29.02 - 69.08                      30.70 - 67.70                      20.63 - 70.52                   
#>     WEEK 1 DAY 8                                                                                                                   
#>       n                           134               134              134               134              132               132      
#>       Mean (SD)              51.609 (8.125)   1.861 (11.836)    49.473 (8.207)   -1.400 (9.925)    49.756 (7.597)   0.638 (11.167) 
#>       Median                     52.257            2.303            48.810           -1.054            50.172            0.943     
#>       Min - Max              31.45 - 71.21    -27.99 - 33.15    27.07 - 67.32    -28.15 - 26.63    32.09 - 68.46    -28.01 - 29.36 
#>     WEEK 2 DAY 15                                                                                                                  
#>       n                           134               134              134               134              132               132      
#>       Mean (SD)              50.202 (7.515)   0.454 (11.009)    49.538 (6.968)   -1.334 (10.167)   51.764 (8.117)   2.646 (11.843) 
#>       Median                     50.659           -0.705            49.681           -0.746            51.613            2.760     
#>       Min - Max              35.15 - 67.74    -33.93 - 25.41    32.30 - 73.16    -29.69 - 29.02    30.10 - 73.30    -24.87 - 32.77 
#>     WEEK 3 DAY 22                                                                                                                  
#>       n                           134               134              134               134              132               132      
#>       Mean (SD)              50.077 (8.206)   0.329 (11.065)    49.878 (8.007)   -0.994 (10.913)   49.668 (8.247)   0.550 (11.114) 
#>       Median                     49.476            1.145            49.644           -1.092            49.245            0.487     
#>       Min - Max              28.96 - 68.65    -29.27 - 24.08    32.73 - 70.25    -28.29 - 21.57    35.19 - 75.20    -27.25 - 31.98 
#>     WEEK 4 DAY 29                                                                                                                  
#>       n                           134               134              134               134              132               132      
#>       Mean (SD)              50.667 (7.657)   0.919 (11.764)    49.787 (8.548)   -1.086 (11.790)   49.929 (7.450)   0.811 (11.526) 
#>       Median                     50.701            1.389            49.776           -1.706            49.793            0.063     
#>       Min - Max              29.72 - 70.80    -25.86 - 29.20    24.44 - 68.94    -28.18 - 37.26    28.39 - 68.64    -25.78 - 32.55 
#>     WEEK 5 DAY 36                                                                                                                  
#>       n                           134               134              134               134              132               132      
#>       Mean (SD)              50.329 (8.319)   0.581 (12.284)    50.455 (8.104)   -0.418 (11.688)   50.050 (7.275)   0.933 (10.808) 
#>       Median                     49.811           -0.496            50.827           -0.286            49.692            0.861     
#>       Min - Max              33.37 - 74.95    -30.07 - 33.36    34.61 - 77.96    -23.07 - 30.92    31.14 - 73.43    -23.82 - 39.14 
#>   Weight                                                                                                                           
#>     SCREENING                                                                                                                      
#>       n                           134                0               134                0               132                0       
#>       Mean (SD)              49.927 (7.443)       NE (NE)       49.736 (8.291)       NE (NE)       50.305 (9.057)       NE (NE)    
#>       Median                     48.686             NE              49.275             NE              49.745             NE       
#>       Min - Max              33.72 - 65.91        NE - NE       25.82 - 71.47        NE - NE       26.04 - 69.99        NE - NE    
#>     BASELINE                                                                                                                       
#>       n                           134                                134                                132                        
#>       Mean (SD)              50.726 (7.856)                     49.880 (8.342)                     49.377 (7.489)                  
#>       Median                     49.780                             50.674                             49.568                      
#>       Min - Max              33.33 - 71.90                      27.90 - 69.82                      29.39 - 65.05                   
#>     WEEK 1 DAY 8                                                                                                                   
#>       n                           134               134              134               134              132               132      
#>       Mean (SD)              49.230 (7.681)   -1.496 (11.399)   49.096 (7.612)   -0.785 (11.361)   49.572 (8.396)   0.196 (12.142) 
#>       Median                     50.507           -0.215            48.891           -2.667            49.864           -0.451     
#>       Min - Max              26.02 - 64.53    -42.93 - 21.34    31.45 - 66.13    -25.00 - 23.52    29.73 - 69.07    -31.02 - 32.60 
#>     WEEK 2 DAY 15                                                                                                                  
#>       n                           134               134              134               134              132               132      
#>       Mean (SD)              50.893 (8.178)   0.167 (11.238)    50.022 (7.807)   0.142 (12.504)    49.741 (7.937)   0.364 (10.827) 
#>       Median                     50.344            2.024            50.217           -0.224            49.762            1.317     
#>       Min - Max              32.10 - 70.09    -30.03 - 22.84    28.70 - 72.11    -29.23 - 34.60    28.09 - 73.67    -28.00 - 26.59 
#>     WEEK 3 DAY 22                                                                                                                  
#>       n                           134               134              134               134              132               132      
#>       Mean (SD)              48.962 (7.994)   -1.765 (11.362)   49.573 (8.348)   -0.307 (12.454)   50.446 (8.084)   1.069 (10.807) 
#>       Median                     48.868           -1.295            48.507           -0.982            50.805           -0.190     
#>       Min - Max              31.68 - 75.41    -29.72 - 24.92    32.44 - 75.98    -36.82 - 34.43    27.81 - 70.85    -28.93 - 31.73 
#>     WEEK 4 DAY 29                                                                                                                  
#>       n                           134               134              134               134              132               132      
#>       Mean (SD)              50.061 (8.083)   -0.665 (11.186)   49.529 (8.097)   -0.352 (10.428)   49.667 (7.509)   0.290 (10.415) 
#>       Median                     50.867           -0.870            49.367           -0.513            49.608           -1.187     
#>       Min - Max              30.09 - 68.88    -29.49 - 24.31    28.99 - 69.47    -24.91 - 26.30    30.79 - 77.45    -23.39 - 29.93 
#>     WEEK 5 DAY 36                                                                                                                  
#>       n                           134               134              134               134              132               132      
#>       Mean (SD)              49.023 (8.110)   -1.703 (11.514)   51.584 (8.486)   1.704 (11.545)    49.872 (7.400)   0.495 (10.871) 
#>       Median                     49.752           -0.777            51.807            1.219            49.590           -0.156     
#>       Min - Max              26.38 - 66.72    -34.78 - 29.74    26.29 - 76.63    -23.99 - 29.33    28.74 - 67.93    -36.15 - 31.04