VST01 Table 1 (Default) Vital Sign Results and change from Baseline By Visit Table 1.

The VST01 table provides an overview of the Vital Sign values and its change from baseline of each respective arm over the course of the trial.

Usage

vst01_1_main(
  adam_db,
  armvar = "ACTARM",
  summaryvars = c(`Value at Visit` = "AVAL", `Change from \nBaseline` = "CHG"),
  visitvar = "AVISIT",
  deco = std_deco("VST01"),
  ...
)

vst01_1_lyt(
  armvar,
  summaryvars,
  summaryvars_lbls,
  visitvar,
  lbl_avisit,
  lbl_param,
  deco,
  ...
)

vst01_1_pre(adam_db, ...)

vst01_1_post(tlg, prune_0 = TRUE, ...)

vst01_1

Format

An object of class chevron_t of length 1.

Arguments

adam_db

(dm) object containing the ADaM datasets

armvar

(character) variable used for column splitting

summaryvars

(vector of character) the variables to be analyzed. For this table, AVAL and CHG by default.

visitvar

(character) typically one of "AVISIT" (Default) or "ATPTN" depending on the type of time point to be displayed.

deco

(character) decoration with title, subtitles and main_footer content

...

not used.

summaryvars_lbls

(vector of character) the label of the variables to be analyzed.

lbl_avisit

(character) label of the visitvar variable.

lbl_param

(character) label of the PARAM variable.

tlg

(TableTree, Listing or ggplot) object typically produced by a main function.

prune_0

(logical) remove 0 count rows

Details

The Analysis Value column, displays the number of patients, the mean, standard deviation, median and range of the analysis value for each visit.

The Change from Baseline column, displays the number of patient and the mean, standard deviation, median and range of changes relative to the baseline.

Remove zero-count rows unless overridden with prune_0 = FALSE.

Split columns by arm, typically ACTARM.

Does not include a total column by default.

Sorted based on factor level; first by PARAM labels in alphabetic order then by chronological time point given by AVISIT. Re-level to customize order

Functions

vst01_1_main(): Main TLG function

vst01_1_lyt(): Layout

vst01_1_pre(): Preprocessing

vst01_1_post(): Postprocessing

Note

adam_db object must contain an advs table with the columns specified in summaryvars.

Examples

run(vst01_1, syn_data)
#>   Parameter                            A: Drug X                      B: Placebo                    C: Combination        
#>     Analysis Visit                            Change from                     Change from                     Change from 
#>                              Value at Visit     Baseline     Value at Visit     Baseline     Value at Visit     Baseline  
#>   ————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————
#>   Diastolic Blood Pressure                                                                                                
#>     SCREENING                                                                                                             
#>       n                           134              0              134              0              132              0      
#>       Mean (SD)                50.0 (7.2)          NA          50.8 (8.4)          NA          50.2 (7.6)          NA     
#>       Median                      49.7             NA             50.1             NA             49.6             NA     
#>       Min - Max               31.7 - 71.2          NA         29.3 - 69.2          NA         26.9 - 70.0          NA     
#>     BASELINE                                                                                                              
#>       n                           134             134             134             134             132             132     
#>       Mean (SD)                48.6 (8.0)      0.0 (0.0)       50.4 (7.9)      0.0 (0.0)       51.1 (7.8)      0.0 (0.0)  
#>       Median                      48.4            0.0             50.2            0.0             50.8            0.0     
#>       Min - Max               27.7 - 64.6      0.0 - 0.0      21.7 - 67.5      0.0 - 0.0      29.7 - 71.4      0.0 - 0.0  
#>     WEEK 1 DAY 8                                                                                                          
#>       n                           134             134             134             134             132             132     
#>       Mean (SD)                50.3 (7.5)      1.7 (10.8)      49.7 (7.7)     -0.8 (10.9)      48.9 (7.9)     -2.3 (10.4) 
#>       Median                      50.1            0.6             49.7            -1.1            47.7            -2.2    
#>       Min - Max               33.0 - 69.0     -20.4 - 29.9    33.7 - 66.5     -25.8 - 28.4    30.4 - 67.0     -22.1 - 22.4
#>     WEEK 2 DAY 15                                                                                                         
#>       n                           134             134             134             134             132             132     
#>       Mean (SD)                50.8 (7.8)      2.2 (11.8)      49.7 (8.4)     -0.7 (12.4)      50.0 (8.3)     -1.1 (11.6) 
#>       Median                      51.4            3.1             50.2            -0.7            51.0            -2.3    
#>       Min - Max               31.9 - 70.3     -23.9 - 39.2    30.6 - 68.1     -28.4 - 37.0    24.8 - 65.6     -29.2 - 30.7
#>     WEEK 3 DAY 22                                                                                                         
#>       n                           134             134             134             134             132             132     
#>       Mean (SD)                50.7 (7.8)      2.1 (10.9)      49.1 (7.9)     -1.3 (11.1)      49.9 (7.7)     -1.2 (10.9) 
#>       Median                      50.2            2.6             49.5            -2.3            50.0            -0.6    
#>       Min - Max               29.1 - 74.9     -25.8 - 30.7    24.5 - 67.1     -28.1 - 33.9    31.5 - 68.5     -29.7 - 27.0
#>     WEEK 4 DAY 29                                                                                                         
#>       n                           134             134             134             134             132             132     
#>       Mean (SD)                50.1 (8.1)      1.5 (10.9)      49.6 (7.1)     -0.8 (10.6)      49.7 (8.0)     -1.4 (11.0) 
#>       Median                      49.3            1.2             49.2            -0.8            49.8            -0.2    
#>       Min - Max               29.4 - 71.4     -30.9 - 24.8    32.7 - 67.0     -33.2 - 35.9    28.7 - 68.4     -35.5 - 28.4
#>     WEEK 5 DAY 36                                                                                                         
#>       n                           134             134             134             134             132             132     
#>       Mean (SD)                50.6 (7.5)      2.0 (11.2)      48.4 (8.0)     -2.1 (11.7)      49.1 (7.6)     -2.0 (10.7) 
#>       Median                      49.4            1.6             47.7            -2.9            49.2            -2.2    
#>       Min - Max               35.1 - 72.2     -25.2 - 31.5    30.4 - 70.6     -31.2 - 29.2    30.6 - 72.9     -30.5 - 34.6
#>   Pulse Rate                                                                                                              
#>     SCREENING                                                                                                             
#>       n                           134              0              134              0              132              0      
#>       Mean (SD)                49.6 (9.0)          NA          49.4 (8.4)          NA          49.3 (8.6)          NA     
#>       Median                      49.9             NA             48.9             NA             48.6             NA     
#>       Min - Max               25.9 - 68.8          NA         30.5 - 80.5          NA         26.5 - 67.5          NA     
#>     BASELINE                                                                                                              
#>       n                           134             134             134             134             132             132     
#>       Mean (SD)                51.9 (8.1)      0.0 (0.0)       50.3 (8.8)      0.0 (0.0)       50.3 (9.4)      0.0 (0.0)  
#>       Median                      51.2            0.0             51.2            0.0             50.7            0.0     
#>       Min - Max               28.5 - 72.9      0.0 - 0.0      26.9 - 67.3      0.0 - 0.0      27.8 - 71.2      0.0 - 0.0  
#>     WEEK 1 DAY 8                                                                                                          
#>       n                           134             134             134             134             132             132     
#>       Mean (SD)                50.1 (8.3)     -1.8 (11.6)      49.3 (7.5)     -1.0 (11.6)      49.8 (7.7)     -0.5 (12.6) 
#>       Median                      49.9            -2.7            48.8            -2.2            49.5            -1.5    
#>       Min - Max               26.7 - 69.4     -26.4 - 29.6    30.7 - 67.8     -29.5 - 32.1    30.1 - 72.9     -25.1 - 33.3
#>     WEEK 2 DAY 15                                                                                                         
#>       n                           134             134             134             134             132             132     
#>       Mean (SD)                49.7 (8.8)     -2.2 (12.4)      50.8 (7.6)      0.6 (11.8)      49.1 (7.7)     -1.2 (11.3) 
#>       Median                      49.2            -3.0            50.3            0.1             49.2            -1.1    
#>       Min - Max               26.0 - 69.6     -37.6 - 26.4    33.3 - 70.3     -28.9 - 37.7    32.2 - 73.6     -33.7 - 26.3
#>     WEEK 3 DAY 22                                                                                                         
#>       n                           134             134             134             134             132             132     
#>       Mean (SD)                50.5 (7.4)     -1.4 (11.1)      49.9 (7.8)     -0.4 (11.0)      49.8 (7.2)     -0.5 (12.1) 
#>       Median                      50.6            -1.1            49.4            -0.8            49.6            -2.3    
#>       Min - Max               27.7 - 73.0     -30.0 - 26.5    27.8 - 71.0     -26.5 - 33.4    27.9 - 67.5     -28.7 - 28.0
#>     WEEK 4 DAY 29                                                                                                         
#>       n                           134             134             134             134             132             132     
#>       Mean (SD)                49.0 (8.3)     -2.9 (11.1)      50.0 (8.4)     -0.3 (12.0)      51.0 (7.7)      0.7 (11.6) 
#>       Median                      48.8            -3.0            49.9            -0.4            51.4            0.7     
#>       Min - Max               32.3 - 78.4     -28.4 - 30.4    30.4 - 79.8     -29.1 - 28.3    31.4 - 70.3     -30.6 - 25.1
#>     WEEK 5 DAY 36                                                                                                         
#>       n                           134             134             134             134             132             132     
#>       Mean (SD)                48.4 (8.5)     -3.5 (11.6)      50.3 (7.2)      0.0 (11.2)      49.7 (7.6)     -0.5 (12.4) 
#>       Median                      47.7            -4.0            50.3            -1.5            49.1            -2.7    
#>       Min - Max               25.0 - 75.7     -41.3 - 31.5    29.4 - 67.7     -24.9 - 27.3    35.3 - 73.7     -24.6 - 34.0
#>   Respiratory Rate                                                                                                        
#>     SCREENING                                                                                                             
#>       n                           134              0              134              0              132              0      
#>       Mean (SD)                49.2 (8.2)          NA          50.3 (8.3)          NA          49.6 (8.4)          NA     
#>       Median                      49.3             NA             49.9             NA             50.1             NA     
#>       Min - Max               26.0 - 74.4          NA         31.4 - 69.2          NA         24.3 - 67.9          NA     
#>     BASELINE                                                                                                              
#>       n                           134             134             134             134             132             132     
#>       Mean (SD)                49.4 (8.1)      0.0 (0.0)       51.1 (8.2)      0.0 (0.0)       50.8 (8.4)      0.0 (0.0)  
#>       Median                      49.8            0.0             50.5            0.0             51.1            0.0     
#>       Min - Max               29.3 - 69.3      0.0 - 0.0      31.2 - 74.5      0.0 - 0.0      26.2 - 71.0      0.0 - 0.0  
#>     WEEK 1 DAY 8                                                                                                          
#>       n                           134             134             134             134             132             132     
#>       Mean (SD)                49.5 (8.0)      0.1 (11.6)      50.0 (7.9)     -1.1 (11.5)      51.0 (7.8)      0.2 (11.1) 
#>       Median                      49.3            0.2             49.1            -1.6            51.1            0.4     
#>       Min - Max               24.1 - 66.1     -31.6 - 32.1    25.0 - 67.7     -29.1 - 30.9    29.9 - 69.8     -23.6 - 27.7
#>     WEEK 2 DAY 15                                                                                                         
#>       n                           134             134             134             134             132             132     
#>       Mean (SD)                49.7 (7.7)      0.3 (9.7)       50.4 (8.7)     -0.8 (12.1)      49.3 (8.5)     -1.4 (12.0) 
#>       Median                      49.2            1.4             49.7            -1.6            50.2            -1.6    
#>       Min - Max               30.8 - 71.3     -24.9 - 27.6    28.5 - 77.0     -29.0 - 28.9    22.4 - 67.6     -43.8 - 27.0
#>     WEEK 3 DAY 22                                                                                                         
#>       n                           134             134             134             134             132             132     
#>       Mean (SD)                50.6 (8.2)      1.2 (11.7)      50.2 (6.9)     -1.0 (10.1)      49.1 (7.3)     -1.6 (11.1) 
#>       Median                      50.5            0.7             49.7            -0.7            49.6            -0.9    
#>       Min - Max               33.2 - 70.8     -33.5 - 33.3    32.3 - 67.8     -29.6 - 22.2    27.4 - 63.3     -25.4 - 33.6
#>     WEEK 4 DAY 29                                                                                                         
#>       n                           134             134             134             134             132             132     
#>       Mean (SD)                49.4 (7.7)      0.0 (10.9)      49.9 (7.8)     -1.2 (11.0)      50.8 (7.8)      0.1 (11.1) 
#>       Median                      48.9            -1.4            48.9            -2.4            50.4            -0.6    
#>       Min - Max               25.9 - 69.7     -25.9 - 40.3    33.7 - 68.5     -25.0 - 29.2    31.9 - 72.0     -27.6 - 31.1
#>     WEEK 5 DAY 36                                                                                                         
#>       n                           134             134             134             134             132             132     
#>       Mean (SD)                49.9 (7.7)      0.5 (11.5)      50.6 (8.2)     -0.5 (12.0)      49.9 (7.9)     -0.8 (10.6) 
#>       Median                      49.8            -0.2            50.3            -0.6            49.6            -1.7    
#>       Min - Max               34.3 - 70.7     -30.3 - 37.2    28.4 - 69.8     -29.4 - 23.8    29.8 - 68.4     -21.3 - 22.3
#>   Systolic Blood Pressure                                                                                                 
#>     SCREENING                                                                                                             
#>       n                           134              0              134              0              132              0      
#>       Mean (SD)                49.6 (8.3)          NA          50.3 (8.3)          NA          50.9 (7.8)          NA     
#>       Median                      49.6             NA             50.2             NA             50.8             NA     
#>       Min - Max               24.0 - 70.9          NA         26.2 - 79.1          NA         27.6 - 67.4          NA     
#>     BASELINE                                                                                                              
#>       n                           134             134             134             134             132             132     
#>       Mean (SD)                49.4 (8.5)      0.0 (0.0)       50.2 (8.5)      0.0 (0.0)       48.5 (7.2)      0.0 (0.0)  
#>       Median                      48.3            0.0             50.0            0.0             49.2            0.0     
#>       Min - Max               24.3 - 71.1      0.0 - 0.0      24.4 - 71.1      0.0 - 0.0      26.2 - 63.4      0.0 - 0.0  
#>     WEEK 1 DAY 8                                                                                                          
#>       n                           134             134             134             134             132             132     
#>       Mean (SD)                50.7 (9.2)      1.3 (12.8)      49.3 (8.7)     -1.0 (11.2)      49.6 (8.0)      1.2 (10.2) 
#>       Median                      49.5            0.1             48.3            -1.7            49.8            0.1     
#>       Min - Max               29.8 - 79.0     -29.5 - 37.6    33.0 - 74.0     -27.6 - 24.8    25.5 - 68.6     -22.8 - 31.3
#>     WEEK 2 DAY 15                                                                                                         
#>       n                           134             134             134             134             132             132     
#>       Mean (SD)                49.4 (8.5)     -0.0 (11.9)      49.4 (7.0)     -0.9 (11.0)      49.3 (7.7)      0.8 (9.7)  
#>       Median                      49.5            -0.2            49.4            -0.9            48.3            -0.2    
#>       Min - Max               24.2 - 65.4     -39.1 - 32.7    27.6 - 69.2     -37.1 - 33.9    29.4 - 74.7     -24.9 - 29.0
#>     WEEK 3 DAY 22                                                                                                         
#>       n                           134             134             134             134             132             132     
#>       Mean (SD)                49.7 (7.8)      0.3 (11.4)      49.0 (7.8)     -1.2 (12.1)      50.5 (8.3)      2.0 (11.1) 
#>       Median                      50.1            1.5             48.8            -1.2            51.7            2.0     
#>       Min - Max               31.3 - 70.3     -27.5 - 27.1    15.6 - 66.2     -35.6 - 36.4    29.6 - 69.4     -29.7 - 30.6
#>     WEEK 4 DAY 29                                                                                                         
#>       n                           134             134             134             134             132             132     
#>       Mean (SD)                49.4 (7.6)      0.0 (11.5)      49.4 (8.6)     -0.8 (12.6)      50.7 (7.4)      2.2 (10.5) 
#>       Median                      49.8            0.6             49.9            -1.3            50.8            3.0     
#>       Min - Max               22.9 - 69.6     -32.3 - 28.5    22.1 - 72.1     -29.7 - 38.1    29.7 - 65.1     -28.8 - 34.1
#>     WEEK 5 DAY 36                                                                                                         
#>       n                           134             134             134             134             132             132     
#>       Mean (SD)                49.4 (7.9)     -0.0 (11.9)      50.2 (8.7)     -0.0 (11.8)      50.4 (8.2)      1.9 (10.5) 
#>       Median                      49.6            0.7             49.9            0.2             50.8            2.4     
#>       Min - Max               26.1 - 69.7     -39.2 - 33.9    27.9 - 74.2     -27.7 - 36.2    34.3 - 74.5     -23.7 - 32.4
#>   Temperature                                                                                                             
#>     SCREENING                                                                                                             
#>       n                           134              0              134              0              132              0      
#>       Mean (SD)                48.9 (9.4)          NA          50.1 (7.9)          NA          50.2 (8.4)          NA     
#>       Median                      49.6             NA             50.1             NA             49.8             NA     
#>       Min - Max               29.0 - 75.2          NA         29.9 - 73.4          NA         30.5 - 72.3          NA     
#>     BASELINE                                                                                                              
#>       n                           134             134             134             134             132             132     
#>       Mean (SD)                49.7 (8.3)      0.0 (0.0)       50.9 (7.4)      0.0 (0.0)       49.1 (8.2)      0.0 (0.0)  
#>       Median                      50.4            0.0             51.5            0.0             48.7            0.0     
#>       Min - Max               29.0 - 69.1      0.0 - 0.0      30.7 - 67.7      0.0 - 0.0      20.6 - 70.5      0.0 - 0.0  
#>     WEEK 1 DAY 8                                                                                                          
#>       n                           134             134             134             134             132             132     
#>       Mean (SD)                51.6 (8.1)      1.9 (11.8)      49.5 (8.2)      -1.4 (9.9)      49.8 (7.6)      0.6 (11.2) 
#>       Median                      52.3            2.3             48.8            -1.1            50.2            0.9     
#>       Min - Max               31.5 - 71.2     -28.0 - 33.2    27.1 - 67.3     -28.1 - 26.6    32.1 - 68.5     -28.0 - 29.4
#>     WEEK 2 DAY 15                                                                                                         
#>       n                           134             134             134             134             132             132     
#>       Mean (SD)                50.2 (7.5)      0.5 (11.0)      49.5 (7.0)     -1.3 (10.2)      51.8 (8.1)      2.6 (11.8) 
#>       Median                      50.7            -0.7            49.7            -0.7            51.6            2.8     
#>       Min - Max               35.1 - 67.7     -33.9 - 25.4    32.3 - 73.2     -29.7 - 29.0    30.1 - 73.3     -24.9 - 32.8
#>     WEEK 3 DAY 22                                                                                                         
#>       n                           134             134             134             134             132             132     
#>       Mean (SD)                50.1 (8.2)      0.3 (11.1)      49.9 (8.0)     -1.0 (10.9)      49.7 (8.2)      0.6 (11.1) 
#>       Median                      49.5            1.1             49.6            -1.1            49.2            0.5     
#>       Min - Max               29.0 - 68.7     -29.3 - 24.1    32.7 - 70.3     -28.3 - 21.6    35.2 - 75.2     -27.3 - 32.0
#>     WEEK 4 DAY 29                                                                                                         
#>       n                           134             134             134             134             132             132     
#>       Mean (SD)                50.7 (7.7)      0.9 (11.8)      49.8 (8.5)     -1.1 (11.8)      49.9 (7.5)      0.8 (11.5) 
#>       Median                      50.7            1.4             49.8            -1.7            49.8            0.1     
#>       Min - Max               29.7 - 70.8     -25.9 - 29.2    24.4 - 68.9     -28.2 - 37.3    28.4 - 68.6     -25.8 - 32.5
#>     WEEK 5 DAY 36                                                                                                         
#>       n                           134             134             134             134             132             132     
#>       Mean (SD)                50.3 (8.3)      0.6 (12.3)      50.5 (8.1)     -0.4 (11.7)      50.1 (7.3)      0.9 (10.8) 
#>       Median                      49.8            -0.5            50.8            -0.3            49.7            0.9     
#>       Min - Max               33.4 - 75.0     -30.1 - 33.4    34.6 - 78.0     -23.1 - 30.9    31.1 - 73.4     -23.8 - 39.1
#>   Weight                                                                                                                  
#>     SCREENING                                                                                                             
#>       n                           134              0              134              0              132              0      
#>       Mean (SD)                49.9 (7.4)          NA          49.7 (8.3)          NA          50.3 (9.1)          NA     
#>       Median                      48.7             NA             49.3             NA             49.7             NA     
#>       Min - Max               33.7 - 65.9          NA         25.8 - 71.5          NA         26.0 - 70.0          NA     
#>     BASELINE                                                                                                              
#>       n                           134             134             134             134             132             132     
#>       Mean (SD)                50.7 (7.9)      0.0 (0.0)       49.9 (8.3)      0.0 (0.0)       49.4 (7.5)      0.0 (0.0)  
#>       Median                      49.8            0.0             50.7            0.0             49.6            0.0     
#>       Min - Max               33.3 - 71.9      0.0 - 0.0      27.9 - 69.8      0.0 - 0.0      29.4 - 65.1      0.0 - 0.0  
#>     WEEK 1 DAY 8                                                                                                          
#>       n                           134             134             134             134             132             132     
#>       Mean (SD)                49.2 (7.7)     -1.5 (11.4)      49.1 (7.6)     -0.8 (11.4)      49.6 (8.4)      0.2 (12.1) 
#>       Median                      50.5            -0.2            48.9            -2.7            49.9            -0.5    
#>       Min - Max               26.0 - 64.5     -42.9 - 21.3    31.5 - 66.1     -25.0 - 23.5    29.7 - 69.1     -31.0 - 32.6
#>     WEEK 2 DAY 15                                                                                                         
#>       n                           134             134             134             134             132             132     
#>       Mean (SD)                50.9 (8.2)      0.2 (11.2)      50.0 (7.8)      0.1 (12.5)      49.7 (7.9)      0.4 (10.8) 
#>       Median                      50.3            2.0             50.2            -0.2            49.8            1.3     
#>       Min - Max               32.1 - 70.1     -30.0 - 22.8    28.7 - 72.1     -29.2 - 34.6    28.1 - 73.7     -28.0 - 26.6
#>     WEEK 3 DAY 22                                                                                                         
#>       n                           134             134             134             134             132             132     
#>       Mean (SD)                49.0 (8.0)     -1.8 (11.4)      49.6 (8.3)     -0.3 (12.5)      50.4 (8.1)      1.1 (10.8) 
#>       Median                      48.9            -1.3            48.5            -1.0            50.8            -0.2    
#>       Min - Max               31.7 - 75.4     -29.7 - 24.9    32.4 - 76.0     -36.8 - 34.4    27.8 - 70.9     -28.9 - 31.7
#>     WEEK 4 DAY 29                                                                                                         
#>       n                           134             134             134             134             132             132     
#>       Mean (SD)                50.1 (8.1)     -0.7 (11.2)      49.5 (8.1)     -0.4 (10.4)      49.7 (7.5)      0.3 (10.4) 
#>       Median                      50.9            -0.9            49.4            -0.5            49.6            -1.2    
#>       Min - Max               30.1 - 68.9     -29.5 - 24.3    29.0 - 69.5     -24.9 - 26.3    30.8 - 77.4     -23.4 - 29.9
#>     WEEK 5 DAY 36                                                                                                         
#>       n                           134             134             134             134             132             132     
#>       Mean (SD)                49.0 (8.1)     -1.7 (11.5)      51.6 (8.5)      1.7 (11.5)      49.9 (7.4)      0.5 (10.9) 
#>       Median                      49.8            -0.8            51.8            1.2             49.6            -0.2    
#>       Min - Max               26.4 - 66.7     -34.8 - 29.7    26.3 - 76.6     -24.0 - 29.3    28.7 - 67.9     -36.1 - 31.0