VST02 Table 2 (Supplementary) Vital Sign Abnormalities Table 2.

Assessments Outside Normal Limits Among Subject Without Abnormality at Baseline.

Usage

vst02_2_main(
  adam_db,
  arm_var = "ACTARM",
  lbl_vs_assessment = "Assessment",
  lbl_vs_abnormality = "Abnormality",
  lbl_overall = NULL,
  deco = std_deco("VST02_2"),
  ...
)

vst02_2_lyt(
  arm_var,
  lbl_vs_assessment,
  lbl_vs_abnormality,
  lbl_overall,
  deco,
  ...
)

vst02_2_pre(adam_db, ...)

vst02_2_post(tlg, prune_0 = FALSE, ...)

vst02_2

Format

An object of class chevron_t of length 1.

Arguments

adam_db

(dm) object containing the ADaM datasets

arm_var

(character) variable used for column splitting

lbl_vs_assessment

(character) the label of the assessment variable.

lbl_vs_abnormality

(character) the label of the abnormality variable.

lbl_overall

(character) label used for overall column, if set to NULL the overall column is omitted

deco

(character) decoration with title, subtitles and main_footer content

...

not used.

tlg

(TableTree, Listing or ggplot) object typically produced by a main function.

prune_0

(logical) remove 0 count rows

Details

Only count LOW or HIGH values.

Results of "LOW LOW" are treated as the same as "LOW", and "HIGH HIGH" the same as "HIGH".

Does not include a total column by default.

Does not remove zero-count rows unless overridden with prune_0 = TRUE.

Functions

vst02_2_main(): Main TLG function

vst02_2_lyt(): Layout

vst02_2_pre(): Preprocessing

vst02_2_post(): Postprocessing

Note

adam_db object must contain an advs table with the "PARAM", "ANRIND" and "BNRIND" columns.

Examples

run(vst02_2, syn_data)
#>   Assessment                   A: Drug X        B: Placebo     C: Combination
#>     Abnormality                 (N=134)          (N=134)          (N=132)    
#>   ———————————————————————————————————————————————————————————————————————————
#>   Diastolic Blood Pressure                                                   
#>     Low                      55/120 (45.8%)   60/121 (49.6%)   53/118 (44.9%)
#>     High                     55/113 (48.7%)   56/123 (45.5%)   52/119 (43.7%)
#>   Pulse Rate                                                                 
#>     Low                      66/125 (52.8%)   54/121 (44.6%)   55/118 (46.6%)
#>     High                     52/119 (43.7%)   59/119 (49.6%)   38/117 (32.5%)
#>   Respiratory Rate                                                           
#>     Low                       59/123 (48%)    56/118 (47.5%)   50/118 (42.4%)
#>     High                     65/124 (52.4%)   58/127 (45.7%)   57/118 (48.3%)
#>   Systolic Blood Pressure                                                    
#>     Low                      53/116 (45.7%)   55/120 (45.8%)   58/120 (48.3%)
#>     High                     65/121 (53.7%)   53/125 (42.4%)   56/124 (45.2%)
#>   Temperature                                                                
#>     Low                      51/119 (42.9%)   53/119 (44.5%)   64/115 (55.7%)
#>     High                     61/128 (47.7%)   53/121 (43.8%)   59/122 (48.4%)
#>   Weight                                                                     
#>     Low                      59/120 (49.2%)   54/117 (46.2%)   63/119 (52.9%)
#>     High                     58/123 (47.2%)   61/121 (50.4%)   55/113 (48.7%)