AET01 Table 1 (Default) Overview of Deaths and Adverse Events Summary Table 1.

Overview of death and summary of adverse events.

aet01_1(
  adam_db,
  armvar = .study$actualarm,
  lbl_overall = .study$lbl_overall,
  prune_0 = FALSE,
  deco = std_deco("AET01"),
  safety_var = .study$safety_var,
  lbl_safety_var = var_labels_for(adam_db$adae, safety_var),
  .study = list(actualarm = "ACTARM", lbl_overall = NULL, safety_var = c("FATAL",
    "SER", "SERWD", "SERDSM", "RELSER", "WD", "DSM", "REL", "RELWD", "RELDSM", "SEV"))
)

aet01_1_lyt(
  armvar = .study$actualarm,
  lbl_overall = .study$lbl_overall,
  deco = std_deco("AET01"),
  safety_var = .study$safety_var,
  lbl_safety_var = .study$lbl_safety_var,
  .study = list(actualarm = "ACTARM", lbl_overall = NULL, safety_var = c("FATAL",
    "SER", "SERWD", "SERDSM", "RELSER", "WD", "DSM", "REL", "RELWD", "RELDSM", "CTC35",
    "CTC45", "SEV"), lbl_safety_var = c("FATAL", "SER", "SERWD", "SERDSM", "RELSER",
    "WD", "DSM", "REL", "RELWD", "RELDSM", "CTC35", "CTC45", "SEV"))
)

Arguments

adam_db

(dm) object containing the ADaM datasets

armvar

(character) variable used for column splitting

lbl_overall

(character) label used for overall column, if set to NULL the overall column is omitted

prune_0

(logical) remove 0 count rows

deco

(character) decoration with title, subtitles and main_footer content

safety_var

(character) the safety variables to be summarized.

lbl_safety_var

(character) the labels of the safety variables to be summarized.

.study

(list) with default values for the arguments of the function

Details

Does not remove rows with zero counts by default.

Functions

aet01_1_lyt:

Examples

library(dm)
#> 
#> Attaching package: ‘dm’
#> The following object is masked from ‘package:stats’:
#> 
#>     filter

db <- syn_test_data() %>%
  preprocess_data("aet01_1")

aet01_1(db, armvar = "ARM")
#> Overview of Deaths and Adverse Events: {Specify Population}
#> Protocol: {{protocol}}, Snapshot: {{snapshot}}, Snapshot Date: {{snapshot-date}}, Cutoff Date: {{cutoff-date}}
#> 
#> ———————————————————————————————————————————————————————————————————————————————————————————————————————
#>                                                               A: Drug X    B: Placebo    C: Combination
#>                                                                (N=134)       (N=134)        (N=132)    
#> ———————————————————————————————————————————————————————————————————————————————————————————————————————
#> Total number of patients with at least one AE                 122 (91%)    123 (91.8%)    120 (90.9%)  
#> Total number of AEs                                              609           622            703      
#> Total number of deaths                                       22 (16.4%)    26 (19.4%)      19 (14.4%)  
#> Total number of patients withdrawn from study due to an AE    6 (4.5%)      1 (0.7%)        2 (1.5%)   
#> Total number of patients with at least one                                                             
#>   AE with fatal outcome                                      76 (56.7%)    70 (52.2%)      75 (56.8%)  
#>   Serious AE                                                 104 (77.6%)   101 (75.4%)      99 (75%)   
#>   Serious AE leading to withdrawal from treatment              8 (6%)        8 (6%)         7 (5.3%)   
#>   Serious AE leading to dose modification/interruption       36 (26.9%)    31 (23.1%)      26 (19.7%)  
#>   Related Serious AE                                         76 (56.7%)    70 (52.2%)      75 (56.8%)  
#>   AE leading to withdrawal from treatment                    23 (17.2%)    24 (17.9%)      27 (20.5%)  
#>   AE leading to dose modification/interruption               72 (53.7%)    65 (48.5%)      76 (57.6%)  
#>   Related AE                                                 105 (78.4%)   108 (80.6%)    109 (82.6%)  
#>   Related AE leading to withdrawal from treatment             6 (4.5%)      9 (6.7%)       13 (9.8%)   
#>   Related AE leading to dose modification/interruption       31 (23.1%)    31 (23.1%)      32 (24.2%)  
#>   Severe AE (at greatest intensity)                          91 (67.9%)    90 (67.2%)      93 (70.5%)  
#> ———————————————————————————————————————————————————————————————————————————————————————————————————————
#> 
#> Investigator text for AEs encoded using MedDRA version xx.x.  Percentages are based on N in the column headings.
#> Multiple occurrences of the same AE in one individual are counted only once except for "Total number of AEs" row in which multiple occurrences of the same AE are counted separately. Standard footnote defining AEs included in table (e.g., includes AEs with onset from first dose of study drug through 7 days after last dose of study drug).
aet01_1_lyt(
  armvar = "ACTARM",
  lbl_overall = NULL,
  deco = std_deco("AET01")
)
#> $lyt_adae
#> A Pre-data Table Layout
#> 
#> Column-Split Structure:
#> ACTARM (lvls) 
#> 
#> Row-Split Structure:
#> USUBJID (** analysis **) 
#> 
#> 
#> $lyt_adsl
#> A Pre-data Table Layout
#> 
#> Column-Split Structure:
#> ACTARM (lvls) 
#> 
#> Row-Split Structure:
#> USUBJID:USUBJID (** multivar analysis **) 
#> 
#>