AET01 Table 1 (Default) Overview of Deaths and Adverse Events Summary Table 1.

AET01 Table 1 (Default) Overview of Deaths and Adverse Events Summary Table 1.

Usage

aet01_1_main(
  adam_db,
  armvar = "ACTARM",
  lbl_overall = NULL,
  deco = std_deco("AET01"),
  safety_var = c("FATAL", "SER", "SERWD", "SERDSM", "RELSER", "WD", "DSM", "REL",
    "RELWD", "RELDSM", "SEV"),
  lbl_safety_var = NULL,
  ...
)

aet01_1_lyt(armvar, lbl_overall, deco, safety_var, lbl_safety_var, ...)

aet01_1_pre(adam_db, ...)

aet01_1_check(
  adam_db,
  req_tables = c("adsl", "adae"),
  armvar = "ACTARM",
  safety_var = c("FATAL", "SER", "SERWD", "SERDSM", "RELSER", "WD", "DSM", "REL",
    "RELWD", "RELDSM", "SEV"),
  ...
)

aet01_post(tlg, prune_0 = FALSE, deco = std_deco("AET01"), ...)

aet01_1

Format

An object of class chevron_t of length 1.

Arguments

adam_db

(dm) object containing the ADaM datasets

armvar

(character) variable used for column splitting

lbl_overall

(character) label used for overall column, if set to NULL the overall column is omitted

deco

(character) decoration with title, subtitles and main_footer content

safety_var

(character) the safety variables to be summarized.

lbl_safety_var

(character) the labels of the safety variables to be summarized.

...

not used

req_tables

(character) names of the required tables.

tlg

(TableTree, Listing or ggplot) object typically produced by a main function.

prune_0

(logical) remove 0 count rows

Details

Does not remove rows with zero counts by default.

Functions

aet01_1_main(): Main TLG function

aet01_1_lyt(): Layout

aet01_1_pre(): Preprocessing

aet01_1_check(): Checks

aet01_post(): Postprocessing

Note

adam_db object must contain an adsl table with the "DTHFL" and "DCSREAS" columns.

adam_db object must contain an adae table with the columns passed to safety_var.

Examples

run(aet01_1, syn_data, armvar = "ARM")
#>                                                                 A: Drug X    B: Placebo    C: Combination
#>                                                                  (N=134)       (N=134)        (N=132)    
#>   ———————————————————————————————————————————————————————————————————————————————————————————————————————
#>   Total number of patients with at least one AE                122 (91.0%)   123 (91.8%)    120 (90.9%)  
#>   Total number of AEs                                              609           622            703      
#>   Total number of deaths                                       25 (18.7%)    23 (17.2%)      22 (16.7%)  
#>   Total number of patients withdrawn from study due to an AE    3 (2.2%)      6 (4.5%)        5 (3.8%)   
#>   Total number of patients with at least one                                                             
#>     AE with fatal outcome                                      76 (56.7%)    70 (52.2%)      75 (56.8%)  
#>     Serious AE                                                 104 (77.6%)   101 (75.4%)     99 (75.0%)  
#>     Serious AE leading to withdrawal from treatment             9 (6.7%)      6 (4.5%)       11 (8.3%)   
#>     Serious AE leading to dose modification/interruption       22 (16.4%)    26 (19.4%)      29 (22.0%)  
#>     Related Serious AE                                         76 (56.7%)    70 (52.2%)      75 (56.8%)  
#>     AE leading to withdrawal from treatment                    27 (20.1%)    26 (19.4%)      30 (22.7%)  
#>     AE leading to dose modification/interruption               66 (49.3%)    76 (56.7%)      74 (56.1%)  
#>     Related AE                                                 105 (78.4%)   108 (80.6%)    109 (82.6%)  
#>     Related AE leading to withdrawal from treatment             6 (4.5%)      12 (9.0%)       8 (6.1%)   
#>     Related AE leading to dose modification/interruption       29 (21.6%)    38 (28.4%)      38 (28.8%)  
#>     Severe AE (at greatest intensity)                          91 (67.9%)    90 (67.2%)      93 (70.5%)