AET02 Table 2 (Supplementary) Adverse Events by System Organ Class, High Level Term and Preferred Term Table 2

The AET02_2 table provides an overview of the number of patients experiencing advert events and the number of advert events categorized by Body System, High Level Term and Dictionary-Derived Term.

aet02_2(
  adam_db,
  armvar = .study$armvar,
  lbl_overall = .study$lbl_overall,
  prune_0 = TRUE,
  deco = std_deco("AET02"),
  .study = list(armvar = "ACTARM", lbl_overall = NULL)
)

aet02_2_lyt(
  armvar = .study$armvar,
  lbl_overall = .study$lbl_overall,
  lbl_AEBODSYS = "AEBODSYS",
  lbl_AEHLT = "AEHLT",
  lbl_AEDECOD = "AEDECOD",
  deco = std_deco("AET02"),
  .study = list(armvar = "ACTARM", lbl_overall = NULL)
)

Arguments

adam_db
(dm) object containing the ADaM datasets

armvar
(character) variable used for column splitting

lbl_overall
(character) label used for overall column, if set to NULL the overall column is omitted

prune_0
(logical) remove 0 count rows

deco
(character) decoration with title, subtitles and main_footer content

.study
(list) with default values for the arguments of the function

lbl_AEBODSYS
(string) text label for AEBODSYS.

lbl_AEHLT
(string) text label for AEHLT.

lbl_AEDECOD
(string) text label for AEDECOD.

Details

Numbers represent absolute numbers of patients and fraction of N, or absolute number of event when specified.

Remove zero-count rows unless overridden with prune_0 = FALSE.

Split columns by arm.

Does not include a total column by default.

Sort Body System or Organ Class, High Level Term and Dictionary-Derived Term hierarchically by highest overall frequencies.

Functions

aet02_2_lyt:

Examples

library(dm)
library(rtables)

db <- syn_test_data() %>%
  preprocess_data("aet02_2")

aet02_2(db) %>% head(15)
#> Adverse Events: {Specify Population}
#> Protocol: {{protocol}}, Snapshot: {{snapshot}}, Snapshot Date: {{snapshot-date}}, Cutoff Date: {{cutoff-date}}
#> 
#> ———————————————————————————————————————————————————————————————————————————————————————————————————————
#>                                                                A: Drug X    B: Placebo   C: Combination
#>                                                                  (N=97)      (N=103)        (N=102)    
#> ———————————————————————————————————————————————————————————————————————————————————————————————————————
#> Total number of patients with at least one adverse event       97 (100%)    103 (100%)     102 (100%)  
#> Overall total number of events                                    469          513            599      
#> cl A.1                                                                                                 
#>   Total number of patients with at least one adverse event     62 (63.9%)   64 (62.1%)     77 (75.5%)  
#>   Total number of events                                          107          109            140      
#>   hlt A.1.1.1                                                                                          
#>     Total number of patients with at least one adverse event   62 (63.9%)   64 (62.1%)     77 (75.5%)  
#>     Total number of events                                        107          109            140      
#>     dcd A.1.1.1.1                                              42 (43.3%)    36 (35%)      54 (52.9%)  
#>     dcd A.1.1.1.2                                              37 (38.1%)   44 (42.7%)     43 (42.2%)  
#> cl B.2                                                                                                 
#>   Total number of patients with at least one adverse event     59 (60.8%)   59 (57.3%)     73 (71.6%)  
#>   Total number of events                                           95          110            125      
#>   hlt B.2.2.3                                                                                          
#>     Total number of patients with at least one adverse event   38 (39.2%)   42 (40.8%)     42 (41.2%)  
#> ———————————————————————————————————————————————————————————————————————————————————————————————————————
#> 
#> Investigator text for AEs encoded using MedDRA version xx.x.  Percentages are based on N in the column headings. For frequency counts by preferred term, multiple occurrences of the same AE in an individual are counted only once.  For frequency counts of "Total number of events" rows, multiple occurrences of the same AE in an individual are counted separately. Standard footnote defining AEs included in table (e.g., includes AEs with onset from first dose of study drug through 7 days after last dose of study drug).

# Additional Examples
aet02_2(db, lbl_overall = "All Patients") %>% head()
#> Adverse Events: {Specify Population}
#> Protocol: {{protocol}}, Snapshot: {{snapshot}}, Snapshot Date: {{snapshot-date}}, Cutoff Date: {{cutoff-date}}
#> 
#> ————————————————————————————————————————————————————————————————————————————————————————————————————————————————————
#>                                                              A: Drug X    B: Placebo   C: Combination   All Patients
#>                                                                (N=97)      (N=103)        (N=102)         (N=302)   
#> ————————————————————————————————————————————————————————————————————————————————————————————————————————————————————
#> Total number of patients with at least one adverse event     97 (100%)    103 (100%)     102 (100%)      302 (100%) 
#> Overall total number of events                                  469          513            599             1581    
#> cl A.1                                                                                                              
#>   Total number of patients with at least one adverse event   62 (63.9%)   64 (62.1%)     77 (75.5%)     203 (67.2%) 
#>   Total number of events                                        107          109            140             356     
#>   hlt A.1.1.1                                                                                                       
#> ————————————————————————————————————————————————————————————————————————————————————————————————————————————————————
#> 
#> Investigator text for AEs encoded using MedDRA version xx.x.  Percentages are based on N in the column headings. For frequency counts by preferred term, multiple occurrences of the same AE in an individual are counted only once.  For frequency counts of "Total number of events" rows, multiple occurrences of the same AE in an individual are counted separately. Standard footnote defining AEs included in table (e.g., includes AEs with onset from first dose of study drug through 7 days after last dose of study drug).

db_m <- db %>%
  dm_zoom_to(adae) %>%
  mutate(AEBODSYS = with_label(AEBODSYS, "MedDRA System Organ Class")) %>%
  dm_update_zoomed()

aet02_2(db_m) %>% head()
#> Adverse Events: {Specify Population}
#> Protocol: {{protocol}}, Snapshot: {{snapshot}}, Snapshot Date: {{snapshot-date}}, Cutoff Date: {{cutoff-date}}
#> 
#> —————————————————————————————————————————————————————————————————————————————————————————————————————
#>                                                              A: Drug X    B: Placebo   C: Combination
#>                                                                (N=97)      (N=103)        (N=102)    
#> —————————————————————————————————————————————————————————————————————————————————————————————————————
#> Total number of patients with at least one adverse event     97 (100%)    103 (100%)     102 (100%)  
#> Overall total number of events                                  469          513            599      
#> cl A.1                                                                                               
#>   Total number of patients with at least one adverse event   62 (63.9%)   64 (62.1%)     77 (75.5%)  
#>   Total number of events                                        107          109            140      
#>   hlt A.1.1.1                                                                                        
#> —————————————————————————————————————————————————————————————————————————————————————————————————————
#> 
#> Investigator text for AEs encoded using MedDRA version xx.x.  Percentages are based on N in the column headings. For frequency counts by preferred term, multiple occurrences of the same AE in an individual are counted only once.  For frequency counts of "Total number of events" rows, multiple occurrences of the same AE in an individual are counted separately. Standard footnote defining AEs included in table (e.g., includes AEs with onset from first dose of study drug through 7 days after last dose of study drug).
aet02_2_lyt(
  armvar = "ACTARM",
  lbl_overall = NULL,
  lbl_AEBODSYS = "Body System or Organ Class",
  lbl_AEHLT = "High Level Term",
  lbl_AEDECOD = "Dictionary-Derived Term",
  deco = std_deco("AET02")
)
#> A Pre-data Table Layout
#> 
#> Column-Split Structure:
#> ACTARM (lvls) 
#> 
#> Row-Split Structure:
#> AEBODSYS (lvls) -> AEHLT (lvls) -> AEDECOD (** analysis **) 
#>