AET02 Table 3 (Supplementary) Adverse Events by Dictionary-Derived Term Table 3.

The AET02_3 table provides an overview of the number of patients experiencing adverse events and the number of adverse events categorized by Dictionary-Derived Term.

Usage

aet02_3_lyt(
  armvar,
  lbl_overall,
  lbl_aedecod = "MedDRA Preferred Term",
  deco,
  ...
)

aet02_3_pre(adam_db, ...)

aet02_3_post(tlg, prune_0 = TRUE, ...)

aet02_3

Format

An object of class chevron_t of length 1.

Arguments

armvar

(character) variable used for column splitting

lbl_overall

(character) label used for overall column, if set to NULL the overall column is omitted

lbl_aedecod

(character) text label for AEDECOD.

deco

(character) decoration with title, subtitles and main_footer content

...

not used.

adam_db

(dm) object containing the ADaM datasets

tlg

(TableTree, Listing or ggplot) object typically produced by a main function.

prune_0

(logical) remove 0 count rows

Functions

aet02_3_lyt(): Layout

aet02_3_pre(): Preprocessing

aet02_3_post(): Postprocessing

Examples

run(aet02_3, syn_data)
#>   MedDRA Preferred Term                                       A: Drug X    B: Placebo    C: Combination
#>                                                                (N=134)       (N=134)        (N=132)    
#>   —————————————————————————————————————————————————————————————————————————————————————————————————————
#>   Total number of patients with at least one adverse event   122 (91.0%)   123 (91.8%)    120 (90.9%)  
#>   Total number of events                                         609           622            703      
#>   dcd D.2.1.5.3                                              47 (35.1%)    58 (43.3%)      57 (43.2%)  
#>   dcd A.1.1.1.1                                              50 (37.3%)    45 (33.6%)      63 (47.7%)  
#>   dcd B.2.2.3.1                                              48 (35.8%)    54 (40.3%)      51 (38.6%)  
#>   dcd A.1.1.1.2                                              48 (35.8%)    48 (35.8%)      50 (37.9%)  
#>   dcd B.2.1.2.1                                              49 (36.6%)    44 (32.8%)      52 (39.4%)  
#>   dcd D.1.1.1.1                                              50 (37.3%)    42 (31.3%)      51 (38.6%)  
#>   dcd D.1.1.4.2                                              48 (35.8%)    42 (31.3%)      50 (37.9%)  
#>   dcd B.1.1.1.1                                              47 (35.1%)    49 (36.6%)      43 (32.6%)  
#>   dcd C.2.1.2.1                                              35 (26.1%)    48 (35.8%)      55 (41.7%)  
#>   dcd C.1.1.1.3                                              43 (32.1%)    46 (34.3%)      43 (32.6%)