AET03 Table 1 (Default) Advert Events by Greatest Intensity Table 1

An adverse events table categorized by System Organ Class, Dictionary-Derived Term and Greatest intensity.

aet03_1(
  adam_db,
  armvar = .study$armvar,
  prune_0 = TRUE,
  lbl_overall = .study$lbl_overall,
  deco = std_deco("AET03"),
  .study = list(armvar = "ACTARM", lbl_overall = NULL)
)

aet03_1_lyt(
  armvar = .study$armvar,
  lbl_AEBODSYS = "",
  lbl_AEDECOD = "",
  severity_grade = .study$severity_grade,
  lbl_overall = .study$lbl_overall,
  deco = std_deco("AET03"),
  .study = list(armvar = "ACTARM", lbl_overall = NULL, severity_grade = c("MILD",
    "MODERATE", "SEVERE", "LIFE THREATENING"))
)

Arguments

adam_db
(dm) object containing the ADaM datasets

armvar
(character) variable used for column splitting

prune_0
(logical) remove 0 count rows

lbl_overall
(character) label used for overall column, if set to NULL the overall column is omitted

deco
(character) decoration with title, subtitles and main_footer content

.study
(list) with default values for the arguments of the function

lbl_AEBODSYS
(string) text label for AEBODSYS.

lbl_AEDECOD
(string) text label for AEDECOD.

severity_grade
(vector of strings) describing the severity levels present in the data set.

Details

Default Adverse Events by Greatest Intensity table

Numbers represent absolute numbers of patients and fraction of N

Remove zero-count rows unless overridden with prune_0 = FALSE.

Split columns by arm

Does not include a total column by default

Sort by Body System or Organ Class (SOC) and Dictionary-Derived Term (PT)

Functions

aet03_1_lyt:

Examples

library(dm)
library(rtables)

db <- syn_test_data() %>%
  preprocess_data("aet03_1")

aet03_1(db)
#> Adverse Events by Greatest Intensity: {Specify Population}
#> Protocol: {{protocol}}, Snapshot: {{snapshot}}, Snapshot Date: {{snapshot-date}}, Cutoff Date: {{cutoff-date}}
#> 
#> ————————————————————————————————————————————————————————————————————
#> MedDRA System Organ Class   A: Drug X    B: Placebo   C: Combination
#>   MedDRA Preferred Term      (N=122)      (N=123)        (N=120)    
#> ————————————————————————————————————————————————————————————————————
#> - Any Intensity -           122 (100%)   123 (100%)     120 (100%)  
#> MILD                         7 (5.7%)     9 (7.3%)       4 (3.3%)   
#> MODERATE                    24 (19.7%)   24 (19.5%)     23 (19.2%)  
#> SEVERE                      91 (74.6%)   90 (73.2%)     93 (77.5%)  
#> cl A.1                                                              
#>   - Any Intensity -         78 (63.9%)    75 (61%)      89 (74.2%)  
#>   MILD                      30 (24.6%)    27 (22%)      39 (32.5%)  
#>   MODERATE                  48 (39.3%)    48 (39%)      50 (41.7%)  
#>   dcd A.1.1.1.1                                                     
#>     - Any Intensity -        50 (41%)    45 (36.6%)     63 (52.5%)  
#>     MILD                     50 (41%)    45 (36.6%)     63 (52.5%)  
#>   dcd A.1.1.1.2                                                     
#>     - Any Intensity -       48 (39.3%)    48 (39%)      50 (41.7%)  
#>     MODERATE                48 (39.3%)    48 (39%)      50 (41.7%)  
#> cl B.2                                                              
#>   - Any Intensity -         79 (64.8%)   74 (60.2%)     85 (70.8%)  
#>   MILD                      30 (24.6%)   30 (24.4%)     33 (27.5%)  
#>   MODERATE                  49 (40.2%)   44 (35.8%)     52 (43.3%)  
#>   dcd B.2.2.3.1                                                     
#>     - Any Intensity -       48 (39.3%)   54 (43.9%)     51 (42.5%)  
#>     MILD                    48 (39.3%)   54 (43.9%)     51 (42.5%)  
#>   dcd B.2.1.2.1                                                     
#>     - Any Intensity -       49 (40.2%)   44 (35.8%)     52 (43.3%)  
#>     MODERATE                49 (40.2%)   44 (35.8%)     52 (43.3%)  
#> cl D.1                                                              
#>   - Any Intensity -         79 (64.8%)   67 (54.5%)     80 (66.7%)  
#>   MODERATE                  29 (23.8%)   25 (20.3%)     29 (24.2%)  
#>   SEVERE                     50 (41%)    42 (34.1%)     51 (42.5%)  
#>   dcd D.1.1.1.1                                                     
#>     - Any Intensity -        50 (41%)    42 (34.1%)     51 (42.5%)  
#>     SEVERE                   50 (41%)    42 (34.1%)     51 (42.5%)  
#>   dcd D.1.1.4.2                                                     
#>     - Any Intensity -       48 (39.3%)   42 (34.1%)     50 (41.7%)  
#>     MODERATE                48 (39.3%)   42 (34.1%)     50 (41.7%)  
#> cl D.2                                                              
#>   - Any Intensity -         47 (38.5%)   58 (47.2%)     57 (47.5%)  
#>   MILD                      47 (38.5%)   58 (47.2%)     57 (47.5%)  
#>   dcd D.2.1.5.3                                                     
#>     - Any Intensity -       47 (38.5%)   58 (47.2%)     57 (47.5%)  
#>     MILD                    47 (38.5%)   58 (47.2%)     57 (47.5%)  
#> cl B.1                                                              
#>   - Any Intensity -         47 (38.5%)   49 (39.8%)     43 (35.8%)  
#>   SEVERE                    47 (38.5%)   49 (39.8%)     43 (35.8%)  
#>   dcd B.1.1.1.1                                                     
#>     - Any Intensity -       47 (38.5%)   49 (39.8%)     43 (35.8%)  
#>     SEVERE                  47 (38.5%)   49 (39.8%)     43 (35.8%)  
#> cl C.2                                                              
#>   - Any Intensity -         35 (28.7%)    48 (39%)      55 (45.8%)  
#>   MODERATE                  35 (28.7%)    48 (39%)      55 (45.8%)  
#>   dcd C.2.1.2.1                                                     
#>     - Any Intensity -       35 (28.7%)    48 (39%)      55 (45.8%)  
#>     MODERATE                35 (28.7%)    48 (39%)      55 (45.8%)  
#> cl C.1                                                              
#>   - Any Intensity -         43 (35.2%)   46 (37.4%)     43 (35.8%)  
#>   SEVERE                    43 (35.2%)   46 (37.4%)     43 (35.8%)  
#>   dcd C.1.1.1.3                                                     
#>     - Any Intensity -       43 (35.2%)   46 (37.4%)     43 (35.8%)  
#>     SEVERE                  43 (35.2%)   46 (37.4%)     43 (35.8%)  
#> ————————————————————————————————————————————————————————————————————
#> 
#> Investigator text for AEs encoded using MedDRA version xx.x.  All counts represent patients. Multiple occurrences of the same AE in one individual are counted once at the greatest intensity for this preferred term. To the SOC Overall row counts, a patient contributes only with the AE occuring with the greatest intensity within the SOC. Standard footnote defining AEs included in table (e.g., includes AEs with onset from first dose of study drug through 7 days after last dose of study drug).
aet03_1(db, lbl_overall = "All Patients")
#> Adverse Events by Greatest Intensity: {Specify Population}
#> Protocol: {{protocol}}, Snapshot: {{snapshot}}, Snapshot Date: {{snapshot-date}}, Cutoff Date: {{cutoff-date}}
#> 
#> ———————————————————————————————————————————————————————————————————————————————————
#> MedDRA System Organ Class   A: Drug X    B: Placebo   C: Combination   All Patients
#>   MedDRA Preferred Term      (N=122)      (N=123)        (N=120)         (N=365)   
#> ———————————————————————————————————————————————————————————————————————————————————
#> - Any Intensity -           122 (100%)   123 (100%)     120 (100%)      365 (100%) 
#> MILD                         7 (5.7%)     9 (7.3%)       4 (3.3%)       20 (5.5%)  
#> MODERATE                    24 (19.7%)   24 (19.5%)     23 (19.2%)      71 (19.5%) 
#> SEVERE                      91 (74.6%)   90 (73.2%)     93 (77.5%)     274 (75.1%) 
#> cl A.1                                                                             
#>   - Any Intensity -         78 (63.9%)    75 (61%)      89 (74.2%)     242 (66.3%) 
#>   MILD                      30 (24.6%)    27 (22%)      39 (32.5%)      96 (26.3%) 
#>   MODERATE                  48 (39.3%)    48 (39%)      50 (41.7%)      146 (40%)  
#>   dcd A.1.1.1.1                                                                    
#>     - Any Intensity -        50 (41%)    45 (36.6%)     63 (52.5%)     158 (43.3%) 
#>     MILD                     50 (41%)    45 (36.6%)     63 (52.5%)     158 (43.3%) 
#>   dcd A.1.1.1.2                                                                    
#>     - Any Intensity -       48 (39.3%)    48 (39%)      50 (41.7%)      146 (40%)  
#>     MODERATE                48 (39.3%)    48 (39%)      50 (41.7%)      146 (40%)  
#> cl B.2                                                                             
#>   - Any Intensity -         79 (64.8%)   74 (60.2%)     85 (70.8%)     238 (65.2%) 
#>   MILD                      30 (24.6%)   30 (24.4%)     33 (27.5%)      93 (25.5%) 
#>   MODERATE                  49 (40.2%)   44 (35.8%)     52 (43.3%)     145 (39.7%) 
#>   dcd B.2.2.3.1                                                                    
#>     - Any Intensity -       48 (39.3%)   54 (43.9%)     51 (42.5%)     153 (41.9%) 
#>     MILD                    48 (39.3%)   54 (43.9%)     51 (42.5%)     153 (41.9%) 
#>   dcd B.2.1.2.1                                                                    
#>     - Any Intensity -       49 (40.2%)   44 (35.8%)     52 (43.3%)     145 (39.7%) 
#>     MODERATE                49 (40.2%)   44 (35.8%)     52 (43.3%)     145 (39.7%) 
#> cl D.1                                                                             
#>   - Any Intensity -         79 (64.8%)   67 (54.5%)     80 (66.7%)     226 (61.9%) 
#>   MODERATE                  29 (23.8%)   25 (20.3%)     29 (24.2%)      83 (22.7%) 
#>   SEVERE                     50 (41%)    42 (34.1%)     51 (42.5%)     143 (39.2%) 
#>   dcd D.1.1.1.1                                                                    
#>     - Any Intensity -        50 (41%)    42 (34.1%)     51 (42.5%)     143 (39.2%) 
#>     SEVERE                   50 (41%)    42 (34.1%)     51 (42.5%)     143 (39.2%) 
#>   dcd D.1.1.4.2                                                                    
#>     - Any Intensity -       48 (39.3%)   42 (34.1%)     50 (41.7%)     140 (38.4%) 
#>     MODERATE                48 (39.3%)   42 (34.1%)     50 (41.7%)     140 (38.4%) 
#> cl D.2                                                                             
#>   - Any Intensity -         47 (38.5%)   58 (47.2%)     57 (47.5%)     162 (44.4%) 
#>   MILD                      47 (38.5%)   58 (47.2%)     57 (47.5%)     162 (44.4%) 
#>   dcd D.2.1.5.3                                                                    
#>     - Any Intensity -       47 (38.5%)   58 (47.2%)     57 (47.5%)     162 (44.4%) 
#>     MILD                    47 (38.5%)   58 (47.2%)     57 (47.5%)     162 (44.4%) 
#> cl B.1                                                                             
#>   - Any Intensity -         47 (38.5%)   49 (39.8%)     43 (35.8%)     139 (38.1%) 
#>   SEVERE                    47 (38.5%)   49 (39.8%)     43 (35.8%)     139 (38.1%) 
#>   dcd B.1.1.1.1                                                                    
#>     - Any Intensity -       47 (38.5%)   49 (39.8%)     43 (35.8%)     139 (38.1%) 
#>     SEVERE                  47 (38.5%)   49 (39.8%)     43 (35.8%)     139 (38.1%) 
#> cl C.2                                                                             
#>   - Any Intensity -         35 (28.7%)    48 (39%)      55 (45.8%)     138 (37.8%) 
#>   MODERATE                  35 (28.7%)    48 (39%)      55 (45.8%)     138 (37.8%) 
#>   dcd C.2.1.2.1                                                                    
#>     - Any Intensity -       35 (28.7%)    48 (39%)      55 (45.8%)     138 (37.8%) 
#>     MODERATE                35 (28.7%)    48 (39%)      55 (45.8%)     138 (37.8%) 
#> cl C.1                                                                             
#>   - Any Intensity -         43 (35.2%)   46 (37.4%)     43 (35.8%)     132 (36.2%) 
#>   SEVERE                    43 (35.2%)   46 (37.4%)     43 (35.8%)     132 (36.2%) 
#>   dcd C.1.1.1.3                                                                    
#>     - Any Intensity -       43 (35.2%)   46 (37.4%)     43 (35.8%)     132 (36.2%) 
#>     SEVERE                  43 (35.2%)   46 (37.4%)     43 (35.8%)     132 (36.2%) 
#> ———————————————————————————————————————————————————————————————————————————————————
#> 
#> Investigator text for AEs encoded using MedDRA version xx.x.  All counts represent patients. Multiple occurrences of the same AE in one individual are counted once at the greatest intensity for this preferred term. To the SOC Overall row counts, a patient contributes only with the AE occuring with the greatest intensity within the SOC. Standard footnote defining AEs included in table (e.g., includes AEs with onset from first dose of study drug through 7 days after last dose of study drug).
aet03_1_lyt(
  armvar = "ACTARM",
  lbl_AEBODSYS = "Body System or Organ Class",
  lbl_AEDECOD = "Dictionary-Derived Term",
  lbl_overall = NULL,
  deco = std_deco("AET03")
)
#> A Pre-data Table Layout
#> 
#> Column-Split Structure:
#> ACTARM (lvls) 
#> 
#> Row-Split Structure:
#> AEBODSYS (lvls) -> AEDECOD (lvls) 
#>