AET03 Table 1 (Default) Advert Events by Greatest Intensity Table 1

An adverse events table categorized by System Organ Class, Dictionary-Derived Term and Greatest intensity.

aet03_1(
  adsl,
  adae,
  armvar = .study$armvar,
  lbl_overall = .study$lbl_overall,
  prune_0 = TRUE,
  deco = std_deco("AET03"),
  .study = list(armvar = "ACTARM", lbl_overall = "")
)

aet03_1_lyt(
  armvar = .study$armvar,
  lbl_overall = .study$lbl_overall,
  lbl_AEBODSYS = "",
  lbl_AEDECOD = "",
  lbl_AESEV = "",
  gradation = .study$gradation,
  deco = std_deco("AET03"),
  .study = list(armvar = "ACTARM", lbl_overall = "", gradation = c("MILD", "MODERATE",
    "SEVERE", "LIFE THREATENING"))
)

Arguments

adsl
(data.frame) Subject Level Analysis Dataset

adae
(data.frame) Adverse Events Analysis Dataset

armvar
(string) variable used for column splitting

lbl_overall
(string) label used for overall column, if set to NULL the overall column is omitted

prune_0
(logical) remove 0 count rows

deco
(list) decoration with title, subtitles and main_footer content

.study
(list) with default values for the arguments of the function

lbl_AEBODSYS
(string) text label for AEBODSYS.

lbl_AEDECOD
(string) text label for AEDECOD.

lbl_AESEV
(string) text label for AESEV.

gradation
(vector of strings) describing the severity levels present in the dataset.

Value

Details

Default Adverse Events by Greatest Intensity table

Numbers represent absolute numbers of patients and fraction of N

Remove zero-count rows unless overridden with prune_0 = FALSE.

Split columns by arm

Does not include a total column by default

Sort by Body System or Organ Class (SOC) and Dictionary-Derived Term (PT)

Functions

aet03_1_lyt:

Examples

library(scda)
library(dplyr)
sd <- synthetic_cdisc_data("rcd_2021_03_22")
adsl <- sd$adsl
adae <- sd$adae %>%
 mutate(ANL01FL = 'Y')

aet03_1(adsl, adae)

#> Adverse Events by Greatest Intensity: {Specify Population}
#> Protocol: {{protocol}}, Snapshot: {{snapshot}}, Snapshot Date: {{snapshot-date}}, Cutoff Date: {{cutoff-date}}
#> 
#> ----------------------------------------------------------------------
#> Body System or Organ Class                                            
#>   Dictionary-Derived Term    A: Drug X    B: Placebo    C: Combination
#>   Severity/Intensity          (N=134)       (N=134)        (N=132)    
#> ----------------------------------------------------------------------
#> - Any Intensity -            122 (91%)    123 (91.8%)    120 (90.9%)  
#> MODERATE                     24 (17.9%)   24 (17.9%)      23 (17.4%)  
#> MILD                          7 (5.2%)     9 (6.7%)         4 (3%)    
#> SEVERE                       91 (67.9%)   90 (67.2%)      93 (70.5%)  
#> cl A.1                                                                
#>   - Any Intensity -          78 (58.2%)    75 (56%)       89 (67.4%)  
#>   MODERATE                   48 (35.8%)   48 (35.8%)      50 (37.9%)  
#>   MILD                       30 (22.4%)   27 (20.1%)      39 (29.5%)  
#>   dcd A.1.1.1.1                                                       
#>     - Any Intensity -        50 (37.3%)   45 (33.6%)      63 (47.7%)  
#>     MILD                     50 (37.3%)   45 (33.6%)      63 (47.7%)  
#>   dcd A.1.1.1.2                                                       
#>     - Any Intensity -        48 (35.8%)   48 (35.8%)      50 (37.9%)  
#>     MODERATE                 48 (35.8%)   48 (35.8%)      50 (37.9%)  
#> cl B.2                                                                
#>   - Any Intensity -           79 (59%)    74 (55.2%)      85 (64.4%)  
#>   MODERATE                   49 (36.6%)   44 (32.8%)      52 (39.4%)  
#>   MILD                       30 (22.4%)   30 (22.4%)       33 (25%)   
#>   dcd B.2.2.3.1                                                       
#>     - Any Intensity -        48 (35.8%)   54 (40.3%)      51 (38.6%)  
#>     MILD                     48 (35.8%)   54 (40.3%)      51 (38.6%)  
#>   dcd B.2.1.2.1                                                       
#>     - Any Intensity -        49 (36.6%)   44 (32.8%)      52 (39.4%)  
#>     MODERATE                 49 (36.6%)   44 (32.8%)      52 (39.4%)  
#> cl D.1                                                                
#>   - Any Intensity -           79 (59%)     67 (50%)       80 (60.6%)  
#>   MODERATE                   29 (21.6%)   25 (18.7%)       29 (22%)   
#>   SEVERE                     50 (37.3%)   42 (31.3%)      51 (38.6%)  
#>   dcd D.1.1.1.1                                                       
#>     - Any Intensity -        50 (37.3%)   42 (31.3%)      51 (38.6%)  
#>     SEVERE                   50 (37.3%)   42 (31.3%)      51 (38.6%)  
#>   dcd D.1.1.4.2                                                       
#>     - Any Intensity -        48 (35.8%)   42 (31.3%)      50 (37.9%)  
#>     MODERATE                 48 (35.8%)   42 (31.3%)      50 (37.9%)  
#> cl D.2                                                                
#>   - Any Intensity -          47 (35.1%)   58 (43.3%)      57 (43.2%)  
#>   MILD                       47 (35.1%)   58 (43.3%)      57 (43.2%)  
#>   dcd D.2.1.5.3                                                       
#>     - Any Intensity -        47 (35.1%)   58 (43.3%)      57 (43.2%)  
#>     MILD                     47 (35.1%)   58 (43.3%)      57 (43.2%)  
#> cl B.1                                                                
#>   - Any Intensity -          47 (35.1%)   49 (36.6%)      43 (32.6%)  
#>   SEVERE                     47 (35.1%)   49 (36.6%)      43 (32.6%)  
#>   dcd B.1.1.1.1                                                       
#>     - Any Intensity -        47 (35.1%)   49 (36.6%)      43 (32.6%)  
#>     SEVERE                   47 (35.1%)   49 (36.6%)      43 (32.6%)  
#> cl C.2                                                                
#>   - Any Intensity -          35 (26.1%)   48 (35.8%)      55 (41.7%)  
#>   MODERATE                   35 (26.1%)   48 (35.8%)      55 (41.7%)  
#>   dcd C.2.1.2.1                                                       
#>     - Any Intensity -        35 (26.1%)   48 (35.8%)      55 (41.7%)  
#>     MODERATE                 35 (26.1%)   48 (35.8%)      55 (41.7%)  
#> cl C.1                                                                
#>   - Any Intensity -          43 (32.1%)   46 (34.3%)      43 (32.6%)  
#>   SEVERE                     43 (32.1%)   46 (34.3%)      43 (32.6%)  
#>   dcd C.1.1.1.3                                                       
#>     - Any Intensity -        43 (32.1%)   46 (34.3%)      43 (32.6%)  
#>     SEVERE                   43 (32.1%)   46 (34.3%)      43 (32.6%)  
#> ----------------------------------------------------------------------
#> 
#> Investigator text for AEs encoded using MedDRA version xx.x.  All counts represent patients. Multiple occurrences of the same AE in one individual are counted once at the greatest intensity for this preferred term. To the SOC Overall row counts, a patient contributes only with the AE occuring with the greatest intensity within the SOC. Standard footnote defining AEs included in table (e.g., includes AEs with onset from first dose of study drug through 7 days after last dose of study drug).

aet03_1(adsl, adae, lbl_overall = "All Patients")

#> Adverse Events by Greatest Intensity: {Specify Population}
#> Protocol: {{protocol}}, Snapshot: {{snapshot}}, Snapshot Date: {{snapshot-date}}, Cutoff Date: {{cutoff-date}}
#> 
#> -------------------------------------------------------------------------------------
#> Body System or Organ Class                                                           
#>   Dictionary-Derived Term    A: Drug X    B: Placebo    C: Combination   All Patients
#>   Severity/Intensity          (N=134)       (N=134)        (N=132)         (N=400)   
#> -------------------------------------------------------------------------------------
#> - Any Intensity -            122 (91%)    123 (91.8%)    120 (90.9%)     365 (91.2%) 
#> MODERATE                     24 (17.9%)   24 (17.9%)      23 (17.4%)      71 (17.8%) 
#> MILD                          7 (5.2%)     9 (6.7%)         4 (3%)         20 (5%)   
#> SEVERE                       91 (67.9%)   90 (67.2%)      93 (70.5%)     274 (68.5%) 
#> cl A.1                                                                               
#>   - Any Intensity -          78 (58.2%)    75 (56%)       89 (67.4%)     242 (60.5%) 
#>   MODERATE                   48 (35.8%)   48 (35.8%)      50 (37.9%)     146 (36.5%) 
#>   MILD                       30 (22.4%)   27 (20.1%)      39 (29.5%)       96 (24%)  
#>   dcd A.1.1.1.1                                                                      
#>     - Any Intensity -        50 (37.3%)   45 (33.6%)      63 (47.7%)     158 (39.5%) 
#>     MILD                     50 (37.3%)   45 (33.6%)      63 (47.7%)     158 (39.5%) 
#>   dcd A.1.1.1.2                                                                      
#>     - Any Intensity -        48 (35.8%)   48 (35.8%)      50 (37.9%)     146 (36.5%) 
#>     MODERATE                 48 (35.8%)   48 (35.8%)      50 (37.9%)     146 (36.5%) 
#> cl B.2                                                                               
#>   - Any Intensity -           79 (59%)    74 (55.2%)      85 (64.4%)     238 (59.5%) 
#>   MODERATE                   49 (36.6%)   44 (32.8%)      52 (39.4%)     145 (36.2%) 
#>   MILD                       30 (22.4%)   30 (22.4%)       33 (25%)       93 (23.2%) 
#>   dcd B.2.2.3.1                                                                      
#>     - Any Intensity -        48 (35.8%)   54 (40.3%)      51 (38.6%)     153 (38.2%) 
#>     MILD                     48 (35.8%)   54 (40.3%)      51 (38.6%)     153 (38.2%) 
#>   dcd B.2.1.2.1                                                                      
#>     - Any Intensity -        49 (36.6%)   44 (32.8%)      52 (39.4%)     145 (36.2%) 
#>     MODERATE                 49 (36.6%)   44 (32.8%)      52 (39.4%)     145 (36.2%) 
#> cl D.1                                                                               
#>   - Any Intensity -           79 (59%)     67 (50%)       80 (60.6%)     226 (56.5%) 
#>   MODERATE                   29 (21.6%)   25 (18.7%)       29 (22%)       83 (20.8%) 
#>   SEVERE                     50 (37.3%)   42 (31.3%)      51 (38.6%)     143 (35.8%) 
#>   dcd D.1.1.1.1                                                                      
#>     - Any Intensity -        50 (37.3%)   42 (31.3%)      51 (38.6%)     143 (35.8%) 
#>     SEVERE                   50 (37.3%)   42 (31.3%)      51 (38.6%)     143 (35.8%) 
#>   dcd D.1.1.4.2                                                                      
#>     - Any Intensity -        48 (35.8%)   42 (31.3%)      50 (37.9%)      140 (35%)  
#>     MODERATE                 48 (35.8%)   42 (31.3%)      50 (37.9%)      140 (35%)  
#> cl D.2                                                                               
#>   - Any Intensity -          47 (35.1%)   58 (43.3%)      57 (43.2%)     162 (40.5%) 
#>   MILD                       47 (35.1%)   58 (43.3%)      57 (43.2%)     162 (40.5%) 
#>   dcd D.2.1.5.3                                                                      
#>     - Any Intensity -        47 (35.1%)   58 (43.3%)      57 (43.2%)     162 (40.5%) 
#>     MILD                     47 (35.1%)   58 (43.3%)      57 (43.2%)     162 (40.5%) 
#> cl B.1                                                                               
#>   - Any Intensity -          47 (35.1%)   49 (36.6%)      43 (32.6%)     139 (34.8%) 
#>   SEVERE                     47 (35.1%)   49 (36.6%)      43 (32.6%)     139 (34.8%) 
#>   dcd B.1.1.1.1                                                                      
#>     - Any Intensity -        47 (35.1%)   49 (36.6%)      43 (32.6%)     139 (34.8%) 
#>     SEVERE                   47 (35.1%)   49 (36.6%)      43 (32.6%)     139 (34.8%) 
#> cl C.2                                                                               
#>   - Any Intensity -          35 (26.1%)   48 (35.8%)      55 (41.7%)     138 (34.5%) 
#>   MODERATE                   35 (26.1%)   48 (35.8%)      55 (41.7%)     138 (34.5%) 
#>   dcd C.2.1.2.1                                                                      
#>     - Any Intensity -        35 (26.1%)   48 (35.8%)      55 (41.7%)     138 (34.5%) 
#>     MODERATE                 35 (26.1%)   48 (35.8%)      55 (41.7%)     138 (34.5%) 
#> cl C.1                                                                               
#>   - Any Intensity -          43 (32.1%)   46 (34.3%)      43 (32.6%)      132 (33%)  
#>   SEVERE                     43 (32.1%)   46 (34.3%)      43 (32.6%)      132 (33%)  
#>   dcd C.1.1.1.3                                                                      
#>     - Any Intensity -        43 (32.1%)   46 (34.3%)      43 (32.6%)      132 (33%)  
#>     SEVERE                   43 (32.1%)   46 (34.3%)      43 (32.6%)      132 (33%)  
#> -------------------------------------------------------------------------------------
#> 
#> Investigator text for AEs encoded using MedDRA version xx.x.  All counts represent patients. Multiple occurrences of the same AE in one individual are counted once at the greatest intensity for this preferred term. To the SOC Overall row counts, a patient contributes only with the AE occuring with the greatest intensity within the SOC. Standard footnote defining AEs included in table (e.g., includes AEs with onset from first dose of study drug through 7 days after last dose of study drug).

aet03_1_lyt(armvar = "ACTARM",
 lbl_overall = "",
 lbl_AEBODSYS = "Body System or Organ Class",
 lbl_AESEV = "Severity/Intensity",
 lbl_AEDECOD = "Dictionary-Derived Term",
 deco = std_deco("AET03"))

#> A Pre-data Table Layout
#> 
#> Column-Split Structure:
#> ACTARM (lvls) 
#> 
#> Row-Split Structure:
#> AEBODSYS (lvls) -> AEDECOD (lvls) 
#>