chevron package

Provide standard tables, listings, and graphs (TLGs) libraries used in clinical trials. This package implements a structure to reformat the data with 'dunlin', create reporting tables using 'rtables' and 'tern' with standardized input arguments to enable quick generation of standard outputs. In addition, it also provides comprehensive data checks and script generation functionality.

See also

Author

Maintainer: Liming Li liming.li@roche.com

Authors:

Benoit Falquet benoit.falquet@roche.com

Xiaoli Duan xiaoli.duan@roche.com

Other contributors:

Adrian Waddell waddell.adrian@gene.com [contributor]

Chenkai Lv chenkai.lv@roche.com [contributor]

Pawel Rucki pawel.rucki@roche.com [contributor]

Tim Barnett timothy.barnett@roche.com [contributor]

Tian Fang tian.fang@roche.com [contributor]

F. Hoffmann-La Roche AG [copyright holder, funder]