chevron_t

chevron_t, a subclass of chevron_tlg with specific validation criteria to handle table creation

chevron_l, a subclass of chevron_tlg with specific validation criteria to handle listing creation

chevron_g, a subclass of chevron_tlg with specific validation criteria to handle graph creation

Usage

chevron_t(
  main = function(adam_db, ...) build_table(basic_table(), adam_db[[1]]),
  preprocess = function(adam_db, ...) adam_db,
  postprocess = report_null,
  adam_datasets = NA_character_,
  ...
)

chevron_l(
  main = function(adam_db, ...) data.frame(),
  preprocess = function(adam_db, ...) adam_db,
  postprocess = function(tlg, ...) tlg,
  adam_datasets = NA_character_,
  ...
)

chevron_g(
  main = function(adam_db, ...) ggplot2::ggplot(),
  preprocess = function(adam_db, ...) adam_db,
  postprocess = function(tlg, ...) tlg,
  adam_datasets = NA_character_,
  ...
)

Arguments

main

(function) returning a tlg, with adam_db as first argument. Typically one of the _main function of chevron.

preprocess

(function) returning a pre-processed list of data.frames, with adam_db as first argument. Typically one of the _pre function of chevron.

postprocess

(function) returning a post-processed tlg, with tlg as first argument.

adam_datasets

(character) representing the names of the tables from an ADaM data set required for tlg creation.

...

not used

Slots

main

(function) returning a tlg. Typically one of the *_main function from chevron.

preprocess

(function) returning a pre-processed list of data.frames amenable to tlg creation. Typically one of the *_pre function from chevron.

postprocess

(function) returning a post-processed tlg.

adam_datasets

(character) representing the name of the tables from an AdAM dataset required for tlg creation.

Note

To ensure the correct execution of the workflow additional validation criteria are:

the first argument of the main function must be adam_db, the input list of data.frames to pre-process. The ... argument is mandatory.

the first argument of the preprocess function must be adam_db, the input list of data.frames to create tlg output. The ... argument is mandatory.

the first argument of the postprocess function must be tlg, the input TableTree object to post-process. The ... argument is mandatory.

Examples

chevron_t_obj <- chevron_t()
chevron_t_obj <- chevron_t(postprocess = function(tlg, indent, ...) {
  rtables::table_inset(tlg) <- indent
  tlg
})

chevron_l_obj <- chevron_l()

chevron_g_obj <- chevron_g()
chevron_g_obj <- chevron_g(
  postprocess = function(tlg, title, ...) tlg + ggplot2::labs(main = title)
)