CMT01A Table 3 (Supplementary) Concomitant Medication by Medication Class and Preferred Name (Total number of treatments per medication class suppressed).

A concomitant medication table with the number of subjects and the total number of treatments by medication class sorted alphabetically and medication name sorted by frequencies presented without the total number of treatments per medication.

cmt01a_3(
  adam_db,
  armvar = .study$planarm,
  medcat_var = "ATC2",
  lbl_medcat_var = "ATC Class Level 2",
  medname_var = "CMDECOD",
  lbl_medname_var = "Other Treatment",
  lbl_overall = .study$lbl_overall,
  prune_0 = TRUE,
  deco = std_deco("CMT01A"),
  .study = list(planarm = "ARM", lbl_overall = NULL)
)

cmt01a_3_lyt(
  armvar = .study$planarm,
  lbl_overall = .study$lbl_overall,
  medcat_var = "ATC2",
  lbl_medcat_var = "ATC Class Level 2",
  medname_var = "CMDECOD",
  lbl_medname_var = "Other Treatment",
  deco = std_deco("CMT01A"),
  .study = list(planarm = "ARM", lbl_overall = NULL)
)

cmt01a_3_pre(adam_db)

Arguments

adam_db

(dm) object containing the ADaM datasets

armvar

(character) variable used for column splitting

medcat_var

(character) the variable defining the medication category. By default ATC2.

lbl_medcat_var

(character) the label for the medication category.

medname_var

(character) the variable defining the medication name. By default CMDECOD.

lbl_medname_var

(character) the label for the medication name.

lbl_overall

(character) label used for overall column, if set to NULL the overall column is omitted

prune_0

(logical) remove 0 count rows

deco

(character) decoration with title, subtitles and main_footer content

.study

(list) with default values for the arguments of the function

Details

Data should be filtered for concomitant medication. (ATIREL == "CONCOMITANT").

Numbers represent absolute numbers of subjects and fraction of N, or absolute numbers when specified.

Remove zero-count rows unless overridden with prune_0 = FALSE.

Split columns by arm.

Does not include a total column by default.

Sort by medication class alphabetically and within medication class by decreasing total number of patients with the specific medication.

Functions

cmt01a_3_lyt: cmt01a_3 Layout

cmt01a_3_pre: cmt01a_3 Preprocessing

Examples

library(dm)

db <- syn_test_data() %>%
  dm_zoom_to("adcm") %>%
  filter(ATIREL == "CONCOMITANT") %>%
  dm_update_zoomed() %>%
  cmt01a_3_pre()

cmt01a_3(adam_db = db)
#> Concomitant Medications by Medication Class and Preferred Name: {Specify Population}
#> Protocol: {{protocol}}
#> 
#> ———————————————————————————————————————————————————————————————————————————————————————————————————
#> ATC Class Level 2                                         A: Drug X    B: Placebo    C: Combination
#>   Other Treatment                                          (N=134)       (N=134)        (N=132)    
#> ———————————————————————————————————————————————————————————————————————————————————————————————————
#> Total number of patients with at least one treatment     117 (87.3%)   116 (86.6%)    116 (87.9%)  
#> Total number of treatments                                   415           414            460      
#> ATCCLAS2 A                                                                                         
#>   Total number of patients with at least one treatment    75 (56%)      79 (59%)       81 (61.4%)  
#>   medname A_2/3                                          53 (39.6%)    50 (37.3%)      56 (42.4%)  
#>   medname A_3/3                                          45 (33.6%)    54 (40.3%)      48 (36.4%)  
#> ATCCLAS2 A p2                                                                                      
#>   Total number of patients with at least one treatment    75 (56%)      79 (59%)       81 (61.4%)  
#>   medname A_2/3                                          53 (39.6%)    50 (37.3%)      56 (42.4%)  
#>   medname A_3/3                                          45 (33.6%)    54 (40.3%)      48 (36.4%)  
#> ATCCLAS2 A p3                                                                                      
#>   Total number of patients with at least one treatment    75 (56%)      79 (59%)       81 (61.4%)  
#>   medname A_2/3                                          53 (39.6%)    50 (37.3%)      56 (42.4%)  
#>   medname A_3/3                                          45 (33.6%)    54 (40.3%)      48 (36.4%)  
#> ATCCLAS2 B                                                                                         
#>   Total number of patients with at least one treatment   83 (61.9%)    74 (55.2%)      88 (66.7%)  
#>   medname B_1/4                                          52 (38.8%)    57 (42.5%)      59 (44.7%)  
#>   medname B_4/4                                          50 (37.3%)    45 (33.6%)      55 (41.7%)  
#> ATCCLAS2 C                                                                                         
#>   Total number of patients with at least one treatment   82 (61.2%)    84 (62.7%)      89 (67.4%)  
#>   medname C_2/2                                          52 (38.8%)    58 (43.3%)      60 (45.5%)  
#>   medname C_1/2                                          51 (38.1%)    50 (37.3%)      56 (42.4%)  
#> ATCCLAS2 C p2                                                                                      
#>   Total number of patients with at least one treatment   82 (61.2%)    84 (62.7%)      89 (67.4%)  
#>   medname C_2/2                                          52 (38.8%)    58 (43.3%)      60 (45.5%)  
#>   medname C_1/2                                          51 (38.1%)    50 (37.3%)      56 (42.4%)  
#> ATCCLAS2 C p3                                                                                      
#>   Total number of patients with at least one treatment   51 (38.1%)    50 (37.3%)      56 (42.4%)  
#>   medname C_1/2                                          51 (38.1%)    50 (37.3%)      56 (42.4%)  
#> ———————————————————————————————————————————————————————————————————————————————————————————————————
#> 
#> Treatments are coded using the WHODrug Global B3 Format dictionary {version}.  Treatments can appear under multiple
#> classes. Multiple uses of a specific medication for a patient were counted once in the frequency for the medication.
#> Likewise, multiple uses within a specific medication class for a patient were counted once in the frequency for the
#> medication class. For frequency counts in "Total number of treatments", multiple uses of the same medication for a
#> patient were counted separately. Standard footnote defining medications included in table (e.g., includes concomitant
#> medications used at any time from first dose of study drug through 7 days after last dose of study drug).
cmt01a_3_lyt(
  armvar = "ARM",
  lbl_overall = NULL,
  deco = std_deco("CMT01A")
)
#> A Pre-data Table Layout
#> 
#> Column-Split Structure:
#> ARM (lvls) 
#> 
#> Row-Split Structure:
#> ATC2 (lvls) -> CMDECOD (** analysis **) 
#> 
syn_test_data() %>%
  cmt01a_3_pre()
#> ── Metadata ────────────────────────────────────────────────────────────────────
#> Tables: `adsl`, `adae`, `adaette`, `adcm`, `addv`, … (15 total)
#> Columns: 846
#> Primary keys: 2
#> Foreign keys: 1