DMT01 Table 1 (Default) Demographics and Baseline Characteristics Table

For each variable, summary statistics are by default based on the number of patients in the corresponding n row.

dmt01_1(
  ad_bl,
  armvar = .study$armvar,
  summaryvars = .study$summaryvars,
  lbl_overall = .study$lbl_overall,
  prune_0 = TRUE,
  deco = std_deco("DMT01"),
  .study = list(armvar = "ARM", summaryvars = c("AAGE", "AGEGR1", "SEX", "ETHNIC",
    "RACE", "BWGHTSI"), lbl_overall = "All Patients")
)

dmt01_1_lyt(
  armvar = .study$armvar,
  summaryvars = .study$summaryvars,
  summaryvars_lbls = var_labels_for(ad_bl, summaryvars),
  lbl_overall = .study$lbl_overall,
  deco = std_deco("DMT01"),
  .study = list(armvar = "ARM", summaryvars = c("AAGE", "AGEGR1", "SEX", "ETHNIC",
    "RACE", "BWGHTSI"), overall_col = "All Patients")
)

Arguments

ad_bl
subject baseline dataset with one row per patient

armvar
variable used for column splitting

summaryvars
(vector of strings) variables summarized in demographic table.

lbl_overall
label used for overall column, if set to NULL the overall column is omitted

prune_0
remove 0 count rows

deco
decoration list with title, subtitles and main_footer content

.study
a list with default values for the arguments of the function

summaryvars_lbls
(vector of strings) labels corresponding to the analyzed variables.

Value

Details

Default demographic and characteristics table

If not specified otherwise, numbers represent absolute numbers of patients and fraction of N

Remove zero-count rows

Split columns by arm (planned or actual / code or description)

Include a total column by default

Functions

dmt01_1_lyt:

Examples

library(scda)
adsl <- synthetic_cdisc_data("rcd_2021_03_22")$adsl

dmt01_1(ad_bl = adsl, summaryvars = c("AGE", "RACE", "SEX"))

#> Main Title
#> Protocol: {{protocol}}, Snapshot: {{snapshot}}, Snapshot Date: {{snapshot-date}}, Cutoff Date: {{cutoff-date}}
#> 
#> -----------------------------------------------------------------------------------------------------
#>                                               A: Drug X    B: Placebo   C: Combination   All Patients
#>                                                (N=134)      (N=134)        (N=132)         (N=400)   
#> -----------------------------------------------------------------------------------------------------
#> Age                                                                                                  
#>   n                                              134          134            132             400     
#>   Mean (SD)                                   33.8 (6.6)   35.4 (7.9)     35.4 (7.7)      34.9 (7.4) 
#>   Median                                          33           35             35              34     
#>   Min - Max                                    21 - 50      21 - 62        20 - 69         20 - 69   
#> Race                                                                                                 
#>   n                                              134          134            132             400     
#>   ASIAN                                       68 (50.7%)    67 (50%)      73 (55.3%)      208 (52%)  
#>   BLACK OR AFRICAN AMERICAN                   31 (23.1%)   28 (20.9%)     32 (24.2%)      91 (22.8%) 
#>   WHITE                                       27 (20.1%)   26 (19.4%)     21 (15.9%)      74 (18.5%) 
#>   AMERICAN INDIAN OR ALASKA NATIVE              8 (6%)     11 (8.2%)       6 (4.5%)       25 (6.2%)  
#>   MULTIPLE                                        0         1 (0.7%)          0            1 (0.2%)  
#>   NATIVE HAWAIIAN OR OTHER PACIFIC ISLANDER       0         1 (0.7%)          0            1 (0.2%)  
#> Sex                                                                                                  
#>   n                                              134          134            132             400     
#>   F                                            79 (59%)    82 (61.2%)      70 (53%)      231 (57.8%) 
#>   M                                            55 (41%)    52 (38.8%)      62 (47%)      169 (42.2%) 
#> -----------------------------------------------------------------------------------------------------
#> 
#> Footnotes go here

dmt01_1(
 ad_bl = adsl,
 armvar = "ARMCD",
 summaryvars = c("AGE", "RACE")
)

#> Main Title
#> Protocol: {{protocol}}, Snapshot: {{snapshot}}, Snapshot Date: {{snapshot-date}}, Cutoff Date: {{cutoff-date}}
#> 
#> -------------------------------------------------------------------------------------------------
#>                                                 ARM A        ARM B        ARM C      All Patients
#>                                                (N=134)      (N=134)      (N=132)       (N=400)   
#> -------------------------------------------------------------------------------------------------
#> Age                                                                                              
#>   n                                              134          134          132           400     
#>   Mean (SD)                                   33.8 (6.6)   35.4 (7.9)   35.4 (7.7)    34.9 (7.4) 
#>   Median                                          33           35           35            34     
#>   Min - Max                                    21 - 50      21 - 62      20 - 69       20 - 69   
#> Race                                                                                             
#>   n                                              134          134          132           400     
#>   ASIAN                                       68 (50.7%)    67 (50%)    73 (55.3%)    208 (52%)  
#>   BLACK OR AFRICAN AMERICAN                   31 (23.1%)   28 (20.9%)   32 (24.2%)    91 (22.8%) 
#>   WHITE                                       27 (20.1%)   26 (19.4%)   21 (15.9%)    74 (18.5%) 
#>   AMERICAN INDIAN OR ALASKA NATIVE              8 (6%)     11 (8.2%)     6 (4.5%)     25 (6.2%)  
#>   MULTIPLE                                        0         1 (0.7%)        0          1 (0.2%)  
#>   NATIVE HAWAIIAN OR OTHER PACIFIC ISLANDER       0         1 (0.7%)        0          1 (0.2%)  
#> -------------------------------------------------------------------------------------------------
#> 
#> Footnotes go here