DST01 Table 1 (Default) Patient Disposition Table 1

The DST01 Disposition Table provides an overview of patients study completion. For patients who discontinued the study a reason is provided.

dst01_1(
  adsl,
  armvar = .study$armvar,
  lbl_overall = .study$lbl_overall,
  status = .study$status,
  reason = .study$reason,
  prune_0 = TRUE,
  deco = std_deco("DST01"),
  .study = list(armvar = "ARM", lbl_overall = "All patients", status = "EOSSTT", reason
    = "DCSREAS")
)

dst01_1_lyt(
  armvar = .study$armvar,
  lbl_overall = .study$lbl_overall,
  status = .study$status,
  reason = .study$reason,
  deco = std_deco("DST01"),
  completed_lbl = "COMPLETED",
  ongoing_lbl = "ONGOING",
  discontinued_lbl = "DISCONTINUED",
  .study = list(armvar = "ARM", lbl_overall = "All patients", status = "EOSSTT", reason
    = "DCSREAS")
)

Arguments

adsl
(data.frame) Subject Level Analysis Dataset

armvar
(string) variable. Usually one of ARM, ACTARM, TRT01A, or TRT01A.

lbl_overall
(string) label used for overall column, if set to NULL the overall column is omitted

status
(string) variable used to define patient status. Default is EOSSTT, however can also be a variable name with the pattern EOPxxSTT where xx must be substituted by 2 digits referring to the analysis period.

reason
(string) variable used to define reason for patient withdrawal. Default is DCSREAS, however can also be a variable with the pattern DCPxxRS where xx must be substituted by 2 digits referring to the analysis period.

prune_0
(logical) remove 0 count rows

deco
(list) decoration with title, subtitles and main_footer content

.study
(list) with default values for the arguments of the function

completed_lbl
(string) associated with completed study and found in the columns given by status. By Default COMPLETED.

ongoing_lbl
(string) associated with ongoing treatment and found in the columns given by status_treatment. By Default `ONGOING.

discontinued_lbl
(string) associated with discontinued study and found in the columns given by status. By Default DISCONTINUED.

Value

Details

Default patient disposition table summarizing the reasons for patients withdrawal.

Numbers represent absolute numbers of patients and fraction of N.

Remove zero-count rows.

Split columns by arm.

Include a total column by default.

Sort withdrawal reasons by alphabetic order.

Functions

dst01_1_lyt:

Examples

library(scda)
library(dplyr)

sd <- synthetic_cdisc_data("rcd_2021_03_22")
adsl <- sd$adsl

dst01_1(adsl)

#>                                    A: Drug X    B: Placebo    C: Combination   All patients
#>                                     (N=134)       (N=134)        (N=132)         (N=400)   
#> -------------------------------------------------------------------------------------------
#> Completed Study                   69 (51.49%)   69 (51.49%)    72 (54.55%)     210 (52.5%) 
#> Ongoing Study                     27 (20.15%)   22 (16.42%)    21 (15.91%)      70 (17.5%) 
#> Discontinued Study                38 (28.4%)    43 (32.1%)      39 (29.5%)      120 (30%)  
#>   ADVERSE EVENT                    6 (4.5%)      1 (0.7%)        2 (1.5%)        9 (2.2%)  
#>   DEATH                           22 (16.4%)    26 (19.4%)      19 (14.4%)      67 (16.8%) 
#>   LACK OF EFFICACY                 2 (1.5%)      1 (0.7%)        2 (1.5%)        5 (1.2%)  
#>   PHYSICIAN DECISION                   0         1 (0.7%)        7 (5.3%)         8 (2%)   
#>   PROTOCOL VIOLATION                4 (3%)       7 (5.2%)        6 (4.5%)       17 (4.2%)  
#>   WITHDRAWAL BY PARENT/GUARDIAN    2 (1.5%)      5 (3.7%)        3 (2.3%)       10 (2.5%)  
#>   WITHDRAWAL BY SUBJECT            2 (1.5%)      2 (1.5%)           0             4 (1%)   

dst01_1(adsl, lbl_overall = "")

#>                                    A: Drug X    B: Placebo    C: Combination
#>                                     (N=134)       (N=134)        (N=132)    
#> ----------------------------------------------------------------------------
#> Completed Study                   69 (51.49%)   69 (51.49%)    72 (54.55%)  
#> Ongoing Study                     27 (20.15%)   22 (16.42%)    21 (15.91%)  
#> Discontinued Study                38 (28.4%)    43 (32.1%)      39 (29.5%)  
#>   ADVERSE EVENT                    6 (4.5%)      1 (0.7%)        2 (1.5%)   
#>   DEATH                           22 (16.4%)    26 (19.4%)      19 (14.4%)  
#>   LACK OF EFFICACY                 2 (1.5%)      1 (0.7%)        2 (1.5%)   
#>   PHYSICIAN DECISION                   0         1 (0.7%)        7 (5.3%)   
#>   PROTOCOL VIOLATION                4 (3%)       7 (5.2%)        6 (4.5%)   
#>   WITHDRAWAL BY PARENT/GUARDIAN    2 (1.5%)      5 (3.7%)        3 (2.3%)   
#>   WITHDRAWAL BY SUBJECT            2 (1.5%)      2 (1.5%)           0       

adsl$EOSSTT <- as.factor(toupper(adsl$EOSSTT))
dst01_1(adsl)

#>                                    A: Drug X    B: Placebo    C: Combination   All patients
#>                                     (N=134)       (N=134)        (N=132)         (N=400)   
#> -------------------------------------------------------------------------------------------
#> Completed Study                   69 (51.49%)   69 (51.49%)    72 (54.55%)     210 (52.5%) 
#> Ongoing Study                     27 (20.15%)   22 (16.42%)    21 (15.91%)      70 (17.5%) 
#> Discontinued Study                38 (28.4%)    43 (32.1%)      39 (29.5%)      120 (30%)  
#>   ADVERSE EVENT                    6 (4.5%)      1 (0.7%)        2 (1.5%)        9 (2.2%)  
#>   DEATH                           22 (16.4%)    26 (19.4%)      19 (14.4%)      67 (16.8%) 
#>   LACK OF EFFICACY                 2 (1.5%)      1 (0.7%)        2 (1.5%)        5 (1.2%)  
#>   PHYSICIAN DECISION                   0         1 (0.7%)        7 (5.3%)         8 (2%)   
#>   PROTOCOL VIOLATION                4 (3%)       7 (5.2%)        6 (4.5%)       17 (4.2%)  
#>   WITHDRAWAL BY PARENT/GUARDIAN    2 (1.5%)      5 (3.7%)        3 (2.3%)       10 (2.5%)  
#>   WITHDRAWAL BY SUBJECT            2 (1.5%)      2 (1.5%)           0             4 (1%)   

dst01_1_lyt(armvar = "ACTARM",
status = "EOP01STT",
reason = "DCP01RS")

#> [[1]]
#> A Pre-data Table Layout
#> 
#> Column-Split Structure:
#> ACTARM (lvls) 
#>  (all obs) 
#> 
#> Row-Split Structure:
#> EOP01STT (** analysis **) 
#> 
#> 
#> [[2]]
#> A Pre-data Table Layout
#> 
#> Column-Split Structure:
#> ACTARM (lvls) 
#>  (all obs) 
#> 
#> Row-Split Structure:
#> EOP01STT (** analysis **) 
#> EOP01STT (lvls) -> DCP01RS (** analysis **) 
#> 
#>