DST01 Table 1 (Default) Patient Disposition Table 1

The DST01 Disposition Table provides an overview of patients study completion. For patients who discontinued the study a reason is provided.

dst01_1(
  adam_db,
  armvar = .study$planarm,
  status_var = "EOSSTT",
  disc_reason_var = .study$disc_reason_var,
  prune_0 = TRUE,
  lbl_overall = .study$lbl_overall,
  deco = std_deco("DST01"),
  .study = list(planarm = "ARM", lbl_overall = "All Patients", disc_reason_var =
    "DCSREAS")
)

dst01_1_lyt(
  armvar = .study$planarm,
  status = .study$status,
  disc_reason_var = .study$disc_reason_var,
  completed_lbl = "COMPLETED",
  ongoing_lbl = "ONGOING",
  discontinued_lbl = "DISCONTINUED",
  lbl_overall = .study$lbl_overall,
  deco = std_deco("DST01"),
  .study = list(planarm = "ARM", lbl_overall = "All Patients", status = "EOSSTT",
    disc_reason_var = "DCSREAS")
)

Arguments

adam_db

(dm) object containing the ADaM datasets

armvar

(character) variable. Usually one of ARM, ACTARM, TRT01A, or TRT01A.

status_var

(character) variable used to define patient status. Default is EOSSTT, however can also be a variable name with the pattern EOPxxSTT where xx must be substituted by 2 digits referring to the analysis period.

disc_reason_var

(string) variable used to define reason for patient withdrawal. Default is DCSREAS, however can also be a variable with the pattern DCPxxRS where xx must be substituted by 2 digits referring to the analysis period.

prune_0

(logical) remove 0 count rows

lbl_overall

(character) label used for overall column, if set to NULL the overall column is omitted

deco

(character) decoration with title, subtitles and main_footer content

.study

(list) with default values for the arguments of the function

status

(string) variable used to define patient status. Default is EOSSTT, however can also be a variable name with the pattern EOPxxSTT where xx must be substituted by 2 digits referring to the analysis period.

completed_lbl

(string) associated with completed study and found in the columns given by status. By Default COMPLETED.

ongoing_lbl

(string) associated with ongoing treatment and found in the columns given by status_treatment. By Default `ONGOING.

discontinued_lbl

(string) associated with discontinued study and found in the columns given by status. By Default DISCONTINUED.

Details

Default patient disposition table summarizing the reasons for patients withdrawal.

Numbers represent absolute numbers of patients and fraction of N.

Remove zero-count rows.

Split columns by arm.

Include a total column by default.

Sort withdrawal reasons by alphabetic order.

Since the two parts of the tables are pruned differently, the layout function returns a list of layouts, which allows the tables to be constructed and pruned separately before binding.

Functions

dst01_1_lyt:

Examples

library(dm)

db <- syn_test_data() %>%
  preprocess_data("dst01_1")

dst01_1(db)
#> Patient Disposition: {Analysis Population}
#> Protocol: {{protocol}}, Snapshot: {{snapshot}}, Snapshot Date: {{snapshot-date}}, Cutoff Date: {{cutoff-date}}
#> 
#> —————————————————————————————————————————————————————————————————————————————————————————
#>                                   A: Drug X    B: Placebo   C: Combination   All Patients
#>                                    (N=134)      (N=134)        (N=132)         (N=400)   
#> —————————————————————————————————————————————————————————————————————————————————————————
#> Completed Study                   69 (51.5%)   69 (51.5%)     72 (54.5%)     210 (52.5%) 
#> Ongoing                           27 (20.1%)   22 (16.4%)     21 (15.9%)      70 (17.5%) 
#> Discontinued Study                38 (28.4%)   43 (32.1%)     39 (29.5%)     120 (30.0%) 
#>   ADVERSE EVENT                    6 (4.5%)     1 (0.7%)       2 (1.5%)        9 (2.2%)  
#>   DEATH                           22 (16.4%)   26 (19.4%)     19 (14.4%)      67 (16.8%) 
#>   LACK OF EFFICACY                 2 (1.5%)     1 (0.7%)       2 (1.5%)        5 (1.2%)  
#>   PHYSICIAN DECISION                  0         1 (0.7%)       7 (5.3%)         8 (2%)   
#>   PROTOCOL VIOLATION                4 (3%)      7 (5.2%)       6 (4.5%)       17 (4.2%)  
#>   WITHDRAWAL BY PARENT/GUARDIAN    2 (1.5%)     5 (3.7%)       3 (2.3%)       10 (2.5%)  
#>   WITHDRAWAL BY SUBJECT            2 (1.5%)     2 (1.5%)          0             4 (1%)   
dst01_1_lyt(
  armvar = "ACTARM",
  status = "EOP01STT",
  disc_reason_var = "DCP01RS"
)
#> [[1]]
#> A Pre-data Table Layout
#> 
#> Column-Split Structure:
#> ACTARM (lvls) 
#>  (all obs) 
#> 
#> Row-Split Structure:
#> EOP01STT (** analysis **) 
#> 
#> 
#> [[2]]
#> A Pre-data Table Layout
#> 
#> Column-Split Structure:
#> ACTARM (lvls) 
#>  (all obs) 
#> 
#> Row-Split Structure:
#> EOP01STT (** analysis **) 
#> EOP01STT (lvls) -> DCP01RS (** analysis **) 
#> 
#>