DST01 Table 3 (Supplementary) Patient Disposition Table 3.

The DST01_3 Disposition Table provides an overview of patients study treatment status. For patients who discontinued the study, the reason provided is categorized as "Safety" or "Non-Safety" issue.

Usage

dst01_3_main(
  adam_db,
  arm_var = "ARM",
  status_var = "EOSSTT",
  disc_reason_var = "DCSREAS",
  status_treatment_var = "EOTSTT",
  lbl_overall = "All Patients",
  deco = std_deco("DST01"),
  ...
)

dst01_3_lyt(
  arm_var,
  status_treatment_var,
  completed_trt_lbl,
  discontinued_trt_lbl,
  ongoing_trt_lbl,
  lbl_overall,
  deco,
  ...
)

dst01_3_pre(
  adam_db,
  status_var = "EOSSTT",
  disc_reason_var = "DCSREAS",
  status_treatment_var = "EOTSTT",
  ...
)

dst01_3_post(tlg, prune_0 = TRUE, deco = std_deco("DST01"), ...)

dst01_3

Format

An object of class chevron_t of length 1.

Arguments

adam_db

(dm) object containing the ADaM datasets

arm_var

(character) variable used for column splitting

status_var

(character) variable used to define patient status. Default is EOSSTT, however can also be a variable name with the pattern EOPxxSTT where xx must be substituted by 2 digits referring to the analysis period.

disc_reason_var

(character) variable used to define reason for patient withdrawal. Default is DCSREAS, however can also be a variable with the pattern DCPxxRS where xx must be substituted by 2 digits referring to the analysis period.

status_treatment_var

(string) variable used to define the treatment status of the patients. Default is EOTSTT, however can also be a variable with the pattern EOTxxSTT where xx must be substituted by 2 digits referring to the analysis period.

lbl_overall

(character) label used for overall column, if set to NULL the overall column is omitted

deco

(character) decoration with title, subtitles and main_footer content

...

not used.

completed_trt_lbl

(string) associated with completed treatment and found in the columns given by status_treatment_var. By Default COMPLETED.

discontinued_trt_lbl

(string) associated with discontinued treatment and found in the columns given by status_treatment_var. By Default DISCONTINUED.

ongoing_trt_lbl

(string) associated with ongoing treatment and found in the columns given by status_treatment_var. By Default `ONGOING.

tlg

(TableTree, Listing or ggplot) object typically produced by a main function.

prune_0

(logical) remove 0 count rows

Details

Non-standard disposition table summarizing the reasons for patient withdrawal and treatment status.

Withdrawal reasons are grouped into Safety and Non-Safety issues.

Safety issues include Death and Adverse Events.

Numbers represent absolute numbers of patients and fraction of N.

Remove zero-count rows.

Split columns by arm.

Include a total column by default.

Sort withdrawal reasons by alphabetic order.

Functions

dst01_3_main(): Main TLG function

dst01_3_lyt(): Layout

dst01_3_pre(): Preprocessing

dst01_3_post(): Postprocessing

Note

adam_db object must contain an adsl table with the column specified in status, status_treatment and disc_reason_var.

Examples

library(dm)
library(magrittr)

db <- syn_data %>%
  dst01_3_pre()

dst01_3_main(db)
#> [[1]]
#>                          A: Drug X    B: Placebo   C: Combination   All Patients
#>                           (N=134)      (N=134)        (N=132)         (N=400)   
#> ————————————————————————————————————————————————————————————————————————————————
#> Completed Treatment      47 (35.1%)   35 (26.1%)     42 (31.8%)     124 (31.0%) 
#> Ongoing Treatment        39 (29.1%)   46 (34.3%)     46 (34.8%)     131 (32.8%) 
#> Discontinued Treatment   48 (35.8%)   53 (39.6%)     44 (33.3%)     145 (36.2%) 
#> 
#> [[2]]
#>                   A: Drug X    B: Placebo   C: Combination   All Patients
#>                    (N=134)      (N=134)        (N=132)         (N=400)   
#> —————————————————————————————————————————————————————————————————————————
#> Completed Study   68 (50.7%)   66 (49.3%)     73 (55.3%)     207 (51.7%) 
#> 
#> [[3]]
#>                                     A: Drug X    B: Placebo   C: Combination   All Patients
#>                                      (N=134)      (N=134)        (N=132)         (N=400)   
#> ———————————————————————————————————————————————————————————————————————————————————————————
#> Ongoing                             24 (17.9%)   28 (20.9%)     21 (15.9%)      73 (18.2%) 
#> Discontinued Study                  42 (31.3%)   40 (29.9%)     38 (28.8%)     120 (30.0%) 
#>   Safety                            28 (20.9%)   29 (21.6%)     27 (20.5%)      84 (21.0%) 
#>     ADVERSE EVENT                    3 (2.2%)     6 (4.5%)       5 (3.8%)       14 (3.5%)  
#>     DEATH                           25 (18.7%)   23 (17.2%)     22 (16.7%)      70 (17.5%) 
#>     LACK OF EFFICACY                    0            0              0               0      
#>     PHYSICIAN DECISION                  0            0              0               0      
#>     PROTOCOL VIOLATION                  0            0              0               0      
#>     WITHDRAWAL BY PARENT/GUARDIAN       0            0              0               0      
#>     WITHDRAWAL BY SUBJECT               0            0              0               0      
#>   Non-Safety                        14 (10.4%)   11 (8.2%)      11 (8.3%)       36 (9.0%)  
#>     ADVERSE EVENT                       0            0              0               0      
#>     DEATH                               0            0              0               0      
#>     LACK OF EFFICACY                 2 (1.5%)     2 (1.5%)       3 (2.3%)        7 (1.8%)  
#>     PHYSICIAN DECISION               2 (1.5%)     3 (2.2%)       2 (1.5%)        7 (1.8%)  
#>     PROTOCOL VIOLATION               5 (3.7%)     3 (2.2%)       4 (3.0%)       12 (3.0%)  
#>     WITHDRAWAL BY PARENT/GUARDIAN    4 (3.0%)     2 (1.5%)       1 (0.8%)        7 (1.8%)  
#>     WITHDRAWAL BY SUBJECT            1 (0.7%)     1 (0.7%)       1 (0.8%)        3 (0.8%)  
#> 
dst01_3_main(db, lbl_overall = NULL)
#> [[1]]
#>                          A: Drug X    B: Placebo   C: Combination
#>                           (N=134)      (N=134)        (N=132)    
#> —————————————————————————————————————————————————————————————————
#> Completed Treatment      47 (35.1%)   35 (26.1%)     42 (31.8%)  
#> Ongoing Treatment        39 (29.1%)   46 (34.3%)     46 (34.8%)  
#> Discontinued Treatment   48 (35.8%)   53 (39.6%)     44 (33.3%)  
#> 
#> [[2]]
#>                   A: Drug X    B: Placebo   C: Combination
#>                    (N=134)      (N=134)        (N=132)    
#> ——————————————————————————————————————————————————————————
#> Completed Study   68 (50.7%)   66 (49.3%)     73 (55.3%)  
#> 
#> [[3]]
#>                                     A: Drug X    B: Placebo   C: Combination
#>                                      (N=134)      (N=134)        (N=132)    
#> ————————————————————————————————————————————————————————————————————————————
#> Ongoing                             24 (17.9%)   28 (20.9%)     21 (15.9%)  
#> Discontinued Study                  42 (31.3%)   40 (29.9%)     38 (28.8%)  
#>   Safety                            28 (20.9%)   29 (21.6%)     27 (20.5%)  
#>     ADVERSE EVENT                    3 (2.2%)     6 (4.5%)       5 (3.8%)   
#>     DEATH                           25 (18.7%)   23 (17.2%)     22 (16.7%)  
#>     LACK OF EFFICACY                    0            0              0       
#>     PHYSICIAN DECISION                  0            0              0       
#>     PROTOCOL VIOLATION                  0            0              0       
#>     WITHDRAWAL BY PARENT/GUARDIAN       0            0              0       
#>     WITHDRAWAL BY SUBJECT               0            0              0       
#>   Non-Safety                        14 (10.4%)   11 (8.2%)      11 (8.3%)   
#>     ADVERSE EVENT                       0            0              0       
#>     DEATH                               0            0              0       
#>     LACK OF EFFICACY                 2 (1.5%)     2 (1.5%)       3 (2.3%)   
#>     PHYSICIAN DECISION               2 (1.5%)     3 (2.2%)       2 (1.5%)   
#>     PROTOCOL VIOLATION               5 (3.7%)     3 (2.2%)       4 (3.0%)   
#>     WITHDRAWAL BY PARENT/GUARDIAN    4 (3.0%)     2 (1.5%)       1 (0.8%)   
#>     WITHDRAWAL BY SUBJECT            1 (0.7%)     1 (0.7%)       1 (0.8%)   
#> 
run(dst01_3, syn_data)
#>                                       A: Drug X    B: Placebo   C: Combination   All Patients
#>                                        (N=134)      (N=134)        (N=132)         (N=400)   
#>   ———————————————————————————————————————————————————————————————————————————————————————————
#>   Completed Study                     68 (50.7%)   66 (49.3%)     73 (55.3%)     207 (51.7%) 
#>   Ongoing                             24 (17.9%)   28 (20.9%)     21 (15.9%)      73 (18.2%) 
#>   Discontinued Study                  42 (31.3%)   40 (29.9%)     38 (28.8%)     120 (30.0%) 
#>     Safety                            28 (20.9%)   29 (21.6%)     27 (20.5%)      84 (21.0%) 
#>       ADVERSE EVENT                    3 (2.2%)     6 (4.5%)       5 (3.8%)       14 (3.5%)  
#>       DEATH                           25 (18.7%)   23 (17.2%)     22 (16.7%)      70 (17.5%) 
#>     Non-Safety                        14 (10.4%)   11 (8.2%)      11 (8.3%)       36 (9.0%)  
#>       LACK OF EFFICACY                 2 (1.5%)     2 (1.5%)       3 (2.3%)        7 (1.8%)  
#>       PHYSICIAN DECISION               2 (1.5%)     3 (2.2%)       2 (1.5%)        7 (1.8%)  
#>       PROTOCOL VIOLATION               5 (3.7%)     3 (2.2%)       4 (3.0%)       12 (3.0%)  
#>       WITHDRAWAL BY PARENT/GUARDIAN    4 (3.0%)     2 (1.5%)       1 (0.8%)        7 (1.8%)  
#>       WITHDRAWAL BY SUBJECT            1 (0.7%)     1 (0.7%)       1 (0.8%)        3 (0.8%)  
#>   Completed Treatment                 47 (35.1%)   35 (26.1%)     42 (31.8%)     124 (31.0%) 
#>   Ongoing Treatment                   39 (29.1%)   46 (34.3%)     46 (34.8%)     131 (32.8%) 
#>   Discontinued Treatment              48 (35.8%)   53 (39.6%)     44 (33.3%)     145 (36.2%)