EGT01 ECG Assessments and Change from Baseline by Visit Table.

The EGT01 table summarizes several electrocardiogram parameters and their evolution throughout the study.

Usage

egt01_main(
  adam_db,
  arm_var = "ACTARM",
  summaryvars = c("AVAL", "CHG"),
  visitvar = "AVISIT",
  precision = list(),
  default_precision = 2,
  ...
)

egt01_pre(adam_db, ...)

egt01_post(tlg, prune_0 = TRUE, ...)

egt01

Format

An object of class chevron_t of length 1.

Arguments

adam_db

(list of data.frames) object containing the ADaM datasets

arm_var

(string) variable used for column splitting

summaryvars

(character) variables to be analyzed. The label attribute of the corresponding column in adeg table of adam_db is used as label.

visitvar

(string) typically one of "AVISIT" (Default) or "ATPTN" depending on the type of time point to be displayed

precision

(named list of integer) where names are values found in the PARAMCD column and the the values indicate the number of digits that should be represented for min, max and median. Mean and sd are represented with one more decimal of precision.

default_precision

(integer) the default number of digits.

...

not used.

tlg

(TableTree, Listing or ggplot) object typically produced by a main function.

prune_0

(flag) remove 0 count rows

Details

The Analysis Value column, displays the number of patients, the mean, standard deviation, median and range of the analysis value for each visit.

The Change from Baseline column, displays the number of patient and the mean, standard deviation, median and range of changes relative to the baseline.

Remove zero-count rows unless overridden with prune_0 = FALSE.

Split columns by arm, typically ACTARM.

Does not include a total column by default.

Sorted based on factor level; first by PARAM labels in alphabetic order then by chronological time point given by AVISIT. Re-level to customize order.

Functions

egt01_main(): Main TLG function

egt01_pre(): Preprocessing

egt01_post(): Postprocessing

Note

adam_db object must contain an adeg table with a "PARAM" column as well as columns specified in summaryvars and visitvar.

Examples

run(egt01, syn_data, precision = list(
  "HR" = 1,
  "QT" = 1,
  "RR" = 0
))
#>   Parameter                     A: Drug X                          B: Placebo                        C: Combination          
#>     Analysis Visit                      Change from                         Change from                         Change from  
#>                      Value at Visit       Baseline      Value at Visit       Baseline       Value at Visit       Baseline    
#>                          (N=134)          (N=134)           (N=134)           (N=134)           (N=132)           (N=132)    
#>   ———————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————
#>   Heart Rate                                                                                                                 
#>     BASELINE                                                                                                                 
#>       n                    134                                134                                 132                        
#>       Mean (SD)       71.45 (17.93)                      69.83 (20.72)                       69.34 (20.95)                   
#>       Median              72.70                              73.35                               71.96                       
#>       Min - Max        9.1 - 106.9                       13.5 - 115.5                        11.6 - 115.5                    
#>     WEEK 1 DAY 8                                                                                                             
#>       n                    134              134               134               134               132               132      
#>       Mean (SD)       70.19 (20.27)    -1.26 (25.97)     69.43 (20.80)     -0.40 (29.73)     68.65 (18.06)     -0.69 (27.81) 
#>       Median              70.70            -2.20             70.46             -0.52             68.57             -0.76     
#>       Min - Max        8.5 - 127.5      -51.0 - 89.2     16.9 - 129.1      -68.2 - 88.4      16.9 - 115.6      -68.2 - 67.2  
#>     WEEK 2 DAY 15                                                                                                            
#>       n                    134              134               134               134               132               132      
#>       Mean (SD)       70.32 (19.56)    -1.13 (26.26)     71.11 (20.56)     1.28 (31.35)      69.19 (18.06)     -0.15 (28.61) 
#>       Median              70.88            -1.10             72.52             -0.31             69.30             -1.55     
#>       Min - Max       17.1 - 116.3      -85.0 - 67.5      9.2 - 120.5      -73.1 - 81.4      29.6 - 120.5      -67.7 - 66.3  
#>     WEEK 3 DAY 22                                                                                                            
#>       n                    134              134               134               134               132               132      
#>       Mean (SD)       68.25 (20.17)    -3.20 (28.66)     68.95 (20.56)     -0.88 (30.14)     70.88 (20.28)     1.54 (27.28)  
#>       Median              68.79            -2.17             67.80             1.19              70.12             0.14      
#>       Min - Max       13.3 - 131.7      -81.2 - 72.6     24.0 - 130.4      -73.0 - 103.3     20.9 - 116.8      -64.5 - 72.5  
#>     WEEK 4 DAY 29                                                                                                            
#>       n                    134              134               134               134               132               132      
#>       Mean (SD)       70.06 (18.41)    -1.38 (26.59)     71.60 (20.34)     1.77 (30.26)      71.02 (20.59)     1.68 (29.60)  
#>       Median              69.33            -4.54             72.21             -0.88             70.52             3.81      
#>       Min - Max       22.3 - 116.5      -58.1 - 77.3     17.5 - 129.1      -64.3 - 85.3      10.4 - 117.3      -65.9 - 105.7 
#>     WEEK 5 DAY 36                                                                                                            
#>       n                    134              134               134               134               132               132      
#>       Mean (SD)       66.41 (19.74)    -5.04 (27.11)     71.25 (18.92)     1.42 (27.08)      71.26 (18.56)     1.92 (29.59)  
#>       Median              65.40            -6.83             70.86             0.64              73.29             1.74      
#>       Min - Max       23.9 - 110.4      -73.3 - 57.2     25.9 - 125.7      -71.2 - 67.4      23.9 - 117.2      -58.2 - 68.9  
#>   QT Duration                                                                                                                
#>     BASELINE                                                                                                                 
#>       n                    134                                134                                 132                        
#>       Mean (SD)      336.83 (117.96)                    351.00 (98.44)                      352.60 (105.11)                  
#>       Median             344.41                             351.48                              347.96                       
#>       Min - Max       87.0 - 665.4                       98.9 - 641.9                        105.7 - 628.1                   
#>     WEEK 1 DAY 8                                                                                                             
#>       n                    134              134               134               134               132               132      
#>       Mean (SD)      342.57 (101.09)   5.74 (159.94)    363.95 (102.34)   12.94 (140.73)    359.42 (105.16)    6.83 (147.13) 
#>       Median             347.26            -1.78            356.54             13.73            363.84             8.69      
#>       Min - Max       91.6 - 591.4     -346.4 - 452.8    114.9 - 656.5    -317.5 - 416.4     51.9 - 611.9     -473.2 - 358.0 
#>     WEEK 2 DAY 15                                                                                                            
#>       n                    134              134               134               134               132               132      
#>       Mean (SD)      353.24 (93.93)    16.41 (162.10)   345.93 (96.78)    -5.08 (152.51)    335.69 (98.61)    -16.91 (138.45)
#>       Median             351.10            16.72            346.83             -9.68            320.96            -22.27     
#>       Min - Max       138.0 - 587.3    -414.1 - 389.2    146.4 - 556.1    -440.3 - 364.8     104.9 - 562.3    -326.5 - 325.3 
#>     WEEK 3 DAY 22                                                                                                            
#>       n                    134              134               134               134               132               132      
#>       Mean (SD)      370.12 (101.57)   33.28 (152.78)   343.41 (102.58)   -7.59 (138.80)    346.95 (94.96)    -5.65 (144.65) 
#>       Median             378.21            39.59            335.12            -15.89            352.15             9.11      
#>       Min - Max       118.1 - 615.2    -391.7 - 520.1    63.4 - 566.5     -311.3 - 293.8     126.1 - 580.8    -412.1 - 410.0 
#>     WEEK 4 DAY 29                                                                                                            
#>       n                    134              134               134               134               132               132      
#>       Mean (SD)      345.77 (95.97)    8.94 (145.65)    354.38 (108.29)    3.37 (142.39)    341.36 (106.75)   -11.24 (145.65)
#>       Median             340.13             5.58            346.70            -17.13            352.30            -11.87     
#>       Min - Max       110.1 - 616.6    -393.3 - 456.0    80.8 - 687.7     -439.9 - 364.8      4.9 - 570.6     -480.9 - 330.7 
#>     WEEK 5 DAY 36                                                                                                            
#>       n                    134              134               134               134               132               132      
#>       Mean (SD)      358.90 (97.69)    22.07 (155.68)   338.65 (99.26)    -12.36 (130.46)   349.30 (95.77)    -3.30 (136.00) 
#>       Median             351.96             5.89            344.17             0.40             350.32             12.68     
#>       Min - Max       88.4 - 661.1     -353.3 - 539.8    31.3 - 563.9     -338.9 - 352.7     119.0 - 581.8    -311.5 - 295.5 
#>   RR Duration                                                                                                                
#>     BASELINE                                                                                                                 
#>       n                    134                                134                                 132                        
#>       Mean (SD)      1028.4 (286.4)                     1027.5 (324.0)                      1074.6 (277.3)                   
#>       Median             1041.9                             1047.3                              1080.9                       
#>       Min - Max         34 - 1784                          5 - 1877                           290 - 1617                     
#>     WEEK 1 DAY 8                                                                                                             
#>       n                    134              134               134               134               132               132      
#>       Mean (SD)       990.2 (318.7)    -38.3 (446.4)    1061.6 (269.1)     34.1 (401.8)     1039.6 (284.4)     -35.0 (406.0) 
#>       Median              963.9            -86.5            1061.5             66.9             1014.8             -97.8     
#>       Min - Max        111 - 2015       -1015 - 1389      277 - 1712        -942 - 994        498 - 1937        -959 - 909   
#>     WEEK 2 DAY 15                                                                                                            
#>       n                    134              134               134               134               132               132      
#>       Mean (SD)      1013.4 (304.1)    -15.0 (425.8)    1109.4 (318.2)     81.9 (453.1)     1045.2 (277.5)     -29.4 (380.8) 
#>       Median             1040.7             40.5            1117.8             69.3             1034.6             -15.3     
#>       Min - Max        164 - 1677       -1216 - 1053      161 - 2049       -1127 - 1149       253 - 1722        -859 - 872   
#>     WEEK 3 DAY 22                                                                                                            
#>       n                    134              134               134               134               132               132      
#>       Mean (SD)      1014.0 (305.0)    -14.4 (406.5)    1118.3 (296.5)     90.9 (409.6)     1036.6 (268.5)     -38.0 (421.5) 
#>       Median             1027.2            -50.5            1134.6             52.0             1030.7             -60.7     
#>       Min - Max        357 - 1799       -883 - 1080        97 - 1825        -887 - 1166       446 - 1713        -985 - 902   
#>     WEEK 4 DAY 29                                                                                                            
#>       n                    134              134               134               134               132               132      
#>       Mean (SD)      1033.3 (313.9)     4.9 (483.2)     1079.8 (298.9)     52.3 (429.3)     1027.5 (295.9)     -47.1 (378.2) 
#>       Median             1050.4             5.6             1051.6             69.6             1023.4             -49.5     
#>       Min - Max         54 - 1979       -1346 - 1195      348 - 1762        -826 - 1192       342 - 2145       -1002 - 1049  
#>     WEEK 5 DAY 36                                                                                                            
#>       n                    134              134               134               134               132               132      
#>       Mean (SD)      1072.3 (282.9)     43.9 (379.0)    1058.5 (271.0)     31.0 (436.0)     1029.0 (271.2)     -45.6 (405.5) 
#>       Median             1067.3             55.9            1068.9             33.6             1065.3             -34.4     
#>       Min - Max        353 - 2001       -1029 - 1419      209 - 1795        -979 - 1365       436 - 1794        -962 - 1330