EGT01 ECG Parameters and Change from Baseline By Visit Table.

The EGT01 table provides an overview of the ECG values and its change from baseline of each respective arm over the course of the trial.

Usage

egt01_main(
  adam_db,
  dataset = "adeg",
  arm_var = "ACTARM",
  row_split_var = NULL,
  summaryvars = c("AVAL", "CHG"),
  visitvar = "AVISIT",
  precision = list(),
  default_precision = 2,
  page_by = TRUE,
  ...
)

egt01

Format

An object of class chevron_t of length 1.

Source

cfbt01.R

Arguments

adam_db

(list of data.frames) object containing the ADaM datasets

dataset

(string) the name of a table in the adam_db object.

arm_var

(string) variable used for column splitting

row_split_var

(character) row split variable other than PARAMCD.

summaryvars

(character) variables to be analyzed. The label attribute of the corresponding column in table of adam_db is used as label.

visitvar

(string) typically one of "AVISIT" (Default) or "ATPTN" depending on the type of time point to be displayed

precision

(named list of integer) where names are values found in the PARAMCD column and the the values indicate the number of digits that should be represented for min, max and median. Mean and sd are represented with one more decimal of precision.

default_precision

(integer) the default number of digits.

page_by

(flag) indicator whether the parameter row split is by page.

...

not used.

Details

The Analysis Value column, displays the number of patients, the mean, standard deviation, median and range of the analysis value for each visit.

The Change from Baseline column, displays the number of patient and the mean, standard deviation, median and range of changes relative to the baseline.

Remove zero-count rows unless overridden with prune_0 = FALSE.

Split columns by arm, typically ACTARM.

Does not include a total column by default.

Sorted based on factor level; first by PARAM labels in alphabetic order then by chronological time point given by AVISIT. Re-level to customize order

Functions

egt01_main(): Main TLG function

Note

adam_db object must contain table named as dataset with the columns specified in summaryvars.

Examples

run(egt01, syn_data)
#>     Analysis Visit                  A: Drug X                                B: Placebo                              C: Combination             
#>                                              Change from                               Change from                               Change from    
#>                        Value at Visit          Baseline          Value at Visit          Baseline          Value at Visit          Baseline     
#>                           (N=134)              (N=134)              (N=134)              (N=134)              (N=132)              (N=132)      
#>   ——————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————
#>   Heart Rate                                                                                                                                    
#>     BASELINE                                                                                                                                    
#>       n                     134                                       134                                       132                             
#>       Mean (SD)       71.447 (17.932)                           69.829 (20.717)                           69.341 (20.948)                       
#>       Median               72.697                                    73.352                                    71.956                           
#>       Min - Max        9.09 - 106.91                             13.49 - 115.52                            11.63 - 115.49                       
#>     WEEK 1 DAY 8                                                                                                                                
#>       n                     134                  134                  134                  134                  132                  132        
#>       Mean (SD)       70.191 (20.274)      -1.257 (25.966)      69.427 (20.799)      -0.401 (29.735)      68.654 (18.064)      -0.687 (27.807)  
#>       Median               70.703               -2.197               70.462               -0.520               68.573               -0.760      
#>       Min - Max        8.53 - 127.50        -50.97 - 89.16       16.85 - 129.14       -68.18 - 88.38       16.87 - 115.60       -68.25 - 67.20  
#>     WEEK 2 DAY 15                                                                                                                               
#>       n                     134                  134                  134                  134                  132                  132        
#>       Mean (SD)       70.317 (19.561)      -1.130 (26.262)      71.108 (20.556)       1.279 (31.353)      69.193 (18.056)      -0.148 (28.606)  
#>       Median               70.885               -1.105               72.524               -0.311               69.304               -1.548      
#>       Min - Max        17.14 - 116.32       -85.03 - 67.52       9.22 - 120.54        -73.07 - 81.44       29.62 - 120.50       -67.75 - 66.29  
#>     WEEK 3 DAY 22                                                                                                                               
#>       n                     134                  134                  134                  134                  132                  132        
#>       Mean (SD)       68.251 (20.168)      -3.196 (28.657)      68.947 (20.558)      -0.881 (30.142)      70.884 (20.280)       1.543 (27.285)  
#>       Median               68.787               -2.167               67.802               1.195                70.124               0.139       
#>       Min - Max        13.33 - 131.73       -81.20 - 72.57       23.98 - 130.41      -73.03 - 103.31       20.91 - 116.79       -64.51 - 72.52  
#>     WEEK 4 DAY 29                                                                                                                               
#>       n                     134                  134                  134                  134                  132                  132        
#>       Mean (SD)       70.063 (18.410)      -1.385 (26.595)      71.601 (20.336)       1.772 (30.262)      71.020 (20.591)       1.679 (29.597)  
#>       Median               69.330               -4.535               72.211               -0.884               70.523               3.811       
#>       Min - Max        22.30 - 116.51       -58.07 - 77.32       17.53 - 129.06       -64.32 - 85.34       10.35 - 117.30      -65.91 - 105.67  
#>     WEEK 5 DAY 36                                                                                                                               
#>       n                     134                  134                  134                  134                  132                  132        
#>       Mean (SD)       66.406 (19.745)      -5.041 (27.112)      71.252 (18.923)       1.423 (27.084)      71.257 (18.563)       1.916 (29.591)  
#>       Median               65.405               -6.827               70.858               0.637                73.292               1.735       
#>       Min - Max        23.89 - 110.38       -73.26 - 57.24       25.86 - 125.73       -71.15 - 67.37       23.89 - 117.19       -58.23 - 68.86  
#>   QT Duration                                                                                                                                   
#>     BASELINE                                                                                                                                    
#>       n                     134                                       134                                       132                             
#>       Mean (SD)      336.834 (117.958)                          351.004 (98.436)                         352.598 (105.114)                      
#>       Median              344.414                                   351.480                                   347.956                           
#>       Min - Max        86.98 - 665.40                            98.90 - 641.92                           105.68 - 628.14                       
#>     WEEK 1 DAY 8                                                                                                                                
#>       n                     134                  134                  134                  134                  132                  132        
#>       Mean (SD)      342.574 (101.093)     5.741 (159.936)     363.946 (102.338)     12.943 (140.729)    359.424 (105.161)     6.826 (147.129)  
#>       Median              347.264               -1.779              356.539               13.730              363.838               8.685       
#>       Min - Max        91.63 - 591.42      -346.44 - 452.75     114.92 - 656.45      -317.53 - 416.35      51.91 - 611.88      -473.19 - 358.00 
#>     WEEK 2 DAY 15                                                                                                                               
#>       n                     134                  134                  134                  134                  132                  132        
#>       Mean (SD)       353.244 (93.926)     16.410 (162.103)     345.926 (96.783)     -5.078 (152.509)     335.689 (98.605)    -16.909 (138.448) 
#>       Median              351.099               16.722              346.831               -9.681              320.957              -22.267      
#>       Min - Max       138.01 - 587.30      -414.07 - 389.16     146.42 - 556.07      -440.28 - 364.76     104.91 - 562.34      -326.55 - 325.27 
#>     WEEK 3 DAY 22                                                                                                                               
#>       n                     134                  134                  134                  134                  132                  132        
#>       Mean (SD)      370.115 (101.568)     33.282 (152.780)    343.412 (102.585)     -7.592 (138.803)     346.947 (94.965)     -5.652 (144.651) 
#>       Median              378.209               39.586              335.119              -15.888              352.154               9.113       
#>       Min - Max       118.14 - 615.18      -391.72 - 520.09      63.37 - 566.51      -311.28 - 293.76     126.09 - 580.81      -412.11 - 410.01 
#>     WEEK 4 DAY 29                                                                                                                               
#>       n                     134                  134                  134                  134                  132                  132        
#>       Mean (SD)       345.774 (95.970)     8.941 (145.647)     354.377 (108.286)     3.374 (142.390)     341.358 (106.745)    -11.240 (145.648) 
#>       Median              340.134               5.576               346.699              -17.129              352.295              -11.868      
#>       Min - Max       110.12 - 616.58      -393.34 - 456.04      80.82 - 687.69      -439.90 - 364.80      4.95 - 570.61       -480.94 - 330.67 
#>     WEEK 5 DAY 36                                                                                                                               
#>       n                     134                  134                  134                  134                  132                  132        
#>       Mean (SD)       358.903 (97.693)     22.069 (155.681)     338.648 (99.260)    -12.356 (130.457)     349.303 (95.769)     -3.296 (135.997) 
#>       Median              351.959               5.886               344.167               0.405               350.319               12.680      
#>       Min - Max        88.38 - 661.12      -353.30 - 539.84      31.25 - 563.90      -338.85 - 352.75     119.02 - 581.83      -311.45 - 295.53 
#>   RR Duration                                                                                                                                   
#>     BASELINE                                                                                                                                    
#>       n                     134                                       134                                       132                             
#>       Mean (SD)      1028.419 (286.385)                        1027.484 (324.000)                        1074.623 (277.328)                     
#>       Median              1041.860                                  1047.268                                  1080.866                          
#>       Min - Max       34.33 - 1783.71                            5.29 - 1877.19                           289.60 - 1617.06                      
#>     WEEK 1 DAY 8                                                                                                                                
#>       n                     134                  134                  134                  134                  132                  132        
#>       Mean (SD)      990.155 (318.736)    -38.264 (446.403)    1061.587 (269.138)    34.103 (401.759)    1039.642 (284.389)   -34.981 (406.037) 
#>       Median              963.881              -86.501              1061.504              66.914              1014.780             -97.826      
#>       Min - Max       110.82 - 2014.56    -1014.82 - 1389.40    276.83 - 1711.99     -942.16 - 993.82     498.21 - 1937.47     -958.61 - 908.90 
#>     WEEK 2 DAY 15                                                                                                                               
#>       n                     134                  134                  134                  134                  132                  132        
#>       Mean (SD)      1013.371 (304.135)   -15.047 (425.807)    1109.402 (318.227)    81.918 (453.105)    1045.209 (277.506)   -29.414 (380.841) 
#>       Median              1040.693              40.485              1117.764              69.275              1034.637             -15.331      
#>       Min - Max       164.19 - 1677.10    -1216.10 - 1053.15    160.73 - 2048.73    -1127.10 - 1148.61    252.84 - 1722.33     -859.27 - 871.54 
#>     WEEK 3 DAY 22                                                                                                                               
#>       n                     134                  134                  134                  134                  132                  132        
#>       Mean (SD)      1013.987 (304.990)   -14.431 (406.529)    1118.345 (296.527)    90.861 (409.573)    1036.593 (268.521)   -38.030 (421.484) 
#>       Median              1027.232             -50.543              1134.596              52.029              1030.717             -60.720      
#>       Min - Max       357.04 - 1798.65    -882.94 - 1080.26     97.14 - 1825.43     -887.06 - 1166.15     446.02 - 1713.38     -984.79 - 902.37 
#>     WEEK 4 DAY 29                                                                                                                               
#>       n                     134                  134                  134                  134                  132                  132        
#>       Mean (SD)      1033.309 (313.919)    4.890 (483.153)     1079.762 (298.940)    52.278 (429.253)    1027.549 (295.856)   -47.074 (378.235) 
#>       Median              1050.399              5.601               1051.613              69.554              1023.360             -49.519      
#>       Min - Max       54.33 - 1979.43     -1345.93 - 1195.41    347.94 - 1762.04    -826.43 - 1191.83     341.62 - 2144.86    -1002.03 - 1048.66
#>     WEEK 5 DAY 36                                                                                                                               
#>       n                     134                  134                  134                  134                  132                  132        
#>       Mean (SD)      1072.335 (282.901)    43.916 (379.048)    1058.452 (271.012)    30.968 (436.030)    1029.043 (271.211)   -45.580 (405.528) 
#>       Median              1067.330              55.872              1068.950              33.615              1065.256             -34.395      
#>       Min - Max       352.97 - 2000.56    -1028.79 - 1418.57    208.83 - 1794.73    -978.97 - 1365.13     436.28 - 1794.07    -962.18 - 1329.88