EGT01 ECG Parameters and Change from Baseline By Visit Table.

The EGT01 table provides an overview of the ECG values and its change from baseline of each respective arm over the course of the trial.

Usage

egt01_main(
  adam_db,
  dataset = "adeg",
  arm_var = "ACTARM",
  lbl_overall = NULL,
  row_split_var = NULL,
  summaryvars = c("AVAL", "CHG"),
  visitvar = "AVISIT",
  precision = list(default = 2L),
  page_var = "PARAMCD",
  .stats = c("n", "mean_sd", "median", "range"),
  skip = list(CHG = "BASELINE"),
  ...
)

egt01_pre(adam_db, dataset = "adeg", ...)

egt01

Format

An object of class chevron_t of length 1.

Arguments

adam_db

(list of data.frames) object containing the ADaM datasets

dataset

(string) the name of a table in the adam_db object.

arm_var

(string) variable used for column splitting

lbl_overall

(string) label used for overall column, if set to NULL the overall column is omitted

row_split_var

(character) additional row split variables.

summaryvars

(character) variables to be analyzed. The label attribute of the corresponding column in table of adam_db is used as label.

visitvar

(string) typically one of "AVISIT" or user-defined visit incorporating "ATPT".

precision

(named list of integer) where names are values found in the PARAMCD column and the values indicate the number of digits in statistics. If default is set, and parameter precision not specified, the value for default will be used.

page_var

(string) variable name prior to which the row split is by page.

.stats

(character) statistics names, see tern::analyze_vars().

skip

Named (list) of visit values that need to be inhibited.

...

additional arguments like .indent_mods, .labels.

Value

the main function returns an rtables object.

the preprocessing function returns a list of data.frame.

Details

The Analysis Value column, displays the number of patients, the mean, standard deviation, median and range of the analysis value for each visit.

The Change from Baseline column, displays the number of patient and the mean, standard deviation, median and range of changes relative to the baseline.

Remove zero-count rows unless overridden with prune_0 = FALSE.

Split columns by arm, typically ACTARM.

Does not include a total column by default.

Sorted based on factor level; first by PARAM labels in alphabetic order then by chronological time point given by AVISIT. Re-level to customize order

Functions

egt01_main(): Main TLG function

egt01_pre(): Preprocessing

Note

adam_db object must contain table named as dataset with the columns specified in summaryvars.

Examples

run(egt01, syn_data)
#>                                    A: Drug X                                B: Placebo                             C: Combination             
#>                                             Change from                               Change from                              Change from    
#>                       Value at Visit          Baseline          Value at Visit         Baseline          Value at Visit          Baseline     
#>   Analysis Visit          (N=15)               (N=15)               (N=15)              (N=15)               (N=15)               (N=15)      
#>   ————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————
#>   Heart Rate                                                                                                                                  
#>     BASELINE                                                                                                                                  
#>       n                     15                                        15                                       15                             
#>       Mean (SD)      76.594 (17.889)                           69.899 (18.788)                          70.492 (18.175)                       
#>       Median              77.531                                    77.174                                   74.111                           
#>       Min - Max       46.50 - 106.68                            26.42 - 97.69                            45.37 - 115.49                       
#>     WEEK 1 DAY 8                                                                                                                              
#>       n                     15                   15                   15                  15                   15                   15        
#>       Mean (SD)      71.140 (23.441)      -5.454 (25.128)      70.958 (14.877)      1.059 (23.345)      67.450 (18.932)      -3.043 (23.753)  
#>       Median              77.210               -2.152               70.033              -8.403               68.471               0.181       
#>       Min - Max       8.53 - 102.63        -50.97 - 36.54       44.85 - 93.79       -25.34 - 60.50       38.90 - 100.05       -52.20 - 33.13  
#>     WEEK 2 DAY 15                                                                                                                             
#>       n                     15                   15                   15                  15                   15                   15        
#>       Mean (SD)      69.350 (16.083)      -7.244 (28.960)      76.096 (14.958)      6.198 (29.319)      63.694 (12.920)      -6.799 (23.949)  
#>       Median              65.746              -11.369               75.323               0.255               61.076               -4.954      
#>       Min - Max       47.22 - 101.44       -49.59 - 42.91       47.50 - 111.40      -37.51 - 69.34       43.25 - 86.13        -52.70 - 40.76  
#>     WEEK 3 DAY 22                                                                                                                             
#>       n                     15                   15                   15                  15                   15                   15        
#>       Mean (SD)      73.894 (24.576)      -2.700 (32.079)      67.635 (19.114)      -2.263 (29.989)     72.054 (19.308)       1.562 (27.494)  
#>       Median              69.296               5.492                68.468              -2.093               68.686               -5.848      
#>       Min - Max       44.15 - 131.73       -62.53 - 38.19       31.89 - 108.87      -52.26 - 66.81       32.16 - 109.86       -49.61 - 35.23  
#>     WEEK 4 DAY 29                                                                                                                             
#>       n                     15                   15                   15                  15                   15                   15        
#>       Mean (SD)      73.241 (19.256)      -3.353 (29.170)      66.524 (25.487)      -3.374 (36.024)     66.600 (22.839)      -3.892 (24.140)  
#>       Median              68.689               0.232                66.397              -11.730              64.969               -6.827      
#>       Min - Max       33.71 - 111.54       -55.14 - 65.04       19.66 - 111.29      -60.39 - 61.00       10.35 - 100.88       -50.72 - 26.77  
#>     WEEK 5 DAY 36                                                                                                                             
#>       n                     15                   15                   15                  15                   15                   15        
#>       Mean (SD)      61.690 (22.182)      -14.904 (30.330)     60.712 (20.025)      -9.187 (24.587)     72.683 (23.495)       2.191 (26.654)  
#>       Median              57.925              -12.660               60.454              -16.100              77.585               14.635      
#>       Min - Max       23.89 - 103.74       -60.00 - 57.24       32.53 - 102.02      -52.56 - 50.96       31.21 - 105.05       -42.90 - 34.64  
#>   QT Duration                                                                                                                                 
#>     BASELINE                                                                                                                                  
#>       n                     15                                        15                                       15                             
#>       Mean (SD)     335.294 (123.231)                          363.104 (68.160)                         347.311 (86.236)                      
#>       Median             372.731                                   386.316                                  348.254                           
#>       Min - Max      121.28 - 554.97                           214.65 - 445.53                          170.80 - 508.54                       
#>     WEEK 1 DAY 8                                                                                                                              
#>       n                     15                   15                   15                  15                   15                   15        
#>       Mean (SD)      357.361 (85.688)     22.067 (144.166)    415.225 (105.425)    52.121 (144.259)    321.078 (107.553)    -26.233 (129.135) 
#>       Median             344.797               49.432              421.950              62.762              307.962              -17.006      
#>       Min - Max      241.22 - 517.39      -207.23 - 245.36     234.11 - 604.72     -190.70 - 364.94     118.36 - 480.29      -363.11 - 163.67 
#>     WEEK 2 DAY 15                                                                                                                             
#>       n                     15                   15                   15                  15                   15                   15        
#>       Mean (SD)     344.883 (106.793)     9.589 (174.797)      370.548 (80.862)     7.444 (91.301)      354.129 (95.133)     6.818 (142.397)  
#>       Median             312.236               -9.264              388.515              -9.429              365.292               39.930      
#>       Min - Max      187.77 - 501.87      -278.91 - 372.71     204.55 - 514.43     -190.58 - 173.87     200.19 - 493.40      -279.46 - 265.56 
#>     WEEK 3 DAY 22                                                                                                                             
#>       n                     15                   15                   15                  15                   15                   15        
#>       Mean (SD)      342.062 (92.568)     6.768 (151.505)     326.684 (116.421)    -36.420 (145.415)    366.245 (99.106)     18.935 (168.417) 
#>       Median             352.930              -22.771              298.353              -78.409             329.688              -21.584      
#>       Min - Max      199.40 - 476.04      -230.25 - 303.00     151.05 - 561.23     -205.30 - 293.76     249.42 - 580.81      -252.73 - 410.01 
#>     WEEK 4 DAY 29                                                                                                                             
#>       n                     15                   15                   15                  15                   15                   15        
#>       Mean (SD)      371.650 (44.805)     36.356 (139.308)    333.697 (110.377)    -29.407 (125.592)    333.181 (96.466)    -14.130 (107.622) 
#>       Median             375.412               58.958              308.020              -40.987             330.911              -25.820      
#>       Min - Max      302.32 - 451.62      -214.07 - 258.04     183.09 - 531.08     -241.72 - 134.12     126.95 - 488.57      -234.92 - 152.49 
#>     WEEK 5 DAY 36                                                                                                                             
#>       n                     15                   15                   15                  15                   15                   15        
#>       Mean (SD)     345.504 (130.543)     10.210 (198.224)     309.919 (84.624)    -53.185 (105.730)    322.931 (67.801)    -24.380 (117.331) 
#>       Median             355.730              -23.213              306.219              -12.373             341.988              -26.952      
#>       Min - Max       88.38 - 661.12      -271.06 - 539.84     189.01 - 448.58      -256.52 - 91.57     217.51 - 427.16      -291.03 - 171.19 
#>   RR Duration                                                                                                                                 
#>     BASELINE                                                                                                                                  
#>       n                     15                                        15                                       15                             
#>       Mean (SD)     1086.908 (363.811)                        1050.034 (390.444)                       1102.659 (310.359)                     
#>       Median             1116.849                                  1089.193                                 1250.037                          
#>       Min - Max      626.19 - 1653.12                          414.61 - 1721.89                         385.51 - 1430.81                      
#>     WEEK 1 DAY 8                                                                                                                              
#>       n                     15                   15                   15                  15                   15                   15        
#>       Mean (SD)     968.499 (287.811)    -118.409 (546.796)   1041.186 (211.201)   -8.848 (435.281)    948.491 (213.746)    -154.168 (442.882)
#>       Median             961.296              -147.460             1013.786             24.754              965.429              -224.054     
#>       Min - Max      358.92 - 1593.51    -1014.82 - 911.82     714.44 - 1417.52    -618.80 - 847.31     513.35 - 1229.09     -736.69 - 843.58 
#>     WEEK 2 DAY 15                                                                                                                             
#>       n                     15                   15                   15                  15                   15                   15        
#>       Mean (SD)     932.717 (259.634)    -154.191 (331.884)   1139.332 (454.231)   89.298 (582.750)    1021.283 (233.529)   -81.376 (415.781) 
#>       Median             950.533              -205.949             1068.007             -5.449              964.616              -142.180     
#>       Min - Max      409.68 - 1269.35     -649.69 - 473.09     486.51 - 2048.73    -846.72 - 1148.61    667.36 - 1367.25     -647.47 - 616.15 
#>     WEEK 3 DAY 22                                                                                                                             
#>       n                     15                   15                   15                  15                   15                   15        
#>       Mean (SD)     1068.865 (319.540)   -18.043 (513.412)    1110.882 (259.523)   60.848 (432.700)    1105.918 (306.185)    3.259 (516.734)  
#>       Median             1201.998             -65.085              1163.690             51.200              1187.130              30.318      
#>       Min - Max      380.49 - 1551.65     -832.86 - 703.74     621.41 - 1453.29    -887.06 - 822.18     446.02 - 1648.32     -984.79 - 816.30 
#>     WEEK 4 DAY 29                                                                                                                             
#>       n                     15                   15                   15                  15                   15                   15        
#>       Mean (SD)     1087.915 (205.940)    1.008 (403.039)     1161.681 (293.257)   111.647 (460.979)   992.134 (283.177)    -110.525 (334.932)
#>       Median             1084.658             146.611              1055.223             191.008             1028.997             -112.599     
#>       Min - Max      697.59 - 1499.17     -801.16 - 402.97     722.35 - 1762.04    -528.27 - 1191.83    497.14 - 1382.12     -597.95 - 757.99 
#>     WEEK 5 DAY 36                                                                                                                             
#>       n                     15                   15                   15                  15                   15                   15        
#>       Mean (SD)     1016.880 (424.428)   -70.027 (505.078)    1135.131 (224.684)   85.097 (497.679)    1089.527 (238.909)   -13.132 (362.606) 
#>       Median             962.584              -142.925             1158.815             -9.553              1081.015              16.706      
#>       Min - Max      352.97 - 1843.86    -894.83 - 1162.79     714.34 - 1436.68    -843.41 - 992.34     699.72 - 1611.38     -696.03 - 561.53