EXT01 Table 1 (Default) Exposure Summary Table

The EXT01 table provides an overview of the of the exposure of the patients in terms of Total dose administered or missed, and treatment duration.

ext01_1(
  adam_db,
  armvar = .study$armvar,
  summaryvars = "AVAL",
  lbl_overall = .study$lbl_overall,
  prune_0 = TRUE,
  deco = std_deco("EXT01"),
  .study = list(armvar = "ACTARM", lbl_overall = NULL)
)

ext01_1_lyt(
  armvar = .study$armvar,
  summaryvars = .study$analysis_var,
  summaryvars_lbls = .study$lbl_analysis_var,
  lbl_overall = .study$lbl_overall,
  deco = std_deco("EXT01"),
  .study = list(armvar = "ACTARM", lbl_overall = NULL, analysis_var = "AVAL",
    lbl_analysis_var = "Analysis Value")
)

Arguments

adam_db

(dm) object containing the ADaM datasets

armvar

(character) variable used for column splitting

summaryvars

(string) the name of the variable to be analyzed. By default "AVAL".

lbl_overall

(character) label used for overall column, if set to NULL the overall column is omitted

prune_0

(logical) remove 0 count rows

deco

(character) decoration with title, subtitles and main_footer content

.study

(list) with default values for the arguments of the function

summaryvars_lbls

(string) the label associated with the analyzed variable.

Details

Default Exposure table

The n row provides the number of non-missing values. The percentages for categorical variables is based on n. The percentages for Total number of patients with at least one dose modification are based on the number of patients in the corresponding analysis population given by N.

Split columns by arm, typically ACTARM.

Does not include a total column by default.

Sorted by alphabetic order of the PARAM value. Transform to factor and re-level for custom order.

ANL01FL is not relevant subset

Functions

ext01_1_lyt:

Examples

library(dm)

db <- syn_test_data() %>%
  preprocess_data("ext01_1")

ext01_1(db)
#> Study Drug Exposure:  {Specify Population}
#> Protocol: {{protocol}}, Snapshot: {{snapshot}}, Snapshot Date: {{snapshot-date}}, Cutoff Date: {{cutoff-date}}
#> 
#> —————————————————————————————————————————————————————————————————
#>                           A: Drug X   B: Placebo   C: Combination
#>                            (N=134)     (N=134)        (N=132)    
#> —————————————————————————————————————————————————————————————————
#> Total dose administered                                          
#>   n                          134         134            132      
#>   Mean (SD)                 7 (0)       7 (0)          7 (0)     
#>   Median                      7           7              7       
#>   Min - Max                 7 - 7       7 - 7          7 - 7     
#> —————————————————————————————————————————————————————————————————
#> 
#> Treatment duration is the date of the last study drug administration minus the date of the first study drug administration plus one day.  Dose intensity is {definition}. <Percentages for reasons for dose modification are based on total number of patients with at least one dose modification.>