EXT01 Table 1 (Default) Exposure Summary Table.

The EXT01 table provides an overview of the of the exposure of the patients in terms of Total dose administered or missed, and treatment duration.

Usage

ext01_1_main(
  adam_db,
  arm_var = "ACTARM",
  summaryvars = "AVAL",
  lbl_overall = NULL,
  deco = std_deco("EXT01"),
  ...
)

ext01_1_lyt(arm_var, summaryvars, summaryvars_lbls, lbl_overall, deco)

ext01_1_pre(
  adam_db,
  paramcd_order = c("TNDOSE", "DOSE", "NDOSE", "TDOSE"),
  ...
)

ext01_1_post(tlg, prune_0 = TRUE, ...)

ext01_1

Format

An object of class chevron_t of length 1.

Arguments

adam_db

(dm) object containing the ADaM datasets

arm_var

(character) variable used for column splitting

summaryvars

(string) the name of the variable to be analyzed. By default "AVAL".

lbl_overall

(character) label used for overall column, if set to NULL the overall column is omitted

deco

(character) decoration with title, subtitles and main_footer content

...

not used.

summaryvars_lbls

(string) the label associated with the analyzed variable.

paramcd_order

(character) providing the PARAMCD values in the desired order.

tlg

(TableTree, Listing or ggplot) object typically produced by a main function.

prune_0

(logical) remove 0 count rows

Details

Default Exposure table

The n row provides the number of non-missing values. The percentages for categorical variables is based on n. The percentages for Total number of patients with at least one dose modification are based on the number of patients in the corresponding analysis population given by N.

Split columns by arm, typically ACTARM.

Does not include a total column by default.

Sorted by alphabetic order of the PARAM value. Transform to factor and re-level for custom order.

ANL01FL is not relevant subset.

Functions

ext01_1_main(): Main TLG function

ext01_1_lyt(): Layout

ext01_1_pre(): Preprocessing

ext01_1_post(): Postprocessing

Note

adam_db object must contain an adex table with columns specified in summaryvars.

Examples

run(ext01_1, syn_data)
#>                                           A: Drug X        B: Placebo      C: Combination 
#>                                            (N=134)           (N=134)           (N=132)    
#>   ————————————————————————————————————————————————————————————————————————————————————————
#>   Total number of doses administered                                                      
#>     n                                        134               134               132      
#>     Mean (SD)                             7.0 (0.0)         7.0 (0.0)         7.0 (0.0)   
#>     Median                                   7.0               7.0               7.0      
#>     Min - Max                             7.0 - 7.0         7.0 - 7.0         7.0 - 7.0   
#>   Total dose administered                                                                 
#>     n                                        134               134               132      
#>     Mean (SD)                          6655.5 (1209.6)   6401.2 (1269.9)   6776.4 (1212.5)
#>     Median                                 6720.0            6360.0            6720.0     
#>     Min - Max                          4320.0 - 9360.0   4080.0 - 9360.0   4320.0 - 9360.0