General Argument Name Convention

General Argument Name Convention

gen_args(
  armvar,
  lbl_overall,
  prune_0,
  deco,
  .study,
  ad_bl,
  adsl,
  adeg,
  adae,
  adcm,
  admh,
  aag,
  advs,
  adlb,
  adex,
  adsub,
  adrs,
  adtte,
  adhy,
  adsaftte,
  addv,
  adcssrs,
  adqlqc,
  acg,
  adic,
  group
)

Arguments

armvar
(character) variable used for column splitting

lbl_overall
(character) label used for overall column, if set to NULL the overall column is omitted

prune_0
(logical) remove 0 count rows

deco
(character) decoration with title, subtitles and main_footer content

.study
(list) with default values for the arguments of the function

ad_bl
(data.frame) subject baseline Dataset with one row per patient

adsl
(data.frame) Subject Level Analysis Dataset

adeg
(data.frame) ECG Analysis Dataset

adae
(data.frame) Adverse Events Analysis Dataset

adcm
(data.frame) Concomitant Medications Analysis Dataset

admh
(data.frame) Medical History Analysis Dataset

aag
(data.frame) Adverse Event Grouping Defs Dataset

advs
(data.frame) Vital Signs Analysis Dataset

adlb
(data.frame) Laboratory Analysis Dataset

adex
(data.frame) Exposure Analysis Dataset

adsub
(data.frame) Subcategory Analysis Dataset

adrs
(data.frame) Response Analysis Dataset

adtte
(data.frame) Time to Event Analysis Dataset

adhy
(data.frame) Hy's Law Analysis Dataset

adsaftte
(data.frame) Safety Time to Event Analysis Dataset

addv
(data.frame) Protocol Deviations Analysis Dataset

adcssrs
(data.frame) C-SSRS Analysis Dataset

adqlqc
(data.frame) EORTC QLQ-C30 V3 Analysis Dataset

acg
(data.frame) Concomitant Med. Grouping Defs Dataset

adic
(data.frame) Interval Censoring Analysis Dataset

group
nest list for group-dependent data binning

Details

the following arguments are better provided through the study object:

lbl_overall

armvar