LBT01 Laboratory Test Results and Change from Baseline by Visit Table.

The LBT01 table provides an overview of the analysis values and its change from baseline of each respective arm over the course of the trial.

Usage

lbt01_main(
  adam_db,
  arm_var = "ACTARM",
  summaryvars = c("AVAL", "CHG"),
  visitvar = "AVISIT",
  precision = list(),
  default_precision = 2,
  ...
)

lbt01_pre(adam_db, ...)

lbt01_post(tlg, prune_0 = TRUE, ...)

lbt01

Format

An object of class chevron_t of length 1.

Arguments

adam_db

(list of data.frames) object containing the ADaM datasets

arm_var

(string) variable used for column splitting

summaryvars

(character) variables to be analyzed. The label attribute of the corresponding column in adlb table of adam_db is used as label.

visitvar

(string) the type of time point to use. Typically one of "AVISIT" (Default) or "ATPTN".

precision

(named list of integer) where names are values found in the PARAMCD column and the the values indicate the number of digits that should be represented for min, max and median. Mean and sd are represented with one more decimal of precision.

default_precision

(integer) the default number of digits.

...

not used.

tlg

(TableTree, Listing or ggplot) object typically produced by a main function.

prune_0

(flag) remove 0 count rows

Details

The Analysis Value column, displays the number of patients, the mean, standard deviation, median and range of the analysis value for each visit.

The Change from Baseline column, displays the number of patient and the mean, standard deviation, median and range of changes relative to the baseline.

Remove zero-count rows unless overridden with prune_0 = FALSE.

Split columns by arm, typically ACTARM.

Does not include a total column by default.

Sorted based on factor level; first by PARAM labels in alphabetic order then by chronological time point given by AVISIT. Re-level to customize order.

Functions

lbt01_main(): Main TLG function

lbt01_pre(): Preprocessing

lbt01_post(): Postprocessing

Note

adam_db object must contain an adlb table with columns specified in summaryvars.

Examples

run(lbt01, syn_data, precision = list(
  "ALT" = 0,
  "CRP" = 1
))
#>   Parameter                                         A: Drug X                        B: Placebo                      C: Combination         
#>     Analysis Visit                                         Change from                       Change from                       Change from  
#>                                          Value at Visit      Baseline      Value at Visit      Baseline      Value at Visit      Baseline   
#>                                             (N=2412)         (N=2412)         (N=2412)         (N=2412)         (N=2376)         (N=2376)   
#>   ——————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————
#>   Alanine Aminotransferase Measurement                                                                                                      
#>     BASELINE                                                                                                                                
#>       n                                       134                               134                               132                       
#>       Mean (SD)                            19.8 (4.2)                        20.2 (4.2)                        20.5 (3.9)                   
#>       Median                                  19.8                              20.1                              20.4                      
#>       Min - Max                              7 - 30                            8 - 35                            9 - 29                     
#>     WEEK 1 DAY 8                                                                                                                            
#>       n                                       134              134              134              134              132              132      
#>       Mean (SD)                            19.3 (4.0)       -0.5 (5.9)       20.2 (4.0)       0.1 (6.0)        20.6 (3.9)       0.1 (5.5)   
#>       Median                                  19.2             -0.5             20.1             -0.7             20.4             0.0      
#>       Min - Max                              9 - 27          -13 - 20          6 - 29          -12 - 17         10 - 31          -14 - 16   
#>     WEEK 2 DAY 15                                                                                                                           
#>       n                                       134              134              134              134              132              132      
#>       Mean (SD)                            19.7 (4.2)       -0.1 (6.0)       20.1 (4.3)       -0.0 (6.3)       19.2 (3.6)       -1.2 (5.5)  
#>       Median                                  19.2             0.0              20.0             -0.4             19.6             -0.5     
#>       Min - Max                              7 - 31          -13 - 17          7 - 31          -18 - 14          8 - 27          -14 - 16   
#>     WEEK 3 DAY 22                                                                                                                           
#>       n                                       134              134              134              134              132              132      
#>       Mean (SD)                            20.1 (3.8)       0.3 (5.8)        19.8 (3.9)       -0.3 (5.6)       19.4 (3.9)       -1.0 (5.6)  
#>       Median                                  20.0             0.5              19.9             -0.2             18.8             -1.1     
#>       Min - Max                             12 - 29          -16 - 16         12 - 28          -19 - 15         10 - 28          -12 - 11   
#>     WEEK 4 DAY 29                                                                                                                           
#>       n                                       134              134              134              134              132              132      
#>       Mean (SD)                            20.4 (4.6)       0.6 (6.5)        19.6 (4.3)       -0.5 (6.3)       19.8 (4.0)       -0.6 (5.3)  
#>       Median                                  19.8             0.4              19.2             -1.2             19.9             -0.6     
#>       Min - Max                             10 - 34          -17 - 19         11 - 32          -17 - 17          8 - 29          -12 - 15   
#>     WEEK 5 DAY 36                                                                                                                           
#>       n                                       134              134              134              134              132              132      
#>       Mean (SD)                            20.4 (3.9)       0.6 (6.1)        19.9 (4.2)       -0.3 (6.3)       20.0 (4.2)       -0.5 (5.6)  
#>       Median                                  20.7             0.9              20.1             0.7              20.5             -1.0     
#>       Min - Max                             11 - 30          -16 - 16         10 - 29          -19 - 13          7 - 28          -16 - 17   
#>   C-Reactive Protein Measurement                                                                                                            
#>     BASELINE                                                                                                                                
#>       n                                       134                               134                               132                       
#>       Mean (SD)                           0.97 (0.24)                       1.00 (0.20)                       1.01 (0.21)                   
#>       Median                                  0.99                              1.00                              1.00                      
#>       Min - Max                            0.5 - 1.6                         0.5 - 1.6                         0.5 - 1.6                    
#>     WEEK 1 DAY 8                                                                                                                            
#>       n                                       134              134              134              134              132              132      
#>       Mean (SD)                           1.05 (0.20)      0.07 (0.31)      1.01 (0.22)      0.00 (0.29)      1.01 (0.23)      0.00 (0.33)  
#>       Median                                  1.03             0.07             1.03             0.02             1.02            -0.01     
#>       Min - Max                            0.5 - 1.6        -0.7 - 1.0       0.4 - 1.4        -0.8 - 0.9       0.4 - 1.5        -0.8 - 0.8  
#>     WEEK 2 DAY 15                                                                                                                           
#>       n                                       134              134              134              134              132              132      
#>       Mean (SD)                           0.99 (0.21)      0.02 (0.31)      1.02 (0.19)      0.02 (0.27)      0.98 (0.21)      -0.03 (0.28) 
#>       Median                                  1.01             0.01             1.04             0.02             0.97            -0.04     
#>       Min - Max                            0.5 - 1.5        -0.9 - 0.8       0.5 - 1.4        -0.6 - 0.7       0.3 - 1.5        -0.8 - 0.5  
#>     WEEK 3 DAY 22                                                                                                                           
#>       n                                       134              134              134              134              132              132      
#>       Mean (SD)                           1.00 (0.21)      0.03 (0.31)      0.98 (0.19)      -0.02 (0.27)     0.99 (0.19)      -0.01 (0.30) 
#>       Median                                  1.00             0.08             0.97             0.00             0.99            -0.03     
#>       Min - Max                            0.4 - 1.5        -0.9 - 0.8       0.5 - 1.4        -0.8 - 0.7       0.5 - 1.6        -0.7 - 0.9  
#>     WEEK 4 DAY 29                                                                                                                           
#>       n                                       134              134              134              134              132              132      
#>       Mean (SD)                           1.04 (0.20)      0.07 (0.31)      0.99 (0.21)      -0.02 (0.28)     0.99 (0.19)      -0.01 (0.29) 
#>       Median                                  1.06             0.08             0.97            -0.03             1.00             0.02     
#>       Min - Max                            0.5 - 1.5        -0.8 - 1.0       0.4 - 1.4        -0.6 - 0.6       0.6 - 1.5        -0.8 - 0.8  
#>     WEEK 5 DAY 36                                                                                                                           
#>       n                                       134              134              134              134              132              132      
#>       Mean (SD)                           0.99 (0.22)      0.02 (0.33)      1.02 (0.19)      0.02 (0.28)      0.98 (0.19)      -0.03 (0.29) 
#>       Median                                  0.98            -0.01             1.01             0.01             0.98            -0.04     
#>       Min - Max                            0.4 - 1.5        -1.0 - 0.8       0.6 - 1.5        -0.6 - 0.8       0.6 - 1.6        -0.8 - 0.8  
#>   Immunoglobulin A Measurement                                                                                                              
#>     BASELINE                                                                                                                                
#>       n                                       134                               134                               132                       
#>       Mean (SD)                          1.996 (0.372)                     1.987 (0.415)                     2.015 (0.453)                  
#>       Median                                 1.934                             1.964                             1.987                      
#>       Min - Max                           1.19 - 2.80                       0.79 - 3.07                       0.80 - 3.00                   
#>     WEEK 1 DAY 8                                                                                                                            
#>       n                                       134              134              134              134              132              132      
#>       Mean (SD)                          1.970 (0.404)    -0.026 (0.534)   2.056 (0.409)    0.069 (0.562)    2.038 (0.421)    0.022 (0.645) 
#>       Median                                 1.989            -0.007           2.026            0.077            2.055            0.023     
#>       Min - Max                           0.97 - 2.97      -1.42 - 1.08     1.06 - 3.23      -1.27 - 1.78     0.81 - 3.05      -1.49 - 1.55 
#>     WEEK 2 DAY 15                                                                                                                           
#>       n                                       134              134              134              134              132              132      
#>       Mean (SD)                          2.036 (0.393)    0.040 (0.576)    1.994 (0.417)    0.007 (0.583)    1.969 (0.374)    -0.046 (0.602)
#>       Median                                 1.989            0.068            2.034            0.047            1.978            -0.010    
#>       Min - Max                           1.17 - 3.09      -1.29 - 1.61     0.89 - 2.99      -1.48 - 1.38     0.97 - 2.75      -1.64 - 1.45 
#>     WEEK 3 DAY 22                                                                                                                           
#>       n                                       134              134              134              134              132              132      
#>       Mean (SD)                          1.974 (0.402)    -0.022 (0.519)   1.999 (0.396)    0.012 (0.577)    2.048 (0.391)    0.033 (0.602) 
#>       Median                                 1.964            0.023            1.957            0.039            2.056            0.013     
#>       Min - Max                           0.71 - 2.81      -1.44 - 1.21     0.75 - 2.89      -1.60 - 1.21     0.99 - 2.99      -1.81 - 1.23 
#>     WEEK 4 DAY 29                                                                                                                           
#>       n                                       134              134              134              134              132              132      
#>       Mean (SD)                          1.962 (0.384)    -0.035 (0.543)   1.955 (0.381)    -0.032 (0.551)   1.979 (0.420)    -0.037 (0.622)
#>       Median                                 2.025            -0.001           1.945            0.028            1.993            -0.068    
#>       Min - Max                           0.80 - 2.73      -1.53 - 1.33     1.07 - 2.81      -1.38 - 1.09     0.99 - 2.95      -1.35 - 1.59 
#>     WEEK 5 DAY 36                                                                                                                           
#>       n                                       134              134              134              134              132              132      
#>       Mean (SD)                          1.983 (0.384)    -0.013 (0.505)   2.018 (0.433)    0.031 (0.635)    1.965 (0.427)    -0.050 (0.609)
#>       Median                                 1.961            -0.012           1.987            0.071            2.009            0.034     
#>       Min - Max                           1.04 - 3.07      -1.69 - 1.28     0.92 - 3.35      -1.38 - 1.64     0.62 - 2.88      -1.60 - 1.32