LBT01 Lab Results and Change from Baseline by Visit Table.

The LBT01 table provides an overview of the Lab values and its change from baseline of each respective arm over the course of the trial.

Usage

lbt01_main(
  adam_db,
  dataset = "adlb",
  arm_var = "ACTARM",
  row_split_var = NULL,
  summaryvars = c("AVAL", "CHG"),
  visitvar = "AVISIT",
  precision = list(),
  default_precision = 2,
  page_by = TRUE,
  ...
)

lbt01_pre(adam_db, dataset = "adlb", ...)

lbt01

Format

An object of class chevron_t of length 1.

Source

cfbt01.R

Arguments

adam_db

(list of data.frames) object containing the ADaM datasets

dataset

(string) the name of a table in the adam_db object.

arm_var

(string) variable used for column splitting

row_split_var

(character) row split variable other than PARAMCD.

summaryvars

(character) variables to be analyzed. The label attribute of the corresponding column in table of adam_db is used as label.

visitvar

(string) typically one of "AVISIT" (Default) or "ATPTN" depending on the type of time point to be displayed

precision

(named list of integer) where names are values found in the PARAMCD column and the the values indicate the number of digits that should be represented for min, max and median. Mean and sd are represented with one more decimal of precision.

default_precision

(integer) the default number of digits.

page_by

(flag) indicator whether the parameter row split is by page.

...

not used.

Details

The Analysis Value column, displays the number of patients, the mean, standard deviation, median and range of the analysis value for each visit.

The Change from Baseline column, displays the number of patient and the mean, standard deviation, median and range of changes relative to the baseline.

Remove zero-count rows unless overridden with prune_0 = FALSE.

Split columns by arm, typically ACTARM.

Does not include a total column by default.

Sorted based on factor level; first by PARAM labels in alphabetic order then by chronological time point given by AVISIT. Re-level to customize order

Functions

lbt01_main(): Main TLG function

lbt01_pre(): Preprocessing

Note

adam_db object must contain table named as dataset with the columns specified in summaryvars.

Examples

run(lbt01, syn_data)
#>     Analysis Visit                                  A: Drug X                        B: Placebo                      C: Combination         
#>                                                            Change from                       Change from                       Change from  
#>                                          Value at Visit      Baseline      Value at Visit      Baseline      Value at Visit      Baseline   
#>                                             (N=134)          (N=134)          (N=134)          (N=134)          (N=132)          (N=132)    
#>   ——————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————
#>   Alanine Aminotransferase Measurement                                                                                                      
#>     BASELINE                                                                                                                                
#>       n                                       134                               134                               132                       
#>       Mean (SD)                          19.786 (4.161)                    20.153 (4.171)                    20.451 (3.911)                 
#>       Median                                 19.786                            20.077                            20.387                     
#>       Min - Max                           6.99 - 30.43                      8.11 - 34.56                      8.82 - 28.72                  
#>     WEEK 1 DAY 8                                                                                                                            
#>       n                                       134              134              134              134              132              132      
#>       Mean (SD)                          19.301 (3.981)   -0.485 (5.850)   20.221 (3.974)   0.068 (6.036)    20.553 (3.895)   0.102 (5.461) 
#>       Median                                 19.209           -0.511           20.090           -0.725           20.402           0.025     
#>       Min - Max                           8.86 - 27.32    -12.52 - 19.70    5.84 - 28.75    -12.29 - 17.41    9.87 - 30.70    -13.69 - 15.54
#>     WEEK 2 DAY 15                                                                                                                           
#>       n                                       134              134              134              134              132              132      
#>       Mean (SD)                          19.704 (4.233)   -0.082 (6.003)   20.123 (4.265)   -0.030 (6.263)   19.247 (3.592)   -1.204 (5.485)
#>       Median                                 19.168           0.038            19.986           -0.371           19.608           -0.482    
#>       Min - Max                           7.17 - 30.53    -13.15 - 17.17    7.19 - 30.54    -18.37 - 14.39    8.11 - 26.68    -14.28 - 15.65
#>     WEEK 3 DAY 22                                                                                                                           
#>       n                                       134              134              134              134              132              132      
#>       Mean (SD)                          20.130 (3.757)   0.344 (5.761)    19.837 (3.871)   -0.316 (5.604)   19.428 (3.944)   -1.023 (5.550)
#>       Median                                 20.044           0.451            19.873           -0.178           18.838           -1.065    
#>       Min - Max                          11.52 - 29.49    -16.04 - 16.26   11.86 - 28.24    -18.52 - 15.00   10.20 - 28.49    -12.33 - 11.10
#>     WEEK 4 DAY 29                                                                                                                           
#>       n                                       134              134              134              134              132              132      
#>       Mean (SD)                          20.351 (4.583)   0.565 (6.506)    19.632 (4.346)   -0.521 (6.293)   19.823 (3.995)   -0.627 (5.311)
#>       Median                                 19.763           0.353            19.151           -1.186           19.884           -0.646    
#>       Min - Max                           9.88 - 34.50    -17.00 - 18.79   11.49 - 31.98    -16.58 - 17.23    7.76 - 29.31    -12.13 - 15.30
#>     WEEK 5 DAY 36                                                                                                                           
#>       n                                       134              134              134              134              132              132      
#>       Mean (SD)                          20.418 (3.925)   0.632 (6.053)    19.859 (4.223)   -0.294 (6.315)   19.990 (4.169)   -0.460 (5.570)
#>       Median                                 20.718           0.926            20.113           0.728            20.503           -0.990    
#>       Min - Max                          10.95 - 30.17    -16.03 - 16.09   10.32 - 29.06    -19.04 - 12.64    7.41 - 27.78    -15.62 - 16.61
#>   C-Reactive Protein Measurement                                                                                                            
#>     BASELINE                                                                                                                                
#>       n                                       134                               134                               132                       
#>       Mean (SD)                          0.974 (0.236)                     1.002 (0.197)                     1.005 (0.210)                  
#>       Median                                 0.989                             1.003                             0.996                      
#>       Min - Max                           0.47 - 1.63                       0.50 - 1.58                       0.51 - 1.56                   
#>     WEEK 1 DAY 8                                                                                                                            
#>       n                                       134              134              134              134              132              132      
#>       Mean (SD)                          1.047 (0.202)    0.073 (0.312)    1.007 (0.220)    0.005 (0.289)    1.006 (0.234)    0.001 (0.327) 
#>       Median                                 1.031            0.074            1.031            0.015            1.017            -0.008    
#>       Min - Max                           0.46 - 1.57      -0.71 - 1.03     0.42 - 1.43      -0.78 - 0.93     0.44 - 1.53      -0.80 - 0.75 
#>     WEEK 2 DAY 15                                                                                                                           
#>       n                                       134              134              134              134              132              132      
#>       Mean (SD)                          0.994 (0.207)    0.020 (0.314)    1.022 (0.186)    0.020 (0.267)    0.978 (0.206)    -0.027 (0.278)
#>       Median                                 1.011            0.012            1.038            0.021            0.967            -0.035    
#>       Min - Max                           0.48 - 1.48      -0.85 - 0.76     0.52 - 1.44      -0.55 - 0.68     0.27 - 1.51      -0.76 - 0.50 
#>     WEEK 3 DAY 22                                                                                                                           
#>       n                                       134              134              134              134              132              132      
#>       Mean (SD)                          1.002 (0.208)    0.028 (0.309)    0.982 (0.188)    -0.020 (0.266)   0.995 (0.191)    -0.010 (0.301)
#>       Median                                 0.998            0.078            0.969            0.004            0.987            -0.029    
#>       Min - Max                           0.42 - 1.48      -0.92 - 0.79     0.52 - 1.44      -0.77 - 0.75     0.50 - 1.57      -0.72 - 0.86 
#>     WEEK 4 DAY 29                                                                                                                           
#>       n                                       134              134              134              134              132              132      
#>       Mean (SD)                          1.040 (0.203)    0.067 (0.306)    0.987 (0.205)    -0.015 (0.278)   0.994 (0.190)    -0.011 (0.289)
#>       Median                                 1.056            0.079            0.970            -0.032           1.004            0.018     
#>       Min - Max                           0.54 - 1.53      -0.78 - 1.04     0.43 - 1.43      -0.62 - 0.59     0.55 - 1.46      -0.82 - 0.82 
#>     WEEK 5 DAY 36                                                                                                                           
#>       n                                       134              134              134              134              132              132      
#>       Mean (SD)                          0.993 (0.219)    0.019 (0.334)    1.021 (0.190)    0.019 (0.282)    0.976 (0.191)    -0.029 (0.290)
#>       Median                                 0.981            -0.006           1.006            0.011            0.980            -0.037    
#>       Min - Max                           0.40 - 1.49      -1.00 - 0.84     0.58 - 1.51      -0.64 - 0.83     0.55 - 1.59      -0.84 - 0.79 
#>   Immunoglobulin A Measurement                                                                                                              
#>     BASELINE                                                                                                                                
#>       n                                       134                               134                               132                       
#>       Mean (SD)                          1.996 (0.372)                     1.987 (0.415)                     2.015 (0.453)                  
#>       Median                                 1.934                             1.964                             1.987                      
#>       Min - Max                           1.19 - 2.80                       0.79 - 3.07                       0.80 - 3.00                   
#>     WEEK 1 DAY 8                                                                                                                            
#>       n                                       134              134              134              134              132              132      
#>       Mean (SD)                          1.970 (0.404)    -0.026 (0.534)   2.056 (0.409)    0.069 (0.562)    2.038 (0.421)    0.022 (0.645) 
#>       Median                                 1.989            -0.007           2.026            0.077            2.055            0.023     
#>       Min - Max                           0.97 - 2.97      -1.42 - 1.08     1.06 - 3.23      -1.27 - 1.78     0.81 - 3.05      -1.49 - 1.55 
#>     WEEK 2 DAY 15                                                                                                                           
#>       n                                       134              134              134              134              132              132      
#>       Mean (SD)                          2.036 (0.393)    0.040 (0.576)    1.994 (0.417)    0.007 (0.583)    1.969 (0.374)    -0.046 (0.602)
#>       Median                                 1.989            0.068            2.034            0.047            1.978            -0.010    
#>       Min - Max                           1.17 - 3.09      -1.29 - 1.61     0.89 - 2.99      -1.48 - 1.38     0.97 - 2.75      -1.64 - 1.45 
#>     WEEK 3 DAY 22                                                                                                                           
#>       n                                       134              134              134              134              132              132      
#>       Mean (SD)                          1.974 (0.402)    -0.022 (0.519)   1.999 (0.396)    0.012 (0.577)    2.048 (0.391)    0.033 (0.602) 
#>       Median                                 1.964            0.023            1.957            0.039            2.056            0.013     
#>       Min - Max                           0.71 - 2.81      -1.44 - 1.21     0.75 - 2.89      -1.60 - 1.21     0.99 - 2.99      -1.81 - 1.23 
#>     WEEK 4 DAY 29                                                                                                                           
#>       n                                       134              134              134              134              132              132      
#>       Mean (SD)                          1.962 (0.384)    -0.035 (0.543)   1.955 (0.381)    -0.032 (0.551)   1.979 (0.420)    -0.037 (0.622)
#>       Median                                 2.025            -0.001           1.945            0.028            1.993            -0.068    
#>       Min - Max                           0.80 - 2.73      -1.53 - 1.33     1.07 - 2.81      -1.38 - 1.09     0.99 - 2.95      -1.35 - 1.59 
#>     WEEK 5 DAY 36                                                                                                                           
#>       n                                       134              134              134              134              132              132      
#>       Mean (SD)                          1.983 (0.384)    -0.013 (0.505)   2.018 (0.433)    0.031 (0.635)    1.965 (0.427)    -0.050 (0.609)
#>       Median                                 1.961            -0.012           1.987            0.071            2.009            0.034     
#>       Min - Max                           1.04 - 3.07      -1.69 - 1.28     0.92 - 3.35      -1.38 - 1.64     0.62 - 2.88      -1.60 - 1.32