LBT01 Lab Results and Change from Baseline by Visit Table.

The LBT01 table provides an overview of the Lab values and its change from baseline of each respective arm over the course of the trial.

Usage

lbt01_main(
  adam_db,
  dataset = "adlb",
  arm_var = "ACTARM",
  lbl_overall = NULL,
  row_split_var = NULL,
  summaryvars = c("AVAL", "CHG"),
  visitvar = "AVISIT",
  precision = list(default = 2L),
  page_var = "PARAMCD",
  .stats = c("n", "mean_sd", "median", "range"),
  skip = list(CHG = "BASELINE"),
  ...
)

lbt01_pre(adam_db, dataset = "adlb", ...)

lbt01

Format

An object of class chevron_t of length 1.

Arguments

adam_db

(list of data.frames) object containing the ADaM datasets

dataset

(string) the name of a table in the adam_db object.

arm_var

(string) variable used for column splitting

lbl_overall

(string) label used for overall column, if set to NULL the overall column is omitted

row_split_var

(character) additional row split variables.

summaryvars

(character) variables to be analyzed. The label attribute of the corresponding column in table of adam_db is used as label.

visitvar

(string) typically one of "AVISIT" or user-defined visit incorporating "ATPT".

precision

(named list of integer) where names are values found in the PARAMCD column and the values indicate the number of digits in statistics. If default is set, and parameter precision not specified, the value for default will be used.

page_var

(string) variable name prior to which the row split is by page.

.stats

(character) statistics names, see tern::analyze_vars().

skip

Named (list) of visit values that need to be inhibited.

...

additional arguments like .indent_mods, .labels.

Value

the main function returns an rtables object.

the preprocessing function returns a list of data.frame.

Details

The Analysis Value column, displays the number of patients, the mean, standard deviation, median and range of the analysis value for each visit.

The Change from Baseline column, displays the number of patient and the mean, standard deviation, median and range of changes relative to the baseline.

Remove zero-count rows unless overridden with prune_0 = FALSE.

Split columns by arm, typically ACTARM.

Does not include a total column by default.

Sorted based on factor level; first by PARAM labels in alphabetic order then by chronological time point given by AVISIT. Re-level to customize order

Functions

lbt01_main(): Main TLG function

lbt01_pre(): Preprocessing

Note

adam_db object must contain table named as dataset with the columns specified in summaryvars.

Examples

run(lbt01, syn_data)
#>                                                     A: Drug X                        B: Placebo                      C: Combination         
#>                                                            Change from                       Change from                       Change from  
#>                                          Value at Visit      Baseline      Value at Visit      Baseline      Value at Visit      Baseline   
#>   Analysis Visit                             (N=15)           (N=15)           (N=15)           (N=15)           (N=15)           (N=15)    
#>   ——————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————
#>   Alanine Aminotransferase Measurement                                                                                                      
#>     BASELINE                                                                                                                                
#>       n                                        15                                15                                15                       
#>       Mean (SD)                          20.262 (4.982)                    19.371 (4.687)                    21.724 (3.857)                 
#>       Median                                 19.216                            19.781                            22.900                     
#>       Min - Max                          13.77 - 30.43                     11.58 - 25.59                     12.57 - 27.69                  
#>     WEEK 1 DAY 8                                                                                                                            
#>       n                                        15               15               15               15               15               15      
#>       Mean (SD)                          19.227 (4.492)   -1.036 (7.113)   21.622 (3.015)   2.251 (6.830)    20.630 (3.950)   -1.094 (4.425)
#>       Median                                 18.666           -2.148           21.790           2.901            20.570           -0.398    
#>       Min - Max                          12.12 - 27.32    -12.07 - 9.07    16.69 - 26.32    -6.73 - 14.36    13.21 - 27.31     -7.84 - 7.84 
#>     WEEK 2 DAY 15                                                                                                                           
#>       n                                        15               15               15               15               15               15      
#>       Mean (SD)                          19.458 (4.211)   -0.804 (6.309)   21.103 (3.831)   1.732 (5.981)    17.164 (3.089)   -4.559 (4.438)
#>       Median                                 18.988           -2.571           20.200           2.499            16.740           -3.463    
#>       Min - Max                          12.96 - 26.92    -12.00 - 13.14   13.42 - 29.93    -9.97 - 14.39    12.94 - 22.80    -10.95 - 1.01 
#>     WEEK 3 DAY 22                                                                                                                           
#>       n                                        15               15               15               15               15               15      
#>       Mean (SD)                          19.795 (4.272)   -0.467 (6.304)   20.822 (3.234)   1.451 (6.839)    20.165 (3.805)   -1.559 (4.042)
#>       Median                                 18.489           0.542            21.541           2.138            20.612           -2.878    
#>       Min - Max                          13.51 - 26.07    -16.04 - 7.43    14.18 - 26.58    -9.21 - 15.00    14.01 - 25.74     -7.48 - 6.92 
#>     WEEK 4 DAY 29                                                                                                                           
#>       n                                        15               15               15               15               15               15      
#>       Mean (SD)                          19.517 (3.640)   -0.745 (6.200)   19.824 (3.943)   0.453 (5.486)    19.923 (5.026)   -1.801 (5.268)
#>       Median                                 17.927           1.239            19.875           -1.242           20.213           -1.254    
#>       Min - Max                          14.23 - 26.40    -13.17 - 7.57    13.56 - 26.43     -6.52 - 8.87    11.32 - 26.59    -12.13 - 6.95 
#>     WEEK 5 DAY 36                                                                                                                           
#>       n                                        15               15               15               15               15               15      
#>       Mean (SD)                          19.952 (2.895)   -0.310 (6.188)   19.495 (3.459)   0.124 (6.533)    18.854 (3.392)   -2.869 (4.309)
#>       Median                                 20.300           0.443            19.900           0.822            19.358           -3.248    
#>       Min - Max                          14.40 - 24.56    -16.03 - 6.90    10.76 - 25.16    -12.63 - 11.81   12.69 - 23.49    -11.07 - 6.79 
#>   C-Reactive Protein Measurement                                                                                                            
#>     BASELINE                                                                                                                                
#>       n                                        15                                15                                15                       
#>       Mean (SD)                          0.971 (0.246)                     1.026 (0.136)                     0.995 (0.172)                  
#>       Median                                 1.045                             1.073                             0.997                      
#>       Min - Max                           0.54 - 1.41                       0.73 - 1.19                       0.64 - 1.32                   
#>     WEEK 1 DAY 8                                                                                                                            
#>       n                                        15               15               15               15               15               15      
#>       Mean (SD)                          1.042 (0.261)    0.072 (0.402)    1.025 (0.161)    -0.001 (0.229)   0.972 (0.235)    -0.023 (0.312)
#>       Median                                 1.052            0.040            1.038            -0.001           0.983            -0.019    
#>       Min - Max                           0.58 - 1.57      -0.49 - 1.03     0.68 - 1.35      -0.41 - 0.37     0.57 - 1.39      -0.39 - 0.75 
#>     WEEK 2 DAY 15                                                                                                                           
#>       n                                        15               15               15               15               15               15      
#>       Mean (SD)                          1.025 (0.268)    0.055 (0.286)    0.956 (0.228)    -0.070 (0.229)   1.010 (0.189)    0.015 (0.255) 
#>       Median                                 1.026            0.048            0.973            -0.100           0.983            -0.066    
#>       Min - Max                           0.67 - 1.46      -0.34 - 0.57     0.59 - 1.32      -0.40 - 0.26     0.67 - 1.37      -0.28 - 0.50 
#>     WEEK 3 DAY 22                                                                                                                           
#>       n                                        15               15               15               15               15               15      
#>       Mean (SD)                          1.006 (0.218)    0.035 (0.308)    0.954 (0.154)    -0.072 (0.191)   1.036 (0.162)    0.042 (0.267) 
#>       Median                                 0.982            0.014            0.926            -0.050           1.061            0.021     
#>       Min - Max                           0.61 - 1.43      -0.71 - 0.56     0.71 - 1.26      -0.44 - 0.25     0.81 - 1.33      -0.45 - 0.45 
#>     WEEK 4 DAY 29                                                                                                                           
#>       n                                        15               15               15               15               15               15      
#>       Mean (SD)                          0.984 (0.185)    0.014 (0.303)    0.953 (0.233)    -0.073 (0.291)   1.032 (0.198)    0.038 (0.337) 
#>       Median                                 1.006            -0.046           0.897            -0.157           0.959            -0.043    
#>       Min - Max                           0.70 - 1.25      -0.46 - 0.61     0.60 - 1.42      -0.41 - 0.59     0.80 - 1.46      -0.51 - 0.82 
#>     WEEK 5 DAY 36                                                                                                                           
#>       n                                        15               15               15               15               15               15      
#>       Mean (SD)                          0.943 (0.186)    -0.028 (0.334)   0.989 (0.155)    -0.037 (0.226)   0.956 (0.158)    -0.039 (0.220)
#>       Median                                 0.953            -0.065           0.960            -0.099           0.993            -0.049    
#>       Min - Max                           0.60 - 1.25      -0.81 - 0.71     0.71 - 1.29      -0.36 - 0.56     0.66 - 1.26      -0.44 - 0.34 
#>   Immunoglobulin A Measurement                                                                                                              
#>     BASELINE                                                                                                                                
#>       n                                        15                                15                                15                       
#>       Mean (SD)                          1.843 (0.341)                     2.074 (0.452)                     1.890 (0.486)                  
#>       Median                                 1.815                             1.978                             1.750                      
#>       Min - Max                           1.32 - 2.76                       1.25 - 2.95                       1.15 - 2.65                   
#>     WEEK 1 DAY 8                                                                                                                            
#>       n                                        15               15               15               15               15               15      
#>       Mean (SD)                          2.029 (0.343)    0.186 (0.561)    2.261 (0.422)    0.187 (0.565)    1.884 (0.430)    -0.006 (0.738)
#>       Median                                 1.979            0.257            2.288            0.284            1.832            -0.099    
#>       Min - Max                           1.56 - 2.67      -1.10 - 1.08     1.54 - 3.02      -0.85 - 1.05     1.07 - 2.52      -1.44 - 1.14 
#>     WEEK 2 DAY 15                                                                                                                           
#>       n                                        15               15               15               15               15               15      
#>       Mean (SD)                          2.006 (0.361)    0.163 (0.559)    2.010 (0.332)    -0.064 (0.368)   1.944 (0.325)    0.054 (0.597) 
#>       Median                                 1.944            0.140            2.051            -0.045           1.976            0.171     
#>       Min - Max                           1.59 - 2.88      -0.84 - 1.04     1.25 - 2.55      -0.83 - 0.57     1.32 - 2.52      -0.79 - 0.95 
#>     WEEK 3 DAY 22                                                                                                                           
#>       n                                        15               15               15               15               15               15      
#>       Mean (SD)                          2.001 (0.382)    0.158 (0.468)    2.001 (0.500)    -0.073 (0.572)   2.040 (0.389)    0.150 (0.578) 
#>       Median                                 1.958            0.214            2.174            -0.022           2.068            -0.030    
#>       Min - Max                           1.34 - 2.80      -0.80 - 0.94     0.75 - 2.68      -1.13 - 0.96     1.37 - 2.70      -0.57 - 1.17 
#>     WEEK 4 DAY 29                                                                                                                           
#>       n                                        15               15               15               15               15               15      
#>       Mean (SD)                          1.988 (0.231)    0.145 (0.445)    1.697 (0.404)    -0.377 (0.605)   2.113 (0.432)    0.223 (0.828) 
#>       Median                                 2.080            0.081            1.697            -0.284           2.240            0.583     
#>       Min - Max                           1.44 - 2.25      -0.87 - 0.88     1.10 - 2.43      -1.32 - 0.90     1.31 - 2.68      -1.02 - 1.50 
#>     WEEK 5 DAY 36                                                                                                                           
#>       n                                        15               15               15               15               15               15      
#>       Mean (SD)                          2.087 (0.179)    0.243 (0.397)    1.935 (0.442)    -0.139 (0.636)   1.933 (0.386)    0.043 (0.662) 
#>       Median                                 2.102            0.379            1.832            -0.176           1.924            0.100     
#>       Min - Max                           1.81 - 2.41      -0.81 - 0.78     1.33 - 2.72      -1.17 - 0.85     1.11 - 2.55      -1.23 - 1.15