LBT01 Table 1 (Default) Laboratory Test Results and Change from Baseline by Visit.

The LBT01 table provides an overview of the analysis values and its change from baseline of each respective arm over the course of the trial.

Usage

lbt01_1_main(
  adam_db,
  armvar = "ACTARM",
  summaryvars = c("AVAL", "CHG"),
  summaryvars_lbls = c("Value at Visit", "Change from \nBaseline"),
  visitvar = "AVISIT",
  prune_0 = TRUE,
  deco = std_deco("LBT01"),
  ...
)

lbt01_1_lyt(
  armvar,
  summaryvars,
  summaryvars_lbls,
  visitvar,
  lbl_avisit,
  lbl_param,
  deco,
  ...
)

lbt01_1_pre(adam_db, ...)

lbt01_1

Format

An object of class chevron_t of length 1.

Arguments

adam_db

(dm) object containing the ADaM datasets

armvar

(character) variable used for column splitting

summaryvars

(vector of character) the variables to be analyzed. For this table, AVAL and CHG by default.

summaryvars_lbls

(vector of character) the label of the variables to be analyzed.

visitvar

(character) typically one of "AVISIT" (Default) or "ATPTN" depending on the type of time point to be displayed.

prune_0

(logical) remove 0 count rows

deco

(character) decoration with title, subtitles and main_footer content

...

not used.

lbl_avisit

(character) label of the visitvar variable.

lbl_param

(character) label of the PARAM variable.

Details

The Analysis Value column, displays the number of patients, the mean, standard deviation, median and range of the analysis value for each visit.

The Change from Baseline column, displays the number of patient and the mean, standard deviation, median and range of changes relative to the baseline.

Remove zero-count rows unless overridden with prune_0 = FALSE.

Split columns by arm, typically ACTARM.

Does not include a total column by default.

Sorted based on factor level; first by PARAM labels in alphabetic order then by chronological time point given by AVISIT. Re-level to customize order.

Functions

lbt01_1_main(): Main TLG function

lbt01_1_lyt(): Layout

lbt01_1_pre(): Preprocessing

Note

adam_db object must contain an adlb table with columns specified in summaryvars.

Examples

run(lbt01_1, syn_test_data())
#> Parameter                                                                                                                           
#>   Analysis Visit                                 A: Drug X                      B: Placebo                    C: Combination        
#>                                                         Change from                     Change from                     Change from 
#>                                        Value at Visit     Baseline     Value at Visit     Baseline     Value at Visit     Baseline  
#> ————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————
#> Alanine Aminotransferase Measurement                                                                                                
#>   BASELINE                                                                                                                          
#>     n                                       134             134             134             134             132             132     
#>     Mean (SD)                            19.8 (4.2)      0.0 (0.0)       20.2 (4.2)      0.0 (0.0)       20.5 (3.9)      0.0 (0.0)  
#>     Median                                  19.8            0.0             20.1            0.0             20.4            0.0     
#>     Min - Max                            7.0 - 30.4      0.0 - 0.0       8.1 - 34.6      0.0 - 0.0       8.8 - 28.7      0.0 - 0.0  
#>   WEEK 1 DAY 8                                                                                                                      
#>     n                                       134             134             134             134             132             132     
#>     Mean (SD)                            19.3 (4.0)      -0.5 (5.9)      20.2 (4.0)      0.1 (6.0)       20.6 (3.9)      0.1 (5.5)  
#>     Median                                  19.2            -0.5            20.1            -0.7            20.4            0.0     
#>     Min - Max                            8.9 - 27.3     -12.5 - 19.7     5.8 - 28.8     -12.3 - 17.4     9.9 - 30.7     -13.7 - 15.5
#>   WEEK 2 DAY 15                                                                                                                     
#>     n                                       134             134             134             134             132             132     
#>     Mean (SD)                            19.7 (4.2)      -0.1 (6.0)      20.1 (4.3)      -0.0 (6.3)      19.2 (3.6)      -1.2 (5.5) 
#>     Median                                  19.2            0.0             20.0            -0.4            19.6            -0.5    
#>     Min - Max                            7.2 - 30.5     -13.2 - 17.2     7.2 - 30.5     -18.4 - 14.4     8.1 - 26.7     -14.3 - 15.6
#>   WEEK 3 DAY 22                                                                                                                     
#>     n                                       134             134             134             134             132             132     
#>     Mean (SD)                            20.1 (3.8)      0.3 (5.8)       19.8 (3.9)      -0.3 (5.6)      19.4 (3.9)      -1.0 (5.6) 
#>     Median                                  20.0            0.5             19.9            -0.2            18.8            -1.1    
#>     Min - Max                           11.5 - 29.5     -16.0 - 16.3    11.9 - 28.2     -18.5 - 15.0    10.2 - 28.5     -12.3 - 11.1
#>   WEEK 4 DAY 29                                                                                                                     
#>     n                                       134             134             134             134             132             132     
#>     Mean (SD)                            20.4 (4.6)      0.6 (6.5)       19.6 (4.3)      -0.5 (6.3)      19.8 (4.0)      -0.6 (5.3) 
#>     Median                                  19.8            0.4             19.2            -1.2            19.9            -0.6    
#>     Min - Max                            9.9 - 34.5     -17.0 - 18.8    11.5 - 32.0     -16.6 - 17.2     7.8 - 29.3     -12.1 - 15.3
#>   WEEK 5 DAY 36                                                                                                                     
#>     n                                       134             134             134             134             132             132     
#>     Mean (SD)                            20.4 (3.9)      0.6 (6.1)       19.9 (4.2)      -0.3 (6.3)      20.0 (4.2)      -0.5 (5.6) 
#>     Median                                  20.7            0.9             20.1            0.7             20.5            -1.0    
#>     Min - Max                           10.9 - 30.2     -16.0 - 16.1    10.3 - 29.1     -19.0 - 12.6     7.4 - 27.8     -15.6 - 16.6
#> C-Reactive Protein Measurement                                                                                                      
#>   BASELINE                                                                                                                          
#>     n                                       134             134             134             134             132             132     
#>     Mean (SD)                            1.0 (0.2)       0.0 (0.0)       1.0 (0.2)       0.0 (0.0)       1.0 (0.2)       0.0 (0.0)  
#>     Median                                  1.0             0.0             1.0             0.0             1.0             0.0     
#>     Min - Max                            0.5 - 1.6       0.0 - 0.0       0.5 - 1.6       0.0 - 0.0       0.5 - 1.6       0.0 - 0.0  
#>   WEEK 1 DAY 8                                                                                                                      
#>     n                                       134             134             134             134             132             132     
#>     Mean (SD)                            1.0 (0.2)       0.1 (0.3)       1.0 (0.2)       0.0 (0.3)       1.0 (0.2)       0.0 (0.3)  
#>     Median                                  1.0             0.1             1.0             0.0             1.0             -0.0    
#>     Min - Max                            0.5 - 1.6       -0.7 - 1.0      0.4 - 1.4       -0.8 - 0.9      0.4 - 1.5       -0.8 - 0.8 
#>   WEEK 2 DAY 15                                                                                                                     
#>     n                                       134             134             134             134             132             132     
#>     Mean (SD)                            1.0 (0.2)       0.0 (0.3)       1.0 (0.2)       0.0 (0.3)       1.0 (0.2)       -0.0 (0.3) 
#>     Median                                  1.0             0.0             1.0             0.0             1.0             -0.0    
#>     Min - Max                            0.5 - 1.5       -0.9 - 0.8      0.5 - 1.4       -0.6 - 0.7      0.3 - 1.5       -0.8 - 0.5 
#>   WEEK 3 DAY 22                                                                                                                     
#>     n                                       134             134             134             134             132             132     
#>     Mean (SD)                            1.0 (0.2)       0.0 (0.3)       1.0 (0.2)       -0.0 (0.3)      1.0 (0.2)       -0.0 (0.3) 
#>     Median                                  1.0             0.1             1.0             0.0             1.0             -0.0    
#>     Min - Max                            0.4 - 1.5       -0.9 - 0.8      0.5 - 1.4       -0.8 - 0.7      0.5 - 1.6       -0.7 - 0.9 
#>   WEEK 4 DAY 29                                                                                                                     
#>     n                                       134             134             134             134             132             132     
#>     Mean (SD)                            1.0 (0.2)       0.1 (0.3)       1.0 (0.2)       -0.0 (0.3)      1.0 (0.2)       -0.0 (0.3) 
#>     Median                                  1.1             0.1             1.0             -0.0            1.0             0.0     
#>     Min - Max                            0.5 - 1.5       -0.8 - 1.0      0.4 - 1.4       -0.6 - 0.6      0.6 - 1.5       -0.8 - 0.8 
#>   WEEK 5 DAY 36                                                                                                                     
#>     n                                       134             134             134             134             132             132     
#>     Mean (SD)                            1.0 (0.2)       0.0 (0.3)       1.0 (0.2)       0.0 (0.3)       1.0 (0.2)       -0.0 (0.3) 
#>     Median                                  1.0             -0.0            1.0             0.0             1.0             -0.0    
#>     Min - Max                            0.4 - 1.5       -1.0 - 0.8      0.6 - 1.5       -0.6 - 0.8      0.6 - 1.6       -0.8 - 0.8 
#> Immunoglobulin A Measurement                                                                                                        
#>   BASELINE                                                                                                                          
#>     n                                       134             134             134             134             132             132     
#>     Mean (SD)                            2.0 (0.4)       0.0 (0.0)       2.0 (0.4)       0.0 (0.0)       2.0 (0.5)       0.0 (0.0)  
#>     Median                                  1.9             0.0             2.0             0.0             2.0             0.0     
#>     Min - Max                            1.2 - 2.8       0.0 - 0.0       0.8 - 3.1       0.0 - 0.0       0.8 - 3.0       0.0 - 0.0  
#>   WEEK 1 DAY 8                                                                                                                      
#>     n                                       134             134             134             134             132             132     
#>     Mean (SD)                            2.0 (0.4)       -0.0 (0.5)      2.1 (0.4)       0.1 (0.6)       2.0 (0.4)       0.0 (0.6)  
#>     Median                                  2.0             -0.0            2.0             0.1             2.1             0.0     
#>     Min - Max                            1.0 - 3.0       -1.4 - 1.1      1.1 - 3.2       -1.3 - 1.8      0.8 - 3.0       -1.5 - 1.6 
#>   WEEK 2 DAY 15                                                                                                                     
#>     n                                       134             134             134             134             132             132     
#>     Mean (SD)                            2.0 (0.4)       0.0 (0.6)       2.0 (0.4)       0.0 (0.6)       2.0 (0.4)       -0.0 (0.6) 
#>     Median                                  2.0             0.1             2.0             0.0             2.0             -0.0    
#>     Min - Max                            1.2 - 3.1       -1.3 - 1.6      0.9 - 3.0       -1.5 - 1.4      1.0 - 2.8       -1.6 - 1.4 
#>   WEEK 3 DAY 22                                                                                                                     
#>     n                                       134             134             134             134             132             132     
#>     Mean (SD)                            2.0 (0.4)       -0.0 (0.5)      2.0 (0.4)       0.0 (0.6)       2.0 (0.4)       0.0 (0.6)  
#>     Median                                  2.0             0.0             2.0             0.0             2.1             0.0     
#>     Min - Max                            0.7 - 2.8       -1.4 - 1.2      0.8 - 2.9       -1.6 - 1.2      1.0 - 3.0       -1.8 - 1.2 
#>   WEEK 4 DAY 29                                                                                                                     
#>     n                                       134             134             134             134             132             132     
#>     Mean (SD)                            2.0 (0.4)       -0.0 (0.5)      2.0 (0.4)       -0.0 (0.6)      2.0 (0.4)       -0.0 (0.6) 
#>     Median                                  2.0             -0.0            1.9             0.0             2.0             -0.1    
#>     Min - Max                            0.8 - 2.7       -1.5 - 1.3      1.1 - 2.8       -1.4 - 1.1      1.0 - 3.0       -1.3 - 1.6 
#>   WEEK 5 DAY 36                                                                                                                     
#>     n                                       134             134             134             134             132             132     
#>     Mean (SD)                            2.0 (0.4)       -0.0 (0.5)      2.0 (0.4)       0.0 (0.6)       2.0 (0.4)       -0.0 (0.6) 
#>     Median                                  2.0             -0.0            2.0             0.1             2.0             0.0     
#>     Min - Max                            1.0 - 3.1       -1.7 - 1.3      0.9 - 3.4       -1.4 - 1.6      0.6 - 2.9       -1.6 - 1.3