LBT01 Table 1 (Default) Laboratory Test Results and Change from Baseline by Visit

The LBT01 table provides an overview of the analysis values and its change from baseline of each respective arm over the course of the trial.

lbt01_1(
  adam_db,
  armvar = .study$actualarm,
  summaryvars = c("AVAL", "CHG"),
  summaryvars_lbls = c("Value at Visit", "Change from \nBaseline"),
  visitvar = "AVISIT",
  prune_0 = TRUE,
  deco = std_deco("LBT01"),
  .study = list(actualarm = "ACTARM")
)

lbt01_1_lyt(
  armvar = .study$actualarm,
  summaryvars = c("AVAL", "CHG"),
  summaryvars_lbls = c("Value at Visit", "Change from \nBaseline"),
  visitvar = .study$visitvar,
  lbl_avisit = "Visit",
  lbl_param = "Parameter",
  deco = std_deco("LBT01"),
  .study = list(actualarm = "ACTARM", visitvar = "AVISIT")
)

Arguments

adam_db

(dm) object containing the ADaM datasets

armvar

(character) variable used for column splitting

summaryvars

(vector of character) the variables to be analyzed. For this table, AVAL and CHG by default.

summaryvars_lbls

(vector of character) the label of the variables to be analyzed.

visitvar

(character) typically one of "AVISIT" (Default) or "ATPTN" depending on the type of time point to be displayed.

prune_0

(logical) remove 0 count rows

deco

(character) decoration with title, subtitles and main_footer content

.study

(list) with default values for the arguments of the function

lbl_avisit

(character) label of the visitvar variable.

lbl_param

(character) label of the PARAM variable.

Details

The Analysis Value column, displays the number of patients, the mean, standard deviation, median and range of the analysis value for each visit.

The Change from Baseline column, displays the number of patient and the mean, standard deviation, median and range of changes relative to the baseline.

Remove zero-count rows unless overridden with prune_0 = FALSE.

Split columns by arm, typically ACTARM.

Does not include a total column by default.

Sorted based on factor level; first by PARAM labels in alphabetic order then by chronological time point given by AVISIT. Re-level to customize order

Functions

lbt01_1_lyt:

Examples

library(dm)

db <- syn_test_data() %>%
  preprocess_data("lbt01_1")

lbt01_1(db)
#> Laboratory Test Results and Change from Baseline by Visit, {Specific Lab Test}:  {Specify Population}
#> Protocol: {{protocol}}, Snapshot: {{snapshot}}, Snapshot Date: {{snapshot-date}}, Cutoff Date: {{cutoff-date}}
#> 
#> ————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————
#> Parameter                                                                                                                           
#>   Analysis Visit                                 A: Drug X                      B: Placebo                    C: Combination        
#>                                                         Change from                     Change from                     Change from 
#>                                        Value at Visit     Baseline     Value at Visit     Baseline     Value at Visit     Baseline  
#> ————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————
#> Alanine Aminotransferase Measurement                                                                                                
#>   BASELINE                                                                                                                          
#>     n                                       134             134             134             134             132             132     
#>     Mean (SD)                            49.6 (8.3)      0.0 (0.0)       50.3 (8.3)      0.0 (0.0)       50.9 (7.8)      0.0 (0.0)  
#>     Median                                  49.6            0.0             50.2            0.0             50.8            0.0     
#>     Min - Max                           24.0 - 70.9      0.0 - 0.0      26.2 - 79.1      0.0 - 0.0      27.6 - 67.4      0.0 - 0.0  
#>   WEEK 1 DAY 8                                                                                                                      
#>     n                                       134             134             134             134             132             132     
#>     Mean (SD)                            48.6 (8.0)     -1.0 (11.7)      50.4 (7.9)      0.1 (12.1)      51.1 (7.8)      0.2 (10.9) 
#>     Median                                  48.4            -1.0            50.2            -1.4            50.8            0.1     
#>     Min - Max                           27.7 - 64.6     -25.0 - 39.4    21.7 - 67.5     -24.6 - 34.8    29.7 - 71.4     -27.4 - 31.1
#>   WEEK 2 DAY 15                                                                                                                     
#>     n                                       134             134             134             134             132             132     
#>     Mean (SD)                            49.4 (8.5)     -0.2 (12.0)      50.2 (8.5)     -0.1 (12.5)      48.5 (7.2)     -2.4 (11.0) 
#>     Median                                  48.3            0.1             50.0            -0.7            49.2            -1.0    
#>     Min - Max                           24.3 - 71.1     -26.3 - 34.3    24.4 - 71.1     -36.7 - 28.8    26.2 - 63.4     -28.6 - 31.3
#>   WEEK 3 DAY 22                                                                                                                     
#>     n                                       134             134             134             134             132             132     
#>     Mean (SD)                            50.3 (7.5)      0.7 (11.5)      49.7 (7.7)     -0.6 (11.2)      48.9 (7.9)     -2.0 (11.1) 
#>     Median                                  50.1            0.9             49.7            -0.4            47.7            -2.1    
#>     Min - Max                           33.0 - 69.0     -32.1 - 32.5    33.7 - 66.5     -37.0 - 30.0    30.4 - 67.0     -24.7 - 22.2
#>   WEEK 4 DAY 29                                                                                                                     
#>     n                                       134             134             134             134             132             132     
#>     Mean (SD)                            50.7 (9.2)      1.1 (13.0)      49.3 (8.7)     -1.0 (12.6)      49.6 (8.0)     -1.3 (10.6) 
#>     Median                                  49.5            0.7             48.3            -2.4            49.8            -1.3    
#>     Min - Max                           29.8 - 79.0     -34.0 - 37.6    33.0 - 74.0     -33.2 - 34.5    25.5 - 68.6     -24.3 - 30.6
#>   WEEK 5 DAY 36                                                                                                                     
#>     n                                       134             134             134             134             132             132     
#>     Mean (SD)                            50.8 (7.8)      1.3 (12.1)      49.7 (8.4)     -0.6 (12.6)      50.0 (8.3)     -0.9 (11.1) 
#>     Median                                  51.4            1.9             50.2            1.5             51.0            -2.0    
#>     Min - Max                           31.9 - 70.3     -32.1 - 32.2    30.6 - 68.1     -38.1 - 25.3    24.8 - 65.6     -31.2 - 33.2
#> C-Reactive Protein Measurement                                                                                                      
#>   BASELINE                                                                                                                          
#>     n                                       134             134             134             134             132             132     
#>     Mean (SD)                            48.9 (9.4)      0.0 (0.0)       50.1 (7.9)      0.0 (0.0)       50.2 (8.4)      0.0 (0.0)  
#>     Median                                  49.6            0.0             50.1            0.0             49.8            0.0     
#>     Min - Max                           29.0 - 75.2      0.0 - 0.0      29.9 - 73.4      0.0 - 0.0      30.5 - 72.3      0.0 - 0.0  
#>   WEEK 1 DAY 8                                                                                                                      
#>     n                                       134             134             134             134             132             132     
#>     Mean (SD)                            51.9 (8.1)      2.9 (12.5)      50.3 (8.8)      0.2 (11.6)      50.3 (9.4)      0.0 (13.1) 
#>     Median                                  51.2            2.9             51.2            0.6             50.7            -0.3    
#>     Min - Max                           28.5 - 72.9     -28.2 - 41.2    26.9 - 67.3     -31.3 - 37.2    27.8 - 71.2     -32.0 - 30.1
#>   WEEK 2 DAY 15                                                                                                                     
#>     n                                       134             134             134             134             132             132     
#>     Mean (SD)                            49.7 (8.3)      0.8 (12.6)      50.9 (7.4)      0.8 (10.7)      49.1 (8.2)     -1.1 (11.1) 
#>     Median                                  50.4            0.5             51.5            0.8             48.7            -1.4    
#>     Min - Max                           29.0 - 69.1     -34.1 - 30.4    30.7 - 67.7     -22.2 - 27.2    20.6 - 70.5     -30.2 - 19.8
#>   WEEK 3 DAY 22                                                                                                                     
#>     n                                       134             134             134             134             132             132     
#>     Mean (SD)                            50.1 (8.3)      1.1 (12.4)      49.3 (7.5)     -0.8 (10.6)      49.8 (7.7)     -0.4 (12.1) 
#>     Median                                  49.9            3.1             48.8            0.2             49.5            -1.2    
#>     Min - Max                           26.7 - 69.4     -36.7 - 31.8    30.7 - 67.8     -30.7 - 29.9    30.1 - 72.9     -28.8 - 34.6
#>   WEEK 4 DAY 29                                                                                                                     
#>     n                                       134             134             134             134             132             132     
#>     Mean (SD)                            51.6 (8.1)      2.7 (12.2)      49.5 (8.2)     -0.6 (11.1)      49.8 (7.6)     -0.5 (11.6) 
#>     Median                                  52.3            3.2             48.8            -1.3            50.2            0.7     
#>     Min - Max                           31.5 - 71.2     -31.3 - 41.6    27.1 - 67.3     -24.9 - 23.5    32.1 - 68.5     -33.0 - 32.8
#>   WEEK 5 DAY 36                                                                                                                     
#>     n                                       134             134             134             134             132             132     
#>     Mean (SD)                            49.7 (8.8)      0.8 (13.4)      50.8 (7.6)      0.8 (11.3)      49.1 (7.7)     -1.1 (11.6) 
#>     Median                                  49.2            -0.3            50.3            0.5             49.2            -1.5    
#>     Min - Max                           26.0 - 69.6     -40.0 - 33.6    33.3 - 70.3     -25.6 - 33.1    32.2 - 73.6     -33.5 - 31.6
#> Immunoglobulin A Measurement                                                                                                        
#>   BASELINE                                                                                                                          
#>     n                                       134             134             134             134             132             132     
#>     Mean (SD)                            49.9 (7.4)      0.0 (0.0)       49.7 (8.3)      0.0 (0.0)       50.3 (9.1)      0.0 (0.0)  
#>     Median                                  48.7            0.0             49.3            0.0             49.7            0.0     
#>     Min - Max                           33.7 - 65.9      0.0 - 0.0      25.8 - 71.5      0.0 - 0.0      26.0 - 70.0      0.0 - 0.0  
#>   WEEK 1 DAY 8                                                                                                                      
#>     n                                       134             134             134             134             132             132     
#>     Mean (SD)                            49.4 (8.1)     -0.5 (10.7)      51.1 (8.2)      1.4 (11.2)      50.8 (8.4)      0.4 (12.9) 
#>     Median                                  49.8            -0.1            50.5            1.5             51.1            0.5     
#>     Min - Max                           29.3 - 69.3     -28.5 - 21.6    31.2 - 74.5     -25.4 - 35.7    26.2 - 71.0     -29.7 - 31.0
#>   WEEK 2 DAY 15                                                                                                                     
#>     n                                       134             134             134             134             132             132     
#>     Mean (SD)                            50.7 (7.9)      0.8 (11.5)      49.9 (8.3)      0.1 (11.7)      49.4 (7.5)     -0.9 (12.0) 
#>     Median                                  49.8            1.4             50.7            0.9             49.6            -0.2    
#>     Min - Max                           33.3 - 71.9     -25.8 - 32.3    27.9 - 69.8     -29.5 - 27.6    29.4 - 65.1     -32.8 - 28.9
#>   WEEK 3 DAY 22                                                                                                                     
#>     n                                       134             134             134             134             132             132     
#>     Mean (SD)                            49.5 (8.0)     -0.4 (10.4)      50.0 (7.9)      0.2 (11.5)      51.0 (7.8)      0.7 (12.0) 
#>     Median                                  49.3            0.5             49.1            0.8             51.1            0.3     
#>     Min - Max                           24.1 - 66.1     -28.7 - 24.2    25.0 - 67.7     -32.1 - 24.1    29.9 - 69.8     -36.2 - 24.6
#>   WEEK 4 DAY 29                                                                                                                     
#>     n                                       134             134             134             134             132             132     
#>     Mean (SD)                            49.2 (7.7)     -0.7 (10.9)      49.1 (7.6)     -0.6 (11.0)      49.6 (8.4)     -0.7 (12.4) 
#>     Median                                  50.5            -0.0            48.9            0.6             49.9            -1.4    
#>     Min - Max                           26.0 - 64.5     -30.5 - 26.5    31.5 - 66.1     -27.6 - 21.9    29.7 - 69.1     -26.9 - 31.8
#>   WEEK 5 DAY 36                                                                                                                     
#>     n                                       134             134             134             134             132             132     
#>     Mean (SD)                            49.7 (7.7)     -0.3 (10.1)      50.4 (8.7)      0.6 (12.7)      49.3 (8.5)     -1.0 (12.2) 
#>     Median                                  49.2            -0.2            49.7            1.4             50.2            0.7     
#>     Min - Max                           30.8 - 71.3     -33.8 - 25.6    28.5 - 77.0     -27.5 - 32.7    22.4 - 67.6     -32.0 - 26.5