RMPT01Duration of Exposure for Risk Management Plan Table.

The RMPT01 table provides an overview of duration of exposure.

Usage

rmpt01_main(adam_db, summaryvars = c("AVALCAT1", "AVAL"), parcat = NULL, ...)

rmpt01_pre(adam_db, parcat = NULL, ...)

rmpt01_post(tlg, prune_0 = FALSE, ...)

rmpt01

Format

An object of class chevron_t of length 1.

Arguments

adam_db

(list of data.frames) object containing the ADaM datasets

summaryvars

(character) variables to be analyzed. The label attribute of the corresponding columns in adex table of adam_db is used as label.

parcat

(string) the name of the variable initiating a new row split.

...

not used.

tlg

(TableTree, Listing or ggplot) object typically produced by a main function.

prune_0

(flag) remove 0 count rows

Details

Person time is the sum of exposure across all patients.

Summary statistics are by default based on the number of patients in the corresponding N row (number of non-missing values).

Does not remove zero-count rows unless overridden with prune_0 = TRUE.

Functions

rmpt01_main(): Main TLG function

rmpt01_pre(): Preprocessing

rmpt01_post(): Postprocessing

Note

adam_db object must contain an adex table with "PARAMCD" and the columns specified by anl_vars which is denoted as c("AVALCAT1", "AVAL") by default.

Examples

run(rmpt01, syn_data)
#>                                                Patients     Person time*
#>                                                (N=400)        (N=400)   
#>   ——————————————————————————————————————————————————————————————————————
#>   Total Number of Patients and Person Time   400 (100.0%)      49700    
#>     < 1 month                                 39 (9.8%)         728     
#>     1 to <3 months                           111 (27.8%)        6418    
#>     3 to <6 months                           136 (34.0%)       17645    
#>     >=6 months                               114 (28.5%)       24909    

run(rmpt01, syn_data, parcat = "PARCAT2")
#>   Parameter Category    Patients     Person time*
#>                          (N=400)       (N=400)   
#>   ———————————————————————————————————————————————
#>   Drug A               217 (54.2%)      27693    
#>     < 1 month           22 (5.5%)        402     
#>     1 to <3 months     55 (13.8%)        3139    
#>     3 to <6 months     73 (18.2%)        9543    
#>     >=6 months         67 (16.8%)       14609    
#>   Drug B               183 (45.8%)      22007    
#>     < 1 month           17 (4.2%)        326     
#>     1 to <3 months     56 (14.0%)        3279    
#>     3 to <6 months     63 (15.8%)        8102    
#>     >=6 months         47 (11.8%)       10300