TTET01 Binary Outcomes Summary

TTET01 template may be used to summarize any binary outcome or response variable at a single time point. Typical application for oncology

Usage

ttet01_main(
  adam_db,
  dataset = "adtte",
  arm_var = "ARM",
  ref_group = NULL,
  summarize_event = TRUE,
  perform_analysis = "unstrat",
  strata = NULL,
  pval_method = "log-rank",
  conf_level = 0.95,
  conf_type = "log-log",
  quantiles = c(0.25, 0.75),
  ties = "efron",
  timepoint = c(6, 12),
  method = "both",
  ...
)

ttet01_pre(adam_db, dataset = "adtte", ...)

ttet01_post(tlg, prune_0 = TRUE, ...)

ttet01

Format

An object of class chevron_t of length 1.

Arguments

adam_db

(list of data.frames) object containing the ADaM datasets

dataset

(string) the name of a table in the adam_db object.

arm_var

(string) variable used for column splitting

ref_group

(string) The name of the reference group, the value should be identical to the values in arm_var, if not specified, it will by default use the first level or value of arm_var.

summarize_event

(flag) should the event description be displayed, default is TRUE

perform_analysis

(string) option to display statistical comparisons using stratified analyses, or unstratified analyses, or both, e.g. c("unstrat", "strat"). Only unstratified will be displayed by default

strata

(string) stratification factors, e.g. strata = c("STRATA1", "STRATA2"), by default as NULL

pval_method

(string) p-value method for testing hazard ratio = 1. Default method is "log-rank", can also be set to "wald" or "likelihood".

conf_level

(numeric) the level of confidence interval, default is 0.95.

conf_type

(string) confidence interval type. Options are "plain" (default), "log", "log-log", see more in survival::survfit(). Note option "none" is no longer supported.

quantiles

(numeric) of length two to specify the quantiles of survival time.

ties

(string) specifying the method for tie handling. Default is "efron", can also be set to "breslow" or "exact". see more in survival::coxph()

timepoint

(numeric) survival time point of interest.

method

(string) either "surv" (survival estimations), "surv_diff" (difference in survival with the control) or "both".

...

not used.

tlg

(TableTree, Listing or ggplot) object typically produced by a main function.

prune_0

(flag) remove 0 count rows

Details

No overall value.

Functions

ttet01_main(): Main TLG function

ttet01_pre(): Preprocessing

ttet01_post(): Postprocessing

Examples

library(dplyr)
library(dunlin)

syn_data2 <- log_filter(syn_data, PARAMCD == "PFS", "adtte")
run(ttet01, syn_data2)
#>                                       A: Drug X        B: Placebo      C: Combination
#>                                        (N=134)           (N=134)          (N=132)    
#>   ———————————————————————————————————————————————————————————————————————————————————
#>   Patients with event (%)             86 (64.2%)       90 (67.2%)        92 (69.7%)  
#>     Earliest contributing event                                                      
#>       Death                           38 (44.2%)       48 (53.3%)        41 (44.6%)  
#>       Disease Progression             48 (55.8%)       42 (46.7%)        51 (55.4%)  
#>   Patients without event (%)          48 (35.8%)       44 (32.8%)        40 (30.3%)  
#>   Time to Event (MONTHS)                                                             
#>     Median                               16.9             19.7              12.3     
#>       95% CI                         (12.8, 19.0)     (13.2, 25.2)      (8.4, 14.9)  
#>     25% and 75%-ile                   7.7, 30.0         7.2, 34.1        4.5, 25.3   
#>     Range (censored)                 0.3 to 65.9       0.4 to 83.0      1.0 to 81.8  
#>     Range (event)                    0.5 to 85.8       0.1 to 58.7      0.2 to 77.8  
#>   Unstratified Analysis                                                              
#>     p-value (log-rank)                                   0.4808            0.0429    
#>     Hazard Ratio                                          0.90              1.36     
#>     95% CI                                            (0.67, 1.21)      (1.01, 1.82) 
#>   6 MONTHS                                                                           
#>     Patients remaining at risk            87               92                77      
#>     Event Free Rate (%)                 77.27             77.85            67.70     
#>     95% CI                          (68.76, 83.74)   (69.53, 84.15)    (58.64, 75.20)
#>     Difference in Event Free Rate                         0.57             -9.57     
#>       95% CI                                         (-9.82, 10.97)    (-20.70, 1.57)
#>       p-value (Z-test)                                   0.9137            0.0921    
#>   12 MONTHS                                                                          
#>     Patients remaining at risk            65               68                51      
#>     Event Free Rate (%)                 62.83             62.69            50.16     
#>     95% CI                          (53.36, 70.91)   (53.31, 70.71)    (40.69, 58.89)
#>     Difference in Event Free Rate                         -0.14            -12.68    
#>       95% CI                                         (-12.53, 12.25)   (-25.38, 0.03)
#>       p-value (Z-test)                                   0.9822            0.0505    
run(ttet01, syn_data2,
  summarize_event = FALSE, perform_analysis = c("unstrat", "strat"),
  strata = c("STRATA1", "STRATA2")
)
#>                                       A: Drug X        B: Placebo      C: Combination
#>                                        (N=134)           (N=134)          (N=132)    
#>   ———————————————————————————————————————————————————————————————————————————————————
#>   Patients with event (%)             86 (64.2%)       90 (67.2%)        92 (69.7%)  
#>   Patients without event (%)          48 (35.8%)       44 (32.8%)        40 (30.3%)  
#>   Time to Event (MONTHS)                                                             
#>     Median                               16.9             19.7              12.3     
#>       95% CI                         (12.8, 19.0)     (13.2, 25.2)      (8.4, 14.9)  
#>     25% and 75%-ile                   7.7, 30.0         7.2, 34.1        4.5, 25.3   
#>     Range (censored)                 0.3 to 65.9       0.4 to 83.0      1.0 to 81.8  
#>     Range (event)                    0.5 to 85.8       0.1 to 58.7      0.2 to 77.8  
#>   Unstratified Analysis                                                              
#>     p-value (log-rank)                                   0.4808            0.0429    
#>     Hazard Ratio                                          0.90              1.36     
#>     95% CI                                            (0.67, 1.21)      (1.01, 1.82) 
#>   Stratified Analysis                                                                
#>     p-value (log-rank)                                   0.8068            0.0252    
#>     Hazard Ratio                                          0.96              1.42     
#>     95% CI                                            (0.71, 1.31)      (1.04, 1.92) 
#>   6 MONTHS                                                                           
#>     Patients remaining at risk            87               92                77      
#>     Event Free Rate (%)                 77.27             77.85            67.70     
#>     95% CI                          (68.76, 83.74)   (69.53, 84.15)    (58.64, 75.20)
#>     Difference in Event Free Rate                         0.57             -9.57     
#>       95% CI                                         (-9.82, 10.97)    (-20.70, 1.57)
#>       p-value (Z-test)                                   0.9137            0.0921    
#>   12 MONTHS                                                                          
#>     Patients remaining at risk            65               68                51      
#>     Event Free Rate (%)                 62.83             62.69            50.16     
#>     95% CI                          (53.36, 70.91)   (53.31, 70.71)    (40.69, 58.89)
#>     Difference in Event Free Rate                         -0.14            -12.68    
#>       95% CI                                         (-12.53, 12.25)   (-25.38, 0.03)
#>       p-value (Z-test)                                   0.9822            0.0505