VST01 Vital Sign Results and change from Baseline By Visit Table.

The VST01 table provides an overview of the Vital Sign values and its change from baseline of each respective arm over the course of the trial.

Usage

egt01_pre(adam_db, dataset = "adeg", ...)

vst01_main(
  adam_db,
  dataset = "advs",
  arm_var = "ACTARM",
  row_split_var = NULL,
  summaryvars = c("AVAL", "CHG"),
  visitvar = "AVISIT",
  precision = list(),
  default_precision = 2,
  page_by = TRUE,
  ...
)

vst01_pre(adam_db, dataset = "advs", ...)

vst01

Format

An object of class chevron_t of length 1.

Source

cfbt01.R

Arguments

adam_db

(list of data.frames) object containing the ADaM datasets

dataset

(string) the name of a table in the adam_db object.

...

not used.

arm_var

(string) variable used for column splitting

row_split_var

(character) row split variable other than PARAMCD.

summaryvars

(character) variables to be analyzed. The label attribute of the corresponding column in table of adam_db is used as label.

visitvar

(string) typically one of "AVISIT" (Default) or "ATPTN" depending on the type of time point to be displayed

precision

(named list of integer) where names are values found in the PARAMCD column and the the values indicate the number of digits that should be represented for min, max and median. Mean and sd are represented with one more decimal of precision.

default_precision

(integer) the default number of digits.

page_by

(flag) indicator whether the parameter row split is by page.

Details

The Analysis Value column, displays the number of patients, the mean, standard deviation, median and range of the analysis value for each visit.

The Change from Baseline column, displays the number of patient and the mean, standard deviation, median and range of changes relative to the baseline.

Remove zero-count rows unless overridden with prune_0 = FALSE.

Split columns by arm, typically ACTARM.

Does not include a total column by default.

Sorted based on factor level; first by PARAM labels in alphabetic order then by chronological time point given by AVISIT. Re-level to customize order

Functions

egt01_pre(): Preprocessing

vst01_main(): Main TLG function

vst01_pre(): Preprocessing

Note

adam_db object must contain table named as dataset with the columns specified in summaryvars.

Examples

library(dunlin)
proc_data <- log_filter(
  syn_data,
  PARAMCD %in% c("DIABP", "SYSBP"), "advs"
)
run(vst01, proc_data)
#>   Analysis Visit                        A: Drug X                          B: Placebo                       C: Combination         
#>                                                 Change from                        Change from                        Change from  
#>                              Value at Visit      Baseline       Value at Visit      Baseline       Value at Visit      Baseline    
#>                                 (N=134)           (N=134)          (N=134)           (N=134)          (N=132)           (N=132)    
#>   —————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————
#>   Diastolic Blood Pressure                                                                                                         
#>     SCREENING                                                                                                                      
#>       n                           134                0               134                0               132                0       
#>       Mean (SD)              49.968 (7.197)       NE (NE)       50.753 (8.429)       NE (NE)       50.192 (7.627)       NE (NE)    
#>       Median                     49.706             NE              50.087             NE              49.641             NE       
#>       Min - Max              31.69 - 71.19        NE - NE       29.26 - 69.21        NE - NE       26.89 - 69.98        NE - NE    
#>     BASELINE                                                                                                                       
#>       n                           134                                134                                132                        
#>       Mean (SD)              48.602 (7.962)                     50.442 (7.948)                     51.107 (7.790)                  
#>       Median                     48.418                             50.179                             50.804                      
#>       Min - Max              27.71 - 64.64                      21.68 - 67.51                      29.75 - 71.40                   
#>     WEEK 1 DAY 8                                                                                                                   
#>       n                           134               134              134               134              132               132      
#>       Mean (SD)              50.260 (7.514)   1.658 (10.774)    49.674 (7.743)   -0.767 (10.947)   48.855 (7.888)   -2.251 (10.380)
#>       Median                     50.088            0.607            49.747           -1.114            47.677           -2.229     
#>       Min - Max              33.04 - 68.98    -20.37 - 29.94    33.71 - 66.49    -25.80 - 28.37    30.39 - 66.99    -22.06 - 22.44 
#>     WEEK 2 DAY 15                                                                                                                  
#>       n                           134               134              134               134              132               132      
#>       Mean (SD)              50.836 (7.850)   2.234 (11.752)    49.718 (8.445)   -0.724 (12.444)   49.981 (8.338)   -1.126 (11.639)
#>       Median                     51.437            3.126            50.227           -0.709            51.007           -2.271     
#>       Min - Max              31.89 - 70.34    -23.88 - 39.23    30.64 - 68.11    -28.40 - 36.97    24.82 - 65.57    -29.24 - 30.72 
#>     WEEK 3 DAY 22                                                                                                                  
#>       n                           134               134              134               134              132               132      
#>       Mean (SD)              50.712 (7.815)   2.110 (10.865)    49.092 (7.941)   -1.350 (11.104)   49.944 (7.662)   -1.162 (10.901)
#>       Median                     50.157            2.611            49.533           -2.317            50.025           -0.617     
#>       Min - Max              29.07 - 74.89    -25.78 - 30.66    24.50 - 67.11    -28.09 - 33.90    31.52 - 68.55    -29.72 - 26.98 
#>     WEEK 4 DAY 29                                                                                                                  
#>       n                           134               134              134               134              132               132      
#>       Mean (SD)              50.074 (8.071)   1.472 (10.885)    49.624 (7.102)   -0.818 (10.634)   49.707 (8.048)   -1.399 (11.047)
#>       Median                     49.269            1.202            49.193           -0.757            49.844           -0.182     
#>       Min - Max              29.41 - 71.36    -30.92 - 24.76    32.71 - 66.98    -33.18 - 35.95    28.75 - 68.45    -35.47 - 28.36 
#>     WEEK 5 DAY 36                                                                                                                  
#>       n                           134               134              134               134              132               132      
#>       Mean (SD)              50.569 (7.531)   1.967 (11.206)    48.366 (7.959)   -2.075 (11.736)   49.091 (7.570)   -2.015 (10.658)
#>       Median                     49.422            1.573            47.721           -2.871            49.237           -2.247     
#>       Min - Max              35.06 - 72.21    -25.22 - 31.47    30.42 - 70.61    -31.23 - 29.16    30.59 - 72.91    -30.45 - 34.57 
#>   Systolic Blood Pressure                                                                                                          
#>     SCREENING                                                                                                                      
#>       n                           134                0               134                0               132                0       
#>       Mean (SD)              49.572 (8.322)       NE (NE)       50.306 (8.343)       NE (NE)       50.902 (7.822)       NE (NE)    
#>       Median                     49.571             NE              50.155             NE              50.774             NE       
#>       Min - Max              23.97 - 70.85        NE - NE       26.22 - 79.12        NE - NE       27.63 - 67.45        NE - NE    
#>     BASELINE                                                                                                                       
#>       n                           134                                134                                132                        
#>       Mean (SD)              49.407 (8.467)                     50.245 (8.531)                     48.494 (7.184)                  
#>       Median                     48.335                             49.971                             49.217                      
#>       Min - Max              24.34 - 71.06                      24.38 - 71.07                      26.22 - 63.35                   
#>     WEEK 1 DAY 8                                                                                                                   
#>       n                           134               134              134               134              132               132      
#>       Mean (SD)              50.702 (9.166)   1.295 (12.814)    49.263 (8.693)   -0.982 (11.221)   49.647 (7.991)   1.153 (10.185) 
#>       Median                     49.526            0.129            48.302           -1.678            49.768            0.085     
#>       Min - Max              29.75 - 78.99    -29.51 - 37.65    32.99 - 73.97    -27.62 - 24.77    25.52 - 68.62    -22.79 - 31.30 
#>     WEEK 2 DAY 15                                                                                                                  
#>       n                           134               134              134               134              132               132      
#>       Mean (SD)              49.369 (8.469)   -0.038 (11.936)   49.377 (7.009)   -0.869 (10.968)   49.260 (7.683)    0.766 (9.716) 
#>       Median                     49.489           -0.243            49.386           -0.929            48.289           -0.216     
#>       Min - Max              24.18 - 65.44    -39.12 - 32.72    27.62 - 69.23    -37.10 - 33.85    29.39 - 74.75    -24.92 - 28.99 
#>     WEEK 3 DAY 22                                                                                                                  
#>       n                           134               134              134               134              132               132      
#>       Mean (SD)              49.682 (7.776)   0.275 (11.383)    49.034 (7.844)   -1.211 (12.123)   50.544 (8.303)   2.050 (11.059) 
#>       Median                     50.136            1.479            48.794           -1.190            51.672            1.951     
#>       Min - Max              31.30 - 70.31    -27.49 - 27.15    15.58 - 66.17    -35.56 - 36.36    29.62 - 69.40    -29.72 - 30.62 
#>     WEEK 4 DAY 29                                                                                                                  
#>       n                           134               134              134               134              132               132      
#>       Mean (SD)              49.424 (7.637)   0.017 (11.473)    49.400 (8.640)   -0.846 (12.599)   50.657 (7.395)   2.162 (10.495) 
#>       Median                     49.783            0.599            49.927           -1.288            50.823            2.965     
#>       Min - Max              22.92 - 69.57    -32.30 - 28.46    22.14 - 72.06    -29.71 - 38.11    29.72 - 65.12    -28.75 - 34.15 
#>     WEEK 5 DAY 36                                                                                                                  
#>       n                           134               134              134               134              132               132      
#>       Mean (SD)              49.395 (7.898)   -0.012 (11.876)   50.222 (8.653)   -0.023 (11.828)   50.367 (8.176)   1.873 (10.490) 
#>       Median                     49.565            0.730            49.934            0.192            50.811            2.431     
#>       Min - Max              26.11 - 69.66    -39.25 - 33.93    27.91 - 74.17    -27.70 - 36.16    34.27 - 74.52    -23.70 - 32.42