VST01 Vital Sign Results and change from Baseline By Visit Table.

The VST01 table provides an overview of the Vital Sign values and its change from baseline of each respective arm over the course of the trial.

Usage

vst01_main(
  adam_db,
  dataset = "advs",
  arm_var = "ACTARM",
  lbl_overall = NULL,
  row_split_var = NULL,
  summaryvars = c("AVAL", "CHG"),
  visitvar = "AVISIT",
  precision = list(default = 2L),
  page_var = "PARAMCD",
  .stats = c("n", "mean_sd", "median", "range"),
  skip = list(CHG = "BASELINE"),
  ...
)

vst01_pre(adam_db, dataset = "advs", ...)

vst01

Format

An object of class chevron_t of length 1.

Arguments

adam_db

(list of data.frames) object containing the ADaM datasets

dataset

(string) the name of a table in the adam_db object.

arm_var

(string) variable used for column splitting

lbl_overall

(string) label used for overall column, if set to NULL the overall column is omitted

row_split_var

(character) additional row split variables.

summaryvars

(character) variables to be analyzed. The label attribute of the corresponding column in table of adam_db is used as label.

visitvar

(string) typically one of "AVISIT" or user-defined visit incorporating "ATPT".

precision

(named list of integer) where names are values found in the PARAMCD column and the values indicate the number of digits in statistics. If default is set, and parameter precision not specified, the value for default will be used.

page_var

(string) variable name prior to which the row split is by page.

.stats

(character) statistics names, see tern::analyze_vars().

skip

Named (list) of visit values that need to be inhibited.

...

additional arguments like .indent_mods, .labels.

Value

the main function returns an rtables object.

the preprocessing function returns a list of data.frame.

Details

The Analysis Value column, displays the number of patients, the mean, standard deviation, median and range of the analysis value for each visit.

The Change from Baseline column, displays the number of patient and the mean, standard deviation, median and range of changes relative to the baseline.

Remove zero-count rows unless overridden with prune_0 = FALSE.

Split columns by arm, typically ACTARM.

Does not include a total column by default.

Sorted based on factor level; first by PARAM labels in alphabetic order then by chronological time point given by AVISIT. Re-level to customize order

Functions

vst01_main(): Main TLG function

vst01_pre(): Preprocessing

Note

adam_db object must contain table named as dataset with the columns specified in summaryvars.

Examples

library(dunlin)

proc_data <- log_filter(
  syn_data,
  PARAMCD %in% c("DIABP", "SYSBP"), "advs"
)
run(vst01, proc_data)
#>                                          A: Drug X                          B: Placebo                        C: Combination          
#>                                                  Change from                         Change from                         Change from  
#>                              Value at Visit       Baseline       Value at Visit       Baseline       Value at Visit       Baseline    
#>   Analysis Visit                 (N=15)            (N=15)            (N=15)            (N=15)            (N=15)            (N=15)     
#>   ————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————
#>   Diastolic Blood Pressure                                                                                                            
#>     SCREENING                                                                                                                         
#>       n                            15                 0                15                 0                15                 0       
#>       Mean (SD)              47.754 (6.827)        NE (NE)       52.552 (8.234)        NE (NE)       52.587 (5.051)        NE (NE)    
#>       Median                     47.973              NE              54.453              NE              53.344              NE       
#>       Min - Max               32.28 - 58.80        NE - NE        34.09 - 62.76        NE - NE        43.31 - 60.87        NE - NE    
#>     BASELINE                                                                                                                          
#>       n                            15                                  15                                  15                         
#>       Mean (SD)              48.453 (8.983)                      53.244 (6.030)                      51.260 (7.901)                   
#>       Median                     47.331                              53.581                              51.140                       
#>       Min - Max               34.23 - 64.64                       43.38 - 62.65                       36.42 - 64.62                   
#>     WEEK 1 DAY 8                                                                                                                      
#>       n                            15                15                15                15                15                15       
#>       Mean (SD)              49.591 (8.543)    1.138 (11.242)    51.644 (6.469)    -1.601 (8.747)    50.330 (7.611)    -0.929 (10.007)
#>       Median                     46.979            -1.626            53.081             1.291            51.223            -0.990     
#>       Min - Max               37.02 - 62.15    -13.13 - 19.07     38.36 - 63.15    -21.18 - 11.45     38.02 - 61.47    -22.06 - 16.72 
#>     WEEK 2 DAY 15                                                                                                                     
#>       n                            15                15                15                15                15                15       
#>       Mean (SD)              49.903 (5.791)     1.450 (8.476)    48.989 (6.919)    -4.255 (8.332)    47.709 (6.784)    -3.551 (10.892)
#>       Median                     50.599             0.692            49.800            -3.891            48.715            -4.062     
#>       Min - Max               38.79 - 59.12    -15.80 - 19.03     31.52 - 60.32    -22.06 - 10.10     35.38 - 56.99    -29.24 - 18.62 
#>     WEEK 3 DAY 22                                                                                                                     
#>       n                            15                15                15                15                15                15       
#>       Mean (SD)              49.640 (10.443)   1.187 (13.731)    46.794 (6.770)    -6.451 (8.753)    47.834 (5.424)    -3.425 (8.685) 
#>       Median                     50.458            -0.109            46.497            -5.754            49.359            -6.430     
#>       Min - Max               29.07 - 74.89    -19.15 - 30.66     37.12 - 57.89     -21.22 - 9.01     39.40 - 56.17    -15.16 - 13.06 
#>     WEEK 4 DAY 29                                                                                                                     
#>       n                            15                15                15                15                15                15       
#>       Mean (SD)              51.360 (8.909)    2.907 (12.296)    48.587 (7.780)    -4.657 (10.369)   53.335 (7.367)     2.075 (8.753) 
#>       Median                     49.267             1.004            48.954            -6.717            53.022             3.186     
#>       Min - Max               35.24 - 69.30    -15.37 - 24.76     35.34 - 62.63    -22.84 - 19.25     37.51 - 63.01    -13.58 - 16.60 
#>     WEEK 5 DAY 36                                                                                                                     
#>       n                            15                15                15                15                15                15       
#>       Mean (SD)              47.289 (7.415)    -1.164 (11.549)   49.156 (8.281)    -4.088 (11.037)   48.127 (6.561)    -3.133 (7.931) 
#>       Median                     46.319            -1.354            49.620            -4.528            47.550            -1.866     
#>       Min - Max               35.42 - 65.64    -19.45 - 18.94     37.79 - 65.06    -21.75 - 15.14     38.76 - 65.42    -17.79 - 11.64 
#>   Systolic Blood Pressure                                                                                                             
#>     SCREENING                                                                                                                         
#>       n                            15                 0                15                 0                15                 0       
#>       Mean (SD)              50.524 (9.964)        NE (NE)       48.742 (9.375)        NE (NE)       53.448 (7.713)        NE (NE)    
#>       Median                     48.432              NE              49.563              NE              55.800              NE       
#>       Min - Max               37.54 - 70.85        NE - NE        33.15 - 61.19        NE - NE        35.15 - 65.39        NE - NE    
#>     BASELINE                                                                                                                          
#>       n                            15                                  15                                  15                         
#>       Mean (SD)              48.917 (8.422)                      52.206 (7.663)                      44.329 (6.177)                   
#>       Median                     47.976                              50.399                              43.480                       
#>       Min - Max               35.92 - 63.83                       36.85 - 69.87                       35.88 - 55.61                   
#>     WEEK 1 DAY 8                                                                                                                      
#>       n                            15                15                15                15                15                15       
#>       Mean (SD)              49.034 (7.281)    0.118 (12.328)    49.648 (7.885)    -2.558 (10.315)   49.845 (10.051)    5.516 (8.410) 
#>       Median                     45.853            -2.350            49.750            -4.912            50.426             7.279     
#>       Min - Max               38.46 - 62.80    -18.78 - 19.87     37.12 - 62.87    -17.67 - 16.38     32.64 - 63.18    -11.55 - 21.18 
#>     WEEK 2 DAY 15                                                                                                                     
#>       n                            15                15                15                15                15                15       
#>       Mean (SD)              50.697 (10.111)   1.781 (11.712)    48.917 (6.464)    -3.289 (12.641)   44.922 (7.731)     0.593 (8.516) 
#>       Median                     48.793             2.483            48.304            -1.084            46.824             2.724     
#>       Min - Max               33.60 - 64.41    -20.66 - 23.99     32.77 - 57.63    -37.10 - 15.27     29.39 - 61.47    -18.72 - 13.40 
#>     WEEK 3 DAY 22                                                                                                                     
#>       n                            15                15                15                15                15                15       
#>       Mean (SD)              47.129 (8.417)    -1.787 (10.523)   44.842 (9.839)    -7.365 (12.156)   52.240 (6.714)     7.911 (9.955) 
#>       Median                     47.910            -1.677            44.495            -4.628            52.734             7.321     
#>       Min - Max               33.85 - 59.67    -19.31 - 14.54     15.58 - 58.38    -35.56 - 11.49     40.87 - 64.82     -9.47 - 22.94 
#>     WEEK 4 DAY 29                                                                                                                     
#>       n                            15                15                15                15                15                15       
#>       Mean (SD)              50.984 (9.702)    2.068 (12.168)    50.001 (10.412)   -2.206 (11.420)   51.404 (8.276)    7.075 (10.011) 
#>       Median                     51.783             4.255            51.045             1.095            55.334             6.775     
#>       Min - Max               38.70 - 66.08    -15.94 - 20.88     33.06 - 67.92    -21.01 - 19.90     32.28 - 60.16    -15.84 - 24.23 
#>     WEEK 5 DAY 36                                                                                                                     
#>       n                            15                15                15                15                15                15       
#>       Mean (SD)              48.748 (9.820)    -0.169 (11.595)   47.553 (7.967)    -4.653 (11.661)   51.113 (7.249)     6.784 (9.493) 
#>       Median                     47.844             0.827            46.869            -1.284            51.106             8.107     
#>       Min - Max               29.55 - 69.66    -18.39 - 18.76     35.50 - 63.63    -27.59 - 13.23     34.27 - 60.01    -14.02 - 21.11